Cardano.Api.Extra

Description

Module containing extra Api functionality needed by the wallet.

Documentation

withShelleyBasedTx :: InAnyShelleyBasedEra Tx -> (forall era. IsShelleyBasedEra era => Tx era -> a) -> a

Apply an era-parameterized function to an existentially-wrapped tx.

inAnyCardanoEra :: IsCardanoEra era => Tx era -> InAnyCardanoEra Tx

Helper function for more easily creating an existential InAnyCardanoEra Tx.

asAnyShelleyBasedEra :: InAnyCardanoEra a -> Maybe (InAnyShelleyBasedEra a)

Downcast an existentially wrapped tx.