Cardano.Api.Gen

Documentation

genAddressAny :: Gen AddressAny

genAddressByron :: Gen (Address ByronAddr)

genAddressInEra :: CardanoEra era -> Gen (AddressInEra era)

genAddressShelley :: Gen (Address ShelleyAddr)

genAlphaNum :: Gen Char

genAssetIdNoAda :: Gen AssetId

genAssetName :: Gen AssetName

genByronKeyWitness :: Gen (KeyWitness ByronEra)

genCertIx :: Gen CertIx

genCostModel :: Gen CostModel

genCostModels :: Gen (Map AnyPlutusScriptVersion CostModel)

genEncodingBoundaryLovelace :: Gen Lovelace

genEpochNo :: Gen EpochNo

genExecutionUnitPrices :: Gen ExecutionUnitPrices

genExecutionUnits :: Gen ExecutionUnits

genExtraKeyWitnesses :: CardanoEra era -> Gen (TxExtraKeyWitnesses era)

genLovelace :: Gen Lovelace

genMIRPot :: Gen MIRPot

genMIRTarget :: Gen MIRTarget

genNat :: Gen Natural

genNetworkId :: Gen NetworkId

genNetworkMagic :: Gen NetworkMagic

genPaymentCredential :: Gen PaymentCredential

genPlutusScript :: PlutusScriptVersion lang -> Gen (PlutusScript lang)

genPolicyId :: Gen PolicyId

genPoolId :: Gen PoolId

genProtocolParameters :: Gen ProtocolParameters

genProtocolParametersUpdate :: Gen ProtocolParametersUpdate

genPtr :: Gen Ptr

genRational :: Gen Rational

genRationalInt64 :: Gen Rational

genScript :: ScriptLanguage lang -> Gen (Script lang)

genScriptData :: Gen ScriptData

genScriptHash :: Gen ScriptHash

genScriptInAnyLang :: Gen ScriptInAnyLang

genScriptInEra :: CardanoEra era -> Gen (ScriptInEra era)

genScriptValidity :: Gen ScriptValidity

genScriptWitnessStake :: ScriptLanguageInEra lang era -> Gen (ScriptWitness WitCtxStake era)

genSeed :: Int -> Gen Seed

genShelleyHash :: Gen (Hash Blake2b_256 EraIndependentTxBody)

genShelleyWitnessSigningKey :: Gen ShelleyWitnessSigningKey

genSignedQuantity :: Gen Quantity

Generate a positive or negative quantity.

genSignedValue :: Gen Value

Generate a Value which could represent the balance of a partial transaction, where both ada and other assets can be included, and quantities can be both positive and negative.

genSigningKey :: Key keyrole => AsType keyrole -> Gen (SigningKey keyrole)

genSimpleScript :: SimpleScriptVersion lang -> Gen (SimpleScript lang)

genSlotNo :: Gen SlotNo

genSlotNo32 :: Gen SlotNo32

genStakeAddress :: Gen StakeAddress

genStakeAddressReference :: Gen StakeAddressReference

genStakeCredential :: Gen StakeCredential

genStakePoolMetadata :: Gen StakePoolMetadata

genStakePoolMetadataReference :: Gen StakePoolMetadataReference

genStakePoolParameters :: Gen StakePoolParameters

genStakePoolRelay :: Gen StakePoolRelay

genTtl :: Gen SlotNo

genTx :: Gen (InAnyCardanoEra Tx)

genTxAuxScripts :: CardanoEra era -> Gen (TxAuxScripts era)

genTxBody :: IsCardanoEra era => CardanoEra era -> Gen (TxBody era)

genTxBodyContent :: CardanoEra era -> Gen (TxBodyContent BuildTx era)

genTxBodyForBalancing :: IsCardanoEra era => CardanoEra era -> Gen (TxBody era)

Similar to genTxBody, but with a distribution better suitable for testing balancing.

genTxCertificate :: Gen Certificate

genTxCertificates :: CardanoEra era -> Gen (TxCertificates BuildTx era)

genTxFee :: CardanoEra era -> Gen (TxFee era)

genTxForBalancing :: forall era. IsCardanoEra era => CardanoEra era -> Gen (Tx era)

genTxId :: Gen TxId

genTxIn :: Gen TxIn

genTxIndex :: Gen TxIx

genTxInEra :: forall era. IsCardanoEra era => CardanoEra era -> Gen (Tx era)

genTxInsCollateral :: CardanoEra era -> Gen (TxInsCollateral era)

genTxIx :: Gen TxIx

genTxMetadata :: Gen TxMetadata

genTxMetadataInEra :: CardanoEra era -> Gen (TxMetadataInEra era)

genTxMetadataValue :: Gen TxMetadataValue

genTxMintValue :: forall era. CardanoEra era -> Gen (TxMintValue BuildTx era)

genTxOut :: CardanoEra era -> Gen (TxOut ctx era)

genTxOutDatum :: CardanoEra era -> Gen (TxOutDatum ctx era)

genTxOutValue :: CardanoEra era -> Gen (TxOutValue era)

genTxReturnCollateral :: CardanoEra era -> Gen (TxReturnCollateral ctx era)

genTxScriptValidity :: CardanoEra era -> Gen (TxScriptValidity era)

genTxTotalCollateral :: CardanoEra era -> Gen (TxTotalCollateral era)

genTxValidityLowerBound :: CardanoEra era -> Gen (TxValidityLowerBound era)

genTxValidityRange :: CardanoEra era -> Gen (TxValidityLowerBound era, TxValidityUpperBound era)

genTxValidityUpperBound :: CardanoEra era -> Gen (TxValidityUpperBound era)

genTxWithdrawals :: CardanoEra era -> Gen (TxWithdrawals BuildTx era)

genUnsignedQuantity :: Gen Quantity

genUpdateProposal :: CardanoEra era -> Gen (TxUpdateProposal era)

genValueForMinting :: Gen Value

Generate a Value suitable for minting, i.e. non-ADA asset ID and a positive or negative quantity.

genValueForTxOut :: Gen Value

Generate a Value suitable for usage in a transaction output, i.e. any asset ID and a positive quantity.

genVerificationKey :: Key keyrole => AsType keyrole -> Gen (VerificationKey keyrole)

genVerificationKeyHash :: Key keyrole => AsType keyrole -> Gen (Hash keyrole)

genWithdrawalInfo :: CardanoEra era -> Gen (StakeAddress, Lovelace, BuildTxWith BuildTx (Witness WitCtxStake era))

genWitness :: CardanoEra era -> TxBody era -> Gen (KeyWitness era)

genWitnesses :: CardanoEra era -> TxBody era -> Gen [KeyWitness era]

genWitnessNetworkIdOrByronAddress :: Gen WitnessNetworkIdOrByronAddress

genWitnessStake :: CardanoEra era -> Gen (Witness WitCtxStake era)

Orphan instances

Random Lovelace

Instance details

Methods

randomR :: RandomGen g => (Lovelace, Lovelace) -> g -> (Lovelace, g)

random :: RandomGen g => g -> (Lovelace, g)

randomRs :: RandomGen g => (Lovelace, Lovelace) -> g -> [Lovelace]

randoms :: RandomGen g => g -> [Lovelace]

ToJSON StakePoolMetadata

Instance details

Methods

toJSON :: StakePoolMetadata -> Value

toEncoding :: StakePoolMetadata -> Encoding

toJSONList :: [StakePoolMetadata] -> Value

toEncodingList :: [StakePoolMetadata] -> Encoding