Cardano.DB.Sqlite

Description

An implementation of the DBLayer which uses Persistent and SQLite.

Documentation

newtype SqliteContext

SqliteContext is a function to execute queries.

Constructors

SqliteContext

Fields

runQuery :: forall a. SqlPersistT IO a -> IO a

Run a query with a connection from the pool.

newSqliteContext :: Tracer IO DBLog -> ConnectionPool -> [ManualMigration] -> Migration -> IO (Either MigrationError SqliteContext)

Sets up query logging and timing, runs schema migrations if necessary and provide a safe SqliteContext for interacting with the database.

newInMemorySqliteContext :: Tracer IO DBLog -> [ManualMigration] -> Migration -> ForeignKeysSetting -> IO (IO (), SqliteContext)

data ForeignKeysSetting

Specifies whether or not foreign key constraints are enabled, equivalent to the Sqlite foreign_keys setting.

When foreign key constraints are enabled, the database will enforce referential integrity, and cascading deletes are enabled.

When foreign keys constraints are disabled, the database will not enforce referential integrity, and cascading deletes are disabled.

See the following resource for more information: https://www.sqlite.org/foreignkeys.html#fk_enable

Constructors

ForeignKeysEnabled
Foreign key constraints are enabled.

ForeignKeysDisabled
Foreign key constraints are disabled.

ConnectionPool

type ConnectionPool = Pool (SqlBackend, Connection)

withConnectionPool :: Tracer IO DBLog -> FilePath -> (ConnectionPool -> IO a) -> IO a

Helpers

chunkSize :: Int

Maximum number of variables allowed in a single SQL statement

See also dbChunked.

dbChunked :: forall record b. PersistEntity record => ([record] -> SqlPersistT IO b) -> [record] -> SqlPersistT IO ()

Convert a single DB "updateMany" (or similar) query into multiple updateMany queries with smaller lists of values.

This is to prevent too many variables appearing in the SQL statement. SQLITE_MAX_VARIABLE_NUMBER is 999 by default, and we will get a "too many SQL variables" exception if that is exceeded.

We choose a conservative value chunkSize << 999 because there can be multiple variables per row updated.

dbChunkedFor :: forall record a b. PersistEntity record => ([a] -> SqlPersistT IO b) -> [a] -> SqlPersistT IO ()

Like dbChunked, but generalized for the case where the input list is not the same type as the record.

dbChunked' :: forall record b. PersistEntity record => ([(Key record, record)] -> SqlPersistT IO b) -> [(Key record, record)] -> SqlPersistT IO ()

Like dbChunked, but allows bundling elements with a Key. Useful when used with repsertMany.

handleConstraint :: MonadUnliftIO m => e -> m a -> m (Either e a)

Run an action, and convert any Sqlite constraints exception into the given error result. No other exceptions are handled.

Manual Migration

newtype ManualMigration

Encapsulates a manual migration action (or sequence of actions) to be performed immediately after an SQL connection is initiated.

Constructors

ManualMigration

Fields

executeManualMigration :: Connection -> IO ()

newtype MigrationError

Error type for when migrations go wrong after opening a database.

Constructors

MigrationError

Fields

getMigrationErrorMessage :: Text

data DBField where

Constructors

DBField :: forall record typ. PersistEntity record => EntityField record typ -> DBField

tableName :: DBField -> Text

fieldName :: DBField -> Text

fieldType :: DBField -> Text

Logging

data DBLog

Constructors

MsgMigrations (Either MigrationError Int)

MsgQuery Text Severity

MsgRun BracketLog

MsgCloseSingleConnection FilePath

MsgStartConnectionPool FilePath

MsgStopConnectionPool FilePath

MsgDatabaseReset

MsgIsAlreadyClosed Text

MsgStatementAlreadyFinalized Text

MsgManualMigrationNeeded DBField Text

MsgExpectedMigration DBLog

MsgManualMigrationNotNeeded DBField

MsgUpdatingForeignKeysSetting ForeignKeysSetting

MsgRetryOnBusy Int RetryLog