Cardano.Ledger.Credential.Safe

Description

Re-exports functionality provided by module Credential, but with a safer interface.

Safe Ptr interface

data Ptr

Pointer to a slot number, transaction index and an index in certificate list.

safePtr :: SlotNo32 -> TxIx -> CertIx -> Ptr

Safely constructs a Ptr without silent truncation of slot numbers.

Use toSlotNo32 to convert an ordinary SlotNo to a SlotNo32.

This function should satisfy the following property:

safeUnwrapPtr (safePtr s t c) == (s, t, c)

safeUnwrapPtr :: Ptr -> (SlotNo32, TxIx, CertIx)

Safely deconstructs a Ptr.

Use fromSlotNo32 to convert the returned slot number to a SlotNo.

This function should satisfy the following property:

safeUnwrapPtr (safePtr s t c) == (s, t, c)

Conversions to and from Slot32

newtype SlotNo32

A 32-bit wide slot number.

Constructors

SlotNo32 Word32

toSlotNo32 :: SlotNo -> Maybe SlotNo32

Converts an ordinary SlotNo into a SlotNo32.

Returns Nothing if the slot number could not be converted safely.

fromSlotNo32 :: SlotNo32 -> SlotNo

Converts a SlotNo32 into an ordinary SlotNo.