Cardano.Pool

Description

High-level interface for dealing with stake pools.

Documentation

data StakePoolLayer e m

StakePoolLayer is a thin layer ontop of the DB. It is one value that can easily be passed to the API-server, where it can be used in a simple way.

Constructors

StakePoolLayer

Fields

listStakePools :: ExceptT e m [(StakePool, Maybe StakePoolMetadata)]

knownStakePools :: m [PoolId]

Get a list of known pools that doesn't require fetching things from any registry. This list comes from the registration certificates that have been seen on chain.