Cardano.Pool.DB.Model

Description

An implementation of the production pool database using only pure functions.

These functions and types model the behaviour of the SQLite database backend, and are used for QuickCheck state machine testing, and the MVar database backend.

Model Types

data PoolDatabase

Constructors

PoolDatabase

Fields

pools :: !(Map PoolId [BlockHeader])

Information of what blocks were produced by which stake pools

distributions :: !(Map EpochNo [(PoolId, Quantity "lovelace" Word64)])

Store known stake distributions for epochs

owners :: !(Map PoolId [PoolOwner])

Mapping between pool ids and owners

registrations :: !(Map (CertificatePublicationTime, PoolId) PoolRegistrationCertificate)

On-chain registrations associated with pools

retirements :: !(Map (CertificatePublicationTime, PoolId) PoolRetirementCertificate)

On-chain retirements associated with pools

delisted :: !(Set PoolId)

metadata :: !(Map StakePoolMetadataHash StakePoolMetadata)

Off-chain metadata cached in database

fetchAttempts :: !(Map (StakePoolMetadataUrl, StakePoolMetadataHash) Int)

Metadata (failed) fetch attempts

seed :: !SystemSeed

Store an arbitrary random generator seed

blockHeaders :: [BlockHeader]

Store headers during syncing

settings :: Settings

internalState :: InternalState

Various internal states that need to persist across wallet restarts.

emptyPoolDatabase :: PoolDatabase

Produces an empty model pool production database.

Model Operation Types

type ModelOp a = StateT PoolDatabase (Either PoolErr) a

newtype PoolErr

Constructors

PointAlreadyExists BlockHeader

Model pool database functions

mCleanDatabase :: ModelOp ()

mCleanPoolMetadata :: ModelOp ()

mPutPoolProduction :: BlockHeader -> PoolId -> ModelOp ()

mPutHeader :: BlockHeader -> ModelOp ()

mListHeaders :: Int -> ModelOp [BlockHeader]

mReadPoolProduction :: TimeInterpreter Identity -> EpochNo -> ModelOp (Map PoolId [BlockHeader])

mReadTotalProduction :: ModelOp (Map PoolId (Quantity "block" Word64))

mPutStakeDistribution :: EpochNo -> [(PoolId, Quantity "lovelace" Word64)] -> ModelOp ()

mReadStakeDistribution :: EpochNo -> ModelOp [(PoolId, Quantity "lovelace" Word64)]

mReadPoolMetadata :: ModelOp (Map StakePoolMetadataHash StakePoolMetadata)

mPutPoolRegistration :: CertificatePublicationTime -> PoolRegistrationCertificate -> ModelOp ()

mReadPoolRegistration :: PoolId -> ModelOp (Maybe (CertificatePublicationTime, PoolRegistrationCertificate))

mPutPoolRetirement :: CertificatePublicationTime -> PoolRetirementCertificate -> ModelOp ()

mReadPoolRetirement :: PoolId -> ModelOp (Maybe (CertificatePublicationTime, PoolRetirementCertificate))

mUnfetchedPoolMetadataRefs :: Int -> ModelOp [(PoolId, StakePoolMetadataUrl, StakePoolMetadataHash)]

mPutDelistedPools :: [PoolId] -> ModelOp ()

mPutFetchAttempt :: (StakePoolMetadataUrl, StakePoolMetadataHash) -> ModelOp ()

mPutPoolMetadata :: StakePoolMetadataHash -> StakePoolMetadata -> ModelOp ()

mListPoolLifeCycleData :: EpochNo -> ModelOp [PoolLifeCycleStatus]

mListRegisteredPools :: ModelOp [PoolId]

mListRetiredPools :: EpochNo -> ModelOp [PoolRetirementCertificate]

mReadPoolLifeCycleStatus :: PoolId -> ModelOp PoolLifeCycleStatus

mReadSystemSeed :: PoolDatabase -> IO (StdGen, PoolDatabase)

mRollbackTo :: TimeInterpreter Identity -> SlotNo -> ModelOp ()

mReadCursor :: Int -> ModelOp [BlockHeader]

mRemovePools :: [PoolId] -> ModelOp ()

mReadDelistedPools :: ModelOp [PoolId]

mRemoveRetiredPools :: EpochNo -> ModelOp [PoolRetirementCertificate]

mReadSettings :: ModelOp Settings

mPutSettings :: Settings -> ModelOp ()

mPutLastMetadataGC :: POSIXTime -> ModelOp ()

mReadLastMetadataGC :: ModelOp (Maybe POSIXTime)