Cardano.Pool.DB.Sqlite

Description

An implementation of the DBLayer which uses Persistent and SQLite.

Documentation

newDBLayer

Arguments

:: Tracer IO PoolDbLog
Logging object

-> TimeInterpreter IO
Time interpreter for slot to time conversions

-> SqliteContext
A (thread-) safe wrapper for running db queries.

-> DBLayer IO

Sets up a connection to the SQLite database.

Database migrations are run to create tables if necessary.

If the given file path does not exist, it will be created by the sqlite library.

getDBLayer will provide the actual DBLayer implementation. The database should be closed with destroyDBLayer. If you use withDBLayer then both of these things will be handled for you.

withDBLayer

Arguments

:: Tracer IO PoolDbLog
Logging object.

-> Maybe FilePath
Database file location, or Nothing for in-memory database.

-> TimeInterpreter IO
The time interpreter object.

-> (DBLayer IO -> IO a)
Action to run.

-> IO a

Runs an action with a connection to the SQLite database.

Database migrations are run to create tables if necessary.

If the given file path does not exist, it will be created by the sqlite library.

withDecoratedDBLayer

Arguments

:: DBDecorator IO
The database decorator.

-> Tracer IO PoolDbLog
Logging object

-> Maybe FilePath
Database file location, or Nothing for in-memory database

-> TimeInterpreter IO
The time interpreter object.

-> (DBLayer IO -> IO a)
Action to run.

-> IO a

Runs an action with a connection to the SQLite database.

This function has the same behaviour as withDBLayer, but provides a way to decorate the created DBLayer object with a DBDecorator, useful for instrumenting or monitoring calls to database operations.

newtype DBDecorator a

A decorator for the database layer, useful for instrumenting or monitoring calls to database operations.

Constructors

DBDecorator

Fields

decorateDBLayer :: DBLayer a -> DBLayer a

undecoratedDB :: DBDecorator a

The identity decorator.

Equivalent to an undecorated database.

defaultFilePath

Arguments

:: FilePath
The directory in which the .sqlite file will be located.

-> FilePath

Return the preferred FilePath for the stake pool .sqlite file, given a parent directory.

data DatabaseView

Represents a database view.

Constructors

DatabaseView

Fields

databaseViewName :: Text

A name for the view.

databaseViewDefinition :: Text

A select query to generate the view.

createViews :: [ManualMigration]