Cardano.Pool.DB.Sqlite.TH

Description

Auto-generated Sqlite & Persistent machinery via Template-Haskell. This has been moved into a separate file so that we can treat it slightly differently when computing code-coverage.

Documentation

data InternalState

Constructors

InternalState

Fields

lastGCMetadata :: !(Maybe POSIXTime)

type InternalStateId = Key InternalState

data Settings

Constructors

Settings

Fields

settingsPoolMetadataSource :: !PoolMetadataSource

type SettingsId = Key Settings

data ArbitrarySeed

Constructors

ArbitrarySeed

Fields

seedSeed :: !StdGen

type ArbitrarySeedId = Key ArbitrarySeed

data PoolProduction

Constructors

PoolProduction

Fields

poolProductionPoolId :: !PoolId

poolProductionSlot :: !SlotNo

poolProductionHeaderHash :: !BlockId

poolProductionParentHash :: !BlockId

poolProductionBlockHeight :: !Word32

type PoolProductionId = Key PoolProduction

data BlockHeader

Constructors

BlockHeader

Fields

blockSlot :: !SlotNo

blockHeaderHash :: !BlockId

blockParentHash :: !BlockId

blockHeight :: !Word32

type BlockHeaderId = Key BlockHeader

data StakeDistribution

Constructors

StakeDistribution

Fields

stakeDistributionPoolId :: !PoolId

stakeDistributionEpoch :: !Word64

stakeDistributionStake :: !Word64

type StakeDistributionId = Key StakeDistribution

data PoolOwner

Constructors

PoolOwner

Fields

poolOwnerPoolId :: !PoolId

poolOwnerSlot :: !SlotNo

poolOwnerSlotInternalIndex :: !Word64

poolOwnerOwner :: !PoolOwner

poolOwnerIndex :: !Word8

type PoolOwnerId = Key PoolOwner

data PoolRegistration

Constructors

PoolRegistration

Fields

poolRegistrationPoolId :: !PoolId

poolRegistrationSlot :: !SlotNo

poolRegistrationSlotInternalIndex :: !Word64

poolRegistrationMarginNumerator :: !Word64

poolRegistrationMarginDenominator :: !Word64

poolRegistrationCost :: !Word64

poolRegistrationPledge :: !Word64

poolRegistrationMetadataUrl :: !(Maybe StakePoolMetadataUrl)

poolRegistrationMetadataHash :: !(Maybe StakePoolMetadataHash)

type PoolRegistrationId = Key PoolRegistration

data PoolDelistment

Constructors

PoolDelistment

Fields

delistedPoolId :: !PoolId

type PoolDelistmentId = Key PoolDelistment

data PoolRetirement

Constructors

PoolRetirement

Fields

poolRetirementPoolId :: !PoolId

poolRetirementSlot :: !SlotNo

poolRetirementSlotInternalIndex :: !Word64

poolRetirementEpoch :: !Word64

type PoolRetirementId = Key PoolRetirement

data PoolMetadata

Constructors

PoolMetadata

Fields

poolMetadataHash :: !StakePoolMetadataHash

poolMetadataName :: !Text

poolMetadataTicker :: !StakePoolTicker

poolMetadataDescription :: !(Maybe Text)

poolMetadataHomepage :: !Text

type PoolMetadataId = Key PoolMetadata

data PoolMetadataFetchAttempts

Constructors

PoolMetadataFetchAttempts

Fields

poolFetchAttemptsMetadataHash :: !StakePoolMetadataHash

poolFetchAttemptsMetadataUrl :: !StakePoolMetadataUrl

poolFetchAttemptsRetryAfter :: !UTCTime

poolFetchAttemptsRetryCount :: !Word8

type PoolMetadataFetchAttemptsId = Key PoolMetadataFetchAttempts

migrateAll :: Migration

entityDefListFormigrateAll :: [EntityDef]

fk_registration_pool_id :: PoolOwner -> Key PoolRegistration