Cardano.Pool.DB.Sqlite.Types

Description

This module contains instances and types necessary for storing wallets in a SQL database with Persistent.

It's in a separate module due to the GHC stage restriction.

The ToJSON/FromJSON and Read instance orphans exist due to class constraints on Persistent functions.

Documentation

sqlSettings' :: MkPersistSettings

Settings for generating the Persistent types.

newtype BlockId

Constructors

BlockId

Fields

getBlockId :: Hash "BlockHeader"

Orphan instances

Read PoolId

Instance details

Methods

readsPrec :: Int -> ReadS PoolId

readList :: ReadS [PoolId]

readPrec :: ReadPrec PoolId

readListPrec :: ReadPrec [PoolId]

FromJSON EpochNo

Instance details

Methods

parseJSON :: Value -> Parser EpochNo

parseJSONList :: Value -> Parser [EpochNo]

FromJSON SlotNo

Instance details

Methods

parseJSON :: Value -> Parser SlotNo

parseJSONList :: Value -> Parser [SlotNo]

FromJSON SlotId

Instance details

Methods

parseJSON :: Value -> Parser SlotId

parseJSONList :: Value -> Parser [SlotId]

FromJSON PoolId

Instance details

Methods

parseJSON :: Value -> Parser PoolId

parseJSONList :: Value -> Parser [PoolId]

ToJSON EpochNo

Instance details

Methods

toJSON :: EpochNo -> Value

toEncoding :: EpochNo -> Encoding

toJSONList :: [EpochNo] -> Value

toEncodingList :: [EpochNo] -> Encoding

ToJSON SlotNo

Instance details

Methods

toJSON :: SlotNo -> Value

toEncoding :: SlotNo -> Encoding

toJSONList :: [SlotNo] -> Value

toEncodingList :: [SlotNo] -> Encoding

ToJSON SlotId

Instance details

Methods

toJSON :: SlotId -> Value

toEncoding :: SlotId -> Encoding

toJSONList :: [SlotId] -> Value

toEncodingList :: [SlotId] -> Encoding

ToJSON PoolId

Instance details

Methods

toJSON :: PoolId -> Value

toEncoding :: PoolId -> Encoding

toJSONList :: [PoolId] -> Value

toEncodingList :: [PoolId] -> Encoding

PersistField SlotId

Instance details

Methods

toPersistValue :: SlotId -> PersistValue

fromPersistValue :: PersistValue -> Either Text SlotId

PersistField PoolId

Instance details

Methods

toPersistValue :: PoolId -> PersistValue

fromPersistValue :: PersistValue -> Either Text PoolId

PersistFieldSql SlotId

Instance details

Methods

sqlType :: Proxy SlotId -> SqlType

PersistFieldSql PoolId

Instance details

Methods

sqlType :: Proxy PoolId -> SqlType

FromHttpApiData SlotId

Instance details

Methods

parseUrlPiece :: Text -> Either Text SlotId

parseHeader :: ByteString -> Either Text SlotId

parseQueryParam :: Text -> Either Text SlotId

PathPiece SlotId

Instance details

Methods

fromPathPiece :: Text -> Maybe SlotId

toPathPiece :: SlotId -> Text

PathPiece PoolId

Instance details

Methods

fromPathPiece :: Text -> Maybe PoolId

toPathPiece :: PoolId -> Text

ToHttpApiData SlotId

Instance details

Methods

toUrlPiece :: SlotId -> Text

toEncodedUrlPiece :: SlotId -> Builder

toHeader :: SlotId -> ByteString

toQueryParam :: SlotId -> Text