Cardano.Pool.Metadata

Description

HTTP-client(s) for fetching stake pool metadata from remote servers (directly from pool operators, or from smash).

Fetch

fetchFromRemote :: Tracer IO StakePoolMetadataFetchLog -> [UrlBuilder] -> Manager -> PoolId -> StakePoolMetadataUrl -> StakePoolMetadataHash -> IO (Maybe StakePoolMetadata)

data StakePoolMetadataFetchLog

Constructors

MsgFetchPoolMetadata StakePoolMetadataHash URI

MsgFetchPoolMetadataSuccess StakePoolMetadataHash StakePoolMetadata

MsgFetchPoolMetadataFailure StakePoolMetadataHash String

MsgFetchPoolMetadataFallback URI Bool

MsgFetchSMASH URI

MsgFetchDelistedPoolsFailure String

MsgFetchDelistedPoolsSuccess [PoolId]

MsgFetchHealthCheckFailure String

MsgFetchHealthCheckSuccess HealthStatusSMASH

fetchDelistedPools :: Tracer IO StakePoolMetadataFetchLog -> URI -> Manager -> IO (Maybe [PoolId])

healthCheck :: Tracer IO StakePoolMetadataFetchLog -> URI -> Manager -> IO (Maybe HealthStatusSMASH)

Gets the health status from the SMASH server. Returns Nothing if the server is unreachable.

isHealthyStatus :: HealthStatusSMASH -> Bool

toHealthCheckSMASH :: Maybe HealthStatusSMASH -> HealthCheckSMASH

Convert the result of healthCheck, which represents the server response to our own HealthCheckSMASH type, which is a superset of it.

data HealthStatusSMASH

Parses the SMASH HealthCheck type from the SMASH API.

Constructors

HealthStatusSMASH

Fields

status :: Text

version :: Text

Construct URLs

type UrlBuilder = PoolId -> StakePoolMetadataUrl -> StakePoolMetadataHash -> Either HttpException URI

A type-alias to ease signatures

identityUrlBuilder :: UrlBuilder

Simply return a pool metadata url, unchanged

registryUrlBuilder :: URI -> UrlBuilder

Build a URL from a metadata hash compatible with an aggregation registry

re-exports

data Manager

Keeps track of open connections for keep-alive.

If possible, you should share a single Manager between multiple threads and requests.

Since 0.1.0

newManager :: MonadIO m => ManagerSettings -> m Manager

Create a connection manager that supports TLS connections.

defaultManagerSettings :: ManagerSettings

Some default settings, overriding some of the library's default with stricter values.

Types

newtype SMASHPoolId

TODO: import SMASH types

Constructors

SMASHPoolId

Fields

poolId :: Text