Cardano.Pool.Metrics

Description

This module can fold over a blockchain to collect metrics about Stake pools.

It interacts with: - Cardano.Wallet.Network which provides the chain - Cardano.Pool.DB - which can persist the metrics - Cardano.Wallet.Api.Server - which presents the results in an endpoint

Documentation

activityForEpoch :: EpochNo -> State -> Map PoolId Int

For a given epoch, and state, this function returns how many blocks each pool produced.

combineMetrics :: Map PoolId (Quantity "lovelace" Word64) -> Map PoolId (Quantity "block" Natural) -> Either ErrMetricsInconsistency (Map PoolId (Quantity "lovelace" Word64, Quantity "block" Natural))

Combines two different sources of data into one:

A stake-distribution map

A pool-production map

If a pool has produced a block without existing in the stake-distribution, i.e it exists in (2) but not (1), this function will return Left ErrMetricsInconsistencyBlockProducerNotInStakeDistr.

If a pool is in (1) but not (2), it simply means it has produced 0 blocks so far.

data ErrMetricsInconsistency

Constructors

ErrMetricsInconsistencyBlockProducerNotInStakeDistr PoolId (Quantity "block" Natural)

Helper

withinSameTip

Arguments

:: (MonadIO m, Eq header)

=> RetryPolicyM m
Retrying policy to command what to do in case the action wasn't executed within the same tip.

-> (ErrWithinSameTip -> m a)
On failure

-> m header
A getter for that header.

-> (header -> m a)
The action to run

-> m a

Runs an action by making sure that the network header hasn't changed before and after the action. This is useful in order to combine data from different sources and make sure that all data retrieval and computations happens on a same version.

The action must be retry-able (i.e. read-only) for it will be retried if the tip has changed before and after the call.

data ErrWithinSameTip

Constructors

ErrWithinSameTipMaxRetries
Retried too many times

data State

In-memory state keeping track of which pool produced blocks at which slots.

Constructors

State

Fields

tip :: BlockHeader

The blockHeader of the most recently applied block. Used to resume restoration from NetworkLayer.

activity :: Map PoolId [SlotId]

Mapping from pools to the slots where pool produced blocks.

This is needed internally to support rollback, but publicly, only the length of the SlotId-list is likely needed.

applyBlock :: (BlockHeader, PoolId) -> State -> State