Cardano.Pool.Performance

Description

This module defines necessary bits to compute an averaged apparent performance for stake pools. The definition is inspired from the "Design Specification for Delegation and Incentives in Cardano" by Kant et Al, yet deviates slightly from it because of different real constraints:

a) The wallet software doesn't have a reliable access to various epoch details such as the stake distribution for past epochs.

b) The wallet software makes therefore a significant use of the current epoch's data which represents unfinished snapshots of epochs which, because of their non-deterministic nature, may favor leaders elected early in the process more than those not yet elected.

To cope with these two issues, we try to harmonize past known epochs with the ongoing one by averaging over a whole range of epochs, simultaneously. This means computing a ratio of the total blocks produced by a leader on the total number of blocks we could reasonably expect a leader to have produced, across many epochs at once (i.e., doing a ratio of sums, instead of a sum of ratios).

This should be less punitive for leaders that have not yet been elected as part of the ongoing epoch while having still performed reasonably okay in the past.

Documentation

data EpochStats

Constructors

EpochStats

Fields

poolProduction :: !Natural

poolStake :: !Natural

totalStake :: !Natural

totalProduction :: !Natural

apparentPerformance

Arguments

:: [EpochStats]
Epoch statistics, for a given pool

-> Double
Average performance

readPoolsPerformances :: Monad m => DBLayer m -> EpochNo -> m (Map PoolId Double)

Read pool performances of many epochs from the database

Helpers

count :: Map k [a] -> Map k (Quantity any Word64)

Count elements inside a Map