Cardano.Pool.Rank.Likelihood

Description

This module provides a maximum likelhood estimator for pool performance.

Copy & paste of the original implementation in

cardano-ledgererasshelleyimplsrcCardanoLedgerShelleyPoolRank.hs

The copy was made in order to reduce (transitive) dependencies.

Pool performance estimate from historical block production

data BlockProduction

Information about block production of a pool in one epoch.

Constructors

BlockProduction

Fields

blocksProduced :: !Natural

Blocks produced in the given epoch.

stakeRelative :: !Rational

Relative stake of the pool that was relevant for block production. (i.e. from the "set" snapshot).

newtype PerformanceEstimate

This is a estimate of the proportion of allowed blocks a pool will make in the future. It is used for ranking pools in delegation.

Constructors

PerformanceEstimate

Fields

unPerformanceEstimate :: Double

estimatePoolPerformance

Arguments

:: SlottingParameters

-> DecentralizationLevel

-> Seq BlockProduction
Historical block production data. Most recent data comes first. Recent performance weighs more than past performance:

Block production from > 25 epochs ago has less than 10% influence on the likelihoods.

Block production from > 50 epochs ago has less than 1% influence on the likelihoods and can be ignored.

-> PerformanceEstimate

Estimate the performance of a pool from historical block production data.

Assumes that the SlottingParameters are constant through the given history.

Likelihood computations

newtype LogWeight

Constructors

LogWeight

Fields

unLogWeight :: Float

newtype Likelihood

Constructors

Likelihood

Fields

unLikelihood :: StrictSeq LogWeight

likelihood :: Natural -> Double -> EpochSize -> Likelihood

applyDecay :: Float -> Likelihood -> Likelihood

Decay previous likelihood

newtype Histogram

Constructors

Histogram

Fields

unHistogram :: StrictSeq LogWeight

percentile' :: Likelihood -> PerformanceEstimate