Cardano.Wallet

Description

Provides wallet layer functions that are used by API layer. Uses both Cardano.Wallet.DB and Cardano.Wallet.Network to realize its role as being intermediary between the three.

Functions of the wallet layer are often parameterized with variables following the convention below:

s: A state used to keep track of known addresses. Typically, possible values for this parameter are described in AddressDiscovery sub-modules. For instance SeqState or Rnd State.

k: A key derivation scheme intrinsically connected to the underlying discovery state s. This describes how the hierarchical structure of a wallet is defined as well as the relationship between secret keys and public addresses.

Development

Naming Conventions

Components inside a particular context ctx can be called via dedicated lenses (see Cardano.Wallet#Capabilities). These components are extracted from the context in a where clause according to the following naming convention:

db = ctx ^. dbLayer @s \k@ for the DBLayer.

tr = ctx ^. logger for the Logger.

nw = ctx ^. networkLayer for the NetworkLayer.

tl = ctx ^. transactionLayer \k@ for the TransactionLayer.

re = ctx ^. workerRegistry for the WorkerRegistry.

TroubleShooting

• Overlapping instances for HasType (DBLayer IO s k) ctx
    arising from a use of ‘myFunction’
  Matching instances:

Occurs when a particular function is missing a top-level constraint (because it uses another function that demands such constraint). Here, myFunction needs its surrounding context ctx to have a DBLayer but the constraint is missing from its host function.

Fix: Add "HasDBLayer s k" as a class-constraint to the surrounding function.

• Overlapping instances for HasType (DBLayer IO s t0 k0) ctx
    arising from a use of ‘myFunction’
  Matching givens (or their superclasses):

Occurs when a function is called in a context where type-level parameters can be inferred. Here, myFunction is called but it is unclear whether the parameter t0 and k0 of its context are the same as the ones from the function at the call-site.

Fix: Add type-applications at the call-site "myFunction @ctx @s \k@"

WalletLayer

data WalletLayer m s (k :: Depth -> Type -> Type)

Constructors

WalletLayer (Tracer m WalletWorkerLog) (Block, NetworkParameters) (NetworkLayer m Block) (TransactionLayer k SealedTx) (DBLayer m s k)

Capabilities

Each function in the wallet layer is defined in function of a non-specialized context ctx. That context may require some extra capabilities via class-constraints in the function signature. Capabilities are expressed in the form of a "HasXXX" class-constraints sometimes with extra type parameters.

For example:

listWallets
    :: forall ctx s k.
        ( HasDBLayer s k ctx
        )
    => ctx
    -> IO [WalletId]

Requires that the given context has an access to a database layer DBLayer parameterized over the wallet state, a network target and a key derivation scheme. Components are pulled from the context generically (i.e. the concrete ctx must derive Generic) using their associated type. The concrete ctx is therefore expected to be a product-type of all the necessary components.

One can build an interface using only a subset of the wallet layer capabilities and functions, for instance, something to fiddle with wallets and their metadata does not require any networking layer.

type HasDBLayer m s k = HasType (DBLayer m s k)

dbLayer :: forall m s k ctx. HasDBLayer m s k ctx => Lens' ctx (DBLayer m s k)

type HasLogger m msg = HasType (Tracer m msg)

logger :: forall m msg ctx. HasLogger m msg ctx => Lens' ctx (Tracer m msg)

type HasNetworkLayer m = HasType (NetworkLayer m Block)

This module is only interested in one block-, and tx-type. This constraint hides that choice, for some ease of use.

networkLayer :: forall m ctx. HasNetworkLayer m ctx => Lens' ctx (NetworkLayer m Block)

type HasTransactionLayer k = HasType (TransactionLayer k SealedTx)

transactionLayer :: forall k ctx. HasTransactionLayer k ctx => Lens' ctx (TransactionLayer k SealedTx)

type HasGenesisData = HasType (Block, NetworkParameters)

genesisData :: forall ctx. HasGenesisData ctx => Lens' ctx (Block, NetworkParameters)

Interface

Wallet

createWallet :: forall ctx m s k. (MonadUnliftIO m, MonadTime m, HasGenesisData ctx, HasDBLayer m s k ctx, IsOurs s Address, IsOurs s RewardAccount) => ctx -> WalletId -> WalletName -> s -> ExceptT ErrWalletAlreadyExists m WalletId

Initialise and store a new wallet, returning its ID.

createIcarusWallet :: forall ctx s k n. (HasGenesisData ctx, HasDBLayer IO s k ctx, PaymentAddress n k, k ~ IcarusKey, s ~ SeqState n k, Typeable n) => ctx -> WalletId -> WalletName -> (k 'RootK XPrv, Passphrase "encryption") -> ExceptT ErrWalletAlreadyExists IO WalletId

Initialise and store a new legacy Icarus wallet. These wallets are intrinsically sequential, but, in the incentivized testnet, we only have access to the a snapshot of the MainNet.

To work-around this, we scan the genesis block with an arbitrary big gap and resort to a default gap afterwards.

attachPrivateKeyFromPwd :: forall ctx s k. HasDBLayer IO s k ctx => ctx -> WalletId -> (k 'RootK XPrv, Passphrase "user") -> ExceptT ErrNoSuchWallet IO ()

attachPrivateKeyFromPwdHash :: forall ctx s k. HasDBLayer IO s k ctx => ctx -> WalletId -> (k 'RootK XPrv, PassphraseHash) -> ExceptT ErrNoSuchWallet IO ()

The hash here is the output of Scrypt function with the following parameters: - logN = 14 - r = 8 - p = 1 - bytesNumber = 64

getWalletUtxoSnapshot :: forall ctx s k. (HasDBLayer IO s k ctx, HasNetworkLayer IO ctx, HasTransactionLayer k ctx) => ctx -> WalletId -> ExceptT ErrNoSuchWallet IO [(TokenBundle, Coin)]

listUtxoStatistics :: forall ctx s k. HasDBLayer IO s k ctx => ctx -> WalletId -> ExceptT ErrListUTxOStatistics IO UTxOStatistics

List the wallet's UTxO statistics.

readWallet :: forall ctx s k. HasDBLayer IO s k ctx => ctx -> WalletId -> ExceptT ErrNoSuchWallet IO (Wallet s, WalletMetadata, Set Tx)

Retrieve the wallet state for the wallet with the given ID.

deleteWallet :: forall ctx s k. HasDBLayer IO s k ctx => ctx -> WalletId -> ExceptT ErrNoSuchWallet IO ()

Remove an existing wallet. Note that there's no particular work to be done regarding the restoration worker as it will simply terminate on the next tick when noticing that the corresponding wallet is gone.

restoreWallet :: forall ctx s k. (HasNetworkLayer IO ctx, HasDBLayer IO s k ctx, HasLogger IO WalletWorkerLog ctx, IsOurs s Address, IsOurs s RewardAccount, AddressBookIso s, MaybeLight s) => ctx -> WalletId -> ExceptT ErrNoSuchWallet IO ()

Restore a wallet from its current tip.

After the wallet has been restored, this action will continue to fetch newly created blocks and apply them, or roll back to a previous point whenever the chain switches.

updateWallet :: forall ctx s k. HasDBLayer IO s k ctx => ctx -> WalletId -> (WalletMetadata -> WalletMetadata) -> ExceptT ErrNoSuchWallet IO ()

Update a wallet's metadata with the given update function.

updateWalletPassphraseWithOldPassphrase :: forall ctx s k. (HasDBLayer IO s k ctx, WalletKey k) => ctx -> WalletId -> (Passphrase "user", Passphrase "user") -> ExceptT ErrUpdatePassphrase IO ()

Change a wallet's passphrase to the given passphrase.

updateWalletPassphraseWithMnemonic :: forall ctx s k. HasDBLayer IO s k ctx => ctx -> WalletId -> (k 'RootK XPrv, Passphrase "user") -> ExceptT ErrUpdatePassphrase IO ()

walletSyncProgress :: forall ctx s. HasNetworkLayer IO ctx => ctx -> Wallet s -> IO SyncProgress

fetchRewardBalance :: forall ctx s k. HasDBLayer IO s k ctx => ctx -> WalletId -> IO Coin

Fetch the cached reward balance of a given wallet from the database.

manageRewardBalance :: forall ctx s k (n :: NetworkDiscriminant). (HasLogger IO WalletWorkerLog ctx, HasNetworkLayer IO ctx, HasDBLayer IO s k ctx, Typeable s, Typeable n) => Proxy n -> ctx -> WalletId -> IO ()

rollbackBlocks :: forall ctx s k. HasDBLayer IO s k ctx => ctx -> WalletId -> Slot -> ExceptT ErrNoSuchWallet IO ChainPoint

Rewind the UTxO snapshots, transaction history and other information to a the earliest point in the past that is before or is the point of rollback.

checkWalletIntegrity :: forall ctx s k. HasDBLayer IO s k ctx => ctx -> WalletId -> GenesisParameters -> ExceptT ErrCheckWalletIntegrity IO ()

Check whether a wallet is in good shape when restarting a worker.

readNextWithdrawal :: forall ctx k. (HasTransactionLayer k ctx, HasNetworkLayer IO ctx) => ctx -> AnyCardanoEra -> Coin -> IO Coin

Read the current withdrawal capacity of a wallet. Note that, this simply returns 0 if:

a) There's no reward account for this type of wallet. b) The current reward value is too small to be considered (adding it would cost more than its value).

readRewardAccount :: forall ctx s k (n :: NetworkDiscriminant) shelley. (HasDBLayer IO s k ctx, shelley ~ SeqState n ShelleyKey, Typeable n, Typeable s) => ctx -> WalletId -> ExceptT ErrReadRewardAccount IO (RewardAccount, XPub, NonEmpty DerivationIndex)

someRewardAccount :: ToRewardAccount k => SomeMnemonic -> (XPrv, RewardAccount, NonEmpty DerivationIndex)

readPolicyPublicKey :: forall ctx s k (n :: NetworkDiscriminant) shelley. (HasDBLayer IO s k ctx, shelley ~ SeqState n ShelleyKey, Typeable n, Typeable s) => ctx -> WalletId -> ExceptT ErrReadPolicyPublicKey IO (XPub, NonEmpty DerivationIndex)

writePolicyPublicKey :: forall ctx s (n :: NetworkDiscriminant). (HasDBLayer IO s ShelleyKey ctx, s ~ SeqState n ShelleyKey) => ctx -> WalletId -> Passphrase "user" -> ExceptT ErrWritePolicyPublicKey IO (ShelleyKey 'PolicyK XPub)

queryRewardBalance :: forall ctx. HasNetworkLayer IO ctx => ctx -> RewardAccount -> ExceptT ErrFetchRewards IO Coin

Query the node for the reward balance of a given wallet.

Rather than force all callers of readWallet to wait for fetching the account balance (via the NetworkLayer), we expose this function for it.

newtype ErrWalletAlreadyExists

Forbidden operation was executed on an already existing wallet

Constructors

ErrWalletAlreadyExists WalletId

newtype ErrNoSuchWallet

Can't perform given operation because there's no wallet

Constructors

ErrNoSuchWallet WalletId

newtype ErrListUTxOStatistics

Errors that can occur when listing UTxO statistics.

Constructors

ErrListUTxOStatisticsNoSuchWallet ErrNoSuchWallet

data ErrUpdatePassphrase

Errors that can occur when trying to change a wallet's passphrase.

Constructors

ErrUpdatePassphraseNoSuchWallet ErrNoSuchWallet

ErrUpdatePassphraseWithRootKey ErrWithRootKey

newtype ErrFetchRewards

Errors that can occur when fetching the reward balance of a wallet

Constructors

ErrFetchRewardsReadRewardAccount ErrReadRewardAccount

data ErrCheckWalletIntegrity

Constructors

ErrCheckWalletIntegrityNoSuchWallet ErrNoSuchWallet

ErrCheckIntegrityDifferentGenesis (Hash "Genesis") (Hash "Genesis")

newtype ErrWalletNotResponding

Can't perform given operation because the wallet died.

Constructors

ErrWalletNotResponding WalletId

data ErrReadRewardAccount

Constructors

ErrReadRewardAccountNotAShelleyWallet

ErrReadRewardAccountNoSuchWallet ErrNoSuchWallet

data ErrReadPolicyPublicKey

Constructors

ErrReadPolicyPublicKeyNotAShelleyWallet

ErrReadPolicyPublicKeyNoSuchWallet ErrNoSuchWallet

ErrReadPolicyPublicKeyAbsent

data ErrWritePolicyPublicKey

Constructors

ErrWritePolicyPublicKeyNoSuchWallet ErrNoSuchWallet

ErrWritePolicyPublicKeyWithRootKey ErrWithRootKey

data ErrGetPolicyId

Errors that can occur when getting policy id.

Constructors

ErrGetPolicyIdReadPolicyPubliKey ErrReadPolicyPublicKey

ErrGetPolicyIdWrongMintingBurningTemplate

Shared Wallet

updateCosigner :: forall ctx s k n. (s ~ SharedState n k, MkKeyFingerprint k Address, SoftDerivation k, Typeable n, WalletKey k, HasDBLayer IO s k ctx, k ~ SharedKey) => ctx -> WalletId -> k 'AccountK XPub -> Cosigner -> CredentialType -> ExceptT ErrAddCosignerKey IO ()

data ErrAddCosignerKey

Constructors

ErrAddCosignerKeyNoSuchWallet ErrNoSuchWallet
The shared wallet doesn't exist?

ErrAddCosignerKey ErrAddCosigner
Error adding this co-signer to the shared wallet.

data ErrConstructSharedWallet

Constructors

ErrConstructSharedWalletWrongScriptTemplate ErrScriptTemplate
The shared wallet' script template doesn't pass validation

ErrConstructSharedWalletInvalidIndex (ErrInvalidDerivationIndex 'Hardened 'AccountK)
User provided a derivation index outside of the Hard domain

normalizeSharedAddress :: forall n k. (SupportsDiscovery n k, k ~ SharedKey) => SharedState n k -> Address -> Maybe Address

constructSharedTransaction :: forall ctx s k (n :: NetworkDiscriminant). (HasTransactionLayer k ctx, HasDBLayer IO s k ctx, HasNetworkLayer IO ctx, k ~ SharedKey, s ~ SharedState n k) => ctx -> WalletId -> AnyCardanoEra -> TransactionCtx -> SelectionOf TxOut -> ExceptT ErrConstructTx IO SealedTx

Construct an unsigned transaction from a given selection for a shared wallet.

Address

createRandomAddress :: forall ctx s k n. (HasDBLayer IO s k ctx, PaymentAddress n k, RndStateLike s, k ~ ByronKey, AddressBookIso s) => ctx -> WalletId -> Passphrase "user" -> Maybe (Index 'Hardened 'AddressK) -> ExceptT ErrCreateRandomAddress IO (Address, NonEmpty DerivationIndex)

importRandomAddresses :: forall ctx s k. (HasDBLayer IO s k ctx, RndStateLike s, k ~ ByronKey, AddressBookIso s) => ctx -> WalletId -> [Address] -> ExceptT ErrImportRandomAddress IO ()

listAddresses

Arguments

:: forall ctx s k. (HasDBLayer IO s k ctx, CompareDiscovery s, KnownAddresses s)

=> ctx

-> WalletId

-> (s -> Address -> Maybe Address)
A function to normalize address, so that delegated addresses non-delegation addresses found in the transaction history are shown with their delegation settings. Use Just for wallet without delegation settings.

-> ExceptT ErrNoSuchWallet IO [(Address, AddressState, NonEmpty DerivationIndex)]

List all addresses of a wallet with their metadata. Addresses are ordered from the most-recently-discovered to the oldest known.

normalizeDelegationAddress :: forall s k n. (DelegationAddress n k, s ~ SeqState n k) => s -> Address -> Maybe Address

lookupTxIns :: forall ctx s k. (HasDBLayer IO s k ctx, IsOurs s Address) => ctx -> WalletId -> [TxIn] -> ExceptT ErrDecodeTx IO [(TxIn, Maybe (TxOut, NonEmpty DerivationIndex))]

lookupTxOuts :: forall ctx s k. (HasDBLayer IO s k ctx, IsOurs s Address) => ctx -> WalletId -> [TxOut] -> ExceptT ErrDecodeTx IO [(TxOut, Maybe (NonEmpty DerivationIndex))]

data ErrCreateRandomAddress

Constructors

ErrIndexAlreadyExists (Index 'Hardened 'AddressK)

ErrCreateAddrNoSuchWallet ErrNoSuchWallet

ErrCreateAddrWithRootKey ErrWithRootKey

ErrCreateAddressNotAByronWallet

data ErrImportRandomAddress

Constructors

ErrImportAddrNoSuchWallet ErrNoSuchWallet

ErrImportAddr ErrImportAddress

ErrImportAddressNotAByronWallet

newtype ErrImportAddress

Constructors

ErrAddrDoesNotBelong Address

newtype ErrDecodeTx

Errors that can occur when decoding a transaction.

Constructors

ErrDecodeTxNoSuchWallet ErrNoSuchWallet

Payment

getTxExpiry

Arguments

:: TimeInterpreter (ExceptT PastHorizonException IO)
Context for time to slot calculation.

-> Maybe NominalDiffTime
Time to live (TTL) in seconds from now.

-> IO SlotNo

Calculate the transaction expiry slot, given a TimeInterpreter, and an optional TTL in seconds.

If no TTL is provided, a default of 2 hours is used (note: there is no particular reason why we chose that duration).

data SelectAssetsParams s result

Parameters for the selectAssets function.

Constructors

SelectAssetsParams

Fields

outputs :: [TxOut]

pendingTxs :: Set Tx

randomSeed :: Maybe StdGenSeed

txContext :: TransactionCtx

utxoAvailableForCollateral :: Map WalletUTxO TokenBundle

utxoAvailableForInputs :: UTxOSelection WalletUTxO

wallet :: Wallet s

selectionStrategy :: SelectionStrategy

Specifies which selection strategy to use. See SelectionStrategy.

selectAssets :: forall ctx m s k result. (BoundedAddressLength k, HasTransactionLayer k ctx, HasLogger m WalletWorkerLog ctx, MonadRandom m) => ctx -> AnyCardanoEra -> ProtocolParameters -> SelectAssetsParams s result -> (s -> Selection -> result) -> ExceptT ErrSelectAssets m result

Selects assets from a wallet.

This function has the following responsibilities:

selecting inputs from the UTxO set to pay for user-specified outputs;

selecting inputs from the UTxO set to pay for collateral;

producing change outputs to return excess value to the wallet;

balancing a selection to pay for the transaction fee.

When selecting inputs to pay for user-specified outputs, inputs are selected randomly.

By default, the seed used for random selection is derived automatically, from the given MonadRandom context.

However, if a concrete value is specified for the optional $sel:randomSeed:SelectAssetsParams parameter, then that value will be used instead as the seed for random selection.

readWalletUTxOIndex :: forall ctx s k. HasDBLayer IO s k ctx => ctx -> WalletId -> ExceptT ErrNoSuchWallet IO (UTxOIndex WalletUTxO, Wallet s, Set Tx)

Read a wallet checkpoint and index its UTxO, for selectAssets and selectAssetsNoOutputs.

assignChangeAddresses :: forall s. GenChange s => ArgGenChange s -> Selection -> s -> (SelectionOf TxOut, s)

Augments the given outputs with new outputs. These new outputs correspond to change outputs to which new addresses have been assigned. This updates the wallet state as it needs to keep track of new pending change addresses.

assignChangeAddressesAndUpdateDb :: forall ctx s k. (GenChange s, HasDBLayer IO s k ctx, AddressBookIso s) => ctx -> WalletId -> ArgGenChange s -> Selection -> ExceptT ErrSignPayment IO (SelectionOf TxOut)

assignChangeAddressesWithoutDbUpdate :: forall ctx s k. (GenChange s, HasDBLayer IO s k ctx) => ctx -> WalletId -> ArgGenChange s -> Selection -> ExceptT ErrConstructTx IO (SelectionOf TxOut)

selectionToUnsignedTx :: forall s input output change withdrawal. (IsOurs s Address, input ~ (TxIn, TxOut, NonEmpty DerivationIndex), output ~ TxOut, change ~ TxChange (NonEmpty DerivationIndex), withdrawal ~ (RewardAccount, Coin, NonEmpty DerivationIndex)) => Withdrawal -> SelectionOf TxOut -> s -> UnsignedTx input output change withdrawal

buildAndSignTransaction

Arguments

:: forall ctx s k. (HasTransactionLayer k ctx, HasDBLayer IO s k ctx, HasNetworkLayer IO ctx, IsOwned s k)

=> ctx

-> WalletId

-> AnyCardanoEra

-> ((k 'RootK XPrv, Passphrase "encryption") -> (XPrv, Passphrase "encryption"))
Reward account derived from the root key (or somewhere else).

-> Passphrase "user"

-> TransactionCtx

-> SelectionOf TxOut

-> ExceptT ErrSignPayment IO (Tx, TxMeta, UTCTime, SealedTx)

Produce witnesses and construct a transaction from a given selection.

Requires the encryption passphrase in order to decrypt the root private key. Note that this doesn't broadcast the transaction to the network. In order to do so, use submitTx.

signTransaction

Arguments

:: forall k. (WalletKey k, HardDerivation k, Bounded (Index (AddressIndexDerivationType k) 'AddressK))

=> TransactionLayer k SealedTx
The way to interact with the wallet backend

-> AnyCardanoEra
Preferred latest era

-> (Address -> Maybe (k 'AddressK XPrv, Passphrase "encryption"))
The wallets address-key lookup function

-> (k 'RootK XPrv, Passphrase "encryption")
The root key of the wallet

-> UTxO
The total UTxO set of the wallet (i.e. if pending transactions all applied).

-> SealedTx
The transaction to sign

-> SealedTx
The original transaction, with additional signatures added where necessary

constructTransaction :: forall ctx s k (n :: NetworkDiscriminant). (HasTransactionLayer k ctx, HasDBLayer IO s k ctx, HasNetworkLayer IO ctx, Typeable s, Typeable n) => ctx -> WalletId -> AnyCardanoEra -> TransactionCtx -> SelectionOf TxOut -> ExceptT ErrConstructTx IO SealedTx

Construct an unsigned transaction from a given selection.

constructTxMeta :: forall ctx s k. HasDBLayer IO s k ctx => ctx -> WalletId -> TransactionCtx -> [(TxIn, Coin)] -> [TxOut] -> ExceptT ErrSubmitTransaction IO TxMeta

data ErrSelectAssets

Constructors

ErrSelectAssetsPrepareOutputsError (SelectionOutputError WalletSelectionContext)

ErrSelectAssetsNoSuchWallet ErrNoSuchWallet

ErrSelectAssetsAlreadyWithdrawing Tx

ErrSelectAssetsSelectionError (SelectionError WalletSelectionContext)

data ErrSignPayment

Errors that can occur when signing a transaction.

Constructors

ErrSignPaymentMkTx ErrMkTransaction

ErrSignPaymentNoSuchWallet ErrNoSuchWallet

ErrSignPaymentWithRootKey ErrWithRootKey

ErrSignPaymentIncorrectTTL PastHorizonException

data ErrNotASequentialWallet

Constructors

ErrNotASequentialWallet

data ErrWithdrawalNotWorth

Constructors

ErrWithdrawalNotWorth

data ErrConstructTx

Errors that can occur when constructing an unsigned transaction.

Constructors

ErrConstructTxWrongPayload

ErrConstructTxBody ErrMkTransaction

ErrConstructTxNoSuchWallet ErrNoSuchWallet

ErrConstructTxReadRewardAccount ErrReadRewardAccount

ErrConstructTxIncorrectTTL PastHorizonException

ErrConstructTxMultidelegationNotSupported

ErrConstructTxMultiaccountNotSupported

ErrConstructTxWrongMintingBurningTemplate

ErrConstructTxAssetNameTooLong

ErrConstructTxMintOrBurnAssetQuantityOutOfBounds

ErrConstructTxWrongValidityBounds

ErrConstructTxValidityIntervalNotWithinScriptTimelock

ErrConstructTxSharedWalletPending

ErrConstructTxNotImplemented String
Temporary error constructor.

newtype ErrMintBurnAssets

Constructors

ErrMintBurnNotImplemented Text
Temporary error constructor.

data ErrBalanceTx

Errors that can occur when balancing transaction.

Constructors

ErrBalanceTxUpdateError ErrUpdateSealedTx

ErrBalanceTxSelectAssets ErrSelectAssets

ErrBalanceTxMaxSizeLimitExceeded

ErrBalanceTxExistingCollateral

ErrBalanceTxExistingTotalCollateral

ErrBalanceTxExistingReturnCollateral

ErrBalanceTxConflictingNetworks

ErrBalanceTxAssignRedeemers ErrAssignRedeemers

ErrBalanceTxInternalError ErrBalanceTxInternalError

ErrBalanceTxZeroAdaOutput

ErrByronTxNotSupported

data ErrBalanceTxInternalError

Constructors

ErrUnderestimatedFee Coin SealedTx

ErrFailedBalancing Value

newtype ErrUpdateSealedTx

Constructors

ErrExistingKeyWitnesses Int
The SealedTx couldn't not be updated because the *n* existing key-witnesses would have been rendered invalid.

data ErrCannotJoin

Constructors

ErrAlreadyDelegating PoolId

ErrNoSuchPool PoolId

data ErrCannotQuit

Constructors

ErrNotDelegatingOrAboutTo

ErrNonNullRewards Coin

data ErrSubmitTransaction

Errors that can occur when submitting a transaction.

Constructors

ErrSubmitTransactionNoSuchWallet ErrNoSuchWallet

ErrSubmitTransactionForeignWallet

ErrSubmitTransactionPartiallySignedOrNoSignedTx Int Int

ErrSubmitTransactionMultidelegationNotSupported

Migration

createMigrationPlan :: forall ctx k s. (HasDBLayer IO s k ctx, HasNetworkLayer IO ctx, HasTransactionLayer k ctx) => ctx -> AnyCardanoEra -> WalletId -> Withdrawal -> ExceptT ErrCreateMigrationPlan IO MigrationPlan

migrationPlanToSelectionWithdrawals :: MigrationPlan -> Withdrawal -> NonEmpty Address -> Maybe (NonEmpty (SelectionWithoutChange, Withdrawal))

type SelectionWithoutChange = SelectionOf Void

data ErrCreateMigrationPlan

Constructors

ErrCreateMigrationPlanEmpty

ErrCreateMigrationPlanNoSuchWallet ErrNoSuchWallet

Delegation

data PoolRetirementEpochInfo

Constructors

PoolRetirementEpochInfo

Fields

currentEpoch :: EpochNo

The current epoch.

retirementEpoch :: EpochNo

The retirement epoch of a pool.

joinStakePool

Arguments

:: forall ctx s k. (HasDBLayer IO s k ctx, HasNetworkLayer IO ctx, HasLogger IO WalletWorkerLog ctx)

=> ctx

-> EpochNo

-> Set PoolId

-> PoolId

-> PoolLifeCycleStatus

-> WalletId

-> ExceptT ErrStakePoolDelegation IO (DelegationAction, Maybe Coin)
snd is the deposit

quitStakePool :: forall ctx s k. HasDBLayer IO s k ctx => ctx -> WalletId -> Withdrawal -> ExceptT ErrStakePoolDelegation IO DelegationAction

Helper function to factor necessary logic for quitting a stake pool.

guardJoin :: Set PoolId -> WalletDelegation -> PoolId -> Maybe PoolRetirementEpochInfo -> Either ErrCannotJoin ()

guardQuit :: WalletDelegation -> Withdrawal -> Coin -> Either ErrCannotQuit ()

data ErrStakePoolDelegation

Constructors

ErrStakePoolDelegationNoSuchWallet ErrNoSuchWallet

ErrStakePoolJoin ErrCannotJoin

ErrStakePoolQuit ErrCannotQuit

Fee Estimation

data FeeEstimation

Result of a fee estimation process given a wallet and payment order.

Constructors

FeeEstimation

Fields

estMinFee :: Word64

Most coin selections will result in a fee higher than this.

estMaxFee :: Word64

Most coin selections will result in a fee lower than this.

estimateFee :: forall m. Monad m => ExceptT ErrSelectAssets m Coin -> ExceptT ErrSelectAssets m FeeEstimation

Estimate the transaction fee for a given coin selection algorithm by repeatedly running it (100 times) and collecting the results. In the returned FeeEstimation, the minimum fee is that which 90% of the sampled fees are greater than. The maximum fee is the highest fee observed in the samples.

calcMinimumDeposit :: forall ctx s k. (HasDBLayer IO s k ctx, HasNetworkLayer IO ctx) => ctx -> WalletId -> ExceptT ErrSelectAssets IO Coin

Calculate the minimum deposit necessary if a given wallet wanted to delegate to a pool. Said differently, this return either 0, or the value of the key deposit protocol parameters if the wallet has no registered stake key.

calcMinimumCoinValues :: forall ctx k f. (HasTransactionLayer k ctx, HasNetworkLayer IO ctx, Applicative f) => ctx -> AnyCardanoEra -> f TxOut -> IO (f Coin)

Calculate the minimum coin values required for a bunch of specified outputs.

Transaction

forgetTx :: forall ctx s k. HasDBLayer IO s k ctx => ctx -> WalletId -> Hash "Tx" -> ExceptT ErrRemoveTx IO ()

Remove a pending or expired transaction from the transaction history. This happens at the request of the user. If the transaction is already on chain, or is missing from the transaction history, an error will be returned.

If a Pending transaction is removed, but later appears in a block, it will be added back to the transaction history.

listTransactions :: forall ctx s k. (HasDBLayer IO s k ctx, HasNetworkLayer IO ctx) => ctx -> WalletId -> Maybe Coin -> Maybe UTCTime -> Maybe UTCTime -> SortOrder -> ExceptT ErrListTransactions IO [TransactionInfo]

List all transactions and metadata from history for a given wallet.

listAssets :: forall s k ctx. (HasDBLayer IO s k ctx, IsOurs s Address) => ctx -> WalletId -> ExceptT ErrNoSuchWallet IO (Set AssetId)

Extract assets associated with a given wallet from its transaction history.

getTransaction :: forall ctx s k. HasDBLayer IO s k ctx => ctx -> WalletId -> Hash "Tx" -> ExceptT ErrGetTransaction IO TransactionInfo

Get transaction and metadata from history for a given wallet.

submitExternalTx :: forall ctx k. (HasNetworkLayer IO ctx, HasTransactionLayer k ctx, HasLogger IO TxSubmitLog ctx) => ctx -> SealedTx -> ExceptT ErrPostTx IO Tx

Broadcast an externally-signed transaction to the network.

NOTE: external transactions will not be added to the LocalTxSubmission pool, so the user must retry submission themselves.

submitTx :: forall ctx s k. (HasNetworkLayer IO ctx, HasDBLayer IO s k ctx, HasLogger IO WalletWorkerLog ctx) => ctx -> WalletId -> (Tx, TxMeta, SealedTx) -> ExceptT ErrSubmitTx IO ()

Broadcast a (signed) transaction to the network.

balanceTransaction :: forall era m s k ctx. (HasTransactionLayer k ctx, GenChange s, MonadRandom m, HasLogger m WalletWorkerLog ctx, IsShelleyBasedEra era, BoundedAddressLength k) => ctx -> ArgGenChange s -> (ProtocolParameters, ProtocolParameters) -> TimeInterpreter (Either PastHorizonException) -> (UTxOIndex WalletUTxO, Wallet s, Set Tx) -> PartialTx era -> ExceptT ErrBalanceTx m (Tx era)

data PartialTx era

A PartialTx is an an unbalanced SealedTx along with the necessary information to balance it.

The $sel:inputs:PartialTx and $sel:redeemers:PartialTx must match the binary transaction contained in the sealedTx.

Constructors

PartialTx

Fields

tx :: Tx era

inputs :: [(TxIn, TxOut, Maybe (Hash "Datum"))]

redeemers :: [Redeemer]

data LocalTxSubmissionConfig

Parameters for runLocalTxSubmissionPool

Constructors

LocalTxSubmissionConfig

Fields

rateLimit :: DiffTime

Minimum time between checks of pending transactions

blockInterval :: Word64

Resubmission interval, in terms of expected blocks.

defaultLocalTxSubmissionConfig :: LocalTxSubmissionConfig

The current default is to resubmit any pending transaction about once every 10 blocks.

The default rate limit for checking the pending list is 1000ms.

runLocalTxSubmissionPool :: forall ctx s k m. (MonadUnliftIO m, MonadMonotonicTime m, HasLogger IO WalletWorkerLog ctx, HasNetworkLayer m ctx, HasDBLayer m s k ctx) => LocalTxSubmissionConfig -> ctx -> WalletId -> m ()

Continuous process which monitors the chain tip and retries submission of pending transactions as the chain lengthens.

Regardless of the frequency of chain updates, this function won't re-query the database faster than the configured $sel:rateLimit:LocalTxSubmissionConfig.

This only exits if the network layer watchNodeTip function exits.

data ErrMkTransaction

Constructors

ErrMkTransactionNoSuchWallet WalletId

ErrMkTransactionTxBodyError Text
We failed to construct a transaction for some reasons.

ErrMkTransactionInvalidEra AnyCardanoEra
Should never happen, means that that we have programmatically provided an invalid era.

ErrMkTransactionJoinStakePool ErrCannotJoin

ErrMkTransactionQuitStakePool ErrCannotQuit

ErrMkTransactionIncorrectTTL PastHorizonException

data ErrSubmitTx

Errors that can occur when submitting a signed transaction to the network.

Constructors

ErrSubmitTxNetwork ErrPostTx

ErrSubmitTxNoSuchWallet ErrNoSuchWallet

ErrSubmitTxImpossible ErrNoSuchTransaction

data ErrRemoveTx

Can't remove pending or expired transaction.

Constructors

ErrRemoveTxNoSuchWallet ErrNoSuchWallet

ErrRemoveTxNoSuchTransaction ErrNoSuchTransaction

ErrRemoveTxAlreadyInLedger (Hash "Tx")

newtype ErrPostTx

Error while trying to send a transaction

Constructors

ErrPostTxValidationError Text

data ErrListTransactions

Errors that can occur when trying to list transactions.

Constructors

ErrListTransactionsNoSuchWallet ErrNoSuchWallet

ErrListTransactionsStartTimeLaterThanEndTime ErrStartTimeLaterThanEndTime

ErrListTransactionsMinWithdrawalWrong

ErrListTransactionsPastHorizonException PastHorizonException

data ErrGetTransaction

Errors that can occur when trying to get transaction.

Constructors

ErrGetTransactionNoSuchWallet ErrNoSuchWallet

ErrGetTransactionNoSuchTransaction ErrNoSuchTransaction

data ErrNoSuchTransaction

Indicates that the specified transaction hash is not found in the transaction history of the given wallet.

Constructors

ErrNoSuchTransaction WalletId (Hash "Tx")

data ErrStartTimeLaterThanEndTime

Indicates that the specified start time is later than the specified end time.

Constructors

ErrStartTimeLaterThanEndTime

Fields

errStartTime :: UTCTime

errEndTime :: UTCTime

data ErrWitnessTx

Errors that can occur when signing a transaction.

Constructors

ErrWitnessTxSignTx ErrSignTx

ErrWitnessTxNoSuchWallet ErrNoSuchWallet

ErrWitnessTxWithRootKey ErrWithRootKey

ErrWitnessTxIncorrectTTL PastHorizonException

Root Key

withRootKey :: forall ctx s k e a. HasDBLayer IO s k ctx => ctx -> WalletId -> Passphrase "user" -> (ErrWithRootKey -> e) -> (k 'RootK XPrv -> PassphraseScheme -> ExceptT e IO a) -> ExceptT e IO a

Execute an action which requires holding a root XPrv.

withRootKey takes a callback function with two arguments:

The encrypted root private key itself

The underlying passphrase scheme (legacy or new)

Caller are then expected to use preparePassphrase with the given scheme in order to "prepare" the passphrase to be used by other function. This does nothing for the new encryption, but for the legacy encryption with Scrypt, passphrases needed to first be CBOR serialized and blake2b_256 hashed.

@@ withRootKey ctx s k ctx wid pwd OnError $ xprv scheme -> changePassphrase (preparePassphrase scheme pwd) newPwd xprv @@@

derivePublicKey :: forall ctx s k. (HasDBLayer IO s k ctx, SoftDerivation k, GetAccount s k) => ctx -> WalletId -> Role -> DerivationIndex -> ExceptT ErrDerivePublicKey IO (k 'AddressK XPub)

getAccountPublicKeyAtIndex :: forall ctx s k. (HasDBLayer IO s k ctx, WalletKey k, GetPurpose k) => ctx -> WalletId -> Passphrase "user" -> DerivationIndex -> Maybe DerivationIndex -> ExceptT ErrReadAccountPublicKey IO (k 'AccountK XPub)

Retrieve any public account key of a wallet.

readAccountPublicKey :: forall ctx s k. (HasDBLayer IO s k ctx, GetAccount s k) => ctx -> WalletId -> ExceptT ErrReadAccountPublicKey IO (k 'AccountK XPub)

Retrieve current public account key of a wallet.

signMetadataWith :: forall ctx s k n. (HasDBLayer IO s k ctx, HardDerivation k, AddressIndexDerivationType k ~ 'Soft, WalletKey k, s ~ SeqState n k) => ctx -> WalletId -> Passphrase "user" -> (Role, DerivationIndex) -> TxMetadata -> ExceptT ErrSignMetadataWith IO (Signature TxMetadata)

Sign an arbitrary transaction metadata object with a private key belonging to the wallet's account.

This is experimental, and will likely be replaced by a more robust to arbitrary message signing using COSE, or a subset of it.

data ErrWithRootKey

Errors that can occur when trying to perform an operation on a wallet that requires a private key, but where none is attached to the wallet.

Constructors

ErrWithRootKeyNoRootKey WalletId

ErrWithRootKeyWrongPassphrase WalletId ErrWrongPassphrase

ErrWithRootKeyWrongMnemonic WalletId

data ErrWrongPassphrase

Indicate a failure when checking for a given Passphrase match

Constructors

ErrWrongPassphrase

ErrPassphraseSchemeUnsupported PassphraseScheme

data ErrSignMetadataWith

Constructors

ErrSignMetadataWithRootKey ErrWithRootKey
The wallet exists, but there's no root key attached to it

ErrSignMetadataWithNoSuchWallet ErrNoSuchWallet
The wallet doesn't exist?

ErrSignMetadataWithInvalidIndex (ErrInvalidDerivationIndex 'Soft 'AddressK)
User provided a derivation index outside of the Soft domain

data ErrDerivePublicKey

Constructors

ErrDerivePublicKeyNoSuchWallet ErrNoSuchWallet
The wallet doesn't exist?

ErrDerivePublicKeyInvalidIndex (ErrInvalidDerivationIndex 'Soft 'AddressK)
User provided a derivation index outside of the Soft domain

data ErrReadAccountPublicKey

Constructors

ErrReadAccountPublicKeyNoSuchWallet ErrNoSuchWallet
The wallet doesn't exist?

ErrReadAccountPublicKeyInvalidAccountIndex (ErrInvalidDerivationIndex 'Hardened 'AccountK)
User provided a derivation index for account outside of the Hard domain

ErrReadAccountPublicKeyInvalidPurposeIndex (ErrInvalidDerivationIndex 'Hardened 'PurposeK)
User provided a derivation index for purpose outside of the Hard domain

ErrReadAccountPublicKeyRootKey ErrWithRootKey
The wallet exists, but there's no root key attached to it

data ErrInvalidDerivationIndex derivation level

Constructors

ErrIndexOutOfBound (Index derivation level) (Index derivation level) DerivationIndex

Utilities

throttle :: (MonadUnliftIO m, MonadMonotonicTime m) => DiffTime -> (Time -> a -> m ()) -> m (a -> m ())

Return a function to run an action at most once every _interval_.

guardHardIndex :: Monad m => DerivationIndex -> ExceptT (ErrInvalidDerivationIndex 'Hardened level) m (Index 'Hardened whatever)

withNoSuchWallet :: Monad m => WalletId -> m (Maybe a) -> ExceptT ErrNoSuchWallet m a

posAndNegFromCardanoValue :: Value -> (TokenBundle, TokenBundle)

Convert a Value into a positive and negative component. Useful to convert the potentially negative balance of a partial tx into TokenBundles.

Logging

data WalletWorkerLog

Log messages for actions running within a wallet worker context.

Constructors

MsgWallet WalletLog

MsgWalletFollow WalletFollowLog

MsgChainFollow ChainFollowLog

data WalletFollowLog

Log messages arising from the restore and follow process.

Constructors

MsgDiscoveredDelegationCert SlotNo DelegationCertificate

MsgCheckpoint BlockHeader

MsgDiscoveredTxs [(Tx, TxMeta)]

MsgDiscoveredTxsContent [(Tx, TxMeta)]

data WalletLog

Log messages from API server actions running in a wallet worker context.

Constructors

MsgSelectionStart Int [TxOut]

MsgSelectionForBalancingStart Int (BuildableInAnyEra PartialTx)

MsgSelectionError (SelectionError WalletSelectionContext)

MsgSelectionReportSummarized SelectionReportSummarized

MsgSelectionReportDetailed SelectionReportDetailed

MsgMigrationUTxOBefore UTxOStatistics

MsgMigrationUTxOAfter UTxOStatistics

MsgRewardBalanceQuery BlockHeader

MsgRewardBalanceResult (Either ErrFetchRewards Coin)

MsgRewardBalanceNoSuchWallet ErrNoSuchWallet

MsgRewardBalanceExited

MsgTxSubmit TxSubmitLog

MsgIsStakeKeyRegistered Bool

data TxSubmitLog

Constructors

MsgSubmitTx Tx TxMeta SealedTx (BracketLog' (Either ErrSubmitTx ()))

MsgSubmitExternalTx (Hash "Tx") (BracketLog' (Either ErrPostTx Tx))

MsgRetryPostTx (Hash "Tx") (BracketLog' (Either ErrPostTx ()))

MsgProcessPendingPool BracketLog