Cardano.Wallet.Address.Book

Description

Address books consist of a Prologue and discovered addresses (Discoveries).

Documentation

class (Eq (Prologue s), Eq (Discoveries s)) => AddressBookIso s where

FIXME LATER during ADP-1043: Move Prologue and Discoveries closer into address types.

Associated Types

data Prologue s :: Type

Address information contained in the prologue of the address book, such as public keys or the address gap.

data Discoveries s :: Type

Addresses that were collected during discovery on the blockchain.

Methods

addressIso :: Iso' s (Prologue s, Discoveries s)

Isomorphism between the address book type s and its two components.

getPrologue :: AddressBookIso s => s -> Prologue s

getDiscoveries :: AddressBookIso s => s -> Discoveries s

newtype SeqAddressMap (c :: Role) (key :: Depth -> Type -> Type)

Address data from sequential address pool. The phantom type parameter c prevents mixing up the internal with the external pool.

Constructors

SeqAddressMap (Map (KeyFingerprint "payment" key) (Index 'Soft 'AddressK, AddressState))

newtype SharedAddressMap (c :: Role) (key :: Depth -> Type -> Type)

Constructors

SharedAddressMap (Map (KeyFingerprint "payment" key) (Index 'Soft 'ScriptK, AddressState))