Cardano.Wallet.Api.Client

Description

This module provides a half-typed Servant client for the cardano-wallet V2 API.

The functions in this module can be run with "Servant.Client.runClientM".

API Clients

data WalletClient wallet

This data type encapsulates the client functions for all endpoints of the cardano-wallet V2 API.

Constructors

WalletClient

Fields

deleteWallet :: ApiT WalletId -> ClientM ()

getWallet :: ApiT WalletId -> ClientM wallet

getWalletUtxoStatistics :: ApiT WalletId -> ClientM ApiUtxoStatistics

getWalletUtxoSnapshot :: ApiT WalletId -> ClientM ApiWalletUtxoSnapshot

listWallets :: ClientM [wallet]

postWallet :: PostData wallet -> ClientM wallet

putWallet :: ApiT WalletId -> WalletPutData -> ClientM wallet

putWalletPassphrase :: ApiT WalletId -> WalletPutPassphraseFormat wallet -> ClientM NoContent

walletClient :: WalletClient ApiWallet

Produces a WalletClient working against the /wallets API.

byronWalletClient :: WalletClient ApiByronWallet

Produces a WalletClient working against the /wallets API.

data TransactionClient

Constructors

TransactionClient

Fields

listTransactions :: ApiT WalletId -> Maybe Iso8601Time -> Maybe Iso8601Time -> Maybe (ApiT SortOrder) -> Bool -> ClientM [ApiTransactionT Value]

signTransaction :: ApiT WalletId -> ApiSignTransactionPostData -> ClientM ApiSerialisedTransaction

postTransaction :: ApiT WalletId -> PostTransactionOldDataT Value -> ClientM (ApiTransactionT Value)

postTransactionFee :: ApiT WalletId -> PostTransactionFeeOldDataT Value -> ClientM ApiFee

postExternalTransaction :: ApiBytesT 'Base64 SerialisedTx -> ClientM ApiTxId

deleteTransaction :: ApiT WalletId -> ApiTxId -> ClientM NoContent

getTransaction :: ApiT WalletId -> ApiTxId -> TxMetadataSchema -> ClientM (ApiTransactionT Value)

constructTransaction :: ApiT WalletId -> ApiConstructTransactionDataT Value -> ClientM (ApiConstructTransactionT Value)

balanceTransaction :: ApiT WalletId -> ApiBalanceTransactionPostDataT Value -> ClientM ApiSerialisedTransaction

decodeTransaction :: ApiT WalletId -> ApiSerialisedTransaction -> ClientM (ApiDecodedTransactionT Value)

submitTransaction :: ApiT WalletId -> ApiSerialisedTransaction -> ClientM ApiTxId

transactionClient :: TransactionClient

Produces a 'TransactionClient t' working against the /wallets API.

byronTransactionClient :: TransactionClient

Produces a 'TransactionClient n' working against the /byron-wallets API.

data AddressClient

Constructors

AddressClient

Fields

listAddresses :: ApiT WalletId -> Maybe (ApiT AddressState) -> ClientM [Value]

inspectAddress :: Text -> ClientM Value

postRandomAddress :: ApiT WalletId -> ApiPostRandomAddressData -> ClientM (ApiAddressT Value)

putRandomAddress :: ApiT WalletId -> ApiAddressIdT Value -> ClientM NoContent

putRandomAddresses :: ApiT WalletId -> ApiPutAddressesDataT Value -> ClientM NoContent

addressClient :: AddressClient

Produces an 'AddressClient n' working against the /wallets API

byronAddressClient :: AddressClient

Produces an 'AddressClient n' working against the /wallets API

data StakePoolClient apiPool

Constructors

StakePoolClient

Fields

listPools :: Maybe (ApiT Coin) -> ClientM [apiPool]

joinStakePool :: ApiPoolId -> ApiT WalletId -> ApiWalletPassphrase -> ClientM (ApiTransactionT Value)

quitStakePool :: ApiT WalletId -> ApiWalletPassphrase -> ClientM (ApiTransactionT Value)

stakePoolClient :: forall apiPool. FromJSON apiPool => StakePoolClient apiPool

Produces an 'StakePoolsClient n' working against the /stake-pools API

data NetworkClient

Constructors

NetworkClient

Fields

networkInformation :: ClientM ApiNetworkInformation

networkParameters :: ClientM ApiNetworkParameters

networkClock :: Bool -> ClientM ApiNetworkClock

networkClient :: NetworkClient

Produces a NetworkClient