Cardano.Wallet.Api.Link

Description

Type-safe endpoint accessors for the wallet API. Under normal circumstances, one would prefer to use WalletClient from Client and not to bother with endpoints at all.

Yet, in some cases (like in black-box testing), one could want to purposely send malformed requests to specific endpoints. Thus, this module facilitates the construction of valid endpoints that'd be accepted by the server, and for which, users are free to send all sort of data as payload, valid or invalid.

This module is meant to be used via qualified imports and with type-applications since all exposed functions are type ambiguous in a variable style of type WalletStyle.

import qualified Cardano.Wallet.Api.Link as Link

For examples:

>>> Link.deleteWallet @'Shelley myWallet
( "DELETE", "/v2/wallets/2512a00e9653fe49a44a5886202e24d77eeb998f" )

>>> Link.getWallet @('Byron 'Icarus) myWallet
( "GET", "/v2/byron-wallets/2512a00e9653fe49a44a5886202e24d77eeb998f" )

Wallets

deleteWallet :: forall (style :: WalletStyle) w. (Discriminate style, HasType (ApiT WalletId) w) => w -> (Method, Text)

getWallet :: forall (style :: WalletStyle) w. (Discriminate style, HasType (ApiT WalletId) w) => w -> (Method, Text)

listWallets :: forall (style :: WalletStyle). Discriminate style => (Method, Text)

postWallet :: PostWallet style => (Method, Text)

putWallet :: forall (style :: WalletStyle) w. (HasCallStack, Discriminate style, HasType (ApiT WalletId) w) => w -> (Method, Text)

putWalletPassphrase :: forall (style :: WalletStyle) w. (HasCallStack, Discriminate style, HasType (ApiT WalletId) w) => w -> (Method, Text)

getWalletUtxoSnapshot :: forall (style :: WalletStyle) w. (HasCallStack, Discriminate style, HasType (ApiT WalletId) w) => w -> (Method, Text)

getUTxOsStatistics :: forall (style :: WalletStyle) w. (HasCallStack, Discriminate style, HasType (ApiT WalletId) w) => w -> (Method, Text)

createMigrationPlan :: forall (style :: WalletStyle) w. (HasCallStack, Discriminate style, HasType (ApiT WalletId) w) => w -> (Method, Text)

migrateWallet :: forall (style :: WalletStyle) w. (HasCallStack, Discriminate style, HasType (ApiT WalletId) w) => w -> (Method, Text)

WalletKeys

getWalletKey :: forall style w. (HasCallStack, HasType (ApiT WalletId) w, Discriminate style) => w -> Role -> DerivationIndex -> Maybe Bool -> (Method, Text)

signMetadata :: forall w. HasType (ApiT WalletId) w => w -> Role -> DerivationIndex -> (Method, Text)

postAccountKey :: forall style w. (HasCallStack, HasType (ApiT WalletId) w, Discriminate style) => w -> DerivationIndex -> (Method, Text)

getAccountKey :: forall style w. (HasCallStack, HasType (ApiT WalletId) w, Discriminate style) => w -> Maybe KeyFormat -> (Method, Text)

getPolicyKey :: forall style w. (HasCallStack, HasType (ApiT WalletId) w, Discriminate style) => w -> Maybe Bool -> (Method, Text)

postPolicyKey :: forall (style :: WalletStyle) w. (HasCallStack, Discriminate style, HasType (ApiT WalletId) w) => w -> Maybe Bool -> (Method, Text)

postPolicyId :: forall (style :: WalletStyle) w. (HasCallStack, Discriminate style, HasType (ApiT WalletId) w) => w -> (Method, Text)

Addresses

postRandomAddress :: forall w. HasType (ApiT WalletId) w => w -> (Method, Text)

putRandomAddresses :: forall w. HasType (ApiT WalletId) w => w -> (Method, Text)

listAddresses :: forall style w. (HasType (ApiT WalletId) w, Discriminate style) => w -> (Method, Text)

listAddresses' :: forall style w. (HasType (ApiT WalletId) w, Discriminate style) => w -> Maybe AddressState -> (Method, Text)

inspectAddress :: ApiAddressInspectData -> (Method, Text)

postAnyAddress :: (Method, Text)

CoinSelections

selectCoins :: forall style w. (HasCallStack, HasType (ApiT WalletId) w, Discriminate style) => w -> (Method, Text)

Assets

listAssets :: forall w. HasType (ApiT WalletId) w => w -> (Method, Text)

getAsset :: forall w. HasType (ApiT WalletId) w => w -> TokenPolicyId -> TokenName -> (Method, Text)

listByronAssets :: forall w. HasType (ApiT WalletId) w => w -> (Method, Text)

getByronAsset :: forall w. HasType (ApiT WalletId) w => w -> TokenPolicyId -> TokenName -> (Method, Text)

Transactions

createTransactionOld :: forall style w. (HasCallStack, HasType (ApiT WalletId) w, Discriminate style) => w -> (Method, Text)

listTransactions :: forall (style :: WalletStyle) w. (Discriminate style, HasType (ApiT WalletId) w) => w -> (Method, Text)

listTransactions' :: forall (style :: WalletStyle) w. (HasCallStack, Discriminate style, HasType (ApiT WalletId) w) => w -> Maybe Natural -> Maybe Iso8601Time -> Maybe Iso8601Time -> Maybe SortOrder -> (Method, Text)

getTransactionFeeOld :: forall style w. (HasCallStack, HasType (ApiT WalletId) w, Discriminate style) => w -> (Method, Text)

deleteTransaction :: forall (style :: WalletStyle) w t. (HasCallStack, Discriminate style, HasType (ApiT WalletId) w, HasType (ApiT (Hash "Tx")) t) => w -> t -> (Method, Text)

getTransaction :: forall (style :: WalletStyle) w t. (HasCallStack, Discriminate style, HasType (ApiT WalletId) w, HasType (ApiT (Hash "Tx")) t) => w -> t -> (Method, Text)

createUnsignedTransaction :: forall style w. (HasType (ApiT WalletId) w, Discriminate style) => w -> (Method, Text)

signTransaction :: forall style w. (HasCallStack, HasType (ApiT WalletId) w, Discriminate style) => w -> (Method, Text)

balanceTransaction :: forall style w. (HasCallStack, HasType (ApiT WalletId) w, Discriminate style) => w -> (Method, Text)

decodeTransaction :: forall style w. (HasCallStack, HasType (ApiT WalletId) w, Discriminate style) => w -> (Method, Text)

submitTransaction :: forall style w. (HasCallStack, HasType (ApiT WalletId) w, Discriminate style) => w -> (Method, Text)

StakePools

listStakePools :: Maybe Coin -> (Method, Text)

listStakeKeys :: forall w. HasType (ApiT WalletId) w => w -> (Method, Text)

joinStakePool :: forall s w. (HasType (ApiT PoolId) s, HasType (ApiT WalletId) w) => s -> w -> (Method, Text)

quitStakePool :: forall w. HasType (ApiT WalletId) w => w -> (Method, Text)

getDelegationFee :: forall w. HasType (ApiT WalletId) w => w -> (Method, Text)

postPoolMaintenance :: (Method, Text)

getPoolMaintenance :: (Method, Text)

Network

getNetworkInfo :: (Method, Text)

getNetworkParams :: (Method, Text)

getNetworkClock :: (Method, Text)

getNetworkClock'

Arguments

:: Bool
When True, block and force NTP check

-> (Method, Text)

Proxy

postExternalTransaction :: (Method, Text)

Settings

putSettings :: (Method, Text)

getSettings :: (Method, Text)

Utils

getCurrentSMASHHealth :: (Method, Text)

Shared Wallets

patchSharedWallet :: forall w. HasType (ApiT WalletId) w => w -> CredentialType -> (Method, Text)

class PostWallet (style :: WalletStyle)

Minimal complete definition

postWallet

class Discriminate (style :: WalletStyle)

Minimal complete definition

discriminate