Cardano.Wallet.Api.Types

Description

API type representations of various types. We define here pretty much all our user-facing types that are mostly composed with internal / primitive types.

This module also define required API instances (JSON, HttpApiData...) for all those types, making sure to match the specification document:

Wallet API Specification

Wallet Styles

data WalletStyle

Constructors

Shelley

Byron

data ByronWalletStyle

Constructors

Random

Icarus

Trezor

Ledger

type family StyleSymbol (style :: ByronWalletStyle) :: Symbol where ...

Equations

StyleSymbol Random = "random"

StyleSymbol Icarus = "icarus"

StyleSymbol Trezor = "trezor"

StyleSymbol Ledger = "ledger"

type family AllowedMnemonics (style :: k) :: [Nat]

API Types

data ApiAddress (n :: NetworkDiscriminant)

Constructors

ApiAddress

Fields

id :: !(ApiT Address, Proxy n)

state :: !(ApiT AddressState)

data ApiEpochInfo

Constructors

ApiEpochInfo

Fields

epochNumber :: !(ApiT EpochNo)

epochStartTime :: !UTCTime

newtype ApiSelectCoinsData (n :: NetworkDiscriminant)

Constructors

ApiSelectCoinsData

Fields

payments :: NonEmpty (AddressAmount n)

data ApiCoinSelection (n :: NetworkDiscriminant)

Constructors

ApiCoinSelection

Fields

inputs :: !(NonEmpty (ApiCoinSelectionInput n))

outputs :: !(NonEmpty (AddressAmount n))

data ApiCoinSelectionInput (n :: NetworkDiscriminant)

Constructors

ApiCoinSelectionInput

Fields

id :: !(ApiT (Hash "Tx"))

index :: !Word32

address :: !(ApiT Address, Proxy n)

amount :: !(Quantity "lovelace" Natural)

data ApiStakePool

Constructors

ApiStakePool

Fields

id :: !(ApiT PoolId)

metrics :: !ApiStakePoolMetrics

apparentPerformance :: !Double

metadata :: !(Maybe StakePoolMetadata)

cost :: !(Quantity "lovelace" Natural)

margin :: !(Quantity "percent" Percentage)

desirability :: !Double

saturation :: !Double

data ApiStakePoolMetrics

Constructors

ApiStakePoolMetrics

Fields

controlledStake :: !(Quantity "lovelace" Natural)

producedBlocks :: !(Quantity "block" Natural)

data ApiWallet

Constructors

ApiWallet

Fields

id :: !(ApiT WalletId)

addressPoolGap :: !(ApiT AddressPoolGap)

balance :: !(ApiT WalletBalance)

delegation :: !ApiWalletDelegation

name :: !(ApiT WalletName)

passphrase :: !(Maybe (ApiT WalletPassphraseInfo))

state :: !(ApiT SyncProgress)

tip :: !ApiBlockReference

newtype ApiWalletPassphrase

Constructors

ApiWalletPassphrase

Fields

passphrase :: ApiT (Passphrase "encryption")

data ApiUtxoStatistics

Constructors

ApiUtxoStatistics

Fields

total :: !(Quantity "lovelace" Natural)

scale :: !(ApiT BoundType)

distribution :: !(Map Word64 Word64)

data WalletBalance

Constructors

WalletBalance

Fields

available :: !(Quantity "lovelace" Natural)

total :: !(Quantity "lovelace" Natural)

reward :: !(Quantity "lovelace" Natural)

data WalletPostData

Constructors

WalletPostData

Fields

addressPoolGap :: !(Maybe (ApiT AddressPoolGap))

mnemonicSentence :: !(ApiMnemonicT (AllowedMnemonics Shelley))

mnemonicSecondFactor :: !(Maybe (ApiMnemonicT (AllowedMnemonics SndFactor)))

name :: !(ApiT WalletName)

passphrase :: !(ApiT (Passphrase "encryption"))

newtype WalletPutData

Constructors

WalletPutData

Fields

name :: Maybe (ApiT WalletName)

data WalletPutPassphraseData

Constructors

WalletPutPassphraseData

Fields

oldPassphrase :: !(ApiT (Passphrase "encryption-old"))

newPassphrase :: !(ApiT (Passphrase "encryption-new"))

data PostTransactionData n

Constructors

PostTransactionData

Fields

payments :: !(NonEmpty (AddressAmount n))

passphrase :: !(ApiT (Passphrase "encryption"))

newtype PostTransactionFeeData n

Constructors

PostTransactionFeeData

Fields

payments :: NonEmpty (AddressAmount n)

newtype PostExternalTransactionData

Constructors

PostExternalTransactionData

Fields

payload :: ByteString

data ApiTimeReference

Constructors

ApiTimeReference

Fields

time :: !UTCTime

block :: !ApiBlockReference

data ApiTransaction n

Constructors

ApiTransaction

Fields

id :: !(ApiT (Hash "Tx"))

amount :: !(Quantity "lovelace" Natural)

insertedAt :: !(Maybe ApiTimeReference)

pendingSince :: !(Maybe ApiTimeReference)

depth :: !(Quantity "block" Natural)

direction :: !(ApiT Direction)

inputs :: ![ApiTxInput n]

outputs :: ![AddressAmount n]

status :: !(ApiT TxStatus)

newtype ApiFee

Constructors

ApiFee

Fields

amount :: Quantity "lovelace" Natural

newtype ApiTxId

Constructors

ApiTxId

Fields

id :: ApiT (Hash "Tx")

data ApiTxInput n

Constructors

ApiTxInput

Fields

source :: !(Maybe (AddressAmount n))

input :: !(ApiT TxIn)

data AddressAmount (n :: NetworkDiscriminant)

Constructors

AddressAmount

Fields

address :: !(ApiT Address, Proxy n)

amount :: !(Quantity "lovelace" Natural)

data ApiErrorCode

Error codes returned by the API, in the form of snake_cased strings

Constructors

NoSuchWallet

NoSuchTransaction

TransactionNotPending

WalletAlreadyExists

NoRootKey

WrongEncryptionPassphrase

MalformedTxPayload

KeyNotFoundForAddress

NotEnoughMoney

UtxoNotEnoughFragmented

TransactionIsTooBig

InputsDepleted

CannotCoverFee

InvalidCoinSelection

NetworkUnreachable

NetworkMisconfigured

NetworkTipNotFound

CreatedInvalidTransaction

RejectedByCoreNode

BadRequest

NotFound

MethodNotAllowed

NotAcceptable

StartTimeLaterThanEndTime

UnsupportedMediaType

UnexpectedError

NotSynced

NothingToMigrate

NoSuchPool

PoolAlreadyJoined

NotDelegatingTo

InvalidRestorationParameters

RejectedTip

NoSuchEpochNo

InvalidDelegationDiscovery

data ApiNetworkInformation

Constructors

ApiNetworkInformation

Fields

syncProgress :: !(ApiT SyncProgress)

nextEpoch :: !ApiEpochInfo

nodeTip :: !ApiBlockReference

networkTip :: !ApiNetworkTip

data ApiNtpStatus

Constructors

ApiNtpStatus

Fields

status :: !NtpSyncingStatus

offset :: !(Maybe (Quantity "microsecond" Integer))

data NtpSyncingStatus

Constructors

NtpSyncingStatusUnavailable

NtpSyncingStatusPending

NtpSyncingStatusAvailable

newtype ApiNetworkClock

Constructors

ApiNetworkClock

Fields

ntpStatus :: ApiNtpStatus

data ApiBlockReference

Constructors

ApiBlockReference

Fields

epochNumber :: !(ApiT EpochNo)

slotNumber :: !(ApiT SlotNo)

height :: !(Quantity "block" Natural)

data ApiNetworkTip

Constructors

ApiNetworkTip

Fields

epochNumber :: !(ApiT EpochNo)

slotNumber :: !(ApiT SlotNo)

newtype Iso8601Time

Defines a point in time that can be formatted as and parsed from an ISO 8601-compliant string.

Constructors

Iso8601Time

Fields

getIso8601Time :: UTCTime

data ApiEpochNumber

Constructors

ApiEpochNumberLatest

ApiEpochNumber EpochNo

data ApiNetworkParameters

Constructors

ApiNetworkParameters

Fields

genesisBlockHash :: !(ApiT (Hash "Genesis"))

blockchainStartTime :: !(ApiT StartTime)

slotLength :: !(Quantity "second" NominalDiffTime)

epochLength :: !(Quantity "slot" Word32)

epochStability :: !(Quantity "block" Word32)

activeSlotCoefficient :: !(Quantity "percent" Double)

toApiNetworkParameters :: BlockchainParameters -> ApiNetworkParameters

data ApiWalletDelegation

Constructors

ApiWalletDelegation

Fields

active :: !ApiWalletDelegationNext

next :: ![ApiWalletDelegationNext]

data ApiWalletDelegationStatus

Constructors

NotDelegating

Delegating

data ApiWalletDelegationNext

Constructors

ApiWalletDelegationNext

Fields

status :: !ApiWalletDelegationStatus

target :: !(Maybe (ApiT PoolId))

changesAt :: !(Maybe ApiEpochInfo)

data ApiPoolId

Constructors

ApiPoolIdPlaceholder

ApiPoolId PoolId

API Types (Byron)

data ApiByronWallet

Constructors

ApiByronWallet

Fields

id :: !(ApiT WalletId)

balance :: !ApiByronWalletBalance

name :: !(ApiT WalletName)

passphrase :: !(Maybe (ApiT WalletPassphraseInfo))

state :: !(ApiT SyncProgress)

tip :: !ApiBlockReference

data ApiByronWalletBalance

Constructors

ApiByronWalletBalance

Fields

available :: !(Quantity "lovelace" Natural)

total :: !(Quantity "lovelace" Natural)

newtype ApiByronWalletMigrationInfo

Constructors

ApiByronWalletMigrationInfo

Fields

migrationCost :: Quantity "lovelace" Natural

data ByronWalletPostData mw

Constructors

ByronWalletPostData

Fields

mnemonicSentence :: !(ApiMnemonicT mw)

name :: !(ApiT WalletName)

passphrase :: !(ApiT (Passphrase "encryption"))

data SomeByronWalletPostData

Constructors

SomeRandomWallet (ByronWalletPostData (AllowedMnemonics Random))

SomeIcarusWallet (ByronWalletPostData (AllowedMnemonics Icarus))

SomeTrezorWallet (ByronWalletPostData (AllowedMnemonics Trezor))

SomeLedgerWallet (ByronWalletPostData (AllowedMnemonics Ledger))

API Types (Hardware)

data AccountPostData

Constructors

AccountPostData

Fields

name :: !(ApiT WalletName)

accountPublicKey :: !AccountPublicKey

addressPoolGap :: !(Maybe (ApiT AddressPoolGap))

newtype AccountPublicKey

Constructors

AccountPublicKey

Fields

key :: ApiT (ShelleyKey AccountK XPub)

newtype WalletOrAccountPostData

Constructors

WalletOrAccountPostData

Fields

postData :: Either WalletPostData AccountPostData

User-Facing Address Encoding/Decoding

class EncodeAddress (n :: NetworkDiscriminant) where

An abstract class to allow encoding of addresses depending on the target backend used.

Methods

encodeAddress :: Address -> Text

class DecodeAddress (n :: NetworkDiscriminant) where

An abstract class to allow decoding of addresses depending on the target backend used.

Methods

decodeAddress :: Text -> Either TextDecodingError Address

Polymorphic Types

newtype ApiT a

Polymorphic wrapper type to put around primitive types and, 3rd party lib types to avoid defining orphan instances and/or, undesirable instances on primitive types. It helps to keep a nice separation of concerns between the API layer and other modules.

Constructors

ApiT

Fields

getApiT :: a

newtype ApiMnemonicT (sizes :: [Nat])

Representation of mnemonics at the API-level, using a polymorphic type in the lengths of mnemonics that are supported (and an underlying purpose). In practice, mnemonics correspond to passphrases or seeds, and although they're nice to manipulate as mnemonics from a user-perspective, carrying around a list of words doesn't really make sense for the business logic, which prefers manipulating scrubbed bytes directly.

data MyWallet
    { mnemonic :: ApiMnemonicT '[15,18,21,24]
    }

Note that the given Nats **have** to be valid mnemonic sizes, otherwise the underlying code won't even compile, with not-so-friendly error messages.

Also, the internal representation holds a [Text] which contains the list of mnemonic words that was parsed. This is only to be able to implement the ToJSON instances and roundtrip, which is a very dubious argument. In practice, we'll NEVER peek at the mnemonic, output them and whatnot.

Constructors

ApiMnemonicT

Fields

getApiMnemonicT :: SomeMnemonic