Cardano.Wallet.Binary

Description

These are (partial) CBOR decoders for blocks and block headers. Note that we ignore most of the block's and header's content and only retrieve the pieces of information relevant to us, wallet (we do assume a trusted node and therefore, we needn't to care about verifying signatures and blocks themselves).

The format described in the decoders below are the one used in the Byron era of Cardano and will endure in the first stages of Shelley. They are also used by components like the Rust cardano-http-bridge.

Decoding

decodeBlock :: Decoder s Block

decodeBlockHeader :: Decoder s BlockHeader

decodeTx :: Decoder s Tx

decodeTxWitness :: Decoder s TxWitness

decodeSignedTx :: Decoder s (Tx, [TxWitness])

Encoding

encodeTx :: Tx -> Encoding

encodeAddress :: XPub -> Encoding -> Encoding

Encode a public key to a corresponding Cardano Address. The encoding of the attributes part of an address is left out to the caller; This allows for distinguishing between Sequential and Random addresses (the former doesn't have any attributes to encode).

-- Old / Random Addresses
let encodeAttributes = mempty
     <> CBOR.encodeMapLen 1
     <> CBOR.encodeWord8 1
     <> encodeDerivationPath (hdPassphrase rootXPub) accIx addrIx
let addr = encodeAddress xpub encodeAttributes

-- New / Sequential Addresses
let encodeAttributes = mempty <> CBOR.encodeMapLen 0
let addr = encodeAddress xpub encodeAttributes

Note that we are passing the behavior to encode attributes as a parameter here and do not handle multiple cases in encodeAddress itself for multiple reasons:

Inversion of control gives us a nicer implementation overall

Encoding attributes for Random addresses requires more context than just the public key (like the wallet root id and some extra logic for encoding passphrases). This is just scheme-specific and is better left out of this particular function

encodeTxWitness :: TxWitness -> Encoding

encodeSignedTx :: (Tx, [TxWitness]) -> Encoding

encodeProtocolMagic :: ProtocolMagic -> Encoding

Helpers

inspectNextToken :: Decoder s ()

Inspect the next token that has to be decoded and print it to the console as a trace. Useful for debugging Decoders. Example:

    myDecoder :: CBOR.Decoder s MyType
    myDecoder = do
        a <- CBOR.decodeWord64
        inspectNextToken
        [...]

decodeList :: forall s a. Decoder s a -> Decoder s [a]

Decode an list of known length. Very similar to decodeListIndef.

    myDecoder :: CBOR.Decoder s [MyType]
    myDecoder = decodeList decodeOne
      where
        decodeOne :: CBOR.Decoder s MyType

decodeListIndef :: forall s a. Decoder s a -> Decoder s [a]

Decode an arbitrary long list. CBOR introduce a "break" character to mark the end of the list, so we simply decode each item until we encounter a break character.

    myDecoder :: CBOR.Decoder s [MyType]
    myDecoder = decodeListIndef decodeOne
      where
        decodeOne :: CBOR.Decoder s MyType

toByteString :: Encoding -> ByteString