Cardano.Wallet.CoinSelection

Description

This module provides a wallet-specific interface for coin selection.

Coin selection handles the following responsibilities:

selecting inputs from the UTxO set to pay for user-specified outputs;

selecting inputs from the UTxO set to pay for collateral;

producing change outputs to return excess value to the wallet;

balancing a selection to pay for the transaction fee.

Use the performSelection function to perform a coin selection.

Selection contexts

data WalletSelectionContext

A selection context for the wallet.

data WalletUTxO

A type of unique UTxO identifier for the wallet.

Constructors

WalletUTxO

Fields

txIn :: TxIn

address :: Address

Mapping between external (wallet) types and internal types

toExternalUTxO :: (WalletUTxO, TokenBundle) -> (TxIn, TxOut)

toExternalUTxOMap :: Map WalletUTxO TokenBundle -> UTxO

toInternalUTxO :: (TxIn, TxOut) -> (WalletUTxO, TokenBundle)

toInternalUTxOMap :: UTxO -> Map WalletUTxO TokenBundle

Mapping between external (wallet) selections and internal selections.

toExternalSelection :: Selection WalletSelectionContext -> Selection

toInternalSelection :: (change -> TokenBundle) -> SelectionOf change -> Selection WalletSelectionContext

Performing selections

performSelection :: forall m. (HasCallStack, MonadRandom m) => SelectionConstraints -> SelectionParams -> ExceptT (SelectionError WalletSelectionContext) m Selection

Performs a coin selection.

This function has the following responsibilities:

selecting inputs from the UTxO set to pay for user-specified outputs;

selecting inputs from the UTxO set to pay for collateral;

producing change outputs to return excess value to the wallet;

balancing a selection to pay for the transaction fee.

See performSelection for more details.

type Selection = SelectionOf TokenBundle

The default type of selection.

In this type of selection, change values do not have addresses assigned.

data SelectionCollateralRequirement

Indicates the collateral requirement for a selection.

Constructors

SelectionCollateralRequired
Indicates that collateral is required.

SelectionCollateralNotRequired
Indicates that collateral is not required.

data SelectionConstraints

Specifies all constraints required for coin selection.

Selection constraints:

are dependent on the current set of protocol parameters.

are not specific to a given selection.

place limits on the coin selection algorithm, enabling it to produce selections that are acceptable to the ledger.

Constructors

SelectionConstraints

Fields

assessTokenBundleSize :: TokenBundle -> TokenBundleSizeAssessment

Assesses the size of a token bundle relative to the upper limit of what can be included in a transaction output. See documentation for the TokenBundleSizeAssessor type to learn about the expected properties of this field.

certificateDepositAmount :: Coin

Amount that should be taken from/returned back to the wallet for each stake key registration/de-registration in the transaction.

computeMinimumAdaQuantity :: Address -> TokenMap -> Coin

Computes the minimum ada quantity required for a given output.

isBelowMinimumAdaQuantity :: Address -> TokenBundle -> Bool

Returns True if the given TokenBundle has a Coin value that is below the minimum required.

computeMinimumCost :: SelectionSkeleton -> Coin

Computes the minimum cost of a given selection skeleton.

computeSelectionLimit :: [TxOut] -> SelectionLimit

Computes an upper bound for the number of ordinary inputs to select, given a current set of outputs.

maximumCollateralInputCount :: Int

Specifies an inclusive upper bound on the number of unique inputs that can be selected as collateral.

minimumCollateralPercentage :: Natural

Specifies the minimum required amount of collateral as a percentage of the total transaction fee.

maximumLengthChangeAddress :: Address

data SelectionError ctx

Indicates that an error occurred while performing a coin selection.

Constructors

SelectionBalanceErrorOf (SelectionBalanceError ctx)

SelectionCollateralErrorOf (SelectionCollateralError ctx)

SelectionOutputErrorOf (SelectionOutputError ctx)

type SelectionLimit = SelectionLimitOf Int

Specifies a limit to adhere to when performing a selection.

data SelectionLimitOf a

Constructors

NoLimit
Indicates that there is no limit.

MaximumInputLimit a
Indicates a maximum limit on the number of inputs to select.

data SelectionOf change

Represents a balanced selection.

Constructors

Selection

Fields

inputs :: !(NonEmpty (TxIn, TxOut))

Selected inputs.

collateral :: ![(TxIn, TxOut)]

Selected collateral inputs.

outputs :: ![TxOut]

User-specified outputs

change :: ![change]

Generated change outputs.

assetsToMint :: !TokenMap

Assets to mint.

assetsToBurn :: !TokenMap

Assets to burn.

extraCoinSource :: !Coin

An extra source of ada.

extraCoinSink :: !Coin

An extra sink for ada.

data SelectionParams

Specifies all parameters that are specific to a given selection.

Constructors

SelectionParams

Fields

assetsToBurn :: !TokenMap

Specifies a set of assets to burn.

assetsToMint :: !TokenMap

Specifies a set of assets to mint.

extraCoinIn :: !Coin

Specifies extra Coin in.

extraCoinOut :: !Coin

Specifies extra Coin out.

outputsToCover :: ![TxOut]

Specifies a set of outputs that must be paid for.

rewardWithdrawal :: !Coin

Specifies the value of a withdrawal from a reward account.

certificateDepositsTaken :: !Natural

Number of deposits for stake key registrations.

certificateDepositsReturned :: !Natural

Number of deposits from stake key de-registrations.

collateralRequirement :: !SelectionCollateralRequirement

Specifies the collateral requirement for this selection.

utxoAvailableForCollateral :: !(Map WalletUTxO TokenBundle)

Specifies a set of UTxOs that are available for selection as collateral inputs.

This set is allowed to intersect with $sel:utxoAvailableForInputs:SelectionParams, since the ledger does not require that these sets are disjoint.

utxoAvailableForInputs :: !(UTxOSelection WalletUTxO)

Specifies a set of UTxOs that are available for selection as ordinary inputs and optionally, a subset that has already been selected.

Further entries from this set will be selected to cover any deficit.

selectionStrategy :: SelectionStrategy

Specifies which selection strategy to use. See SelectionStrategy.

data SelectionStrategy

Indicates a choice of selection strategy.

A SelectionStrategy determines how much of each asset the selection algorithm will attempt to select from the available UTxO set, relative to the minimum amount necessary to make the selection balance.

The default SelectionStrategy is SelectionStrategyOptimal, which when specified will cause the selection algorithm to attempt to select around twice the minimum possible amount of each asset from the available UTxO set, making it possible to generate change outputs that are roughly the same sizes and shapes as the user-specified outputs.

Specifying SelectionStrategyMinimal will cause the selection algorithm to only select just enough of each asset from the available UTxO set to meet the minimum amount. The selection process will terminate as soon as the minimum amount of each asset is covered.

The "optimal" strategy is recommended for most situations, as using this strategy will help to ensure that a wallet's UTxO distribution can evolve over time to resemble the typical distribution of payments made by the wallet owner. This increases the likelihood that future selections will succeed, and lowers the amortized cost of future transactions.

The "minimal" strategy is recommended only for situations where it is not possible to create a selection with the "optimal" strategy. It is advised to use this strategy only when necessary, as it increases the likelihood of generating change outputs that are much smaller than user-specified outputs. If this strategy is used regularly, the UTxO set can evolve to a state where the distribution no longer resembles the typical distribution of payments made by the user. This increases the likelihood that future selections will not succeed, and increases the amortized cost of future transactions.

Constructors

SelectionStrategyMinimal

SelectionStrategyOptimal

Selection skeletons

data SelectionSkeleton

A skeleton selection that can be used to estimate the cost of a final selection.

Change outputs are deliberately stripped of their asset quantities, as the fee estimation function must be agnostic to the magnitudes of these quantities.

Increasing or decreasing the quantity of a particular asset in a change output must not change the estimated cost of a selection.

Constructors

SelectionSkeleton

Fields

skeletonInputCount :: !Int

skeletonOutputs :: ![TxOut]

skeletonChange :: ![Set AssetId]

emptySkeleton :: SelectionSkeleton

Creates an empty SelectionSkeleton.

Selection errors

data BalanceInsufficientError

Indicates that the balance of available UTxO entries is insufficient to cover the balance required.

See computeUTxOBalanceSufficiency.

Constructors

BalanceInsufficientError

Fields

utxoBalanceAvailable :: !TokenBundle

The balance of $sel:utxoAvailable:SelectionParams.

utxoBalanceRequired :: !TokenBundle

The balance of $sel:outputsToCover:SelectionParams.

data SelectionBalanceError ctx

Represents the set of errors that may occur while performing a selection.

Constructors

BalanceInsufficient BalanceInsufficientError

SelectionLimitReached (SelectionLimitReachedError ctx)

UnableToConstructChange UnableToConstructChangeError

EmptyUTxO

data SelectionCollateralError ctx

Represents an unsuccessful attempt to select collateral.

data SelectionOutputError ctx

Indicates a problem when preparing outputs for a coin selection.

Constructors

SelectionOutputCoinInsufficient (SelectionOutputCoinInsufficientError ctx)

SelectionOutputSizeExceedsLimit (SelectionOutputSizeExceedsLimitError ctx)

SelectionOutputTokenQuantityExceedsLimit (SelectionOutputTokenQuantityExceedsLimitError ctx)

data SelectionOutputCoinInsufficientError ctx

Constructors

SelectionOutputCoinInsufficientError

Fields

minimumExpectedCoin :: Coin

output :: (Address ctx, TokenBundle)

newtype SelectionOutputSizeExceedsLimitError ctx

Constructors

SelectionOutputSizeExceedsLimitError

Fields

outputThatExceedsLimit :: (Address ctx, TokenBundle)

data SelectionOutputTokenQuantityExceedsLimitError ctx

Indicates that a token quantity exceeds the maximum quantity that can appear in a transaction output's token bundle.

Constructors

SelectionOutputTokenQuantityExceedsLimitError

Fields

address :: !(Address ctx)

The address to which this token quantity was to be sent.

asset :: !AssetId

The asset identifier to which this token quantity corresponds.

quantity :: !TokenQuantity

The token quantity that exceeded the bound.

quantityMaxBound :: !TokenQuantity

The maximum allowable token quantity.

data UnableToConstructChangeError

Constructors

UnableToConstructChangeError

Fields

requiredCost :: !Coin

The minimal required cost needed for the transaction to be considered valid. This does not include min Ada values.

shortfall :: !Coin

The additional coin quantity that would be required to cover the selection cost and minimum coin quantity of each change output.

Selection reports

makeSelectionReportDetailed :: Selection -> SelectionReportDetailed

makeSelectionReportSummarized :: Selection -> SelectionReportSummarized

data SelectionReportDetailed

Includes detailed information about a selection.

data SelectionReportSummarized

Includes summarized information about a selection.

Each data point can be serialized as a single line of text.

Selection deltas

balanceMissing :: BalanceInsufficientError -> TokenBundle

Calculate the missing balance from a BalanceInsufficientError.

selectionDelta

Arguments

:: (change -> Coin)
A function to extract the coin value from a change value.

-> SelectionOf change

-> Coin

Computes the ada surplus of a selection, assuming there is a surplus.