Cardano.Wallet.CoinSelection.Internal

Description

This module provides INTERNAL functions and types for coin selection.

It is recommended to import from CoinSelection instead.

Performing selections

performSelection :: (HasCallStack, MonadRandom m, SelectionContext ctx) => PerformSelection m ctx (Selection ctx)

Performs a coin selection.

This function has the following responsibilities:

selecting inputs from the UTxO set to pay for user-specified outputs;

selecting inputs from the UTxO set to pay for collateral;

producing change outputs to return excess value to the wallet;

balancing a selection to pay for the transaction fee.

This function guarantees that given a set of SelectionConstraints cs and SelectionParams ps:

if creation of a selection succeeds, a value s of type Selection will be returned for which the following property holds:

>>> verifySelection cs ps s == VerificationSuccess

if creation of a selection fails, a value e of type SelectionError will be returned for which the following property holds:

>>> verifySelectionError cs ps e == VerificationSuccess

data Selection ctx

Represents a balanced selection.

Constructors

Selection

Fields

inputs :: !(NonEmpty (UTxO ctx, TokenBundle))

Selected inputs.

collateral :: ![(UTxO ctx, Coin)]

Selected collateral inputs.

outputs :: ![(Address ctx, TokenBundle)]

User-specified outputs

change :: ![TokenBundle]

Generated change outputs.

assetsToMint :: !TokenMap

Assets to mint.

assetsToBurn :: !TokenMap

Assets to burn.

extraCoinSource :: !Coin

An extra source of ada.

extraCoinSink :: !Coin

An extra sink for ada.

data SelectionConstraints ctx

Specifies all constraints required for coin selection.

Selection constraints:

are dependent on the current set of protocol parameters.

are not specific to a given selection.

place limits on the coin selection algorithm, enabling it to produce selections that are acceptable to the ledger.

Constructors

SelectionConstraints

Fields

assessTokenBundleSize :: TokenBundle -> TokenBundleSizeAssessment

Assesses the size of a token bundle relative to the upper limit of what can be included in a transaction output. See documentation for the TokenBundleSizeAssessor type to learn about the expected properties of this field.

certificateDepositAmount :: Coin

Amount that should be taken from/returned back to the wallet for each stake key registration/de-registration in the transaction.

computeMinimumAdaQuantity :: Address ctx -> TokenMap -> Coin

Computes the minimum ada quantity required for a given output.

isBelowMinimumAdaQuantity :: Address ctx -> TokenBundle -> Bool

Returns True if the given TokenBundle has a Coin value that is below the minimum required.

computeMinimumCost :: SelectionSkeleton ctx -> Coin

Computes the minimum cost of a given selection skeleton.

computeSelectionLimit :: [(Address ctx, TokenBundle)] -> SelectionLimit

Computes an upper bound for the number of ordinary inputs to select, given a current set of outputs.

maximumCollateralInputCount :: Int

Specifies an inclusive upper bound on the number of unique inputs that can be selected as collateral.

minimumCollateralPercentage :: Natural

Specifies the minimum required amount of collateral as a percentage of the total transaction fee.

maximumOutputAdaQuantity :: Coin

Specifies the largest ada quantity that can appear in the token bundle of an output.

maximumOutputTokenQuantity :: TokenQuantity

Specifies the largest non-ada quantity that can appear in the token bundle of an output.

maximumLengthChangeAddress :: Address ctx

nullAddress :: Address ctx

data SelectionError ctx

Indicates that an error occurred while performing a coin selection.

Constructors

SelectionBalanceErrorOf (SelectionBalanceError ctx)

SelectionCollateralErrorOf (SelectionCollateralError ctx)

SelectionOutputErrorOf (SelectionOutputError ctx)

data SelectionParams ctx

Specifies all parameters that are specific to a given selection.

Constructors

SelectionParams

Fields

assetsToBurn :: !TokenMap

Specifies a set of assets to burn.

assetsToMint :: !TokenMap

Specifies a set of assets to mint.

extraCoinIn :: !Coin

Specifies extra Coin in.

extraCoinOut :: !Coin

Specifies extra Coin out.

outputsToCover :: ![(Address ctx, TokenBundle)]

Specifies a set of outputs that must be paid for.

rewardWithdrawal :: !Coin

Specifies the value of a withdrawal from a reward account.

certificateDepositsTaken :: !Natural

Number of deposits for stake key registrations.

certificateDepositsReturned :: !Natural

Number of deposits from stake key de-registrations.

collateralRequirement :: !SelectionCollateralRequirement

Specifies the collateral requirement for this selection.

utxoAvailableForCollateral :: !(Map (UTxO ctx) Coin)

Specifies a set of UTxOs that are available for selection as collateral inputs.

This set is allowed to intersect with $sel:utxoAvailableForInputs:SelectionParams, since the ledger does not require that these sets are disjoint.

utxoAvailableForInputs :: !(UTxOSelection (UTxO ctx))

Specifies a set of UTxOs that are available for selection as ordinary inputs and optionally, a subset that has already been selected.

Further entries from this set will be selected to cover any deficit.

selectionStrategy :: SelectionStrategy

Specifies which selection strategy to use. See SelectionStrategy.

data SelectionSkeleton ctx

A skeleton selection that can be used to estimate the cost of a final selection.

Change outputs are deliberately stripped of their asset quantities, as the fee estimation function must be agnostic to the magnitudes of these quantities.

Increasing or decreasing the quantity of a particular asset in a change output must not change the estimated cost of a selection.

Constructors

SelectionSkeleton

Fields

skeletonInputCount :: !Int

skeletonOutputs :: ![(Address ctx, TokenBundle)]

skeletonChange :: ![Set AssetId]

Output preparation

prepareOutputsWith :: forall f address. Functor f => (address -> TokenMap -> Coin) -> f (address, TokenBundle) -> f (address, TokenBundle)

Assigns minimal ada quantities to outputs without ada quantities.

This function only modifies outputs that have an ada quantity of zero. Outputs that have non-zero ada quantities will not be modified.

data SelectionOutputError ctx

Indicates a problem when preparing outputs for a coin selection.

Constructors

SelectionOutputCoinInsufficient (SelectionOutputCoinInsufficientError ctx)

SelectionOutputSizeExceedsLimit (SelectionOutputSizeExceedsLimitError ctx)

SelectionOutputTokenQuantityExceedsLimit (SelectionOutputTokenQuantityExceedsLimitError ctx)

data SelectionOutputCoinInsufficientError ctx

Constructors

SelectionOutputCoinInsufficientError

Fields

minimumExpectedCoin :: Coin

output :: (Address ctx, TokenBundle)

newtype SelectionOutputSizeExceedsLimitError ctx

Constructors

SelectionOutputSizeExceedsLimitError

Fields

outputThatExceedsLimit :: (Address ctx, TokenBundle)

data SelectionOutputTokenQuantityExceedsLimitError ctx

Indicates that a token quantity exceeds the maximum quantity that can appear in a transaction output's token bundle.

Constructors

SelectionOutputTokenQuantityExceedsLimitError

Fields

address :: !(Address ctx)

The address to which this token quantity was to be sent.

asset :: !AssetId

The asset identifier to which this token quantity corresponds.

quantity :: !TokenQuantity

The token quantity that exceeded the bound.

quantityMaxBound :: !TokenQuantity

The maximum allowable token quantity.

Verification of post conditions

data VerificationResult

The result of verifying a post condition.

Constructors

VerificationSuccess

VerificationFailure (NonEmpty VerificationFailureReason)

Verification of selections and selection errors

verifySelection :: SelectionContext ctx => VerifySelection ctx

Verifies a Selection for correctness.

This function is provided primarily as a convenience for testing. As such, it's not usually necessary to call this function from ordinary application code, unless you suspect that a Selection is incorrect in some way.

verifySelectionError :: SelectionContext ctx => VerifySelectionError (SelectionError ctx) ctx

Verifies a SelectionError for correctness.

This function is provided primarily as a convenience for testing. As such, it's not usually necessary to call this function from ordinary application code, unless you suspect that a SelectionError is incorrect in some way.

Selection deltas

data SelectionDelta a

Indicates the difference between total input value and total output value of a SelectionResult.

There are two possibilities:

SelectionSurplus

Indicates a surplus, when the total input value is greater than or equal to the total output value.

SelectionDeficit

Indicates a deficit, when the total input value is NOT greater than or equal to the total output value.

Constructors

SelectionSurplus a

SelectionDeficit a

selectionDeltaAllAssets :: Selection ctx -> SelectionDelta TokenBundle

Calculates the selection delta for all assets.

See SelectionDelta.

selectionDeltaCoin :: Selection ctx -> SelectionDelta Coin

Calculates the ada selection delta.

See SelectionDelta.

selectionHasValidSurplus :: SelectionConstraints ctx -> SelectionParams ctx -> Selection ctx -> Bool

Indicates whether or not a selection has a valid surplus.

This function returns True if and only if the selection has a delta that is a *surplus*, and that surplus is greater than or equal to the result of selectionMinimumCost.

See SelectionDelta.

selectionMinimumCost :: SelectionConstraints ctx -> SelectionParams ctx -> Selection ctx -> Coin

Computes the minimum required cost of a selection.

selectionSurplusCoin :: Selection ctx -> Coin

Calculates the ada selection surplus, assuming there is a surplus.

If there is a surplus, then this function returns that surplus. If there is a deficit, then this function returns zero.

Use selectionDeltaCoin if you wish to handle the case where there is a deficit.

Selection collateral

data SelectionCollateralError ctx

Represents an unsuccessful attempt to select collateral.

Constructors

SelectionCollateralError

Fields

largestCombinationAvailable :: Map (UTxO ctx) Coin

The largest combination of coins available.

minimumSelectionAmount :: Coin

A lower bound on the sum of coins to be selected as collateral.

data SelectionCollateralRequirement

Indicates the collateral requirement for a selection.

Constructors

SelectionCollateralRequired
Indicates that collateral is required.

SelectionCollateralNotRequired
Indicates that collateral is not required.

selectionCollateral :: Selection ctx -> Coin

Computes the total amount of collateral within a selection.

selectionCollateralRequired :: SelectionParams ctx -> Bool

Indicates True if and only if collateral is required.

selectionHasSufficientCollateral :: SelectionConstraints ctx -> SelectionParams ctx -> Selection ctx -> Bool

Indicates whether or not a selection has sufficient collateral.

selectionMinimumCollateral :: SelectionConstraints ctx -> SelectionParams ctx -> Selection ctx -> Coin

Computes the minimum required amount of collateral for a selection.

Internal types and functions

data ComputeMinimumCollateralParams

Parameters for computeMinimumCollateral.

Constructors

ComputeMinimumCollateralParams

Fields

minimumCollateralPercentage :: Natural

transactionFee :: Coin

computeMinimumCollateral :: ComputeMinimumCollateralParams -> Coin

Computes the minimum required amount of collateral given a fee and a minimum collateral percentage.

toBalanceConstraintsParams :: forall ctx. (SelectionConstraints ctx, SelectionParams ctx) -> (SelectionConstraints ctx, SelectionParams ctx)

Creates constraints and parameters for performSelection.