Cardano.Wallet.CoinSelection.Internal.Balance

Description

Provides an algorithm for producing a balanced coin selection with change, where the fee is paid for.

This module uses the Random-Round-Robin coin selection algorithm for multi-asset UTxO sets.

See documentation for the performSelection function for more details on how to perform a selection.

Performing a selection

type PerformSelection m f ctx = SelectionConstraints ctx -> SelectionParamsOf f ctx -> m (Either (SelectionBalanceError ctx) (SelectionResultOf f ctx))

performSelection :: forall m ctx. (HasCallStack, MonadRandom m, SelectionContext ctx) => PerformSelection m [] ctx

Performs a coin selection and generates change bundles in one step.

Provided that isUTxOBalanceSufficient returns True for the given selection criteria, this function guarantees to return a SelectionResult for which selectionHasValidSurplus returns True.

performSelectionEmpty :: forall m ctx. Functor m => PerformSelection m NonEmpty ctx -> PerformSelection m [] ctx

Transforms a coin selection function that requires a non-empty list of outputs into a function that accepts an empty list of outputs.

If the original list is already non-empty, this function does not alter the parameters or the result in any way, such that:

params == transformParams params result == transformResult result

If the original list is empty, this function:

applies a balance-preserving transformation to the parameters, adding a single minimal ada-only output to act as a change generation target, such that:

computeUTxOBalanceSufficiencyInfo params == computeUTxOBalanceSufficiencyInfo (transformParams params)

applies an inverse transformation to the result, removing the output, such that:

selectionSurplus result == selectionSurplus (transformResult result)

selectionHasValidSurplus constraints result ==> selectionHasValidSurplus constraints (transformResult result)

data SelectionConstraints ctx

Specifies all constraints required for coin selection.

Selection constraints:

place limits on the coin selection algorithm, enabling it to produce selections that are acceptable to the ledger.

are dependent on the current set of protocol parameters.

are not specific to a given selection.

Constructors

SelectionConstraints

Fields

assessTokenBundleSize :: TokenBundle -> TokenBundleSizeAssessment

Assesses the size of a token bundle relative to the upper limit of what can be included in a transaction output. See documentation for the TokenBundleSizeAssessor type to learn about the expected properties of this field.

computeMinimumAdaQuantity :: Address ctx -> TokenMap -> Coin

Computes the minimum ada quantity required for a given output.

computeMinimumCost :: SelectionSkeleton ctx -> Coin

Computes the minimum cost of a given selection skeleton.

computeSelectionLimit :: [(Address ctx, TokenBundle)] -> SelectionLimit

Computes an upper bound for the number of ordinary inputs to select, given a current set of outputs.

maximumLengthChangeAddress :: Address ctx

maximumOutputAdaQuantity :: Coin

Specifies the largest ada quantity that can appear in the token bundle of an output.

maximumOutputTokenQuantity :: TokenQuantity

Specifies the largest non-ada quantity that can appear in the token bundle of an output.

nullAddress :: Address ctx

type SelectionParams = SelectionParamsOf []

data SelectionParamsOf f ctx

Specifies all parameters that are specific to a given selection.

Constructors

SelectionParams

Fields

outputsToCover :: !(f (Address ctx, TokenBundle))

The complete set of outputs to be covered.

utxoAvailable :: !(UTxOSelection (UTxO ctx))

Specifies a set of UTxOs that are available for selection as inputs and optionally, a subset that has already been selected.

Further entries from this set will be selected to cover any deficit.

extraCoinSource :: !Coin

An extra source of ada.

extraCoinSink :: !Coin

An extra sink for ada.

assetsToMint :: !TokenMap

Assets to mint: these provide input value to a transaction.

By minting tokens, we generally decrease the burden of the selection algorithm, allowing it to select fewer UTxO entries in order to cover the required outputs.

assetsToBurn :: !TokenMap

Assets to burn: these consume output value from a transaction.

By burning tokens, we generally increase the burden of the selection algorithm, requiring it to select more UTxO entries in order to cover the burn.

selectionStrategy :: SelectionStrategy

Specifies which selection strategy to use. See SelectionStrategy.

data SelectionSkeleton ctx

A skeleton selection that can be used to estimate the cost of a final selection.

Change outputs are deliberately stripped of their asset quantities, as the fee estimation function must be agnostic to the magnitudes of these quantities.

Increasing or decreasing the quantity of a particular asset in a change output must not change the estimated cost of a selection.

Constructors

SelectionSkeleton

Fields

skeletonInputCount :: !Int

skeletonOutputs :: ![(Address ctx, TokenBundle)]

skeletonChange :: ![Set AssetId]

type SelectionResult = SelectionResultOf []

data SelectionResultOf f ctx

The result of performing a successful selection.

Constructors

SelectionResult

Fields

inputsSelected :: !(NonEmpty (UTxO ctx, TokenBundle))

A (non-empty) list of inputs selected from $sel:utxoAvailable:SelectionParams.

extraCoinSource :: !Coin

An extra source of ada.

extraCoinSink :: !Coin

An extra sink for ada.

outputsCovered :: !(f (Address ctx, TokenBundle))

A list of outputs covered.

changeGenerated :: ![TokenBundle]

A list of generated change outputs.

assetsToMint :: !TokenMap

The assets to mint.

assetsToBurn :: !TokenMap

The assets to burn.

data SelectionStrategy

Indicates a choice of selection strategy.

A SelectionStrategy determines how much of each asset the selection algorithm will attempt to select from the available UTxO set, relative to the minimum amount necessary to make the selection balance.

The default SelectionStrategy is SelectionStrategyOptimal, which when specified will cause the selection algorithm to attempt to select around twice the minimum possible amount of each asset from the available UTxO set, making it possible to generate change outputs that are roughly the same sizes and shapes as the user-specified outputs.

Specifying SelectionStrategyMinimal will cause the selection algorithm to only select just enough of each asset from the available UTxO set to meet the minimum amount. The selection process will terminate as soon as the minimum amount of each asset is covered.

The "optimal" strategy is recommended for most situations, as using this strategy will help to ensure that a wallet's UTxO distribution can evolve over time to resemble the typical distribution of payments made by the wallet owner. This increases the likelihood that future selections will succeed, and lowers the amortized cost of future transactions.

The "minimal" strategy is recommended only for situations where it is not possible to create a selection with the "optimal" strategy. It is advised to use this strategy only when necessary, as it increases the likelihood of generating change outputs that are much smaller than user-specified outputs. If this strategy is used regularly, the UTxO set can evolve to a state where the distribution no longer resembles the typical distribution of payments made by the user. This increases the likelihood that future selections will not succeed, and increases the amortized cost of future transactions.

Constructors

SelectionStrategyMinimal

SelectionStrategyOptimal

data SelectionBalanceError ctx

Represents the set of errors that may occur while performing a selection.

Constructors

BalanceInsufficient BalanceInsufficientError

SelectionLimitReached (SelectionLimitReachedError ctx)

UnableToConstructChange UnableToConstructChangeError

EmptyUTxO

data BalanceInsufficientError

Indicates that the balance of available UTxO entries is insufficient to cover the balance required.

See computeUTxOBalanceSufficiency.

Constructors

BalanceInsufficientError

Fields

utxoBalanceAvailable :: !TokenBundle

The balance of $sel:utxoAvailable:SelectionParams.

utxoBalanceRequired :: !TokenBundle

The balance of $sel:outputsToCover:SelectionParams.

data UnableToConstructChangeError

Constructors

UnableToConstructChangeError

Fields

requiredCost :: !Coin

The minimal required cost needed for the transaction to be considered valid. This does not include min Ada values.

shortfall :: !Coin

The additional coin quantity that would be required to cover the selection cost and minimum coin quantity of each change output.

Selection limits

type SelectionLimit = SelectionLimitOf Int

Specifies a limit to adhere to when performing a selection.

data SelectionLimitOf a

Constructors

NoLimit
Indicates that there is no limit.

MaximumInputLimit a
Indicates a maximum limit on the number of inputs to select.

selectionLimitExceeded :: IsUTxOSelection s u => s u -> SelectionLimit -> Bool

Indicates whether or not the given selection limit has been exceeded.

data SelectionLimitReachedError ctx

Indicates that the balance of selected UTxO entries was insufficient to cover the balance required while remaining within the selection limit.

Constructors

SelectionLimitReachedError

Fields

utxoBalanceRequired :: !TokenBundle

The UTXO balance required.

inputsSelected :: ![(UTxO ctx, TokenBundle)]

The inputs that could be selected while satisfying the $sel:selectionLimit:RunSelectionParams.

outputsToCover :: !(NonEmpty (Address ctx, TokenBundle))

reduceSelectionLimitBy :: SelectionLimit -> Int -> SelectionLimit

Reduces a selection limit by a given reduction amount.

If the given reduction amount is positive, then this function will reduce the selection limit by that amount.

If the given reduction amount is zero or negative, then this function will return the original limit unchanged.

Querying selections

data SelectionDelta a

Indicates the difference between total input value and total output value of a SelectionResult.

There are two possibilities:

SelectionSurplus

Indicates a surplus, when the total input value is greater than or equal to the total output value.

SelectionDeficit

Indicates a deficit, when the total input value is NOT greater than or equal to the total output value.

Constructors

SelectionSurplus a

SelectionDeficit a

selectionDeltaAllAssets :: Foldable f => SelectionResultOf f ctx -> SelectionDelta TokenBundle

Calculates the selection delta for all assets.

See SelectionDelta.

selectionDeltaCoin :: Foldable f => SelectionResultOf f ctx -> SelectionDelta Coin

Calculates the ada selection delta.

See SelectionDelta.

selectionHasValidSurplus :: Foldable f => SelectionConstraints ctx -> SelectionResultOf f ctx -> Bool

Indicates whether or not a selection result has a valid surplus.

selectionSurplusCoin :: Foldable f => SelectionResultOf f ctx -> Coin

Calculates the ada selection surplus, assuming there is a surplus.

If there is a surplus, then this function returns that surplus. If there is a deficit, then this function returns zero.

Use selectionDeltaCoin if you wish to handle the case where there is a deficit.

selectionMinimumCost :: Foldable f => SelectionConstraints ctx -> SelectionResultOf f ctx -> Coin

Computes the minimum required cost of a selection.

selectionMaximumCost :: Foldable f => SelectionConstraints ctx -> SelectionResultOf f ctx -> Coin

Computes the maximum acceptable cost of a selection.

This function acts as a safety limit to ensure that fees of selections produced by performSelection are not excessively high.

Ideally, we'd always be able to generate selections with fees that are precisely equal to selectionMinimumCost. However, in some situations it may be necessary to exceed this cost very slightly.

This function provides a conservative upper bound to a selection cost that we can reference from within property tests.

See selectionHasValidSurplus.

selectionSkeleton :: Foldable f => SelectionResultOf f ctx -> SelectionSkeleton ctx

Converts a selection into a skeleton.

Querying parameters

data UTxOBalanceSufficiency

Indicates whether the balance of available UTxO entries is sufficient.

See computeUTxOBalanceSufficiency.

Constructors

UTxOBalanceSufficient
Indicates that the UTxO balance is sufficient.

UTxOBalanceInsufficient
Indicates that the UTxO balance is insufficient.

data UTxOBalanceSufficiencyInfo

Gives more information about UTxO balance sufficiency.

See computeUTxOBalanceSufficiencyInfo.

Constructors

UTxOBalanceSufficiencyInfo

Fields

available :: TokenBundle

See computeUTxOBalanceAvailable.

required :: TokenBundle

See computeUTxOBalanceRequired.

difference :: TokenBundle

The difference between $sel:available:UTxOBalanceSufficiencyInfo and $sel:required:UTxOBalanceSufficiencyInfo.

sufficiency :: UTxOBalanceSufficiency

Whether or not the balance is sufficient.

computeBalanceInOut :: Foldable f => SelectionParamsOf f ctx -> (TokenBundle, TokenBundle)

computeDeficitInOut :: Foldable f => SelectionParamsOf f ctx -> (TokenBundle, TokenBundle)

computeUTxOBalanceAvailable :: SelectionParamsOf f ctx -> TokenBundle

Computes the balance of UTxO entries available for selection.

computeUTxOBalanceRequired :: Foldable f => SelectionParamsOf f ctx -> TokenBundle

Computes the balance of UTxO entries required to be selected.

computeUTxOBalanceSufficiency :: Foldable f => SelectionParamsOf f ctx -> UTxOBalanceSufficiency

Computes the UTxO balance sufficiency.

See UTxOBalanceSufficiency.

computeUTxOBalanceSufficiencyInfo :: Foldable f => SelectionParamsOf f ctx -> UTxOBalanceSufficiencyInfo

Computes information about the UTxO balance sufficiency.

See UTxOBalanceSufficiencyInfo.

isUTxOBalanceSufficient :: Foldable f => SelectionParamsOf f ctx -> Bool

Indicates whether or not the UTxO balance is sufficient.

The balance of available UTxO entries is sufficient if (and only if) it is greater than or equal to the required balance.

Running a selection (without making change)

runSelection :: forall m u. (MonadRandom m, Ord u) => RunSelectionParams u -> m (UTxOSelection u)

runSelectionNonEmpty :: (MonadRandom m, Ord u) => RunSelectionParams u -> m (Maybe (UTxOSelectionNonEmpty u))

runSelectionNonEmptyWith :: Monad m => (UTxOSelection u -> m (Maybe (UTxOSelectionNonEmpty u))) -> UTxOSelection u -> m (Maybe (UTxOSelectionNonEmpty u))

data RunSelectionParams u

Parameters for runSelection.

Constructors

RunSelectionParams

Fields

selectionLimit :: SelectionLimit

A limit to adhere to when performing a selection.

utxoAvailable :: UTxOSelection u

UTxO entries available for selection.

minimumBalance :: TokenBundle

Minimum balance to cover.

selectionStrategy :: SelectionStrategy

Specifies which selection strategy to use. See SelectionStrategy.

Running a selection step

runSelectionStep :: forall m state state'. Monad m => SelectionLens m state state' -> state -> m (Maybe state')

Runs just a single step of a coin selection.

It returns an updated state if (and only if) the updated selection represents an improvement over the selection in the previous state.

An improvement, for a given token quantity, is defined in the following way:

If the total selected token quantity of the previous selection had not yet reached 100% of the output token quantity, any additional selection is considered to be an improvement.

If the total selected token quantity of the previous selection had already reached or surpassed 100% of the output token quantity, any additional selection is considered to be an improvement if and only if it takens the total selected token quantity closer to the target token quantity, but not further away.

data SelectionLens m state state'

Provides a lens on the current selection state.

A SelectionLens gives runSelectionStep just the information it needs to make a decision, and no more.

Constructors

SelectionLens

Fields

currentQuantity :: state -> Natural

updatedQuantity :: state' -> Natural

selectQuantity :: state -> m (Maybe state')

minimumQuantity :: Natural

selectionStrategy :: SelectionStrategy

assetSelectionLens :: (MonadRandom m, Ord u) => SelectionLimit -> SelectionStrategy -> (AssetId, TokenQuantity) -> SelectionLens m (UTxOSelection u) (UTxOSelectionNonEmpty u)

coinSelectionLens

Arguments

:: (MonadRandom m, Ord u)

=> SelectionLimit

-> SelectionStrategy

-> Coin
Minimum coin quantity.

-> SelectionLens m (UTxOSelection u) (UTxOSelectionNonEmpty u)

Making change

data MakeChangeCriteria minCoinFor bundleSizeAssessor

Criteria for the makeChange function.

Constructors

MakeChangeCriteria

Fields

minCoinFor :: minCoinFor

A function that computes the minimum required ada quantity for a particular output.

bundleSizeAssessor :: bundleSizeAssessor

A function to assess the size of a token bundle.

requiredCost :: Coin

The minimal (and optimal) delta between the total ada balance of all input bundles and the total ada balance of all output and change bundles, where:

delta = getCoin (fold inputBundles) - getCoin (fold outputBundles) - getCoin (fold changeBundles)

This typically captures fees plus key deposits.

extraCoinSource :: Coin

An extra source of ada.

extraCoinSink :: Coin

An extra sink for ada.

inputBundles :: NonEmpty TokenBundle

Token bundles of selected inputs.

outputBundles :: NonEmpty TokenBundle

Token bundles of original outputs.

assetsToMint :: TokenMap

Assets to mint: these provide input value to a transaction.

assetsToBurn :: TokenMap

Assets to burn: these consume output value from a transaction.

maximumOutputAdaQuantity :: Coin

Specifies the largest ada quantity that can appear in the token bundle of an output.

maximumOutputTokenQuantity :: TokenQuantity

Specifies the largest non-ada quantity that can appear in the token bundle of an output.

makeChange

Arguments

:: MakeChangeCriteria (TokenMap -> Coin) TokenBundleSizeAssessor
Criteria for making change.

-> Either UnableToConstructChangeError [TokenBundle]
Generated change bundles.

Constructs change bundles for a set of selected inputs and outputs.

Returns Nothing if the specified inputs do not provide enough ada to satisfy the minimum delta and minimum ada quantities of the change bundles generated.

This function will generate runtime errors if:

The total balance of all outputs is not less than or equal to the total balance of all inputs.

The total ada balance of all outputs is zero.

Pre-condition (1) should be satisfied by any result produced by the runSelection function.

Pre-condition (2) should be satisfied by assigning a minimum ada quantity to every output token bundle.

makeChangeForCoin

Arguments

:: HasCallStack

=> NonEmpty Coin
A list of weights for the distribution. Conveniently captures both the weights, and the number of elements amongst which the surplus ada quantity should be distributed.

-> Coin
A surplus ada quantity to be distributed.

-> NonEmpty Coin

Constructs a list of ada change outputs based on the given distribution.

If the sum of weights in given distribution is equal to zero, this function throws a runtime error.

The length of the output list is always the same as the length of the input list, and the sum of its quantities is always exactly equal to the Coin value given as the second argument.

makeChangeForUserSpecifiedAsset

Arguments

:: NonEmpty TokenMap
A list of weights for the distribution. Conveniently captures both the weights, and the number of elements amongst which the quantity should be distributed.

-> (AssetId, TokenQuantity)
A surplus token quantity to distribute.

-> NonEmpty TokenMap

Constructs change outputs for a user-specified asset: an asset that was present in the original set of outputs.

If the given asset does not appear in the given distribution, this function returns a list of empty token maps. Otherwise, the given token quantity is partitioned into a list of quantities that are proportional to the weights within the given input distribution, modulo rounding.

The length of the output list is always the same as the the length of the input list, and the sum of its quantities is either zero, or exactly equal to the token quantity in the second argument.

makeChangeForNonUserSpecifiedAsset

Arguments

:: NonEmpty a
Determines the number of change maps to create.

-> (AssetId, NonEmpty TokenQuantity)
An asset quantity to distribute.

-> NonEmpty TokenMap
The resultant change maps.

Constructs change outputs for a non-user-specified asset: an asset that was not present in the original set of outputs.

This function constructs a list of change outputs by preserving the input distribution as much as possible. Note that only the length of the first argument is used.

The length of the output list is always the same as the length of the input list, and the sum of its quantities is always exactly equal to the sum of all token quantities given in the second argument.

The resultant list is sorted into ascending order when maps are compared with the leq function.

makeChangeForNonUserSpecifiedAssets

Arguments

:: NonEmpty a
Determines the number of change maps to create.

-> Map AssetId (NonEmpty TokenQuantity)
A map of asset quantities to distribute.

-> NonEmpty TokenMap
The resultant change maps.

Constructs change outputs for all non-user-specified assets: assets that were not present in the original set of outputs.

The resultant list is sorted into ascending order when maps are compared with the leq function.

assignCoinsToChangeMaps

Arguments

:: HasCallStack

=> Coin
The total quantity of ada available, including any extra source of ada.

-> (TokenMap -> Coin)
A function to calculate the minimum required ada quantity for any token map.

-> NonEmpty (TokenMap, Coin)
A list of pre-computed asset change maps paired with original output coins, sorted into an order that ensures all empty token maps are at the start of the list.

-> Either Coin [TokenBundle]
Resulting change bundles, or the shortfall quantity if there was not enough ada available to assign a minimum ada quantity to all non-empty token maps.

Assigns coin quantities to a list of pre-computed asset change maps.

Each pre-computed asset change map must be paired with the original coin value of its corresponding output.

This function:

expects the list of pre-computed asset change maps to be sorted in an order that ensures all empty token maps are at the start of the list.

attempts to assign a minimum ada quantity to every change map, but iteratively drops empty change maps from the start of the list if the amount of ada is insufficient to cover them all.

continues dropping empty change maps from the start of the list until it is possible to assign a minimum ada value to all remaining entries.

returns a list that is identical in length to the input list if (and only if) it was possible to assign a minimum ada quantity to all change maps.

returns a list that is shorter than the input list if it was only possible to assign a minimum ada quantity to a suffix of the given list.

fails if (and only if) there was not enough ada available to assign the minimum ada quantity to all non-empty change maps.

collateNonUserSpecifiedAssetQuantities

Arguments

:: NonEmpty TokenMap
Token maps of all selected inputs.

-> Set AssetId
Set of all assets in user-specified outputs.

-> Map AssetId (NonEmpty TokenQuantity)

Generates a map of all non-user-specified assets and their quantities.

Each key in the resulting map corresponds to an asset that was NOT included in the original set of user-specified outputs, but that was nevertheless selected during the selection process.

The value associated with each key corresponds to the complete list of all discrete non-zero quantities of that asset present in the selected inputs.

addMintValueToChangeMaps :: (AssetId, TokenQuantity) -> NonEmpty TokenMap -> NonEmpty TokenMap

Adds a minted asset quantity to a list of change maps.

This function always adds the given quantity to the final change map in the given list.

Example:

Suppose we have the following list of change maps:

[ (B, 7)

[(A, 1), (B, 8)

[(A, 2), (B, 9)

[(A, 3), (B, 9)

[(A, 4), (B, 12)

If we add 4 tokens of asset A, we obtain the following result:

[ (B, 7)

[(A, 1), (B, 8)

[(A, 2), (B, 9)

[(A, 3), (B, 9)

[(A, 8), (B, 12)

-- Increased by 4

Provided that the specified change maps are in ascending partial order, this function guarantees that the resulting change maps will also be in ascending partial order.

The length of the given list is preserved in the output list.

addMintValuesToChangeMaps

Arguments

:: TokenMap
Map of minted values

-> NonEmpty TokenMap
Change maps

-> NonEmpty TokenMap
Change maps with minted values

Adds minted values for multiple assets to a list of change maps.

Plural of addMintValueToChangeMaps.

removeBurnValueFromChangeMaps

Arguments

:: (AssetId, TokenQuantity)
Asset quantity reduction target

-> NonEmpty TokenMap
Change maps with quantities of the given asset to be reduced

-> NonEmpty TokenMap
Change maps with reduced quantities of the given asset

Removes a burned asset quantity from a list of change maps.

For a given asset a and reduction target t, this function traverses the given list from left to right, reducing the quantity of asset a in each change map until the reduction target t has been met, or until the list is exhausted.

For each change map m under consideration:

if the quantity q of asset a in map m is less than or equal to the remaining required reduction r, it will be replaced with a zero (effectively eliminating asset a from the map).

if the quantity q of asset a in map m is greater than the remaining required reduction r, it will be replaced with the absolute difference between q and r.

If the total quantity of the given asset in the given change maps is greater than the specified reduction target, the total reduction will be equal to the specified reduction target. Otherwise, the given asset will be completely eliminated from all change maps.

Example:

Suppose we have the following list of change maps:

[ (B, 7)

[(A, 1), (B, 8)

[(A, 2), (B, 9)

[(A, 3), (B, 9)

[(A, 4), (B, 12)

If our target is to reduce the quantity of asset A by 4, then we should obtain the following result:

[ (B, 7)

-- Unable to reduce (already 0)

[ (B, 8)

-- Reduced by 1 (and eliminated from map)

[ (B, 9)

-- Reduced by 2 (and eliminated from map)

[(A, 2), (B, 9)

-- Reduced by 1

[(A, 4), (B, 12)

Provided that the specified change maps are in ascending partial order, this function guarantees that the resulting change maps will also be in ascending partial order.

The length of the given list is preserved in the output list.

removeBurnValuesFromChangeMaps

Arguments

:: TokenMap
Map of burned values

-> NonEmpty TokenMap
Change maps

-> NonEmpty TokenMap
Change maps with burned values removed

Removes burned values for multiple assets from a list of change maps.

Plural of removeBurnValueFromChangeMaps.

reduceTokenQuantities

Arguments

:: TokenQuantity
Reduction target

-> NonEmpty TokenQuantity
List of quantities to reduce

-> NonEmpty TokenQuantity
The list of reduced quantities

Reduces the total value of the given list of token quantities by the given reduction target.

This function traverses the given list of quantities from left to right, reducing each quantity in turn until the total reduction is equal to the given reduction target, or until the list is exhausted.

For each quantity q under consideration:

if q is less than or equal to the remaining required reduction r, it will be replaced with a zero.

if q is greater than the remaining required reduction r, it will be replaced with the absolute difference between q and r.

If the total value in the list is less than the reduction target, the result will be a list of zeros.

Provided the given list is in ascending order, the resulting list is also guaranteed to be in ascending order.

The length of the given list is preserved in the output.

Splitting bundles

splitBundleIfAssetCountExcessive

Arguments

:: TokenBundle
The token bundle suspected to have an excessive number of assets.

-> (TokenBundle -> Bool)
A function that returns True if (and only if) the asset count of the given bundle is excessive.

-> NonEmpty TokenBundle

Splits a bundle into smaller bundles if its asset count is excessive when measured with the given isExcessive indicator function.

Returns a list of smaller bundles for which isExcessive returns False.

splitBundlesWithExcessiveAssetCounts

Arguments

:: NonEmpty TokenBundle
Token bundles.

-> (TokenBundle -> Bool)
A function that returns True if (and only if) the asset count of the given bundle is excessive.

-> NonEmpty TokenBundle

Splits bundles with excessive asset counts into smaller bundles.

Only token bundles where the isExcessive indicator function returns True will be split.

Returns a list of smaller bundles for which isExcessive returns False.

If none of the bundles in the given list has an excessive asset count, this function will return the original list.

splitBundlesWithExcessiveTokenQuantities

Arguments

:: NonEmpty TokenBundle
Token bundles.

-> TokenQuantity
Maximum allowable token quantity.

-> NonEmpty TokenBundle
The partitioned bundles.

Splits bundles with excessive token quantities into smaller bundles.

Only token bundles containing quantities that exceed the maximum token quantity will be split.

If none of the bundles in the given list contain a quantity that exceeds the maximum token quantity, this function will return the original list.

Grouping and ungrouping

groupByKey :: forall k v. Ord k => [(k, v)] -> Map k (NonEmpty v)

ungroupByKey :: forall k v. Map k (NonEmpty v) -> [(k, v)]

Round-robin processing

runRoundRobin :: s -> (s' -> s) -> [s -> Maybe s'] -> s

runRoundRobinM :: Monad m => s -> (s' -> s) -> [s -> m (Maybe s')] -> m s

Utility classes

newtype AssetCount a

Constructors

AssetCount

Fields

unAssetCount :: a

Utility functions

distance :: Natural -> Natural -> Natural

mapMaybe :: (a -> Maybe b) -> NonEmpty a -> [b]

balanceMissing :: BalanceInsufficientError -> TokenBundle

Calculate the missing balance from a BalanceInsufficientError.