Cardano.Wallet.CoinSelection.Internal.Context

Description

This module provides the SelectionContext class, which provides a shared context for types used by coin selection.

Selection contexts

class (Buildable (Address c), Buildable (UTxO c), Ord (Address c), Ord (UTxO c), Show (Address c), Show (UTxO c)) => SelectionContext c

Provides a shared context for types used by coin selection.

Associated Types

type Address c

A target address to which payments can be made.

type UTxO c

A unique identifier for an individual UTxO.