Cardano.Wallet.DB.Layer

Description

An implementation of the DBLayer which uses Persistent and SQLite.

Directory of single-file wallet databases

newDBFactory

Arguments

:: forall s k. (PersistAddressBook s, PersistPrivateKey (k 'RootK), WalletKey k)

=> Tracer IO DBFactoryLog
Logging object

-> DefaultFieldValues
Default database field values, used during migration.

-> TimeInterpreter IO
Time interpreter for slot to time conversions

-> Maybe FilePath
Path to database directory, or Nothing for in-memory database

-> IO (DBFactory IO s k)

Instantiate a DBFactory from a given directory, or in-memory for testing.

findDatabases :: forall k. WalletKey k => Tracer IO DBFactoryLog -> FilePath -> IO [WalletId]

Return all wallet databases that match the specified key type within the specified directory.

data DBFactoryLog

Constructors

MsgFoundDatabase FilePath Text

MsgUnknownDBFile FilePath

MsgRemoving Text

MsgRemovingInUse Text Int

MsgRemovingDatabaseFile Text DeleteSqliteDatabaseLog

MsgWaitingForDatabase Text (Maybe Int)

MsgWalletDB FilePath WalletDBLog

Internal implementation

withDBLayer

Arguments

:: forall s k a. (PersistAddressBook s, PersistPrivateKey (k 'RootK), WalletKey k)

=> Tracer IO WalletDBLog
Logging object

-> DefaultFieldValues
Default database field values, used during migration.

-> FilePath
Path to database file

-> TimeInterpreter IO
Time interpreter for slot to time conversions

-> (DBLayer IO s k -> IO a)
Action to run.

-> IO a

Runs an action with a connection to the SQLite database.

Database migrations are run to create tables if necessary.

If the given file path does not exist, it will be created by the sqlite library.

withDBLayerInMemory

Arguments

:: forall s k a. (PersistAddressBook s, PersistPrivateKey (k 'RootK))

=> Tracer IO WalletDBLog
Logging object

-> TimeInterpreter IO
Time interpreter for slot to time conversions

-> (DBLayer IO s k -> IO a)

-> IO a

Runs an IO action with a new DBLayer backed by a sqlite in-memory database.

newtype WalletDBLog

Constructors

MsgDB DBLog

data CacheBehavior

What to do with regards to caching. This is useful to disable caching in database benchmarks.

Constructors

NoCache

CacheLatestCheckpoint

Unbracketed internal implementation

newDBLayerWith

Arguments

:: forall s k. (PersistAddressBook s, PersistPrivateKey (k 'RootK))

=> CacheBehavior
Option to disable caching.

-> Tracer IO WalletDBLog
Logging

-> TimeInterpreter IO
Time interpreter for slot to time conversions

-> SqliteContext
A (thread-)safe wrapper for query execution.

-> IO (DBLayer IO s k)

Like newDBLayer, but allows to explicitly specify the caching behavior.

newDBLayerInMemory

Arguments

:: forall s k. (PersistAddressBook s, PersistPrivateKey (k 'RootK))

=> Tracer IO WalletDBLog
Logging object

-> TimeInterpreter IO
Time interpreter for slot to time conversions

-> IO (IO (), DBLayer IO s k)

Creates a DBLayer backed by a sqlite in-memory database.

Returns a cleanup function which you should always use exactly once when finished with the DBLayer.

Interfaces

class AddressBookIso s => PersistAddressBook s where

Functions for saving loading the wallet's address book to from SQLite

Methods

insertPrologue :: WalletId -> Prologue s -> SqlPersistT IO ()

insertDiscoveries :: WalletId -> SlotNo -> Discoveries s -> SqlPersistT IO ()

loadPrologue :: WalletId -> SqlPersistT IO (Maybe (Prologue s))

loadDiscoveries :: WalletId -> SlotNo -> SqlPersistT IO (Discoveries s)

Migration Support

data DefaultFieldValues

A set of default field values that can be consulted when performing a database migration.

Constructors

DefaultFieldValues

Fields

defaultActiveSlotCoefficient :: ActiveSlotCoefficient

defaultDesiredNumberOfPool :: Word16

defaultMinimumUTxOValue :: Coin

defaultHardforkEpoch :: Maybe EpochNo

defaultKeyDeposit :: Coin