Cardano.Wallet.DB.Model

Description

An implementation of the wallet database using only pure functions.

These functions and types model the behaviour of the SQLite database backend, and are used for QuickCheck state machine testing, and the MVar database backend.

Model Types

data Database wid s t xprv

Model database, parameterised by the wallet ID type, the wallet AD state type, the target backend, and the private key type.

Tne type parameters exist so that simpler mock types can be used in place of actual wallet types.

Constructors

Database

Fields

wallets :: !(Map wid (WalletDatabase s t xprv))

Wallet-related information.

txs :: Map (Hash "Tx") (Tx t)

In the database, transactions are global and not associated with any particular wallet.

data WalletDatabase s t xprv

Model database record for a single wallet.

Constructors

WalletDatabase

Fields

checkpoints :: ![Wallet s t]

metadata :: !WalletMetadata

txHistory :: !(Map (Hash "Tx") TxMeta)

xprv :: !(Maybe xprv)

emptyDatabase :: Ord wid => Database wid s t xprv

Produces an empty model database.

type TxHistory t = [(Hash "Tx", (Tx t, TxMeta))]

Shorthand for the readTxHistory result type.

type TxHistoryMap t = Map (Hash "Tx") (Tx t, TxMeta)

Shorthand for the putTxHistory argument type.

filterTxHistory :: SortOrder -> Range SlotId -> TxHistory t -> TxHistory t

Apply optional filters on slotId and sort using the default sort order (first time/slotId, then by TxId) to a TxHistory.

Model Operation Types

type ModelOp wid s t xprv a = Database wid s t xprv -> (Either (Err wid) a, Database wid s t xprv)

A database model operation, which is a function that takes a database and returns: * a value, which is a query of the database, or an error; and * a (possibly) modified database.

data Err wid

All of the possible errors that any of the model database functions might return.

Constructors

NoSuchWallet wid

WalletAlreadyExists wid

Model database functions

mCleanDB :: Ord wid => ModelOp wid s t xprv ()

mCreateWallet :: Ord wid => wid -> Wallet s t -> WalletMetadata -> ModelOp wid s t xprv ()

mRemoveWallet :: Ord wid => wid -> ModelOp wid s t xprv ()

mListWallets :: Ord wid => ModelOp wid s t xprv [wid]

mPutCheckpoint :: Ord wid => wid -> Wallet s t -> ModelOp wid s t xprv ()

mReadCheckpoint :: Ord wid => wid -> ModelOp wid s t xprv (Maybe (Wallet s t))

mPutWalletMeta :: Ord wid => wid -> WalletMetadata -> ModelOp wid s t xprv ()

mReadWalletMeta :: Ord wid => wid -> ModelOp wid s t xprv (Maybe WalletMetadata)

mPutTxHistory :: Ord wid => wid -> TxHistoryMap t -> ModelOp wid s t xprv ()

mReadTxHistory :: forall wid s t xprv. Ord wid => wid -> SortOrder -> Range SlotId -> ModelOp wid s t xprv (TxHistory t)

mPutPrivateKey :: Ord wid => wid -> xprv -> ModelOp wid s t xprv ()

mReadPrivateKey :: Ord wid => wid -> ModelOp wid s t xprv (Maybe xprv)