Cardano.Wallet.DB.Pure.Implementation

Description

An implementation of the wallet database using only pure functions.

These functions and types model the behaviour of the SQLite database backend, and are used for QuickCheck state machine testing, and the MVar database backend.

Model Types

data Database wid s xprv

Model database, parameterised by the wallet ID type, the wallet AD state type, the target backend, and the private key type.

Tne type parameters exist so that simpler mock types can be used in place of actual wallet types.

Constructors

Database

Fields

wallets :: !(Map wid (WalletDatabase s xprv))

Wallet-related information.

txs :: !(Map (Hash "Tx") Tx)

In the database, transactions are global and not associated with any particular wallet.

data WalletDatabase s xprv

Model database record for a single wallet.

Constructors

WalletDatabase

Fields

checkpoints :: !(Map SlotNo (Wallet s))

certificates :: !(Map SlotNo (Maybe PoolId))

stakeKeys :: !(Map SlotNo StakeKeyCertificate)

metadata :: !WalletMetadata

txHistory :: !(Map (Hash "Tx") TxMeta)

xprv :: !(Maybe xprv)

genesisParameters :: !GenesisParameters

rewardAccountBalance :: !Coin

submittedTxs :: !(Map (Hash "Tx") (SealedTx, SlotNo))

emptyDatabase :: Ord wid => Database wid s xprv

Produces an empty model database.

type TxHistory = [(Tx, TxMeta)]

Shorthand for the readTxHistory result type.

type TxHistoryMap = Map (Hash "Tx") (Tx, TxMeta)

Shorthand for the putTxHistory argument type.

filterTxHistory :: Maybe Coin -> SortOrder -> Range SlotNo -> TxHistory -> TxHistory

Apply optional filters on slotNo and sort using the default sort order (first time/slotNo, then by TxId) to a TxHistory.

Model Operation Types

type ModelOp wid s xprv a = Database wid s xprv -> (Either (Err wid) a, Database wid s xprv)

A database model operation, which is a function that takes a database and returns: * a value, which is a query of the database, or an error; and * a (possibly) modified database.

data Err wid

All of the possible errors that any of the model database functions might return.

Constructors

NoSuchWallet wid

WalletAlreadyExists wid

NoSuchTx wid (Hash "Tx")

CantRemoveTxInLedger wid (Hash "Tx")

Model database functions

mCleanDB :: Ord wid => ModelOp wid s xprv ()

mInitializeWallet :: forall wid s xprv. Ord wid => wid -> Wallet s -> WalletMetadata -> TxHistory -> GenesisParameters -> ModelOp wid s xprv ()

mRemoveWallet :: Ord wid => wid -> ModelOp wid s xprv ()

mListWallets :: Ord wid => ModelOp wid s xprv [wid]

mPutCheckpoint :: Ord wid => wid -> Wallet s -> ModelOp wid s xprv ()

mReadCheckpoint :: Ord wid => wid -> ModelOp wid s xprv (Maybe (Wallet s))

mListCheckpoints :: Ord wid => wid -> ModelOp wid s xprv [ChainPoint]

mRollbackTo :: Ord wid => wid -> Slot -> ModelOp wid s xprv ChainPoint

mPutWalletMeta :: Ord wid => wid -> WalletMetadata -> ModelOp wid s xprv ()

mReadWalletMeta :: Ord wid => TimeInterpreter Identity -> wid -> ModelOp wid s xprv (Maybe WalletMetadata)

mPutDelegationCertificate :: Ord wid => wid -> DelegationCertificate -> SlotNo -> ModelOp wid s xprv ()

mIsStakeKeyRegistered :: Ord wid => wid -> ModelOp wid s xprv Bool

mPutTxHistory :: forall wid s xprv. Ord wid => wid -> TxHistory -> ModelOp wid s xprv ()

mReadTxHistory :: forall wid s xprv. Ord wid => TimeInterpreter Identity -> wid -> Maybe Coin -> SortOrder -> Range SlotNo -> Maybe TxStatus -> ModelOp wid s xprv [TransactionInfo]

mPutLocalTxSubmission :: Ord wid => wid -> Hash "Tx" -> SealedTx -> SlotNo -> ModelOp wid s xprv ()

mReadLocalTxSubmissionPending :: Ord wid => wid -> ModelOp wid s xprv [LocalTxSubmissionStatus SealedTx]

mUpdatePendingTxForExpiry :: Ord wid => wid -> SlotNo -> ModelOp wid s xprv ()

mRemovePendingOrExpiredTx :: Ord wid => wid -> Hash "Tx" -> ModelOp wid s xprv ()

mPutPrivateKey :: Ord wid => wid -> xprv -> ModelOp wid s xprv ()

mReadPrivateKey :: Ord wid => wid -> ModelOp wid s xprv (Maybe xprv)

mReadGenesisParameters :: Ord wid => wid -> ModelOp wid s xprv (Maybe GenesisParameters)

mPutDelegationRewardBalance :: Ord wid => wid -> Coin -> ModelOp wid s xprv ()

mReadDelegationRewardBalance :: Ord wid => wid -> ModelOp wid s xprv Coin

mCheckWallet :: Ord wid => wid -> ModelOp wid s xprv ()