Cardano.Wallet.DB.Pure.Layer

Description

Dummy implementation of the database-layer, using MVar. This may be good for testing to compare with an implementation on a real data store, or to use when compiling the wallet for targets which don't have SQLite.

Documentation

newDBLayer :: forall m s k. (MonadUnliftIO m, MonadFail m) => TimeInterpreter Identity -> m (DBLayer m s k)

Instantiate a new in-memory "database" layer that simply stores data in a local MVar. Data vanishes if the software is shut down.