Cardano.Wallet.DB.Sqlite

Description

An implementation of the DBLayer which uses Persistent and SQLite.

Documentation

newDBLayer

Arguments

:: (IsOurs s Address, IsOurs s ChimericAccount, NFData s, Show s, PersistState s, PersistPrivateKey (k RootK))

=> Tracer IO DBLog
Logging object

-> DefaultFieldValues
Default database field values, used during migration.

-> Maybe FilePath
Path to database file, or Nothing for in-memory database

-> IO (SqliteContext, DBLayer IO s k)

Sets up a connection to the SQLite database.

Database migrations are run to create tables if necessary.

If the given file path does not exist, it will be created by the sqlite library.

getDBLayer will provide the actual DBLayer implementation. The database should be closed with destroyDBLayer. If you use withDBLayer then both of these things will be handled for you.

newDBFactory

Arguments

:: (IsOurs s Address, IsOurs s ChimericAccount, NFData s, Show s, PersistState s, PersistPrivateKey (k RootK), WalletKey k)

=> Tracer IO DBLog
Logging object

-> DefaultFieldValues
Default database field values, used during migration.

-> Maybe FilePath
Path to database directory, or Nothing for in-memory database

-> IO (DBFactory IO s k)

Instantiate a DBFactory from a given directory

findDatabases :: forall k. WalletKey k => Tracer IO DatabasesStartupLog -> FilePath -> IO [WalletId]

Return all wallet databases that match the specified key type within the specified directory.

withDBLayer

Arguments

:: (IsOurs s Address, IsOurs s ChimericAccount, NFData s, Show s, PersistState s, PersistPrivateKey (k RootK))

=> Tracer IO DBLog
Logging object

-> DefaultFieldValues
Default database field values, used during migration.

-> Maybe FilePath
Path to database directory, or Nothing for in-memory database

-> ((SqliteContext, DBLayer IO s k) -> IO a)
Action to run.

-> IO a

Runs an action with a connection to the SQLite database.

Database migrations are run to create tables if necessary.

If the given file path does not exist, it will be created by the sqlite library.

Interfaces

class PersistState s where

Functions for saving/loading the wallet's address discovery state into SQLite.

Methods

insertState :: (WalletId, SlotId) -> s -> SqlPersistT IO ()

Store the state for a checkpoint.

selectState :: (WalletId, SlotId) -> SqlPersistT IO (Maybe s)

Load the state for a checkpoint.

Migration Support

newtype DefaultFieldValues

A set of default field values that can be consulted when performing a database migration.

Constructors

DefaultFieldValues

Fields

defaultActiveSlotCoefficient :: ActiveSlotCoefficient

Logging

data DatabasesStartupLog

Log messages arising from accessing the wallet repository.

Constructors

MsgFoundDatabase FilePath WalletId

MsgUnknownFile FilePath