Cardano.Wallet.DB.Sqlite.Migration

Description

Old-style manual migrations of the SQLlite database. These migrations are soon to be removed in favor of a file format with version number.

Documentation

data DefaultFieldValues

A set of default field values that can be consulted when performing a database migration.

Constructors

DefaultFieldValues

Fields

defaultActiveSlotCoefficient :: ActiveSlotCoefficient

defaultDesiredNumberOfPool :: Word16

defaultMinimumUTxOValue :: Coin

defaultHardforkEpoch :: Maybe EpochNo

defaultKeyDeposit :: Coin

migrateManually :: WalletKey k => Tracer IO DBLog -> Proxy k -> DefaultFieldValues -> [ManualMigration]

Executes any manual database migration steps that may be required on startup.

newtype SchemaVersion

Constructors

SchemaVersion Natural

currentSchemaVersion :: SchemaVersion

data InvalidDatabaseSchemaVersion

Constructors

InvalidDatabaseSchemaVersion

Fields

expectedVersion :: SchemaVersion

actualVersion :: SchemaVersion