Cardano.Wallet.DB.Sqlite.Schema

Description

Auto-generated Sqlite & Persistent machinery via Template-Haskell. This has been moved into a separate file so that we can treat it slightly differently when computing code-coverage.

More than 6K lines end-up being generated from the instructions below! As a result, we're going to ignore code-coverage on the following module and, no hand-written functions should be written in this module!

Documentation

data Wallet

Constructors

Wallet

Fields

walId :: !WalletId

walCreationTime :: !UTCTime

walName :: !Text

walPassphraseLastUpdatedAt :: !(Maybe UTCTime)

walPassphraseScheme :: !(Maybe PassphraseScheme)

walGenesisHash :: !BlockId

walGenesisStart :: !UTCTime

type WalletId = Key Wallet

data PrivateKey

Constructors

PrivateKey

Fields

privateKeyWalletId :: !WalletId

privateKeyRootKey :: !ByteString

privateKeyHash :: !ByteString

type PrivateKeyId = Key PrivateKey

data TxMeta

Constructors

TxMeta

Fields

txMetaTxId :: !TxId

txMetaWalletId :: !WalletId

txMetaStatus :: !TxStatus

txMetaDirection :: !Direction

txMetaSlot :: !SlotNo

txMetaBlockHeight :: !Word32

txMetaAmount :: !Coin

txMetadata :: !(Maybe TxMetadata)

txMetaSlotExpires :: !(Maybe SlotNo)

txMetaFee :: !(Maybe Word64)

txMetaScriptValidity :: !(Maybe Bool)

type TxMetaId = Key TxMeta

data TxIn

Constructors

TxIn

Fields

txInputTxId :: !TxId

txInputOrder :: !Int

txInputSourceTxId :: !TxId

txInputSourceIndex :: !Word32

txInputSourceAmount :: !Coin

type TxInId = Key TxIn

data TxCollateral

Constructors

TxCollateral

Fields

txCollateralTxId :: !TxId

txCollateralOrder :: !Int

txCollateralSourceTxId :: !TxId

txCollateralSourceIndex :: !Word32

txCollateralSourceAmount :: !Coin

type TxCollateralId = Key TxCollateral

data TxOut

Constructors

TxOut

Fields

txOutputTxId :: !TxId

txOutputIndex :: !Word32

txOutputAddress :: !Address

txOutputAmount :: !Coin

type TxOutId = Key TxOut

data TxOutToken

Constructors

TxOutToken

Fields

txOutTokenTxId :: !TxId

txOutTokenTxIndex :: !Word32

txOutTokenPolicyId :: !TokenPolicyId

txOutTokenName :: !TokenName

txOutTokenQuantity :: !TokenQuantity

type TxOutTokenId = Key TxOutToken

data TxCollateralOut

Constructors

TxCollateralOut

Fields

txCollateralOutTxId :: !TxId

txCollateralOutAddress :: !Address

txCollateralOutAmount :: !Coin

type TxCollateralOutId = Key TxCollateralOut

data TxCollateralOutToken

Constructors

TxCollateralOutToken

Fields

txCollateralOutTokenTxId :: !TxId

txCollateralOutTokenPolicyId :: !TokenPolicyId

txCollateralOutTokenName :: !TokenName

txCollateralOutTokenQuantity :: !TokenQuantity

type TxCollateralOutTokenId = Key TxCollateralOutToken

data TxWithdrawal

Constructors

TxWithdrawal

Fields

txWithdrawalTxId :: !TxId

txWithdrawalAmount :: !Coin

txWithdrawalAccount :: !RewardAccount

type TxWithdrawalId = Key TxWithdrawal

data LocalTxSubmission

Constructors

LocalTxSubmission

Fields

localTxSubmissionTxId :: !TxId

localTxSubmissionWalletId :: !WalletId

localTxSubmissionLastSlot :: !SlotNo

localTxSubmissionTx :: !SealedTx

type LocalTxSubmissionId = Key LocalTxSubmission

data Checkpoint

Constructors

Checkpoint

Fields

checkpointWalletId :: !WalletId

checkpointSlot :: !SlotNo

checkpointHeaderHash :: !BlockId

checkpointParentHash :: !BlockId

checkpointBlockHeight :: !Word32

type CheckpointId = Key Checkpoint

data ProtocolParameters

Constructors

ProtocolParameters

Fields

protocolParametersWalletId :: !WalletId

protocolParametersFeePolicy :: !FeePolicy

protocolParametersTxMaxSize :: !Word16

protocolParametersDecentralizationLevel :: !Percentage

protocolParametersDesiredNumberOfPools :: !Word16

protocolParametersMinimumUtxoValue :: !Coin

protocolParametersHardforkEpoch :: !(Maybe EpochNo)

protocolParametersKeyDeposit :: !Coin

type ProtocolParametersId = Key ProtocolParameters

data StakeKeyCertificate

Constructors

StakeKeyCertificate

Fields

stakeKeyCertWalletId :: !WalletId

stakeKeyCertSlot :: !SlotNo

stakeKeyCertType :: !StakeKeyCertificate

type StakeKeyCertificateId = Key StakeKeyCertificate

data DelegationCertificate

Constructors

DelegationCertificate

Fields

certWalletId :: !WalletId

certSlot :: !SlotNo

certPoolId :: !(Maybe PoolId)

type DelegationCertificateId = Key DelegationCertificate

data DelegationReward

Constructors

DelegationReward

Fields

rewardWalletId :: !WalletId

rewardAccountBalance :: !Word64

type DelegationRewardId = Key DelegationReward

data UTxO

Constructors

UTxO

Fields

utxoWalletId :: !WalletId

utxoSlot :: !SlotNo

utxoInputId :: !TxId

utxoInputIndex :: !Word32

utxoOutputAddress :: !Address

utxoOutputCoin :: !Coin

type UTxOId = Key UTxO

data UTxOToken

Constructors

UTxOToken

Fields

utxoTokenWalletId :: !WalletId

utxoTokenSlot :: !SlotNo

utxoTokenTxId :: !TxId

utxoTokenTxIndex :: !Word32

utxoTokenPolicyId :: !TokenPolicyId

utxoTokenName :: !TokenName

utxoTokenQuantity :: !TokenQuantity

type UTxOTokenId = Key UTxOToken

data SeqState

Constructors

SeqState

Fields

seqStateWalletId :: !WalletId

seqStateExternalGap :: !AddressPoolGap

seqStateInternalGap :: !AddressPoolGap

seqStateAccountXPub :: !ByteString

seqStatePolicyXPub :: !(Maybe ByteString)

seqStateRewardXPub :: !ByteString

seqStateDerivationPrefix :: !DerivationPrefix

type SeqStateId = Key SeqState

data SeqStateAddress

Constructors

SeqStateAddress

Fields

seqStateAddressWalletId :: !WalletId

seqStateAddressSlot :: !SlotNo

seqStateAddressAddress :: !Address

seqStateAddressIndex :: !Word32

seqStateAddressRole :: !Role

seqStateAddressStatus :: !AddressState

type SeqStateAddressId = Key SeqStateAddress

data SeqStatePendingIx

Constructors

SeqStatePendingIx

Fields

seqStatePendingWalletId :: !WalletId

seqStatePendingIxIndex :: !Word32

type SeqStatePendingIxId = Key SeqStatePendingIx

data RndState

Constructors

RndState

Fields

rndStateWalletId :: !WalletId

rndStateAccountIndex :: !Word32

rndStateGen :: !StdGen

rndStateHdPassphrase :: !HDPassphrase

type RndStateId = Key RndState

data RndStateAddress

Constructors

RndStateAddress

Fields

rndStateAddressWalletId :: !WalletId

rndStateAddressSlot :: !SlotNo

rndStateAddressAccountIndex :: !Word32

rndStateAddressIndex :: !Word32

rndStateAddressAddress :: !Address

rndStateAddressStatus :: !AddressState

type RndStateAddressId = Key RndStateAddress

data RndStatePendingAddress

Constructors

RndStatePendingAddress

Fields

rndStatePendingAddressWalletId :: !WalletId

rndStatePendingAddressAccountIndex :: !Word32

rndStatePendingAddressIndex :: !Word32

rndStatePendingAddressAddress :: !Address

type RndStatePendingAddressId = Key RndStatePendingAddress

data SharedState

Constructors

SharedState

Fields

sharedStateWalletId :: !WalletId

sharedStateAccountXPub :: !ByteString

sharedStateScriptGap :: !AddressPoolGap

sharedStatePaymentScript :: !(Script Cosigner)

sharedStateDelegationScript :: !(Maybe (Script Cosigner))

sharedStateDerivationPrefix :: !DerivationPrefix

type SharedStateId = Key SharedState

data SharedStatePendingIx

Constructors

SharedStatePendingIx

Fields

sharedStatePendingWalletId :: !WalletId

sharedStatePendingIxIndex :: !Word32

type SharedStatePendingIxId = Key SharedStatePendingIx

data CosignerKey

Constructors

CosignerKey

Fields

cosignerKeyWalletId :: !WalletId

cosignerKeyCredential :: !CredentialType

cosignerKeyAccountXPub :: !ByteString

cosignerKeyIndex :: !Word8

type CosignerKeyId = Key CosignerKey

migrateAll :: Migration

entityDefListFormigrateAll :: [EntityDef]

cosigner_key :: CosignerKey -> Key Wallet

shared_state_address_pending :: SharedStatePendingIx -> Key Wallet

shared_state :: SharedState -> Key Wallet

rnd_state_pending_address :: RndStatePendingAddress -> Key Wallet

rnd_state_address :: RndStateAddress -> Key Checkpoint

rnd_state :: RndState -> Key Wallet

seq_state_address_pending :: SeqStatePendingIx -> Key Wallet

seq_state_address :: SeqStateAddress -> Key Checkpoint

seq_state :: SeqState -> Key Wallet

utxot :: UTxOToken -> Key Checkpoint

utxo :: UTxO -> Key Checkpoint

delegationReward :: DelegationReward -> Key Wallet

delegationCertificate :: DelegationCertificate -> Key Wallet

stakeKeyRegistration :: StakeKeyCertificate -> Key Wallet

fk_wallet_protocol_parameters :: ProtocolParameters -> Key Wallet

checkpoint :: Checkpoint -> Key Wallet

fk_tx_meta :: LocalTxSubmission -> Key TxMeta

txCollateralOut :: TxCollateralOutToken -> Key TxCollateralOut

txOut :: TxOutToken -> Key TxOut

fk_wallet_tx_meta :: TxMeta -> Key Wallet

fk_wallet_private_key :: PrivateKey -> Key Wallet