Cardano.Wallet.DB.Sqlite.Stores

Description

Store implementations that can store various wallet types in an SQLite database using persistent.

FIXME LATER during ADP-1043:

Inline the contents of Cardano.Wallet.DB.Sqlite.CheckpointsOld into this module

Use Table and Embedding to construct the relevant Store rather than implementing loadS, writeS, updateS in a monadic fashion. Hide the new implementation behind a feature flag, i.e. Cardano.Wallet.DB.Sqlite.StoresNew.

Documentation

mkStoreWallets :: forall s key. (PersistAddressBook s, key ~ WalletId) => Store (SqlPersistT IO) (DeltaMap key (DeltaWalletState s))

Store for WalletState of multiple different wallets.

class AddressBookIso s => PersistAddressBook s where

Functions for saving loading the wallet's address book to from SQLite

Methods

insertPrologue :: WalletId -> Prologue s -> SqlPersistT IO ()

insertDiscoveries :: WalletId -> SlotNo -> Discoveries s -> SqlPersistT IO ()

loadPrologue :: WalletId -> SqlPersistT IO (Maybe (Prologue s))

loadDiscoveries :: WalletId -> SlotNo -> SqlPersistT IO (Discoveries s)

blockHeaderFromEntity :: Checkpoint -> BlockHeader

Testing

mkStoreWallet :: forall s. PersistAddressBook s => WalletId -> Store (SqlPersistT IO) (DeltaWalletState s)

Store for WalletState of a single wallet.