Cardano.Wallet.DB.Sqlite.TH

Description

Auto-generated Sqlite & Persistent machinery via Template-Haskell. This has been moved into a separate file so that we can treat it slightly differently when computing code-coverage.

More than 6K lines end-up being generated from the instructions below! As a result, we're going to ignore code-coverage on the following module and, no hand-written functions should be written in this module!

Documentation

data Wallet

Constructors

Wallet

Fields

walId :: !WalletId

walCreationTime :: !UTCTime

walName :: !Text

walPassphraseLastUpdatedAt :: !(Maybe UTCTime)

walStatus :: !WalletState

walDelegation :: !(Maybe Text)

type WalletId = Key Wallet

data PrivateKey

Constructors

PrivateKey

Fields

privateKeyWalletId :: !WalletId

privateKeyRootKey :: !ByteString

privateKeyHash :: !ByteString

type PrivateKeyId = Key PrivateKey

data TxMeta

Constructors

TxMeta

Fields

txMetaTxId :: !TxId

txMetaWalletId :: !WalletId

txMetaStatus :: !TxStatus

txMetaDirection :: !Direction

txMetaSlotId :: !SlotId

txMetaAmount :: !Natural

type TxMetaId = Key TxMeta

data TxIn

Constructors

TxIn

Fields

txInputTxId :: !TxId

txInputOrder :: !Int

txInputSourceTxId :: !TxId

txInputSourceIndex :: !Word32

txInputSourceAmount :: !(Maybe Coin)

type TxInId = Key TxIn

data TxOut

Constructors

TxOut

Fields

txOutputTxId :: !TxId

txOutputIndex :: !Word32

txOutputAddress :: !Address

txOutputAmount :: !Coin

type TxOutId = Key TxOut

data Checkpoint

Constructors

Checkpoint

Fields

checkpointWalletId :: !WalletId

checkpointSlot :: !SlotId

checkpointParent :: !BlockId

checkpointBlockHeight :: !Word64

type CheckpointId = Key Checkpoint

data UTxO

Constructors

UTxO

Fields

utxoWalletId :: !WalletId

utxoSlot :: !SlotId

utxoSlotSpent :: !(Maybe SlotId)

utxoInputId :: !TxId

utxoInputIndex :: !Word32

utxoOutputAddress :: !Address

utxoOutputCoin :: !Coin

type UTxOId = Key UTxO

data SeqState

Constructors

SeqState

Fields

seqStateWalletId :: !WalletId

seqStateExternalGap :: !AddressPoolGap

seqStateInternalGap :: !AddressPoolGap

seqStateAccountXPub :: !AddressPoolXPub

type SeqStateId = Key SeqState

data SeqStateAddress

Constructors

SeqStateAddress

Fields

seqStateAddressWalletId :: !WalletId

seqStateAddressSlot :: !SlotId

seqStateAddressAddress :: !Address

seqStateAddressIndex :: !Word32

seqStateAddressChangeChain :: !ChangeChain

type SeqStateAddressId = Key SeqStateAddress

data SeqStateCheckpoint

Constructors

SeqStateCheckpoint

Fields

seqStateCheckpointWalletId :: !WalletId

seqStateCheckpointSlot :: !SlotId

type SeqStateCheckpointId = Key SeqStateCheckpoint

data SeqStatePendingIx

Constructors

SeqStatePendingIx

Fields

seqStatePendingIxCheckpointId :: !(Key SeqStateCheckpoint)

seqStatePendingIxIndex :: !Word32

type SeqStatePendingIxId = Key SeqStatePendingIx

data RndState

Constructors

RndState

Fields

rndStateWalletId :: !WalletId

rndStateAccountIndex :: !Word32

type RndStateId = Key RndState

data RndStateAddress

Constructors

RndStateAddress

Fields

rndStateAddressWalletId :: !WalletId

rndStateAddressSlot :: !SlotId

rndStateAddressAccountIndex :: !Word32

rndStateAddressIndex :: !Word32

rndStateAddressAddress :: !Address

type RndStateAddressId = Key RndStateAddress

data RndStateCheckpoint

Constructors

RndStateCheckpoint

Fields

rndStateCheckpointWalletId :: !WalletId

rndStateCheckpointSlot :: !SlotId

rndStateCheckpointGen :: !StdGen

type RndStateCheckpointId = Key RndStateCheckpoint

data RndStatePendingAddress

Constructors

RndStatePendingAddress

Fields

rndStatePendingAddressCheckpointId :: !(Key RndStateCheckpoint)

rndStatePendingAddressAccountIndex :: !Word32

rndStatePendingAddressIndex :: !Word32

rndStatePendingAddressAddress :: !Address

type RndStatePendingAddressId = Key RndStatePendingAddress

migrateAll :: Migration

fk_checkpoint_rnd_state :: RndStateCheckpoint -> Key Checkpoint

fk_wallet_rnd_state :: RndState -> Key Wallet

fk_checkpoint_seq_state :: SeqStateCheckpoint -> Key Checkpoint

fk_wallet_seq_state :: SeqState -> Key Wallet

fk_wallet_utxo :: UTxO -> Key Wallet

fk_wallet_checkpoint :: Checkpoint -> Key Wallet

fk_wallet_tx_meta :: TxMeta -> Key Wallet

fk_wallet_private_key :: PrivateKey -> Key Wallet