Cardano.Wallet.DB.Sqlite.Types

Description

This module contains instances and types necessary for storing wallets in a SQL database with Persistent.

It's in a separate module due to the GHC stage restriction.

The ToJSON/FromJSON and Read instance orphans exist due to class constraints on Persistent functions.

Documentation

sqlSettings' :: MkPersistSettings

Settings for generating the Persistent types.

fromText' :: FromText a => Text -> Either Text a

fromText but with a simpler error type.

aesonFromText :: FromText a => String -> Value -> Parser a

Aeson parser defined in terms of fromText

fromPersistValueFromText :: FromText a => PersistValue -> Either Text a

fromPersistValue defined in terms of fromText

fromPersistValueRead :: Read a => PersistValue -> Either Text a

fromPersistValue defined in terms of the Read class

directionToBool :: Direction -> Bool

directionFromBool :: Bool -> Direction

newtype TxId

Wraps 'Hash Tx' because the persistent entity syntax doesn't seem to support parameterized types.

Constructors

TxId

Fields

getTxId :: Hash "Tx"

newtype BlockId

Constructors

BlockId

Fields

getBlockId :: Hash "BlockHeader"

hashOfNoParent :: Hash "BlockHeader"

Magic value that denotes the hash of the parent of the genesis block (which does not exist). This value is used for serializing the Nothing case of the #parentHeaderHash field.

fromMaybeHash :: Maybe (Hash "BlockHeader") -> BlockId

toMaybeHash :: BlockId -> Maybe (Hash "BlockHeader")

persistSlotNo :: SlotNo -> PersistValue

unPersistSlotNo :: PersistValue -> Either Text SlotNo

mkEpochNo :: Word32 -> Either Text EpochNo

persistEpochNo :: EpochNo -> PersistValue

stdGenFromString :: String -> Either Text StdGen

In random < 1.2 there used to be an instance Read StdGen, but no longer.

The format used to look like this: 5889121503043413025 17512980752375952679

stdGenToString :: StdGen -> String

Equivalent to the old random < 1.2 StdGen Show instance.

newtype HDPassphrase

Constructors

HDPassphrase (Passphrase "addr-derivation-payload")

newtype EitherText a

Newtype to get a MonadFail instance for Either Text.

We need it to use parseTimeM.

Constructors

EitherText

Fields

getEitherText :: Either Text a

Orphan instances

Read SlotNo

Instance details

Methods

readsPrec :: Int -> ReadS SlotNo

readList :: ReadS [SlotNo]

readPrec :: ReadPrec SlotNo

readListPrec :: ReadPrec [SlotNo]

Read PoolOwner

Instance details

Methods

readsPrec :: Int -> ReadS PoolOwner

readList :: ReadS [PoolOwner]

readPrec :: ReadPrec PoolOwner

readListPrec :: ReadPrec [PoolOwner]

Read PoolId

Instance details

Methods

readsPrec :: Int -> ReadS PoolId

readList :: ReadS [PoolId]

readPrec :: ReadPrec PoolId

readListPrec :: ReadPrec [PoolId]

Read StakePoolMetadataUrl

Instance details

Methods

readsPrec :: Int -> ReadS StakePoolMetadataUrl

readList :: ReadS [StakePoolMetadataUrl]

readPrec :: ReadPrec StakePoolMetadataUrl

readListPrec :: ReadPrec [StakePoolMetadataUrl]

Read StakePoolMetadataHash

Instance details

Methods

readsPrec :: Int -> ReadS StakePoolMetadataHash

readList :: ReadS [StakePoolMetadataHash]

readPrec :: ReadPrec StakePoolMetadataHash

readListPrec :: ReadPrec [StakePoolMetadataHash]

Read WalletId

Instance details

Methods

readsPrec :: Int -> ReadS WalletId

readList :: ReadS [WalletId]

readPrec :: ReadPrec WalletId

readListPrec :: ReadPrec [WalletId]

ToJSON RewardAccount

Instance details

Methods

toJSON :: RewardAccount -> Value

toEncoding :: RewardAccount -> Encoding

toJSONList :: [RewardAccount] -> Value

toEncodingList :: [RewardAccount] -> Encoding

ToJSON PoolId

Instance details

Methods

toJSON :: PoolId -> Value

toEncoding :: PoolId -> Encoding

toJSONList :: [PoolId] -> Value

toEncodingList :: [PoolId] -> Encoding

ToJSON StakePoolMetadataUrl

Instance details

Methods

toJSON :: StakePoolMetadataUrl -> Value

toEncoding :: StakePoolMetadataUrl -> Encoding

toJSONList :: [StakePoolMetadataUrl] -> Value

toEncodingList :: [StakePoolMetadataUrl] -> Encoding

ToJSON StakePoolMetadataHash

Instance details

Methods

toJSON :: StakePoolMetadataHash -> Value

toEncoding :: StakePoolMetadataHash -> Encoding

toJSONList :: [StakePoolMetadataHash] -> Value

toEncodingList :: [StakePoolMetadataHash] -> Encoding

ToJSON WalletId

Instance details

Methods

toJSON :: WalletId -> Value

toEncoding :: WalletId -> Encoding

toJSONList :: [WalletId] -> Value

toEncodingList :: [WalletId] -> Encoding

FromJSON RewardAccount

Instance details

Methods

parseJSON :: Value -> Parser RewardAccount

parseJSONList :: Value -> Parser [RewardAccount]

FromJSON PoolId

Instance details

Methods

parseJSON :: Value -> Parser PoolId

parseJSONList :: Value -> Parser [PoolId]

FromJSON StakePoolMetadataUrl

Instance details

Methods

parseJSON :: Value -> Parser StakePoolMetadataUrl

parseJSONList :: Value -> Parser [StakePoolMetadataUrl]

FromJSON StakePoolMetadataHash

Instance details

Methods

parseJSON :: Value -> Parser StakePoolMetadataHash

parseJSONList :: Value -> Parser [StakePoolMetadataHash]

FromJSON WalletId

Instance details

Methods

parseJSON :: Value -> Parser WalletId

parseJSONList :: Value -> Parser [WalletId]

ToHttpApiData SlotNo

Instance details

Methods

toUrlPiece :: SlotNo -> Text

toEncodedUrlPiece :: SlotNo -> Builder

toHeader :: SlotNo -> ByteString

toQueryParam :: SlotNo -> Text

ToHttpApiData RewardAccount

Instance details

Methods

toUrlPiece :: RewardAccount -> Text

toEncodedUrlPiece :: RewardAccount -> Builder

toHeader :: RewardAccount -> ByteString

toQueryParam :: RewardAccount -> Text

ToHttpApiData EpochNo

Instance details

Methods

toUrlPiece :: EpochNo -> Text

toEncodedUrlPiece :: EpochNo -> Builder

toHeader :: EpochNo -> ByteString

toQueryParam :: EpochNo -> Text

ToHttpApiData PoolId

Instance details

Methods

toUrlPiece :: PoolId -> Text

toEncodedUrlPiece :: PoolId -> Builder

toHeader :: PoolId -> ByteString

toQueryParam :: PoolId -> Text

ToHttpApiData StakePoolMetadataUrl

Instance details

Methods

toUrlPiece :: StakePoolMetadataUrl -> Text

toEncodedUrlPiece :: StakePoolMetadataUrl -> Builder

toHeader :: StakePoolMetadataUrl -> ByteString

toQueryParam :: StakePoolMetadataUrl -> Text

ToHttpApiData StakePoolMetadataHash

Instance details

Methods

toUrlPiece :: StakePoolMetadataHash -> Text

toEncodedUrlPiece :: StakePoolMetadataHash -> Builder

toHeader :: StakePoolMetadataHash -> ByteString

toQueryParam :: StakePoolMetadataHash -> Text

ToHttpApiData WalletId

Instance details

Methods

toUrlPiece :: WalletId -> Text

toEncodedUrlPiece :: WalletId -> Builder

toHeader :: WalletId -> ByteString

toQueryParam :: WalletId -> Text

FromHttpApiData SlotNo

Instance details

Methods

parseUrlPiece :: Text -> Either Text SlotNo

parseHeader :: ByteString -> Either Text SlotNo

parseQueryParam :: Text -> Either Text SlotNo

FromHttpApiData RewardAccount

Instance details

Methods

parseUrlPiece :: Text -> Either Text RewardAccount

parseHeader :: ByteString -> Either Text RewardAccount

parseQueryParam :: Text -> Either Text RewardAccount

FromHttpApiData EpochNo

Instance details

Methods

parseUrlPiece :: Text -> Either Text EpochNo

parseHeader :: ByteString -> Either Text EpochNo

parseQueryParam :: Text -> Either Text EpochNo

FromHttpApiData PoolId

Instance details

Methods

parseUrlPiece :: Text -> Either Text PoolId

parseHeader :: ByteString -> Either Text PoolId

parseQueryParam :: Text -> Either Text PoolId

FromHttpApiData StakePoolMetadataUrl

Instance details

Methods

parseUrlPiece :: Text -> Either Text StakePoolMetadataUrl

parseHeader :: ByteString -> Either Text StakePoolMetadataUrl

parseQueryParam :: Text -> Either Text StakePoolMetadataUrl

FromHttpApiData StakePoolMetadataHash

Instance details

Methods

parseUrlPiece :: Text -> Either Text StakePoolMetadataHash

parseHeader :: ByteString -> Either Text StakePoolMetadataHash

parseQueryParam :: Text -> Either Text StakePoolMetadataHash

FromHttpApiData WalletId

Instance details

Methods

parseUrlPiece :: Text -> Either Text WalletId

parseHeader :: ByteString -> Either Text WalletId

parseQueryParam :: Text -> Either Text WalletId

PathPiece SlotNo

Instance details

Methods

fromPathPiece :: Text -> Maybe SlotNo

toPathPiece :: SlotNo -> Text

PathPiece RewardAccount

Instance details

Methods

fromPathPiece :: Text -> Maybe RewardAccount

toPathPiece :: RewardAccount -> Text

PathPiece EpochNo

Instance details

Methods

fromPathPiece :: Text -> Maybe EpochNo

toPathPiece :: EpochNo -> Text

PathPiece PoolId

Instance details

Methods

fromPathPiece :: Text -> Maybe PoolId

toPathPiece :: PoolId -> Text

PathPiece StakePoolMetadataUrl

Instance details

Methods

fromPathPiece :: Text -> Maybe StakePoolMetadataUrl

toPathPiece :: StakePoolMetadataUrl -> Text

PathPiece StakePoolMetadataHash

Instance details

Methods

fromPathPiece :: Text -> Maybe StakePoolMetadataHash

toPathPiece :: StakePoolMetadataHash -> Text

PathPiece WalletId

Instance details

Methods

fromPathPiece :: Text -> Maybe WalletId

toPathPiece :: WalletId -> Text

PersistFieldSql StdGen

Instance details

Methods

sqlType :: Proxy StdGen -> SqlType

PersistFieldSql ScriptHash

Instance details

Methods

sqlType :: Proxy ScriptHash -> SqlType

PersistFieldSql TxMetadata

Instance details

Methods

sqlType :: Proxy TxMetadata -> SqlType

PersistFieldSql SlotNo

Instance details

Methods

sqlType :: Proxy SlotNo -> SqlType

PersistFieldSql POSIXTime

Instance details

Methods

sqlType :: Proxy POSIXTime -> SqlType

PersistFieldSql PassphraseScheme

Instance details

Methods

sqlType :: Proxy PassphraseScheme -> SqlType

PersistFieldSql AddressState

Instance details

Methods

sqlType :: Proxy AddressState -> SqlType

PersistFieldSql Address

Instance details

Methods

sqlType :: Proxy Address -> SqlType

PersistFieldSql TokenQuantity

Instance details

Methods

sqlType :: Proxy TokenQuantity -> SqlType

PersistFieldSql TokenName

Instance details

Methods

sqlType :: Proxy TokenName -> SqlType

PersistFieldSql TokenPolicyId

Instance details

Methods

sqlType :: Proxy TokenPolicyId -> SqlType

PersistFieldSql RewardAccount

Instance details

Methods

sqlType :: Proxy RewardAccount -> SqlType

PersistFieldSql DerivationPrefix

Instance details

Methods

sqlType :: Proxy DerivationPrefix -> SqlType

PersistFieldSql Role

Instance details

Methods

sqlType :: Proxy Role -> SqlType

PersistFieldSql Percentage

Instance details

Methods

sqlType :: Proxy Percentage -> SqlType

PersistFieldSql Coin

Instance details

Methods

sqlType :: Proxy Coin -> SqlType

PersistFieldSql TxScriptValidity

Instance details

Methods

sqlType :: Proxy TxScriptValidity -> SqlType

PersistFieldSql SealedTx

Instance details

Methods

sqlType :: Proxy SealedTx -> SqlType

PersistFieldSql Direction

Instance details

Methods

sqlType :: Proxy Direction -> SqlType

PersistFieldSql TxStatus

Instance details

Methods

sqlType :: Proxy TxStatus -> SqlType

PersistFieldSql PoolMetadataSource

Instance details

Methods

sqlType :: Proxy PoolMetadataSource -> SqlType

PersistFieldSql StakeKeyCertificate

Instance details

Methods

sqlType :: Proxy StakeKeyCertificate -> SqlType

PersistFieldSql EpochNo

Instance details

Methods

sqlType :: Proxy EpochNo -> SqlType

PersistFieldSql FeePolicy

Instance details

Methods

sqlType :: Proxy FeePolicy -> SqlType

PersistFieldSql PoolOwner

Instance details

Methods

sqlType :: Proxy PoolOwner -> SqlType

PersistFieldSql PoolId

Instance details

Methods

sqlType :: Proxy PoolId -> SqlType

PersistFieldSql StakePoolTicker

Instance details

Methods

sqlType :: Proxy StakePoolTicker -> SqlType

PersistFieldSql StakePoolMetadataUrl

Instance details

Methods

sqlType :: Proxy StakePoolMetadataUrl -> SqlType

PersistFieldSql StakePoolMetadataHash

Instance details

Methods

sqlType :: Proxy StakePoolMetadataHash -> SqlType

PersistFieldSql WalletId

Instance details

Methods

sqlType :: Proxy WalletId -> SqlType

PersistFieldSql AddressPoolGap

Instance details

Methods

sqlType :: Proxy AddressPoolGap -> SqlType

PersistFieldSql CredentialType

Instance details

Methods

sqlType :: Proxy CredentialType -> SqlType

PersistField StdGen

Instance details

Methods

toPersistValue :: StdGen -> PersistValue

fromPersistValue :: PersistValue -> Either Text StdGen

PersistField ScriptHash

Instance details

Methods

toPersistValue :: ScriptHash -> PersistValue

fromPersistValue :: PersistValue -> Either Text ScriptHash

PersistField TxMetadata

Instance details

Methods

toPersistValue :: TxMetadata -> PersistValue

fromPersistValue :: PersistValue -> Either Text TxMetadata

PersistField SlotNo

Instance details

Methods

toPersistValue :: SlotNo -> PersistValue

fromPersistValue :: PersistValue -> Either Text SlotNo

PersistField POSIXTime

Instance details

Methods

toPersistValue :: POSIXTime -> PersistValue

fromPersistValue :: PersistValue -> Either Text POSIXTime

PersistField PassphraseScheme

Instance details

Methods

toPersistValue :: PassphraseScheme -> PersistValue

fromPersistValue :: PersistValue -> Either Text PassphraseScheme

PersistField AddressState

Instance details

Methods

toPersistValue :: AddressState -> PersistValue

fromPersistValue :: PersistValue -> Either Text AddressState

PersistField Address

Instance details

Methods

toPersistValue :: Address -> PersistValue

fromPersistValue :: PersistValue -> Either Text Address

PersistField TokenQuantity

Instance details

Methods

toPersistValue :: TokenQuantity -> PersistValue

fromPersistValue :: PersistValue -> Either Text TokenQuantity

PersistField TokenName

Instance details

Methods

toPersistValue :: TokenName -> PersistValue

fromPersistValue :: PersistValue -> Either Text TokenName

PersistField TokenPolicyId

Instance details

Methods

toPersistValue :: TokenPolicyId -> PersistValue

fromPersistValue :: PersistValue -> Either Text TokenPolicyId

PersistField RewardAccount

Instance details

Methods

toPersistValue :: RewardAccount -> PersistValue

fromPersistValue :: PersistValue -> Either Text RewardAccount

PersistField DerivationPrefix

Instance details

Methods

toPersistValue :: DerivationPrefix -> PersistValue

fromPersistValue :: PersistValue -> Either Text DerivationPrefix

PersistField Role

Instance details

Methods

toPersistValue :: Role -> PersistValue

fromPersistValue :: PersistValue -> Either Text Role

PersistField Percentage

Instance details

Methods

toPersistValue :: Percentage -> PersistValue

fromPersistValue :: PersistValue -> Either Text Percentage

PersistField Coin

Instance details

Methods

toPersistValue :: Coin -> PersistValue

fromPersistValue :: PersistValue -> Either Text Coin

PersistField TxScriptValidity

Instance details

Methods

toPersistValue :: TxScriptValidity -> PersistValue

fromPersistValue :: PersistValue -> Either Text TxScriptValidity

PersistField SealedTx

Instance details

Methods

toPersistValue :: SealedTx -> PersistValue

fromPersistValue :: PersistValue -> Either Text SealedTx

PersistField Direction

Instance details

Methods

toPersistValue :: Direction -> PersistValue

fromPersistValue :: PersistValue -> Either Text Direction

PersistField TxStatus

Instance details

Methods

toPersistValue :: TxStatus -> PersistValue

fromPersistValue :: PersistValue -> Either Text TxStatus

PersistField PoolMetadataSource

Instance details

Methods

toPersistValue :: PoolMetadataSource -> PersistValue

fromPersistValue :: PersistValue -> Either Text PoolMetadataSource

PersistField StakeKeyCertificate

Instance details

Methods

toPersistValue :: StakeKeyCertificate -> PersistValue

fromPersistValue :: PersistValue -> Either Text StakeKeyCertificate

PersistField EpochNo

Instance details

Methods

toPersistValue :: EpochNo -> PersistValue

fromPersistValue :: PersistValue -> Either Text EpochNo

PersistField FeePolicy

Instance details

Methods

toPersistValue :: FeePolicy -> PersistValue

fromPersistValue :: PersistValue -> Either Text FeePolicy

PersistField PoolOwner

Instance details

Methods

toPersistValue :: PoolOwner -> PersistValue

fromPersistValue :: PersistValue -> Either Text PoolOwner

PersistField PoolId

Instance details

Methods

toPersistValue :: PoolId -> PersistValue

fromPersistValue :: PersistValue -> Either Text PoolId

PersistField StakePoolTicker

Instance details

Methods

toPersistValue :: StakePoolTicker -> PersistValue

fromPersistValue :: PersistValue -> Either Text StakePoolTicker

PersistField StakePoolMetadataUrl

Instance details

Methods

toPersistValue :: StakePoolMetadataUrl -> PersistValue

fromPersistValue :: PersistValue -> Either Text StakePoolMetadataUrl

PersistField StakePoolMetadataHash

Instance details

Methods

toPersistValue :: StakePoolMetadataHash -> PersistValue

fromPersistValue :: PersistValue -> Either Text StakePoolMetadataHash

PersistField WalletId

Instance details

Methods

toPersistValue :: WalletId -> PersistValue

fromPersistValue :: PersistValue -> Either Text WalletId

PersistField AddressPoolGap

Instance details

Methods

toPersistValue :: AddressPoolGap -> PersistValue

fromPersistValue :: PersistValue -> Either Text AddressPoolGap

PersistField CredentialType

Instance details

Methods

toPersistValue :: CredentialType -> PersistValue

fromPersistValue :: PersistValue -> Either Text CredentialType

ToText ScriptHash

Instance details

Methods

toText :: ScriptHash -> Text

FromText ScriptHash

Instance details

Methods

fromText :: Text -> Either TextDecodingError ScriptHash

PersistFieldSql [PoolOwner]

Instance details

Methods

sqlType :: Proxy [PoolOwner] -> SqlType

PersistFieldSql (Script Cosigner)

Instance details

Methods

sqlType :: Proxy (Script Cosigner) -> SqlType

PersistField [PoolOwner]

Instance details

Methods

toPersistValue :: [PoolOwner] -> PersistValue

fromPersistValue :: PersistValue -> Either Text [PoolOwner]

PersistField (Script Cosigner)

Instance details

Methods

toPersistValue :: Script Cosigner -> PersistValue

fromPersistValue :: PersistValue -> Either Text (Script Cosigner)

FromText [PoolOwner]

Instance details

Methods

fromText :: Text -> Either TextDecodingError [PoolOwner]