Cardano.Wallet.DB.Store.Meta.Model

Description

Pure, low level model for a collection of "meta transactions", i.e. additional data (TxMeta) that the wallet stores for each transaction. Meta transactions are encoded "as" expressed in DB tables.

Documentation

data DeltaTxMetaHistory

All meta-transactions changes, including the addition of new meta-transactions, which has to be done in sync with the transactions store.

Constructors

Manipulate ManipulateTxMetaHistory

Expand TxMetaHistory

data ManipulateTxMetaHistory

Verbs for TxMeta changes that can be issued independently from the transaction store.

Constructors

PruneTxMetaHistory TxId
Remove a meta if it is not in the ledger.

AgeTxMetaHistory SlotNo
Change the state of any meta to Expired if the given slot is equal or after its expiration slot.

RollBackTxMetaHistory SlotNo
Remove all incoming transactions created after the given slot and mark all outgoing transactions after the given slot as Pending.

newtype TxMetaHistory

A collection of TxMeta, indexed by transaction identifier.

Constructors

TxMetaHistory

Fields

relations :: Map TxId TxMeta

mkTxMetaHistory :: WalletId -> [(Tx, TxMeta)] -> TxMetaHistory

Compute a TxMetaHistory for a wallet.