Cardano.Wallet.DB.Store.Transactions.Model

Description

Data type TxHistory for storing a set of transactions. Transactions are encoded "as" expressed in DB tables.

Documentation

data DeltaTxHistory

Verbs to change a TxHistory.

Constructors

Append TxHistory
Add new set of transactions. Overwrites transactions whose id is already present in the TxHistory.

DeleteTx TxId
Try to remove the transaction at the given transaction id.

type TxHistory = TxHistoryF 'Without

Shortcut type for transaction history where inputs are not decorated with their corresponding TxOut.

newtype TxHistoryF f

Transactions history is TxRelationFs indexed by TxId

Constructors

TxHistoryF (Map TxId (TxRelationF f))

data TxRelationF (f :: Decoration)

A low level definition of a transaction covering all transaction content by collecting all related-to-index database rows. Normalization is performed anyway after the first relation level. All values used here are records in the database. Foreign keys are used to group data correctly, but they are not removed from the data.

Constructors

TxRelationF

Fields

ins :: [DecorateWithTxOut f TxIn]

collateralIns :: [DecorateWithTxOut f TxCollateral]

outs :: [(TxOut, [TxOutToken])]

collateralOuts :: Maybe (TxCollateralOut, [TxCollateralOutToken])

withdrawals :: [TxWithdrawal]

tokenCollateralOrd :: TxCollateralOutToken -> (TokenPolicyId, TokenName)

tokenOutOrd :: TxOutToken -> (TokenPolicyId, TokenName)

mkTxHistory :: [Tx] -> TxHistory

Convert high level transactions definition in low level DB history

data Decoration

A kind to index the 2 flavours of a TxRelationF, with or without TxOuts

Constructors

Without

With

data WithTxOut txin

A context that carries a TxOut together with its tokens (this will be needed in the future for the DB Layer to reconstruct TransactionInfo).

Constructors

WithTxOut

Fields

txIn :: txin

context :: Maybe (TxOut, [TxOutToken])

decorateWithTxOuts :: TxHistoryF 'Without -> TxHistoryF 'With

mkTxIn :: TxId -> (Int, (TxIn, Coin)) -> TxIn

mkTxCollateral :: TxId -> (Int, (TxIn, Coin)) -> TxCollateral

mkTxOut

Arguments

:: TxId

-> (Word32, TxOut)
(index, txout)

-> (TxOut, [TxOutToken])

undecorateFromTxOuts :: TxHistoryF 'With -> TxHistoryF 'Without