Cardano.Wallet.DB.Store.Wallets.Model

Description

Pure model for the transactions (Tx) and metadata about them (TxMeta) in a collection of wallets.

Documentation

data DeltaTxWalletsHistory

Verbs to change transactions store and wallet-indexed meta stores.

Constructors

ExpandTxWalletsHistory WalletId [(Tx, TxMeta)]
Add transactions and meta for a wallet.

ChangeTxMetaWalletsHistory WalletId ManipulateTxMetaHistory
Change metas for a wallet.

GarbageCollectTxWalletsHistory
Delete all transactions that have no metas.

RemoveWallet WalletId
Remove all metas of a wallet.

type TxWalletsHistory = (TxHistory, Map WalletId TxMetaHistory)

Transactions history is a shared transactions store together with a set of meta-transactions stores indexed by wallet.

walletsLinkedTransactions :: Map WalletId TxMetaHistory -> Set TxId

mkTransactionInfo :: Monad m => TimeInterpreter m -> BlockHeader -> TxRelationF 'With -> TxMeta -> m TransactionInfo

Compute a high level view of a transaction known as TransactionInfo from a TxMeta and a TxRelationF. Assumes that these data refer to the same TxId, does not check this.