Cardano.Wallet.DB.WalletState

Description

Pure data type which represents the entire wallet state, including all checkpoints.

FIXME during ADP-1043: Actually include everything, e.g. TxHistory, Pending transactions, …

Wallet state

data WalletState s

Wallet state. Currently includes:

Prologue of the address discovery state

Checkpoints of UTxO and of discoveries of the address discovery state.

FIXME during ADP-1043: Include also TxHistory, pending transactions, …, everything.

Constructors

WalletState

Fields

prologue :: !(Prologue s)

checkpoints :: !(Checkpoints (WalletCheckpoint s))

fromGenesis :: AddressBookIso s => Wallet s -> Maybe (WalletState s)

Create a wallet from the genesis block.

getLatest :: AddressBookIso s => WalletState s -> Wallet s

Get the wallet checkpoint with the largest slot number

findNearestPoint :: WalletState s -> Slot -> Maybe Slot

Find the nearest Checkpoint that is either at the given point or before.

WalletCheckpoint (internal use mostly)

data WalletCheckpoint s

Data stored in a single checkpoint. Only includes the UTxO and the Discoveries, but not the Prologue.

Constructors

WalletCheckpoint

Fields

currentTip :: !BlockHeader

utxo :: !UTxO

discoveries :: !(Discoveries s)

toWallet :: AddressBookIso s => Prologue s -> WalletCheckpoint s -> Wallet s

Convert a stored WalletCheckpoint to the legacy Wallet state.

fromWallet :: AddressBookIso s => Wallet s -> (Prologue s, WalletCheckpoint s)

Convert a legacy Wallet state to a Prologue and a WalletCheckpoint

getBlockHeight :: WalletCheckpoint s -> Word32

Helper function: Get the block height of a wallet checkpoint.

getSlot :: WalletCheckpoint s -> Slot

Helper function: Get the Slot of a wallet checkpoint.

Delta types

data DeltaWalletState1 s

Constructors

ReplacePrologue (Prologue s)
Replace the prologue of the address discovery state

UpdateCheckpoints (DeltasCheckpoints (WalletCheckpoint s))
Update the wallet checkpoints.

type DeltaWalletState s = [DeltaWalletState1 s]

Multiple wallets

data DeltaMap key da

Delta type for Map.

Constructors

Insert key (Base da)

Delete key

Adjust key da

newtype ErrNoSuchWallet

Can't perform given operation because there's no wallet

Constructors

ErrNoSuchWallet WalletId

adjustNoSuchWallet :: WalletId -> (ErrNoSuchWallet -> e) -> (w -> Either e (dw, b)) -> Map WalletId w -> (Maybe (DeltaMap WalletId dw), Either e b)

Adjust a specific wallet if it exists or return ErrNoSuchWallet.