Cardano.Wallet.Gen

Description

Shared QuickCheck generators for wallet types.

Our convention is to let each test module define its own Arbitrary orphans. This module allows for code-reuse where desired, by providing generators.

Documentation

genMnemonic :: forall mw ent csz. (ConsistentEntropy ent mw csz, EntropySize mw ~ ent) => Gen (Mnemonic mw)

Generates an arbitrary mnemonic of a size according to the type parameter.

E.g: >>> arbitrary = SomeMnemonic $ genMnemonic @12

genPercentage :: Gen Percentage

shrinkPercentage :: Percentage -> [Percentage]

genLegacyAddress :: Maybe ProtocolMagic -> Gen Address

genBlockHeader :: SlotNo -> Gen BlockHeader

genChainPoint :: Gen ChainPoint

genSlot :: Gen Slot

genActiveSlotCoefficient :: Gen ActiveSlotCoefficient

shrinkActiveSlotCoefficient :: ActiveSlotCoefficient -> [ActiveSlotCoefficient]

genSlotNo :: Gen SlotNo

Don't generate too large slots

shrinkSlotNo :: SlotNo -> [SlotNo]

genNestedTxMetadata :: Gen TxMetadata

Generates a TxMetadata with arbitrary levels of nesting.

genSimpleTxMetadata :: Gen TxMetadata

Generates a TxMetadata containing only simple values, without nesting.

shrinkTxMetadata :: TxMetadata -> [TxMetadata]

genScript :: [a] -> Gen (Script a)

genScriptCosigners :: Gen (Script Cosigner)

genScriptTemplate :: Gen ScriptTemplate

genScriptTemplateEntry :: Gen ApiScriptTemplateEntry

genMockXPub :: Gen XPub

genNatural :: Gen Natural