Cardano.Wallet.Network

Interface

data NetworkLayer m block

Interface for network capabilities.

Constructors

NetworkLayer

Fields

chainSync :: Tracer IO ChainFollowLog -> ChainFollower m ChainPoint BlockHeader (NonEmpty block) -> m ()

Connect to the node and run the ChainSync protocol. The callbacks provided in the ChainFollower argument are used to handle intersection finding, the arrival of new blocks, and rollbacks.

lightSync :: Maybe (ChainFollower m ChainPoint BlockHeader (LightBlocks m Block) -> m ())

Connect to a data source that offers an efficient query Address -> Transactions.

currentNodeTip :: m BlockHeader

Get the current tip from the chain producer

currentNodeEra :: m AnyCardanoEra

Get the era the node is currently in.

currentProtocolParameters :: m ProtocolParameters

Get the last known protocol parameters. In principle, these can only change once per epoch.

currentSlottingParameters :: m SlottingParameters

Get the last known slotting parameters. In principle, these can only change once per era.

watchNodeTip :: (BlockHeader -> m ()) -> m ()

Register a callback for when the node tip changes. This function should never finish, unless the callback throws an exception, which will be rethrown by this function.

postTx :: SealedTx -> ExceptT ErrPostTx m ()

Broadcast a transaction to the chain producer

stakeDistribution :: Coin -> m StakePoolsSummary

getCachedRewardAccountBalance :: RewardAccount -> m Coin

Return the cached reward balance of an account.

If there is no cached value, it will return `Coin 0`, and add the account to the internal set of observed account, such that it will be fetched later.

fetchRewardAccountBalances :: Set RewardAccount -> m (Map RewardAccount Coin)

Fetch the reward account balance of a set of accounts without any caching.

timeInterpreter :: TimeInterpreter (ExceptT PastHorizonException m)

syncProgress :: SlotNo -> m SyncProgress

Compute the ratio between the provided SlotNo and the slot corresponding to the current wall-clock time.

Unlike using syncProgress after retrieving a timeInterpreter, this function will return NotResponding rather than block in the edge case when the era history has not yet been fetched from the node on startup.

Errors

newtype ErrPostTx

Error while trying to send a transaction

Constructors

ErrPostTxValidationError Text

Chain following

data ChainFollower m point tip blocks

A collection of callbacks to use with the chainSync function.

Constructors

ChainFollower

Fields

checkpointPolicy :: Integer -> CheckpointPolicy

The policy for creating and pruning checkpoints that is used by the ChainFollower. The argument of this field is the epochStability.

Exposing this policy here enables any chain synchronizer which does not retrieve full blocks, such as lightSync, to specifically target those block heights at which the ChainFollower intends to create checkpoints.

readChainPoints :: m [point]

Callback for reading the local tip. Used to negotiate the intersection with the node.

A response of [] is interpreted as Origin -- i.e. the chain will be served from genesis.

rollForward :: blocks -> tip -> m ()

Callback for rolling forward.

Implementors _may_ delete old checkpoints while rolling forward.

rollBackward :: point -> m point

Roll back to the requested slot, or further, and return the point actually rolled back to.

Example 1:

If the follower stores checkpoints for all blocks, we can always roll back to the requested point exactly.

-- If
knownSlots follower shouldReturn [0,1,2,3]
let requested = SlotNo 2
-- Then
actual <- rollBackward follower requested
knownSlots follower shouldReturn` [0,1,2]
actual shouldBe SlotNo 2

Note that the slotNos are unlikely to be consecutive in real life, but this doesn't matter, as ouroboros-network asks us to rollback to points, corresponding to blocks.

Example 2:

-- If
knownSlots follower shouldReturn [0,9,10]
let requested = SlotNo 2
-- Then
actual <- rollBackward follower requested
knownSlots follower shouldReturn` [0]
actual shouldBe SlotNo 0

mapChainFollower

Arguments

:: Functor m

=> (point1 -> point2)
Covariant

-> (point2 -> point1)
Contravariant

-> (tip2 -> tip1)
Contravariant

-> (blocks2 -> blocks1)
Contravariant

-> ChainFollower m point1 tip1 blocks1

-> ChainFollower m point2 tip2 blocks2

data ChainFollowLog

Higher level log of a chain follower. Includes computed statistics about synchronization progress.

Constructors

MsgChainSync (ChainSyncLog BlockHeader ChainPoint)

MsgFollowStats (FollowStats Rearview)

MsgStartFollowing

data ChainSyncLog block point

Low-level logs of the ChainSync mini-protocol

Constructors

MsgChainFindIntersect [point]

MsgChainRollForward (NonEmpty block) point

MsgChainRollBackward point Int

MsgChainTip point

MsgLocalTip point

MsgTipDistance Natural

mapChainSyncLog :: (b1 -> b2) -> (p1 -> p2) -> ChainSyncLog b1 p1 -> ChainSyncLog b2 p2

withFollowStatsMonitoring :: Tracer IO ChainFollowLog -> (SlotNo -> IO SyncProgress) -> (Tracer IO (ChainSyncLog BlockHeader ChainPoint) -> IO ()) -> IO ()

Monitors health and statistics by inspecting the messages submitted to a ChainSyncLog tracer.

Statistics are computed in regular time intervals. In order to do that, the monitor runs in separate thread. The results are submitted to the outer ChainFollowLog tracer.

Logging (for testing)

data FollowStats f

Statistics of interest from the follow-function.

The f allows us to use Rearview to keep track of both current and previously logged stats, and perform operations over it in a nice way.

Constructors

FollowStats

Fields

blocksApplied :: !(f Int)

rollbacks :: !(f Int)

localTip :: !(f ChainPoint)

time :: !(f UTCTime)

NOTE: Current time is not updated until flush is called.

prog :: !(f SyncProgress)

NOTE: prog is not updated until flush is called.

data Rearview a

A Rearview consists of a past value and a present value. Useful for keeping track of past logs.

The idea is to 1. Reconstruct a model of the current state using a Trace 2. Sometimes log the difference between the current state and the most recently logged one.

Constructors

Rearview

Fields

past :: !a

Most previously logged state

current :: !a

Not-yet logged state

emptyStats :: UTCTime -> FollowStats Rearview

updateStats :: ChainSyncLog block ChainPoint -> FollowStats Rearview -> FollowStats Rearview

Update the current statistics based on a new log message.