Cardano.Wallet.Network.BlockHeaders

Description

This module provides a data structure and functions for operating on a sequence of block headers.

The BlockHeaders sequence is used for maintaining the global unstable blocks state for the network layer. The unstable blocks state can be compared with the block headers state of chain consumers to determine the intersection point.

Types

newtype BlockHeaders

A list of block headers and their hashes. The last block in this sequence is the network tip. The first block in this sequence is the block of depth k, which is the last unstable block.

Constructors

BlockHeaders

Fields

getBlockHeaders :: Seq BlockHeader

Double-ended queue of block headers, and their IDs.

emptyBlockHeaders :: BlockHeaders

Constuct an empty unstable blocks sequence.

Managing the global unstable blocks state

updateUnstableBlocks

Arguments

:: Monad m

=> Quantity "block" Word32
Maximum number of unstable blocks (k).

-> m (Hash "BlockHeader")
Network Tip

-> (Hash "BlockHeader" -> m BlockHeader)
Fetches block header from its hash

-> BlockHeaders
Current unstable blocks state.

-> m BlockHeaders

Updates the unstable blocks state using the given "fetch" functions.

This attempts to synchronise the local state with that of the node. The node may be on a different chain to the current unstable blocks, so this function handles switching of chains.

For example, this is what it would do when the local tip is a13, but the node's tip is b15, on a different chain.

                                   local tip ↴
                 ┌───┬───  ───┬───┬───┬───┬───┐
 Unstable blocks │a03│..    ..│a10│a11│a12│a13│
                 └───┴───  ───┴───┴───┴───┴───┘
                           ───┬───┬───┬───┬───┬───┬───┐
 Node backend chain        ...│a10│a11│b12│b13│b14│b15│
                           ───┴───┴───┴───┴───┴───┴───┘
                       rollback point ⬏     node tip ⬏

To start with, the node says the tip hash is b15.

Work backwards from tip, fetching blocks and adding them to ac, and removing overlapping blocks from ubs. Overlapping blocks occur when there has been a rollback.

    ubs                     ac
1.  ───┬───┬───┬───┬───┐    ┌───┐
    ...│a10│a11│a12│a13│    │b15│
    ───┴───┴───┴───┴───┘    └───┘
2.  ───┬───┬───┬───┬───┐┌───┬───┐
    ...│a10│a11│a12│a13││b14│b15│
    ───┴───┴───┴───┴───┘└───┴───┘
3.  ───┬───┬───┬───┐┌───┬───┬───┐
    ...│a10│a11│a12││b13│b14│b15│
    ───┴───┴───┴───┘└───┴───┴───┘
4.  ───┬───┬───┐┌───┬───┬───┬───┐
    ...│a10│a11││b12│b13│b14│b15│
    ───┴───┴───┘└───┴───┴───┴───┘

Stop once ubs and ac meet with a block which has the same hash. If they never meet, stop cleanupConfig fetching k blocks.

Finally, to get the new BlockHeaders, append ac to ubs, and limit the length to k.

The new block height is the height of the first block that was fetched.

If any errors occur while this process is running (for example, fetching a block which has been rolled back and lost from the node's state), it will immediately terminate.

Operations

blockHeadersTip :: BlockHeaders -> Maybe BlockHeader

The tip block header of the unstable blocks, if it exists.

blockHeadersBase :: BlockHeaders -> Maybe BlockHeader

The base block header is the oldest block header in the unstable blocks,

blockHeadersAtGenesis :: BlockHeaders -> Bool

Whether we are at genesis or not.

appendBlockHeaders

Arguments

:: Quantity "block" Word32
Maximum length of sequence.

-> BlockHeaders
Current unstable block headers, with rolled back blocks removed.

-> [BlockHeader]
Newly fetched block headers to add.

-> BlockHeaders

Add recently fetched block headers to the unstable blocks. This will drop the oldest block headers to ensure that there are at most k items in the sequence.

dropStartingFromSlotId :: SlotId -> BlockHeaders -> BlockHeaders

Remove unstable blocks which have a slot greater than or equal to the given slot.

dropAfterSlotId :: SlotId -> BlockHeaders -> BlockHeaders

Drop any headers that are (strictly) after the given slot id.

greatestCommonBlockHeader

Arguments

:: BlockHeaders
Node's unstable blocks

-> BlockHeaders
Local wallet unstable blocks

-> Maybe BlockHeader

If the two sequences overlap in terms of slots, return the block header of the last block that is common between the two. Otherwise return Nothing.

greatestCommonBlockHeader works fine if one (or both) list are sparse and different length.

For example: | (1) | (2) | (3) | (4) | (5) Node | abcdefg | efg | abcdefg | abcde | abc Local | cdpqrst | abc | abcdefg | | efg GCBH | Just d | Nothing | Just g | Nothing | Nothing