Cardano.Wallet.Network.Light

Interface

data LightSyncSource m block addr txs

Blockchain data source suitable for the implementation of lightSync.

Constructors

LightSyncSource

Fields

getHeader :: block -> BlockHeader

Get the BlockHeader of a given block.

getTip :: m BlockHeader

Latest tip of the chain.

getBlockHeaderAtHeight :: BlockHeight -> m (Consensual BlockHeader)

Get the BlockHeader at a given block height.

getNextBlockHeader :: BlockHeader -> m (Consensual (Maybe BlockHeader))

Get the next block header.

getBlockHeaderAt :: ChainPoint -> m (Consensual BlockHeader)

Get the full BlockHeader belonging to a given ChainPoint. Return Nothing if the point is not consensus anymore.

getNextBlocks :: ChainPoint -> m (Consensual [block])

Get several blocks immediately following the given Chainpoint.

getAddressTxs :: BlockHeader -> BlockHeader -> addr -> m txs

Transactions for a given address and point range.

type LightBlocks m block addr txs = Either (NonEmpty block) (BlockSummary m addr txs)

hoistLightSyncSource :: (forall a. m a -> n a) -> LightSyncSource m block addr txs -> LightSyncSource n block addr txs

lightSync :: MonadDelay m => Tracer m LightLayerLog -> LightSyncSource m block addr txs -> ChainFollower m ChainPoint BlockHeader (LightBlocks m block addr txs) -> m Void

Drive a ChainFollower using a LightSyncSource. Never returns.

data Consensual a

Consensual represents the result of query on the blockchain. Either the result is a value that is part of the consensus chain, or the result is an indication that the consensus had changed before the entire value could be retrieved.

Constructors

NotConsensual

Consensual a

data LightLayerLog

Constructors

MsgLightRollForward ChainPoint BlockHeader BlockHeader BlockHeader

MsgLightRolledForward BlockHeader

MsgLightRollBackward ChainPoint ChainPoint