Cardano.Wallet.Primitive.AddressDerivation.Common

Description

Use the types from Cardano.Wallet.Primitive.AddressDerivation or the key generation or derivation functions from Cardano.Wallet.Primitive.AddressDerivation.Sequential or Cardano.Wallet.Primitive.AddressDerivation.Random instead of this module.

Documentation

newtype Key (level :: Depth) key

A cryptographic key, with phantom-types to disambiguate key types.

let rootPrivateKey = Key 'RootK XPrv
let accountPubKey = Key 'AccountK XPub
let addressPubKey = Key 'AddressK XPub

Constructors

Key

Fields

getKey :: key

data Depth

Key Depth in the derivation path, according to BIP-0039 / BIP-0044

m | purpose' | cointype' | account' | change | address

We do not manipulate purpose, cointype and change paths directly, so they are left out of the sum type.

Constructors

RootK

AccountK

AddressK