Cardano.Wallet.Primitive.AddressDerivation.Random

Description

Implementation of address derivation for the random scheme, as implemented by the legacy Cardano wallets.

For full documentation of the key derivation schemes, see the Cardano.Crypto.Wallet module, and the implementation in cardano-crypto.

Types

data RndKey (depth :: Depth) key

Material for deriving HD random scheme keys, which can be used for making addresses.

Constructors

RndKey

Fields

getKey :: key

The raw private or public key.

derivationPath :: DerivationPath depth

The address derivation indices for the level of this key.

payloadPassphrase :: Passphrase "addr-derivation-payload"

Used for encryption of payload containing address derivation path.

Generation

unsafeGenerateKeyFromSeed :: DerivationPath depth -> Passphrase "seed" -> Passphrase "encryption" -> RndKey depth XPrv

Generate a new key from seed. Note that the depth is left open so that the caller gets to decide what type of key this is. This is mostly for testing, in practice, seeds are used to represent root keys, and one should use generateKeyFromSeed.

generateKeyFromSeed :: Passphrase "seed" -> Passphrase "encryption" -> RndKey RootK XPrv

Generate a root key from a corresponding seed. The seed should be at least 16 bytes.

minSeedLengthBytes :: Int

The amount of entropy carried by a BIP-39 12-word mnemonic is 16 bytes.

nullKey :: RndKey RootK XPrv

A root key of all zeroes that is used when restoring RndState from the database before a root key has been saved.

Derivation

deriveAccountPrivateKey :: Passphrase "encryption" -> RndKey RootK XPrv -> Index Hardened AccountK -> RndKey AccountK XPrv

Derives account private key from the given root private key, using derivation scheme 1.

NOTE: The caller is expected to provide the corresponding passphrase (and to have checked that the passphrase is valid). Providing a wrong passphrase will not make the function fail but will instead, yield an incorrect new key that doesn't belong to the wallet.

deriveAddressPrivateKey :: Passphrase "encryption" -> RndKey AccountK XPrv -> Index Hardened AddressK -> RndKey AddressK XPrv

Derives address private key from the given account private key, using derivation scheme 1.

NOTE: The caller is expected to provide the corresponding passphrase (and to have checked that the passphrase is valid). Providing a wrong passphrase will not make the function fail but will instead, yield an incorrect new key that doesn't belong to the wallet.