Cardano.Wallet.Primitive.AddressDerivation.Sequential

Description

Implementation of address derivation for the sequential schemes, as implemented by Yoroi/Icarus and cardano-cli.

SeqKey types

newtype SeqKey (depth :: Depth) key

A cryptographic key for sequential-scheme address derivation, with phantom-types to disambiguate key types.

let rootPrivateKey = SeqKey 'RootK XPrv
let accountPubKey = SeqKey 'AccountK XPub
let addressPubKey = SeqKey 'AddressK XPub

Constructors

SeqKey

Fields

getKey :: key

data ChangeChain

Marker for the change chain. In practice, change of a transaction goes onto the addresses generated on the internal chain, whereas the external chain is used for addresses that are part of the advertised targets of a transaction

Constructors

ExternalChain

InternalChain

SeqKey generation and derivation

generateKeyFromSeed

Arguments

:: (Passphrase "seed", Passphrase "generation")
The actual seed and its recovery / generation passphrase

-> Passphrase "encryption"

-> SeqKey RootK XPrv

Generate a root key from a corresponding seed. The seed should be at least 16 bytes.

unsafeGenerateKeyFromSeed

Arguments

:: (Passphrase "seed", Passphrase "generation")
The actual seed and its recovery / generation passphrase

-> Passphrase "encryption"

-> SeqKey depth XPrv

Generate a new key from seed. Note that the depth is left open so that the caller gets to decide what type of key this is. This is mostly for testing, in practice, seeds are used to represent root keys, and one should use generateKeyFromSeed.

minSeedLengthBytes :: Int

The minimum seed length for generateKeyFromSeed and unsafeGenerateKeyFromSeed.

deriveAccountPrivateKey :: Passphrase "encryption" -> SeqKey RootK XPrv -> Index Hardened AccountK -> SeqKey AccountK XPrv

Derives account private key from the given root private key, using derivation scheme 2 (see cardano-crypto package for more details).

NOTE: The caller is expected to provide the corresponding passphrase (and to have checked that the passphrase is valid). Providing a wrong passphrase will not make the function fail but will instead, yield an incorrect new key that doesn't belong to the wallet.

deriveAddressPrivateKey :: Passphrase "encryption" -> SeqKey AccountK XPrv -> ChangeChain -> Index Soft AddressK -> SeqKey AddressK XPrv

Derives address private key from the given account private key, using derivation scheme 2 (see cardano-crypto package for more details).

It is preferred to use deriveAddressPublicKey whenever possible to avoid having to manipulate passphrases and private keys.

NOTE: The caller is expected to provide the corresponding passphrase (and to have checked that the passphrase is valid). Providing a wrong passphrase will not make the function fail but will instead, yield an incorrect new key that doesn't belong to the wallet.

deriveAddressPublicKey :: SeqKey AccountK XPub -> ChangeChain -> Index Soft AddressK -> SeqKey AddressK XPub

Derives address public key from the given account public key, using derivation scheme 2 (see cardano-crypto package for more details).

This is the preferred way of deriving new sequential address public keys.

Passphrase

changePassphraseSeq :: Passphrase "encryption-old" -> Passphrase "encryption-new" -> SeqKey purpose XPrv -> SeqKey purpose XPrv

Re-encrypt a private key using a different passphrase.

*Important**: This function doesn't check that the old passphrase is correct! Caller is expected to have already checked that. Using an incorrect passphrase here will lead to very bad thing.

Storing and retrieving keys

deserializeXPubSeq :: ByteString -> Either String (SeqKey purpose XPub)

The reverse of serializeXPub. This will fail if the input is not hexadecimal string of the correct length.

serializeXPubSeq :: SeqKey purpose XPub -> ByteString

Convert a public key into a hexadecimal string suitable for storing in a text file or database column.