Cardano.Wallet.Primitive.AddressDiscovery.Random

Description

An implementation of address discovery for the random address scheme as used by the legacy Cardano wallets.

State

data RndState (network :: NetworkDiscriminant)

HD random address discovery state and key material for AD.

Constructors

RndState

Fields

hdPassphrase :: Passphrase "addr-derivation-payload"

The HD derivation passphrase

accountIndex :: Index 'Hardened 'AccountK

The account index used for address _generation_ in this wallet. Note that addresses will be _discovered_ from any and all account indices, regardless of this value.

discoveredAddresses :: Map DerivationPath (Address, AddressState)

The addresses which have so far been discovered, and their derivation paths.

pendingAddresses :: Map DerivationPath Address

The addresses which have been generated for use as change addresses, and their derivation paths. Once addresses have been discovered they are removed from this set and added to addresses.

gen :: StdGen

The state of the RNG.

class RndStateLike s

Minimal complete definition

importAddress, addPendingAddress, unavailablePaths, defaultAccountIndex, withRNG

mkRndState :: ByronKey 'RootK XPrv -> Int -> RndState n

Initialize the HD random address discovery state from a root key and RNG seed.

type DerivationPath = (Index 'WholeDomain 'AccountK, Index 'WholeDomain 'AddressK)

Shortcut type alias for HD random address derivation path.

toDerivationIndexes :: DerivationPath -> NonEmpty DerivationIndex

Low-level API

importAddress :: RndStateLike s => Address -> s -> Either ErrImportAddress s

newtype ErrImportAddress

Constructors

ErrAddrDoesNotBelong Address

addPendingAddress :: RndStateLike s => Address -> DerivationPath -> s -> s

deriveRndStateAddress :: forall n. PaymentAddress n ByronKey => ByronKey 'RootK XPrv -> Passphrase "encryption" -> DerivationPath -> Address

Use the key material in RndState to derive a change address.

findUnusedPath :: StdGen -> Index 'Hardened 'AccountK -> Set DerivationPath -> (DerivationPath, StdGen)

Randomly generates an address derivation path for a given account. If the path is already in the "blacklist", it will try generating another.

Note that performance of this function will degrade poorly as more of the account's address space is used up. We may have to improve it later.

unavailablePaths :: RndStateLike s => s -> Set DerivationPath

Returns the set of derivation paths that should not be used for new address generation because they are already in use.

defaultAccountIndex :: RndStateLike s => s -> Index 'Hardened 'AccountK

Default account used for generating new derivation paths.

withRNG :: RndStateLike s => s -> (StdGen -> (a, StdGen)) -> (a, s)

Default random number generator.

Benchmarking

newtype RndAnyState (network :: NetworkDiscriminant) (p :: Nat)

An "unsound" alternative that can be used for benchmarking and stress testing. It re-uses the same underlying structure as the RndState but it discover addresses based on an arbitrary ratio instead of decrypting the derivation path.

The proportion is stored as a type-level parameter so that we don't have to alter the database schema to store it. It simply exists and depends on the caller creating the wallet to define it.

Constructors

RndAnyState

Fields

innerState :: RndState network

mkRndAnyState :: forall (p :: Nat) n. ByronKey 'RootK XPrv -> Int -> RndAnyState n p

Initialize the HD random address discovery state from a root key and RNG seed.

The type parameter is expected to be a ratio of addresses we ought to simply recognize as ours. It is expressed in per-myriad, so "1" means 0.01%, "100" means 1% and 10000 means 100%.