Cardano.Wallet.Primitive.BlockSummary

Description

This module provides the BlockSummary type which summarizes a contiguous sequence of blocks.

Documentation

data BlockSummary m addr txs

A BlockSummary summarizes the data contained in a contiguous sequence of blocks.

However, instead of storing the sequence of blocks of directly as a Haskell list, the BlockSummary only provides a $sel:query:BlockSummary function which looks up all transactions associated to a given addresses. In addition, this query function is monadic, which means that it can call out to an external data source.

Constructors

BlockSummary

Fields

from :: !BlockHeader

to :: !BlockHeader

query :: addr -> m txs

type LightSummary m = BlockSummary m (Either Address RewardAccount) ChainEvents

BlockSummary used for light-mode.

Chain Events

data ChainEvents

BlockEvents, always ordered by slot.

fromBlockEvents :: [BlockEvents] -> ChainEvents

Create ChainEvents from a list of block events (which do not need to be in order.)

toAscBlockEvents :: ChainEvents -> [BlockEvents]

List of BlockEvents, in ascending order. (No duplicate blocks, all transactions within block in order).

Block Events

data BlockEvents

Events (such as txs, delegations) within a single block that are potentially relevant to the wallet. This can be the entire block, or a pre-filtered version of it.

Constructors

BlockEvents

Fields

slot :: !Slot

blockHeight :: !(Quantity "block" Word32)

transactions :: Sublist Tx

(Index of the transaction within the block, transaction data) INVARIANT: The list is ordered by ascending index.

delegations :: Sublist DelegationCertificate

(Index of the delegation within the block, delegation data) INVARIANT: The list is ordered by ascending index.

fromEntireBlock :: Block -> BlockEvents

Get the BlockEvents corresponding to an entire Block.

Sublist

data Sublist a

A data type representing a sublist of a total list. Such a sublist typically arises by filtering and keeps track of the indices of the filtered list elements.

In order to represent sublists of DelegationCertificate, we do not use a single Int, but a pair (Index1,Index2) as index internally. This internal index is not part of the (safe) API of Sublist.

The main purpose of this data type is optimization: When processing whole Block, we want to avoid copying and redecorating the entire list of transactions in that Block; instead, we want to copy a pointer to this list.

filterSublist :: (a -> Bool) -> Sublist a -> Sublist a

Filter a Sublist by a predicate.

wholeList :: [a] -> Sublist a

Construct a Sublist representing the whole list.

Internal & Testing

summarizeOnTxOut :: NonEmpty Block -> LightSummary Identity

For testing: Convert a list of blocks into a BlockSummary. Unfortunately, as TxIn references are not resolved, we can only find transactions with relevant TxOut.

mkChainEvents :: Map Slot BlockEvents -> ChainEvents

mergeSublist :: Sublist a -> Sublist a -> Sublist a

Merge two Sublist assuming that they are sublists of the same list.

unsafeMkSublist :: [((Index1, Index2), a)] -> Sublist a

Construct a Sublist from a list of indexed items.