Cardano.Wallet.Primitive.CoinSelection

Description

Provides the API of Coin Selection algorithm and Fee Calculation This module contains the implementation of adjusting coin selection for a fee. The sender pays for the fee and additional inputs are picked randomly. For more information refer to: https://iohk.io/blog/self-organisation-in-coin-selection/

Coin Selection

data CoinSelection

Constructors

CoinSelection

Fields

inputs :: [(TxIn, TxOut)]

Picked inputs

outputs :: [TxOut]

Picked outputs

change :: [Coin]

Resulting changes

data ErrCoinSelection e

Constructors

ErrNotEnoughMoney Word64 Word64
UTxO exhausted during input selection We record the balance of the UTxO as well as the size of the payment we tried to make.

ErrUtxoNotEnoughFragmented Word64 Word64
UTxO is not enough fragmented for the number of transaction outputs We record the number of UTxO entries as well as the number of the outputs of the transaction.

ErrMaximumInputsReached Word64
When trying to construct a transaction, the max number of allowed inputs was reached.

ErrInputsDepleted
When trying to construct a transaction, the available inputs are depleted even when UTxO is properly fragmented and with enough funds to cover payment

ErrInvalidSelection e
Somewhat, we ended up generating an invalid coin selection because of inputs passed down to the coin selection function, or because a target backend has extra-limitations not covered by our coin selection algorithm.

data CoinSelectionOptions e

Constructors

CoinSelectionOptions

Fields

maximumNumberOfInputs :: Word8 -> Word8

Maximum number of inputs allowed for a given number of outputs

validate :: CoinSelection -> Either e ()

Returns any backend-specific error regarding coin selection

Helpers

shuffle :: [a] -> IO [a]

Shuffles a list of elements.

>>> shuffle (outputs coinSel)
[...]