Cardano.Wallet.Primitive.CoinSelection.Migration

Description

This module contains an algorithm to select coins for migration from legacy wallets to newer wallets.

We want users to be able to migrate their funds from a legacy random wallet to a new sequential wallet. To do this, we have to move funds from a wallet to another by making transactions. Funds are ultimately a sum of many coins (a.k.a UTxOs). In a transaction, we can select a few coins, and send them to addresses, effectively creating new coins / UTxOs doing this.

There are some limitations regarding the number of coins that can be selected at once in a single transaction (theoretically 255 coins, in practice ~170) because there's a transaction max size (in bytes) enforced by the network. Also, there's a direct relationship between the maximum number of inputs we can select, and the maximum number of outputs we can produce (increasing one will decrease the other, and vice-versa).

When making a transaction, coins used as inputs for a transaction becomes unavailable for a while, until the transaction is inserted into the ledger and, make some new coins available as change (very much like when paying with bank notes to a shop, if we give a 20 EUR note to pay for 3 EUR, we can't spend the remaining 17 EUR before we have received the change!). So, a wallet with a small number of UTxO will not be able to make many transactions in parallel and will have to make them sequentially, waiting for the previous ones to be inserted before making new ones (we also say that a wallet is not "fragmented enough").

Documentation

selectCoinsForMigration

Arguments

:: FeeOptions
Fee computation and threshold definition

-> Word8
Maximum number of inputs we can select per transaction

-> UTxO
UTxO to deplete

-> [CoinSelection]

Construct a list of coin selections / transactions to transfer the totality of a user's wallet. The resulting CoinSelection do not contain any outputs, but only change coins (so there's no restriction about how addresses are generated).

It tries to fit as many inputs as possible in a single transaction (fixed by the Word8 maximum number of inputs given as argument.

The fee options are used to balance the coin selections and fix a threshold for dust that is removed from the selections.

idealBatchSize :: CoinSelectionOptions e -> Word8

Try to find a fix "ideal" number of input transactions that would generate rather balanced transactions.