Cardano.Wallet.Primitive.CoinSelection.Random

Description

This module contains the implementation of random input selection algorithm

Documentation

random :: forall m e. MonadRandom m => CoinSelectionOptions e -> NonEmpty TxOut -> UTxO -> ExceptT (ErrCoinSelection e) m (CoinSelection, UTxO)

Random-Improve Algorithm

Randomly select outputs from the UTxO until the payment value is covered. (In the rare case that this fails because the maximum number of transaction inputs has been exceeded, fall back on the largest-first algorithm for this step.)

The algorithm first makes a random selection for each output from the UTxO, processing the biggest output first and proceeding in a descending order. If the selection is not successful largest-first fallback kicks in. If the selection is successful for each output then the improvement is tried for each selection, once again starting from the selection made for the biggest output. The improvement is tried for the next biggest output's selection. An output is considered an improvement when:

(a) It doesn’t exceed a specified upper limit. (b) Adding the new output gets us closer to the ideal change value. (c) It doesn’t exceed a maximum number of transaction inputs.

This algorithm follows three principles:

**Self organisation principle 1**
Random selection has a high probability of picking dust outputs precisely
when there is a lot of dust in the UTxO.

**Self organisation principle 2**
If for each payment request for value x we create a change output roughly
of the same value x, then we will end up with a lot of change outputs in
our UTxO of size x precisely when we have a lot of payment requests of
size x

**Self organisation principle 3**
Searching the UTxO for additional entries to improve our change output is
only useful if the UTxO contains entries that are sufficiently small enough.
But precisely when the UTxO contains many small entries, it is less likely
that a randomly chosen UTxO entry will push the total above the upper bound
we set.