Cardano.Wallet.Primitive.Fee

Description

Provides the API of Coin Selection algorithm and Fee Calculation This module contains the implementation of adjusting coin selection for a fee. The sender pays for the fee and additional inputs are picked randomly. For more information refer to: https://iohk.io/blog/self-organisation-in-coin-selection/

Types

newtype Fee

A Fee, isomorph to Coin but ease type-signatures and readability.

Constructors

Fee

Fields

getFee :: Word64

data FeePolicy

A linear equation a free variable x. Represents the s -> a + b*s function where s can be the transaction size in bytes or, a number of inputs + outputs.

a and b are constant coefficients.

Constructors

LinearFee (Quantity "lovelace" Double) (Quantity "lovelace/x" Double)

Fee Calculation

computeFee :: FeePolicy -> Quantity "byte" Int -> Fee

Compute fee for a given payload. Fee follows a simple linear equation:

    f = a + size * b

where a & b are values fixed by the protocol.

divvyFee :: Fee -> [Coin] -> [(Fee, Coin)]

Proportionally divide the fee over each output.

Pre-condition 1: The given outputs list shouldn't be empty Pre-condition 2: None of the outputs should be null

It returns the a list of pairs (fee, output).

Fee Adjustment

data FeeOptions

Constructors

FeeOptions

Fields

estimate :: CoinSelection -> Fee

Estimate fees based on number of inputs and values of the outputs Some pointers / order of magnitude from the current configuration: a: 155381 # absolute minimal fees per transaction b: 43.946 # additional minimal fees per byte of transaction size

dustThreshold :: Coin

Change addresses below the given threshold will be evicted from the created transaction. Setting dustThreshold to 0 removes output equal to 0

newtype ErrAdjustForFee

Constructors

ErrCannotCoverFee Word64
UTxO exhausted during fee covering We record what amount missed to cover the fee

adjustForFee :: MonadRandom m => FeeOptions -> UTxO -> CoinSelection -> ExceptT ErrAdjustForFee m CoinSelection

Given the coin selection result from a policy run, adjust the outputs for fees, potentially returning additional inputs that we need to cover all fees. We lose the relationship between the transaction outputs and their corresponding inputs/change outputs here. This is a decision we may wish to revisit later. For now however note that since

(a) coin selection tries to establish a particular ratio between payment outputs and change outputs (currently it aims for an average of 1:1)

(b) coin selection currently only generates a single change output per payment output, distributing the fee proportionally across all change outputs is roughly equivalent to distributing it proportionally over the payment outputs (roughly, not exactly, because the 1:1 proportion is best effort only, and may in some cases be wildly different).

Note that for (a) we don't need the ratio to be 1:1, the above reasoning will remain true for any proportion 1:n. For (b) however, if coin selection starts creating multiple outputs, and this number may vary, then losing the connection between outputs and change outputs will mean that that some outputs may pay a larger percentage of the fee (depending on how many change outputs the algorithm happened to choose).