Cardano.Wallet.Primitive.Migration

Description

This module provides a public API for planning wallet migrations.

Use createPlan to create a migration plan.

Creating a migration plan

createPlan :: TxConstraints -> UTxO -> RewardWithdrawal -> MigrationPlan

Creates a migration plan for the given UTxO set and reward withdrawal amount.

See MigrationPlan.

data MigrationPlan

Represents a plan for migrating a UTxO set.

See createPlan to create a migration plan.

Constructors

MigrationPlan

Fields

selections :: ![Selection (TxIn, TxOut)]

A list of generated selections: each selection is the basis for a single transaction.

unselected :: !UTxO

The portion of the UTxO that was not selected.

totalFee :: !Coin

The total fee payable: equal to the sum of the fees of the individual selections.

newtype RewardWithdrawal

Constructors

RewardWithdrawal

Fields

unRewardWithdrawal :: Coin

data Selection input

A selection is the basis for a single transaction.

Use create to create a selection with one or more inputs. Use extend to extend a selection with an additional input. Use verify to verify the correctness of a selection.

Constructors

Selection

Fields

inputIds :: !(NonEmpty input)

The selected inputs.

inputBalance :: !TokenBundle

The total balance of value provided by the inputs.

outputs :: !(NonEmpty TokenBundle)

The outputs, adjusted to pay for the fee.

fee :: !Coin

The actual fee payable for this selection.

feeExcess :: !Coin

The excess over the minimum permissible fee for this selection.

size :: !TxSize

The size of this selection.

rewardWithdrawal :: !Coin

The reward withdrawal amount, if any.