Cardano.Wallet.Primitive.Migration.Planning

Description

This module contains an algorithm for planning migrations at a high level.

It determines how to partition the UTxO set into entries of different types, and in which order to add entries to selections, in order to maximize the number of entries that can be successfully migrated.

Use createPlan to create a migration plan.

Migration planning

createPlan :: TxConstraints -> CategorizedUTxO input -> RewardWithdrawal -> MigrationPlan input

Creates a migration plan for the given categorized UTxO set and reward withdrawal amount.

See MigrationPlan.

data MigrationPlan input

Represents a plan for migrating a set of UTxO entries.

Use createPlan to create a migration plan.

Constructors

MigrationPlan

Fields

selections :: ![Selection input]

A list of generated selections: each selection is the basis for a single transaction.

unselected :: !(CategorizedUTxO input)

The portion of the UTxO that was not selected.

totalFee :: !Coin

The total fee payable: equal to the sum of the fees of the individual selections.

UTxO entry categorization

data CategorizedUTxO input

Constructors

CategorizedUTxO

Fields

supporters :: ![(input, TokenBundle)]

freeriders :: ![(input, TokenBundle)]

ignorables :: ![(input, TokenBundle)]

data UTxOEntryCategory

Constructors

Supporter
A coin or bundle that is capable of paying for its own marginal fee and the base transaction fee.

Freerider
A coin or bundle that is not capable of paying for itself.

Ignorable
A coin that should not be added to a selection, because its value is lower than the marginal fee for an input.

categorizeUTxO :: TxConstraints -> UTxO -> CategorizedUTxO (TxIn, TxOut)

categorizeUTxOEntries :: forall input. TxConstraints -> [(input, TokenBundle)] -> CategorizedUTxO input

categorizeUTxOEntry :: TxConstraints -> TokenBundle -> UTxOEntryCategory

uncategorizeUTxO :: CategorizedUTxO (TxIn, TxOut) -> UTxO

uncategorizeUTxOEntries :: CategorizedUTxO input -> [(input, TokenBundle)]