Cardano.Wallet.Primitive.Mnemonic

Description

This module provides mnemonic (backup phrase) creation, and conversion of a mnemonic to seed for wallet restoration.

The module uses a lot of type-level machinery to ensure that entropy and mnemonic sizes are all compatible and legit. Therefore, it isn't possible to generate an invalid seed by using the smart constructors below, and trying to generate an entropy of an invalid size will result in a runtime error.

Creating Mnemonic (resp. Entropy)

Type families EntropySize, MnemonicWords and CheckSumBits can be used to disambiguate calls to smart constructors in a readable way.

TroubleShooting:

Natural XX is out of bounds for Int: This usually occurs when ones is trying to specify an invalid size for an Entropy or Mnemonic. For example:

>>> genEntropy @42
error:
  • Natural CheckSumBits 42 is out of bounds for Int

This could be the case as well when forgetting to use an adequate type application:

>>> mkEntropy mempty
error:
  • Natural ent is out of bounds for Int

data Mnemonic (mw :: Nat)

A backup-phrase in the form of a non-empty of Mnemonic words Constructor isn't exposed.

data Entropy (n :: Nat)

mkEntropy :: forall ent csz. (ValidEntropySize ent, ValidChecksumSize ent csz) => ByteString -> Either (EntropyError csz) (Entropy ent)

Smart-constructor for the Entropy

>>> mkEntropy @(EntropySize 15) bytes
Entropy {} :: Entropy 160

mkMnemonic :: forall mw ent csz. (ConsistentEntropy ent mw csz, EntropySize mw ~ ent) => [Text] -> Either (MnemonicError csz) (Mnemonic mw)

Smart-constructor for Mnemonic. Requires a type application to disambiguate the mnemonic size:

>>> mkMnemonic @15 sentence
Mnemonic {} :: Mnemonic 15

See also FromMnemonic to build a Mnemonic from lists of words of variable sizes.

Property:

mkMnemonic (mnemonicToText mnemonic) == Right mnemonic

genEntropy :: forall ent csz. (ValidEntropySize ent, ValidChecksumSize ent csz) => IO (Entropy ent)

Generate Entropy of a given size using a random seed.

Example:

>>> genEntropy @(EntropySize 12)
Entropy {} :: Entropy 128

Errors

data MnemonicError csz

This wraps errors from Cardano.Encoding.BIP39

Constructors

ErrMnemonicWords MnemonicWordsError
Wrong number of words in mnemonic.

ErrEntropy (EntropyError csz)
Invalid entropy length or checksum.

ErrDictionary DictionaryError
Invalid word in mnemonic.

newtype MnemonicException csz

This wraps EntropyError of Cardano.Encoding.BIP39

Constructors

UnexpectedEntropyError (EntropyError csz)
Invalid entropy length or checksum

Converting from and to Mnemonic (resp. Entropy)

mnemonicToEntropy :: Mnemonic mw -> Entropy (EntropySize mw)

entropyToMnemonic :: forall mw ent csz. (ValidMnemonicSentence mw, ValidEntropySize ent, ValidChecksumSize ent csz, ent ~ EntropySize mw, mw ~ MnemonicWords ent) => Entropy ent -> Mnemonic mw

Convert an Entropy to a corresponding Mnemonic Sentence. Since Entropy and Mnemonic can only be created through smart-constructors, this function cannot fail and is total.

entropyToBytes :: Entropy n -> ScrubbedBytes

Convert Entropy to a plain bytes.

mnemonicToText :: Mnemonic mw -> [Text]

Convert a Mnemonic to a sentence of English mnemonic words.

Re-export from Crypto.Encoding.BIP39

data EntropyError (csz :: Nat)

Constructors

ErrInvalidEntropyLength Int Int

ErrInvalidEntropyChecksum (Checksum csz) (Checksum csz)

data DictionaryError

Constructors

ErrInvalidDictionaryWord String

data MnemonicWordsError

Constructors

ErrWrongNumberOfWords Int Int

type ValidEntropySize (n :: Nat) = (KnownNat n, NatWithinBound Int n, Elem n (96 ': (128 ': (160 ': (192 ': (224 ': (256 ': ([] :: [Nat]))))))))

type ValidChecksumSize (ent :: Nat) (csz :: Nat) = (KnownNat csz, NatWithinBound Int csz, Elem csz (3 ': (4 ': (5 ': (6 ': (7 ': (8 ': ([] :: [Nat]))))))), CheckSumBits ent ~ csz)

type ConsistentEntropy (ent :: Nat) (mw :: Nat) (csz :: Nat) = (ValidEntropySize ent, ValidChecksumSize ent csz, ValidMnemonicSentence mw, MnemonicWords ent ~ mw)

type family CheckSumBits (n :: Nat) :: Nat where ...

Equations

CheckSumBits 96 = 3

CheckSumBits 128 = 4

CheckSumBits 160 = 5

CheckSumBits 192 = 6

CheckSumBits 224 = 7

CheckSumBits 256 = 8

type family EntropySize (n :: Nat) :: Nat where ...

Equations

EntropySize 9 = 96

EntropySize 12 = 128

EntropySize 15 = 160

EntropySize 18 = 192

EntropySize 21 = 224

EntropySize 24 = 256

type family MnemonicWords (n :: Nat) :: Nat where ...

Equations

MnemonicWords 96 = 9

MnemonicWords 128 = 12

MnemonicWords 160 = 15

MnemonicWords 192 = 18

MnemonicWords 224 = 21

MnemonicWords 256 = 24

Orphan instances

Eq MnemonicWordsError

Instance details

Methods

(==) :: MnemonicWordsError -> MnemonicWordsError -> Bool

(/=) :: MnemonicWordsError -> MnemonicWordsError -> Bool

Eq DictionaryError

Instance details

Methods

(==) :: DictionaryError -> DictionaryError -> Bool

(/=) :: DictionaryError -> DictionaryError -> Bool

Eq (EntropyError czs)

Instance details

Methods

(==) :: EntropyError czs -> EntropyError czs -> Bool

(/=) :: EntropyError czs -> EntropyError czs -> Bool