Cardano.Wallet.Primitive.Passphrase.Types

Description

Passphrases from the user

newtype Passphrase (purpose :: Symbol)

An encapsulated passphrase. The inner format is free, but the wrapper helps readability in function signatures.

Some type parameters in use are:

"user" - a passphrase entered by the user through the API. The FromText instance enforces password length rules.

"lenient" - like "user", except without a minimum length restriction in FromText.`

"encryption" - the user's passphrase, transformed so that it can be used as the key for encrypting wallet keys.

"salt" - the random salt part of a hashed passphrase.

Constructors

Passphrase

Fields

unPassphrase :: ScrubbedBytes

class PassphraseMinLength (purpose :: Symbol) where

Methods

passphraseMinLength :: Proxy purpose -> Int

Minimal Length for a passphrase, for lack of better validations

class PassphraseMaxLength (purpose :: Symbol) where

Methods

passphraseMaxLength :: Proxy purpose -> Int

Maximum length for a passphrase

validatePassphrase :: forall purpose. (PassphraseMaxLength purpose, PassphraseMinLength purpose) => Text -> Either String (Passphrase purpose)

Wallet passphrases stored as hashes

newtype PassphraseHash

Constructors

PassphraseHash

Fields

getPassphraseHash :: ScrubbedBytes

data PassphraseScheme

A type to capture which encryption scheme should be used

Constructors

EncryptWithScrypt
Legacy encryption scheme for passphrases

EncryptWithPBKDF2
Encryption scheme used since cardano-wallet

data WalletPassphraseInfo

Constructors

WalletPassphraseInfo

Fields

lastUpdatedAt :: UTCTime

passphraseScheme :: PassphraseScheme

Error types

data ErrWrongPassphrase

Indicate a failure when checking for a given Passphrase match

Constructors

ErrWrongPassphrase

ErrPassphraseSchemeUnsupported PassphraseScheme