Cardano.Wallet.Primitive.Signing

Description

This module provides functionality for signing transactions.

It relies on the binary CBOR format of transactions, an AddressScheme for deriving address private keys, and cardano-crypto for the actual signing.

Sign transactions

mkStdTx

Arguments

:: (Address -> Maybe (Key AddressK XPrv, Passphrase "encryption"))
A state from which an address private key can be looked up

-> [(TxIn, TxOut)]
Selected inputs

-> [TxOut]
Selected outputs (including change)

-> Either SignTxError (Tx, [TxWitness])

Construct a standard transaction

" Standard " here refers to the fact that we do not deal with redemption, multisignature transactions, etc.

newtype SignTxError

Possible signing error

Constructors

KeyNotFoundForAddress Address
We tried to sign a transaction with inputs that are unknown to us?