Cardano.Wallet.Primitive.SyncProgress

Description

Functionality for calculating SyncProgress of wallets.

Types

data SyncProgress

Constructors

Ready

Syncing !(Quantity "percent" Percentage)

NotResponding

newtype SyncTolerance

Constructors

SyncTolerance NominalDiffTime

mkSyncTolerance :: Int -> SyncTolerance

Construct a SyncTolerance from a number of seconds

Implementations

syncProgress

Arguments

:: (HasCallStack, Monad m)

=> SyncTolerance
A time tolerance inside which we consider ourselves synced

-> TimeInterpreter m
Converts slots to actual time.

-> SlotNo
Slot of latest block consumed

-> RelativeTime
Current wall clock time

-> m SyncProgress

Estimate restoration progress based on:

The slot of the latest block consumed (our progress)

The slot corresponding to the latest wall-clock time (our target)

The estimated progress is the quotient of these two quantities.

In the Cardano consensus protocol, only a fraction of slots contains blocks. Hence, the progress percentage will often be < 100%, as the slot corresponding to the current wall-clock time may not be filled with a block. The sync tolerance should be large enough to accommodate this issue.