Cardano.Wallet.Primitive.Types

Description

This module contains the core primitive of a Wallet. This is roughly a Haskell translation of the Formal Specification for a Cardano Wallet

It doesn't contain any particular business-logic code, but defines a few primitive operations on Wallet core types as well.

Block

data Block

A block on the chain, as the wallet sees it.

Constructors

Block

Fields

header :: !BlockHeader

transactions :: ![Tx]

delegations :: ![DelegationCertificate]

data BlockHeader

Constructors

BlockHeader

Fields

slotNo :: SlotNo

blockHeight :: Quantity "block" Word32

headerHash :: !(Hash "BlockHeader")

parentHeaderHash :: !(Maybe (Hash "BlockHeader"))

isGenesisBlockHeader :: BlockHeader -> Bool

Check whether a block with a given BlockHeader is the genesis block.

data ChainPoint

A point on the blockchain is either the genesis block, or a block with a hash that was created at a particular SlotNo.

TODO:

This type is essentially a copy of the ChainPoint type. We want to import it from there when overhauling our types.

That said, using WithOrigin would not be bad.

BlockHeader is also a good type for rerencing points on the chain, but it's less compatible with the types in ouroboros-network.

Constructors

ChainPointAtGenesis

ChainPoint !SlotNo !(Hash "BlockHeader")

compareSlot :: ChainPoint -> ChainPoint -> Ordering

Compare the slot numbers of two ChainPoints, but where the ChainPointAtGenesis comes before all other slot numbers.

chainPointFromBlockHeader :: BlockHeader -> ChainPoint

Convert a BlockHeader into a ChainPoint.

type Slot = WithOrigin SlotNo

A point in (slot) time, which is either genesis (Origin) or has a slot number (At).

In contrast to ChainPoint, the type Slot does not refer to a point on an actual chain with valid block hashes, but merely to a timeslot which can hold a single block. This implies:

Slot has a linear ordering implemented in the Ord class (where Origin < At slot).

Using Slot in QuickCheck testing requires less context (such as an actual simulated chain.)

data WithOrigin t

Constructors

Origin

At !t

toSlot :: ChainPoint -> Slot

Retrieve the slot of a ChainPoint.

Delegation and stake pools

data CertificatePublicationTime

Represents an abstract notion of a certificate publication time.

Certificates published at later times take precedence over certificates published at earlier times.

Constructors

CertificatePublicationTime

Fields

slotNo :: SlotNo

slotInternalIndex :: Word64

Indicates the relative position of a publication within a slot.

data DelegationCertificate

Represent a delegation certificate.

Constructors

CertDelegateNone RewardAccount

CertDelegateFull RewardAccount PoolId

CertRegisterKey RewardAccount

dlgCertAccount :: DelegationCertificate -> RewardAccount

dlgCertPoolId :: DelegationCertificate -> Maybe PoolId

data PoolLifeCycleStatus

Indicates the current life cycle status of a pool.

Constructors

PoolNotRegistered
Indicates that a pool is not registered.

PoolRegistered PoolRegistrationCertificate
Indicates that a pool is registered BUT NOT marked for retirement. Records the latest registration certificate.

PoolRegisteredAndRetired PoolRegistrationCertificate PoolRetirementCertificate
Indicates that a pool is registered AND ALSO marked for retirement. Records the latest registration and retirement certificates.

data PoolRegistrationCertificate

Pool ownership data from the stake pool registration certificate.

Constructors

PoolRegistrationCertificate

Fields

poolId :: !PoolId

poolOwners :: ![PoolOwner]

poolMargin :: Percentage

poolCost :: Coin

poolPledge :: Coin

poolMetadata :: Maybe (StakePoolMetadataUrl, StakePoolMetadataHash)

data PoolRetirementCertificate

Constructors

PoolRetirementCertificate

Fields

poolId :: !PoolId

retirementEpoch :: !EpochNo

The first epoch when the pool becomes inactive.

data PoolCertificate

Sum-type of pool registration- and retirement- certificates. Mirrors the PoolCert type in cardano-ledger-specs.

Constructors

Registration PoolRegistrationCertificate

Retirement PoolRetirementCertificate

getPoolCertificatePoolId :: PoolCertificate -> PoolId

setPoolCertificatePoolId :: PoolId -> PoolCertificate -> PoolCertificate

getPoolRegistrationCertificate :: PoolLifeCycleStatus -> Maybe PoolRegistrationCertificate

getPoolRetirementCertificate :: PoolLifeCycleStatus -> Maybe PoolRetirementCertificate

data NonWalletCertificate

Constructors

GenesisCertificate

MIRCertificate

data Certificate

Constructors

CertificateOfDelegation DelegationCertificate

CertificateOfPool PoolCertificate

CertificateOther NonWalletCertificate

Network Parameters

data NetworkParameters

Records the complete set of parameters currently in use by the network that are relevant to the wallet.

Constructors

NetworkParameters

Fields

genesisParameters :: GenesisParameters

See GenesisParameters.

slottingParameters :: SlottingParameters

See SlottingParameters.

protocolParameters :: ProtocolParameters

See ProtocolParameters.

data GenesisParameters

Parameters defined by the genesis block.

At present, these values cannot be changed through the update system.

They can only be changed through a soft or hard fork.

Constructors

GenesisParameters

Fields

getGenesisBlockHash :: Hash "Genesis"

Hash of the very first block

getGenesisBlockDate :: StartTime

Start time of the chain.

data SlottingParameters

Constructors

SlottingParameters

Fields

getSlotLength :: SlotLength

Length, in seconds, of a slot.

getEpochLength :: EpochLength

Number of slots in a single epoch.

getActiveSlotCoefficient :: ActiveSlotCoefficient

a.k.a f, in Genesis/Praos, corresponds to the % of active slots (i.e. slots for which someone can be elected as leader).

Determines the value of stabilityWindowShelley.

getSecurityParameter :: Quantity "block" Word32

a.k.a k, used to compute the 'stability window' on the chain (i.e. the longest possible chain fork in slots).

Determines the value of stabilityWindowByron and stabilityWindowShelley.

data ProtocolParameters

Protocol parameters that can be changed through the update system.

Constructors

ProtocolParameters

Fields

decentralizationLevel :: DecentralizationLevel

The current level of decentralization in the network.

txParameters :: TxParameters

Parameters that affect transaction construction.

desiredNumberOfStakePools :: Word16

The current desired number of stakepools in the network. Also known as k parameter.

minimumUTxO :: MinimumUTxO

Represents a way of calculating minimum UTxO values.

stakeKeyDeposit :: Coin

Registering a stake key requires storage on the node and as such needs a deposit. There may be more actions that require deposit (such as registering a stake pool).

eras :: EraInfo EpochNo

Contains information about when each era did start if it has already happened, or otherwise when it will start, if the hard-fork time is confirmed on-chain.

Note: this is not a practical way to tell the current era.

maximumCollateralInputCount :: Word16

Limit on the maximum number of collateral inputs present in a transaction.

minimumCollateralPercentage :: Natural

Specifies the minimum required amount of collateral as a percentage of the total transaction fee.

executionUnitPrices :: Maybe ExecutionUnitPrices

The prices for ExecutionUnits as a fraction of a Lovelace and used to determine the fee for the use of a script within a transaction, based on the ExecutionUnits needed by the use of the script.

currentNodeProtocolParameters :: Maybe ProtocolParameters

Get the last known node's protocol parameters. In principle, these can only change once per epoch.

data TxParameters

Parameters that relate to the construction of transactions.

Constructors

TxParameters

Fields

getFeePolicy :: FeePolicy

Formula for calculating the transaction fee.

getTxMaxSize :: Quantity "byte" Word16

Maximum size of a transaction (soft or hard limit).

getTokenBundleMaxSize :: TokenBundleMaxSize

Maximum size of a serialized TokenBundle (_maxValSize in the Alonzo ledger)

getMaxExecutionUnits :: ExecutionUnits

Max total script execution resources units allowed per tx

newtype TokenBundleMaxSize

The maximum size of a serialized TokenBundle (_maxValSize in the Alonzo ledger)

Constructors

TokenBundleMaxSize

Fields

unTokenBundleMaxSize :: TxSize

data EraInfo info

Represents info about the starting epoch/time of all possible eras.

Field values can be either: - Just pastEpochBoundary - the network forked to this era in the past. - Just futureEpochBoundary - the hard-fork to this era is confirmed, but it hasn't yet occured. - Nothing - the hard-fork to this era is not yet confirmed.

Note: this type is not a practical way to tell what the current era is.

It is expected that there is an order, byron, shelley, allegra, mary, by which the Maybe fields are filled in.

It might be cumbersome to work with this type. But we don't need to. A product of Maybe is both what we can query from the node, and what we need to provide in the wallet API.

Constructors

EraInfo

Fields

byron :: Maybe info

shelley :: Maybe info

allegra :: Maybe info

mary :: Maybe info

alonzo :: Maybe info

babbage :: Maybe info

emptyEraInfo :: EraInfo info

newtype ActiveSlotCoefficient

Constructors

ActiveSlotCoefficient

Fields

unActiveSlotCoefficient :: Double

data DecentralizationLevel

Indicates the current level of decentralization in the network.

According to the Design Specification for Delegation and Incentives in Cardano, the decentralization parameter d is a value in the range '[0, 1]', where:

d = '1' indicates that the network is completely federalized.

d = '0' indicates that the network is completely decentralized.

However, in Cardano Wallet, we represent the decentralization level as a percentage, where:

' 0 %' indicates that the network is completely federalized.

'100 %' indicates that the network is completely decentralized.

$sel:getDecentralizationLevel:DecentralizationLevel :: DecentralizationLevel -> Percentage

getFederationPercentage :: DecentralizationLevel -> Percentage

fromDecentralizationLevel :: Percentage -> DecentralizationLevel

fromFederationPercentage :: Percentage -> DecentralizationLevel

Percentage of federated nodes. Equal to the "decentralization parameter" d from the ledger specification.

newtype EpochLength

Number of slots in a single epoch

Constructors

EpochLength

Fields

unEpochLength :: Word32

newtype EpochNo

Constructors

EpochNo

Fields

unEpochNo :: Word31

unsafeEpochNo :: HasCallStack => Word32 -> EpochNo

Convert the specified value into an EpochNo, or fail if the value is too large.

isValidEpochNo :: EpochNo -> Bool

newtype FeePolicy

A linear equation of a free variable x. Represents the x -> a + b*x function where x can be either a transaction size in bytes or a number of inputs + outputs.

Constructors

LinearFee (LinearFunction Double)

data LinearFunction a

Constructors

LinearFunction

Fields

intercept :: a

slope :: a

data SlotId

A slot identifier is the combination of an epoch and slot.

Constructors

SlotId

Fields

epochNumber :: !EpochNo

slotNumber :: !SlotInEpoch

newtype SlotNo

The 0-based index for the Ourboros time slot.

Constructors

SlotNo

Fields

unSlotNo :: Word64

newtype SlotLength

Duration of a single slot.

Constructors

SlotLength

Fields

unSlotLength :: NominalDiffTime

newtype SlotInEpoch

Constructors

SlotInEpoch

Fields

unSlotInEpoch :: Word32

newtype StartTime

Blockchain start time

Constructors

StartTime UTCTime

stabilityWindowByron :: SlottingParameters -> Quantity "block" Word64

In Byron, this stability window is equal to 2k slots, where _k_ is the $sel:getSecurityParameter:SlottingParameters

stabilityWindowShelley :: SlottingParameters -> Quantity "block" Word64

In Shelley, this stability window is equal to _3k/f_ slots where _k_ is the $sel:getSecurityParameter:SlottingParameters and _f_ is the ActiveSlotCoefficient.

data ExecutionUnits

Constructors

ExecutionUnits

Fields

executionSteps :: Natural

This corresponds roughly to the time to execute a script.

executionMemory :: Natural

This corresponds roughly to the peak memory used during script execution.

data ExecutionUnitPrices

Constructors

ExecutionUnitPrices

Fields

pricePerStep :: Rational

pricePerMemoryUnit :: Rational

Wallet Metadata

data WalletMetadata

Additional information about a wallet that can't simply be derived from the blockchain like Wallet s is.

Whereas Wallet s in Primitive can be updated using applyBlock, WalletMetadata can not*.

) Except for possibly $sel:status:WalletDelegationNext and $sel:delegation:WalletMetadata...

Constructors

WalletMetadata

Fields

name :: !WalletName

creationTime :: !UTCTime

passphraseInfo :: !(Maybe WalletPassphraseInfo)

delegation :: !WalletDelegation

newtype WalletId

Constructors

WalletId

Fields

getWalletId :: Digest Blake2b_160

newtype WalletName

Length-restricted name of a wallet

Constructors

WalletName

Fields

getWalletName :: Text

walletNameMinLength :: Int

Calling 'fromText @WalletName' on shorter string will fail.

walletNameMaxLength :: Int

Calling 'fromText @WalletName' on a longer string will fail.

data WalletDelegation

Constructors

WalletDelegation

Fields

active :: !WalletDelegationStatus

next :: ![WalletDelegationNext]

data WalletDelegationStatus

Constructors

NotDelegating

Delegating !PoolId

data WalletDelegationNext

Constructors

WalletDelegationNext

Fields

changesAt :: !EpochNo

status :: !WalletDelegationStatus

class IsDelegatingTo a where

Methods

isDelegatingTo :: (PoolId -> Bool) -> a -> Bool

Stake Pools

data StakePoolsSummary

Constructors

StakePoolsSummary

Fields

nOpt :: Int

rewards :: Map PoolId Coin

stake :: Map PoolId Percentage

newtype PoolId

Identifies a stake pool. For Jörmungandr a PoolId is the blake2b-256 hash of the stake pool registration certificate.

Constructors

PoolId

Fields

getPoolId :: ByteString

newtype PoolOwner

A stake pool owner, which is a public key encoded in bech32 with prefix ed25519_pk.

Constructors

PoolOwner

Fields

getPoolOwner :: ByteString

poolIdBytesLength :: [Int]

decodePoolIdBech32 :: Text -> Either TextDecodingError PoolId

Decode a Bech32 encoded PoolId.

encodePoolIdBech32 :: PoolId -> Text

Encode PoolId as Bech32 with "pool" hrp.

data StakePoolMetadata

Information about a stake pool.

The metadata information is not used directly by cardano-wallet, but rather passed straight through to API consumers.

Constructors

StakePoolMetadata

Fields

ticker :: StakePoolTicker

Very short human-readable ID for the stake pool.

name :: Text

Name of the stake pool.

description :: Maybe Text

Short description of the stake pool.

homepage :: Text

Absolute URL for the stake pool's homepage link.

newtype StakePoolMetadataHash

A newtype to wrap metadata hash.

NOTE: not using the Hash type as this newtype is primarily for database interop which doesn't quite like DataKinds.

Constructors

StakePoolMetadataHash ByteString

newtype StakePoolMetadataUrl

A newtype to wrap metadata Url, mostly needed for database lookups and signature clarity.

Constructors

StakePoolMetadataUrl Text

newtype StakePoolTicker

Very short name for a stake pool.

Constructors

StakePoolTicker

Fields

unStakePoolTicker :: Text

data StakeKeyCertificate

Constructors

StakeKeyRegistration

StakeKeyDeregistration

data PoolMetadataGCStatus

Constructors

NotApplicable

NotStarted

Restarting POSIXTime

HasRun POSIXTime

Querying

data SortOrder

Represents a sort order, applicable to the results returned by a query.

Constructors

Ascending
Sort in ascending order.

Descending
Sort in descending order.

Ranges

data Range a

Represents a range of values.

A range is defined by two optional bounds:

an inclusive lower bound

an inclusive upper bound

There are four cases:

Value 	Range Represented	Membership Function

Range (Just x) (Just y)	[ x, y ]	\p -> p >= x && p <= y

Range (Just x) Nothing 	[ x, ∞ )	\p -> p >= x

Range Nothing (Just y)	(−∞, y ]	\p -> p <= y

Range Nothing Nothing 	(−∞, ∞ )	\p -> True

Constructors

Range

Fields

inclusiveLowerBound :: Maybe a

inclusiveUpperBound :: Maybe a

data RangeBound a

Represents a range boundary.

Constructors

NegativeInfinity

InclusiveBound a

PositiveInfinity

wholeRange :: Range a

The range that includes everything.

isAfterRange :: Ord a => a -> Range a -> Bool

Returns True if (and only if) the given range has an upper bound and the specified value is greater than the upper bound.

isBeforeRange :: Ord a => a -> Range a -> Bool

Returns True if (and only if) the given range has a lower bound and the specified value is smaller than the lower bound.

isSubrangeOf :: Ord a => Range a -> Range a -> Bool

Returns True if (and only if) the first given range is a subrange of the second given range.

isWithinRange :: Ord a => a -> Range a -> Bool

Returns True if (and only if) the given value is not smaller than the lower bound (if present) of the given range and is not greater than the upper bound (if present) of the given range.

mapRangeLowerBound :: (a -> a) -> Range a -> Range a

Apply a function to the lower bound of a range.

mapRangeUpperBound :: (a -> a) -> Range a -> Range a

Apply a function to the upper bound of a range.

rangeIsFinite :: Range a -> Bool

Returns True if (and only if) the given range has both a lower and upper bound.

rangeIsSingleton :: Eq a => Range a -> Bool

Returns True if (and only if) the range covers exactly one value.

rangeIsValid :: Ord a => Range a -> Bool

Returns True if (and only if) the lower bound of a range is not greater than its upper bound.

rangeHasLowerBound :: Range a -> Bool

Returns True if (and only if) the given range has a lower bound.

rangeHasUpperBound :: Range a -> Bool

Returns True if (and only if) the given range has an upper bound.

rangeLowerBound :: Range a -> RangeBound a

Get the lower bound of a Range.

rangeUpperBound :: Range a -> RangeBound a

Get the upper bound of a Range.

Polymorphic

newtype Signature (what :: Type)

A newtype to wrap raw bytestring representing signed data, captured with a phantom type.

Constructors

Signature

Fields

getSignature :: ByteString

Settings

data Settings

Wallet application settings. These are stored at runtime and potentially mutable.

Constructors

Settings

Fields

poolMetadataSource :: PoolMetadataSource

data SmashServer

A SMASH server is either an absolute http or https url.

Don't export SmashServer constructor, use fromText instance instead.

$sel:unSmashServer:SmashServer :: SmashServer -> URI

data PoolMetadataSource

Source of Stake Pool Metadata aggregation.

Constructors

FetchNone

FetchDirect

FetchSMASH SmashServer

defaultSettings :: Settings

unsafeToPMS :: URI -> PoolMetadataSource

newtype TokenMetadataServer

Constructors

TokenMetadataServer

Fields

unTokenMetadataServer :: URI

InternalState

data InternalState

Various internal states of the pool DB that need to survive wallet restarts. These aren't exposed settings.

Constructors

InternalState

Fields

lastMetadataGC :: Maybe POSIXTime

defaultInternalState :: InternalState

Orphan instances

Buildable Slot

Instance details

Methods

build :: Slot -> Builder