Cardano.Wallet.Primitive.Types.MinimumUTxO

Description

Defines the MinimumUTxO type and related functions.

Types

data MinimumUTxO where

Represents a function for computing minimum UTxO values.

Constructors

MinimumUTxONone

Fields

:: MinimumUTxO

Indicates that there is no minimum UTxO value.

MinimumUTxOConstant

Fields

:: Coin

-> MinimumUTxO

Indicates a constant minimum UTxO value. This constructor is useful for writing tests, where we often want to have precise control over the value that is chosen.

MinimumUTxOForShelleyBasedEraOf

Fields

:: MinimumUTxOForShelleyBasedEra

-> MinimumUTxO

Indicates a Shelley-based era-specific minimum UTxO function.

data MinimumUTxOForShelleyBasedEra where

Represents a minimum UTxO function that is specific to a Shelley-based era.

Constructors

MinimumUTxOForShelleyBasedEra :: ShelleyBasedEra era -> PParams (ShelleyLedgerEra era) -> MinimumUTxOForShelleyBasedEra

Constructor functions

minimumUTxONone :: MinimumUTxO

minimumUTxOConstant :: Coin -> MinimumUTxO

minimumUTxOForShelleyBasedEra :: ShelleyBasedEra era -> PParams (ShelleyLedgerEra era) -> MinimumUTxO