Cardano.Wallet.Primitive.Types.TokenBundle

Description

Provides the TokenBundle type, which combines a Coin (lovelace) value with a map of named token quantities, scoped by token policy.

This module is meant to be imported qualified. For example:

>>> import qualified Cardano.Wallet.Primitive.Types.TokenBundle as TB

Types

data TokenBundle

Combines a Coin (lovelace) value with a map of named token quantities, grouped by token policy.

Constructors

TokenBundle

Fields

coin :: !Coin

tokens :: !TokenMap

data AssetId

A combination of a token policy identifier and a token name that can be used as a compound identifier.

Constructors

AssetId

Fields

tokenPolicyId :: !TokenPolicyId

tokenName :: !TokenName

Construction

empty :: TokenBundle

The empty token bundle.

fromFlatList :: Coin -> [(AssetId, TokenQuantity)] -> TokenBundle

Creates a token bundle from a coin and a flat list of token quantities.

If a token name appears more than once in the list under the same policy, its associated quantities will be added together in the resultant bundle.

fromNestedList :: Coin -> [(TokenPolicyId, NonEmpty (TokenName, TokenQuantity))] -> TokenBundle

Creates a token bundle from a coin and a nested list of token quantities.

If a token name appears more than once in the list under the same policy, its associated quantities will be added together in the resultant bundle.

fromNestedMap :: (Coin, Map TokenPolicyId (NonEmptyMap TokenName TokenQuantity)) -> TokenBundle

fromTokenMap :: TokenMap -> TokenBundle

Deconstruction

toFlatList :: TokenBundle -> (Coin, [(AssetId, TokenQuantity)])

Converts a token bundle to a coin and a flat list of token quantities.

Coins

fromCoin :: Coin -> TokenBundle

Creates a singleton token bundle from an ada Coin value.

toCoin :: TokenBundle -> Maybe Coin

Coerces a token bundle to an ada Coin value.

Returns a coin if (and only if) the token bundle has no other tokens.

isCoin :: TokenBundle -> Bool

Indicates True if (and only if) a token bundle has no tokens other than an ada Coin value.

getCoin :: TokenBundle -> Coin

Gets the current ada Coin value from a token bundle.

If you need to assert that a bundle has no other tokens, consider using the toCoin function instead.

setCoin :: TokenBundle -> Coin -> TokenBundle

Sets the current ada Coin value for a token bundle.

Arithmetic

add :: TokenBundle -> TokenBundle -> TokenBundle

Adds one token bundle to another.

subtract :: TokenBundle -> TokenBundle -> Maybe TokenBundle

Subtracts the second token bundle from the first.

Returns Nothing if the second bundle is not less than or equal to the first bundle when compared with the leq function.

difference :: TokenBundle -> TokenBundle -> TokenBundle

Analogous to Set.difference, return the difference between two token maps.

The following property holds: prop> x leq (x difference y) add y

Note that there's a leq rather than equality, which we'd expect if this was subtraction of integers. I.e.

>>> (0 - 1) + 1
0

whereas

>>> let oneToken = fromFlatList coin [(aid, TokenQuantity 1)]
>>> (mempty `difference` oneToken) `add` oneToken
oneToken

Quantities

getQuantity :: TokenBundle -> AssetId -> TokenQuantity

Gets the quantity associated with a given asset.

If the given bundle does not have an entry for the specified asset, this function returns a value of zero.

hasQuantity :: TokenBundle -> AssetId -> Bool

Returns true if and only if the given bundle has a non-zero quantity for the given asset.

setQuantity :: TokenBundle -> AssetId -> TokenQuantity -> TokenBundle

Sets the quantity associated with a given asset.

If the given quantity is zero, the resultant bundle will not have an entry for the given asset.

Partitioning

equipartitionAssets

Arguments

:: TokenBundle
The token bundle to be partitioned.

-> NonEmpty a
Represents the number of portions in which to partition the bundle.

-> NonEmpty TokenBundle
The partitioned bundles.

Partitions a token bundle into n smaller bundles, where the asset sets of the resultant bundles are disjoint.

In the resultant bundles, the smallest asset set size and largest asset set size will differ by no more than 1.

The ada Coin quantity is equipartitioned across the resulting bundles.

The quantities of each non-ada asset are unchanged.

equipartitionQuantitiesWithUpperBound

Arguments

:: TokenBundle

-> TokenQuantity
Maximum allowable token quantity.

-> NonEmpty TokenBundle
The partitioned bundles.

Partitions a token bundle into n smaller bundles, where the quantity of each token is equipartitioned across the resultant bundles, with the goal that no token quantity in any of the resultant bundles exceeds the given upper bound.

The value n is computed automatically, and is the minimum value required to achieve the goal that no token quantity in any of the resulting bundles exceeds the maximum allowable token quantity.

Ordering

newtype Lexicographic a

Defines a lexicographic ordering.

Constructors

Lexicographic

Fields

unLexicographic :: a

Serialization

newtype Flat a

When used with the Buildable or ToJSON instances, provides a flat serialization style, where token quantities are paired with their asset identifiers.

Constructors

Flat

Fields

getFlat :: a

newtype Nested a

When used with the Buildable or ToJSON instances, provides a nested serialization style, where token quantities are grouped by policy identifier.

Constructors

Nested

Fields

getNested :: a

Queries

getAssets :: TokenBundle -> Set AssetId

Transformations

mapAssetIds :: (AssetId -> AssetId) -> TokenBundle -> TokenBundle

Unsafe operations

unsafeSubtract :: TokenBundle -> TokenBundle -> TokenBundle

Subtracts the second token bundle from the first.

Pre-condition: the second bundle is less than or equal to the first bundle when compared with the leq function.

Throws a run-time exception if the pre-condition is violated.