Cardano.Wallet.Primitive.Types.TokenBundle.Gen

Documentation

genTokenBundleSmallRange :: Gen TokenBundle

genTokenBundleSmallRangePositive :: Gen TokenBundle

genTokenBundle :: Gen TokenBundle

shrinkTokenBundle :: TokenBundle -> [TokenBundle]

shrinkTokenBundleSmallRange :: TokenBundle -> [TokenBundle]

shrinkTokenBundleSmallRangePositive :: TokenBundle -> [TokenBundle]

genTokenBundlePartition :: TokenBundle -> Int -> Gen (NonEmpty TokenBundle)

Partitions a token bundle randomly into a given number of parts.

Satisfies the following properties:

forAll (genTokenBundlePartition b i) $ (==       b) . fold

forAll (genTokenBundlePartition b i) $ (== max 1 i) . length

genTokenBundlePartitionNonNull :: TokenBundle -> Int -> Gen [TokenBundle]

Like genTokenBundlePartition, but with empty values removed from the result.