Cardano.Wallet.Primitive.Types.TokenMap

Description

Provides the TokenMap type, which represents a map of named non-ada token quantities scoped by token policy.

The TokenMap type does not provide a way to store ada quantities. If you also need to store ada quantities, use the TokenBundle type.

This module is meant to be imported qualified. For example:

>>> import qualified Cardano.Wallet.Primitive.Types.TokenMap as TM

Types

data TokenMap

A map of named token quantities, grouped by token policy.

The token map data structure has an important invariant: all token quantities held within a map are non-zero.

This means that:

using the setQuantity function to add a zero-valued quantity to a map is equivalent to applying the identity operation to that map.

using the setQuantity function to change an existing quantity to zero is equivalent to removing that quantity from the map.

As a consequence of this invariant, the token map data structure is always in its canonical form: we can perform an equality check without needing any extra canonicalization steps.

data AssetId

A combination of a token policy identifier and a token name that can be used as a compound identifier.

Constructors

AssetId

Fields

tokenPolicyId :: !TokenPolicyId

tokenName :: !TokenName

Construction

empty :: TokenMap

The empty token map.

singleton :: AssetId -> TokenQuantity -> TokenMap

Creates a singleton token map with just one token quantity.

If the specified token quantity is zero, then the resultant map will be equal to the empty map.

fromFlatList :: [(AssetId, TokenQuantity)] -> TokenMap

Creates a token map from a flat list.

If a token name appears more than once in the list under the same policy, its associated quantities will be added together in the resultant map.

fromNestedList :: [(TokenPolicyId, NonEmpty (TokenName, TokenQuantity))] -> TokenMap

Creates a token map from a nested list.

If a token name appears more than once in the list under the same policy, its associated quantities will be added together in the resultant map.

fromNestedMap :: Map TokenPolicyId (NonEmptyMap TokenName TokenQuantity) -> TokenMap

Creates a token map from a nested map.

Deconstruction

toFlatList :: TokenMap -> [(AssetId, TokenQuantity)]

Converts a token map to a flat list.

toNestedList :: TokenMap -> [(TokenPolicyId, NonEmpty (TokenName, TokenQuantity))]

Converts a token map to a nested list.

toNestedMap :: TokenMap -> Map TokenPolicyId (NonEmptyMap TokenName TokenQuantity)

Converts a token map to a nested map.

Filtering

filter :: (AssetId -> Bool) -> TokenMap -> TokenMap

Arithmetic

add :: TokenMap -> TokenMap -> TokenMap

Adds one token map to another.

subtract :: TokenMap -> TokenMap -> Maybe TokenMap

Subtracts the second token map from the first.

Returns Nothing if the second map is not less than or equal to the first map when compared with the leq function.

difference :: TokenMap -> TokenMap -> TokenMap

Analogous to Set.difference, return the difference between two token maps.

The following property holds: prop> x leq (x difference y) add y

Note that there's a leq rather than equality, which we'd expect if this was subtraction of integers. I.e.

>>> (0 - 1) + 1
0

whereas

>>> let oneToken = singleton aid (TokenQuantity 1)
>>> (mempty `difference` oneToken) `add` oneToken
oneToken

intersection :: TokenMap -> TokenMap -> TokenMap

Computes the intersection of two token maps.

Analogous to Set.intersection.

Example:

>>> m1 = [("a", 1), ("b", 2), ("c", 3)          ]
>>> m2 = [          ("b", 3), ("c", 2), ("d", 1)]
>>> intersection m1 m2
         [          ("b", 2), ("c", 2)          ]

Queries

size :: TokenMap -> Int

Returns the number of unique assets in a token map.

Tests

isEmpty :: TokenMap -> Bool

Returns true if and only if the given map is empty.

isNotEmpty :: TokenMap -> Bool

Returns true if and only if the given map is not empty.

Quantities

getQuantity :: TokenMap -> AssetId -> TokenQuantity

Gets the quantity associated with a given asset.

If the given map does not have an entry for the specified asset, this function returns a value of zero.

setQuantity :: TokenMap -> AssetId -> TokenQuantity -> TokenMap

Updates the quantity associated with a given asset.

If the given quantity is zero, the resultant map will not have an entry for the given asset.

hasQuantity :: TokenMap -> AssetId -> Bool

Returns true if and only if the given map has a non-zero quantity for the given asset.

adjustQuantity :: TokenMap -> AssetId -> (TokenQuantity -> TokenQuantity) -> TokenMap

Uses the specified function to adjust the quantity associated with a given asset.

If the result of adjusting the quantity is equal to zero, the resultant map will not have an entry for the given asset.

removeQuantity :: TokenMap -> AssetId -> TokenMap

Removes the quantity associated with the given asset.

This is equivalent to calling setQuantity with a value of zero.

maximumQuantity :: TokenMap -> TokenQuantity

Get the largest quantity from this map.

Partitioning

equipartitionAssets

Arguments

:: TokenMap
The token map to be partitioned.

-> NonEmpty a
Represents the number of portions in which to partition the token map.

-> NonEmpty TokenMap
The partitioned maps.

Partitions a token map into n smaller maps, where the asset sets of the resultant maps are disjoint.

In the resultant maps, the smallest asset set size and largest asset set size will differ by no more than 1.

The quantities of each asset are unchanged.

equipartitionQuantities

Arguments

:: TokenMap
The map to be partitioned.

-> NonEmpty a
Represents the number of portions in which to partition the map.

-> NonEmpty TokenMap
The partitioned maps.

Partitions a token map into n smaller maps, where the quantity of each token is equipartitioned across the resultant maps.

In the resultant maps, the smallest quantity and largest quantity of a given token will differ by no more than 1.

The resultant list is sorted into ascending order when maps are compared with the leq function.

equipartitionQuantitiesWithUpperBound

Arguments

:: TokenMap

-> TokenQuantity
Maximum allowable token quantity.

-> NonEmpty TokenMap
The partitioned maps.

Partitions a token map into n smaller maps, where the quantity of each token is equipartitioned across the resultant maps, with the goal that no token quantity in any of the resultant maps exceeds the given upper bound.

The value n is computed automatically, and is the minimum value required to achieve the goal that no token quantity in any of the resulting maps exceeds the maximum allowable token quantity.

Ordering

newtype Lexicographic a

Defines a lexicographic ordering.

Constructors

Lexicographic

Fields

unLexicographic :: a

Serialization

newtype Flat a

When used with the Buildable or ToJSON instances, provides a flat serialization style, where token quantities are paired with their asset identifiers.

Constructors

Flat

Fields

getFlat :: a

newtype Nested a

When used with the Buildable or ToJSON instances, provides a nested serialization style, where token quantities are grouped by policy identifier.

Constructors

Nested

Fields

getNested :: a

Queries

getAssets :: TokenMap -> Set AssetId

Transformations

mapAssetIds :: (AssetId -> AssetId) -> TokenMap -> TokenMap

Unsafe operations

unsafeSubtract :: TokenMap -> TokenMap -> TokenMap

Subtracts the second token map from the first.

Pre-condition: the second map is less than or equal to the first map when compared with the leq function.

Throws a run-time exception if the pre-condition is violated.