Cardano.Wallet.Primitive.Types.TokenMap.Gen

Documentation

genAssetId :: Gen AssetId

genAssetIdLargeRange :: Gen AssetId

genTokenMap :: Gen TokenMap

genTokenMapSmallRange :: Gen TokenMap

shrinkAssetId :: AssetId -> [AssetId]

shrinkTokenMap :: TokenMap -> [TokenMap]

newtype AssetIdF

Constructors

AssetIdF AssetId

genTokenMapPartition :: TokenMap -> Int -> Gen (NonEmpty TokenMap)

Partitions a token map randomly into a given number of parts.

Satisfies the following properties:

forAll (genTokenMapPartition m i) $ (== m      ) . fold

forAll (genTokenMapPartition m i) $ (== max 1 i) . length

genTokenMapPartitionNonNull :: TokenMap -> Int -> Gen [TokenMap]

Like genTokenMapPartition, but with empty values removed from the result.