Cardano.Wallet.Primitive.Types.TokenPolicy

Token Policies

newtype TokenPolicyId

Token policy identifiers, represented by the hash of the monetary policy script.

Constructors

UnsafeTokenPolicyId
Construct a TokenPolicyId without any validation.

Fields

unTokenPolicyId :: Hash "TokenPolicy"

Token Names

newtype TokenName

Token names, defined by the monetary policy script.

Constructors

UnsafeTokenName
Construct a TokenName without any validation.

Fields

unTokenName :: ByteString

mkTokenName :: ByteString -> Either String TokenName

Construct a TokenName, validating that the length does not exceed tokenNameMaxLength.

nullTokenName :: TokenName

The empty asset name.

Asset names may be empty, where a monetary policy script only mints a single asset, or where one asset should be considered as the "default" token for the policy.

tokenNameMaxLength :: Int

The maximum length of a valid token name.

Token Fingerprints

newtype TokenFingerprint

Constructors

UnsafeTokenFingerprint

Fields

unTokenFingerprint :: Text

mkTokenFingerprint :: TokenPolicyId -> TokenName -> TokenFingerprint

Construct a fingerprint from a TokenPolicyId and TokenName. The fingerprint is not necessarily unique, but can be used in user-facing interfaces as a comparison mechanism.

Token Metadata

data AssetMetadata

Information about an asset, from a source external to the chain.

Constructors

AssetMetadata

Fields

name :: Text

description :: Text

ticker :: Maybe Text

url :: Maybe AssetURL

logo :: Maybe AssetLogo

decimals :: Maybe AssetDecimals

newtype AssetURL

The validated URL for the asset.

Constructors

AssetURL

Fields

unAssetURL :: URI

newtype AssetLogo

Specify an asset logo as an image data payload

Constructors

AssetLogo

Fields

unAssetLogo :: ByteString

newtype AssetDecimals

Constructors

AssetDecimals

Fields

unAssetDecimals :: Int

validateMetadataDecimals :: AssetDecimals -> Either String AssetDecimals

validateMetadataName :: Text -> Either String Text

validateMetadataTicker :: Text -> Either String Text

validateMetadataDescription :: Text -> Either String Text

validateMetadataURL :: Text -> Either String AssetURL

validateMetadataLogo :: AssetLogo -> Either String AssetLogo