Cardano.Wallet.Primitive.Types.Tx

Description

This module provides the main transaction data types used by the wallet.

Types

data Tx

Primitive Tx-type.

Currently tailored for jormungandr in that inputs are (TxIn, Coin) instead of TxIn. We might have to revisit this when supporting another node.

Constructors

Tx

Fields

txId :: Hash "Tx"

Jörmungandr computes transaction id by hashing the full content of the transaction, which includes witnesses. Therefore, we need either to keep track of the witnesses to be able to re-compute the tx id every time, or, simply keep track of the id itself.

fee :: !(Maybe Coin)

Explicit fee for that transaction, if available. Fee are available explicitly in Shelley, but not in Byron although in Byron they can easily be re-computed from the delta between outputs and inputs.

resolvedInputs :: ![(TxIn, Coin)]

NOTE: Order of inputs matters in the transaction representation. The transaction id is computed from the binary representation of a tx, for which inputs are serialized in a specific order.

resolvedCollateralInputs :: ![(TxIn, Coin)]

NOTE: The order of collateral inputs matters in the transaction representation. The transaction id is computed from the binary representation of a tx, for which collateral inputs are serialized in a specific order.

outputs :: ![TxOut]

NOTE: Order of outputs matters in the transaction representations. Outputs are used as inputs for next transactions which refer to them using their indexes. It matters also for serialization.

collateralOutput :: !(Maybe TxOut)

An output that is only created if a transaction script fails.

withdrawals :: !(Map RewardAccount Coin)

Withdrawals (of funds from a registered reward account) embedded in a transaction. The order does not matter.

metadata :: !(Maybe TxMetadata)

Semi-structured application-specific extension data stored in the transaction on chain.

This is not to be confused with TxMeta, which is information about a transaction derived from the ledger.

See Appendix E of Shelley Ledger: Delegation/Incentives Design Spec.

scriptValidity :: !(Maybe TxScriptValidity)

Tag indicating whether non-native scripts in this transaction passed validation. This is added by the block creator when constructing the block. May be Nothing for pre-Alonzo and pending transactions.

data TxIn

Constructors

TxIn

Fields

inputId :: !(Hash "Tx")

inputIx :: !Word32

data TxOut

Constructors

TxOut

Fields

address :: !Address

tokens :: !TokenBundle

data TxChange derivationPath

Constructors

TxChange

Fields

address :: !Address

amount :: !Coin

assets :: !TokenMap

derivationPath :: derivationPath

data TxMeta

Additional information about a transaction, derived from the transaction and ledger state. This should not be confused with TxMetadata which is application-specific data included with the transaction.

TODO: TxProperties or TxProps would be a good name for this type.

Constructors

TxMeta

Fields

status :: !TxStatus

direction :: !Direction

slotNo :: !SlotNo

blockHeight :: !(Quantity "block" Word32)

amount :: !Coin

Amount seen from the perspective of the wallet. Refers either to a spent value for outgoing transaction, or a received value on incoming transaction.

expiry :: !(Maybe SlotNo)

The slot at which a pending transaction will no longer be accepted into mempools.

newtype TxMetadata

Constructors

TxMetadata (Map Word64 TxMetadataValue)

data TxMetadataValue

Constructors

TxMetaMap [(TxMetadataValue, TxMetadataValue)]

TxMetaList [TxMetadataValue]

TxMetaNumber Integer

TxMetaBytes ByteString

TxMetaText Text

data TxStatus

Constructors

Pending
Created, but not yet in a block.

InLedger
Has been found in a block.

Expired
Time to live (TTL) has passed.

data UnsignedTx input output change withdrawal

An unsigned transaction.

See Tx for a signed transaction.

Constructors

UnsignedTx

Fields

unsignedCollateral :: [input]

unsignedInputs :: [input]

Inputs are *necessarily* non-empty because Cardano requires at least one UTxO input per transaction to prevent replayable transactions. (each UTxO being unique, including at least one UTxO in the transaction body makes it seemingly unique).

However* when used to represent the inputs known by the wallet, in contrast to all inputs, it can be empty.

unsignedOutputs :: [output]

unsignedChange :: [change]

unsignedWithdrawals :: [withdrawal]

data TransactionInfo

Full expanded and resolved information about a transaction, suitable for presentation to the user.

Constructors

TransactionInfo

Fields

txInfoId :: !(Hash "Tx")

Transaction ID of this transaction

txInfoFee :: !(Maybe Coin)

Explicit transaction fee

txInfoInputs :: ![(TxIn, Coin, Maybe TxOut)]

Transaction inputs and (maybe) corresponding outputs of the source. Source information can only be provided for outgoing payments.

txInfoCollateralInputs :: ![(TxIn, Coin, Maybe TxOut)]

Collateral inputs and (maybe) corresponding outputs.

txInfoOutputs :: ![TxOut]

Payment destination.

txInfoCollateralOutput :: !(Maybe TxOut)

An output that is only created if a transaction script fails.

txInfoWithdrawals :: !(Map RewardAccount Coin)

Withdrawals on this transaction.

txInfoMeta :: !TxMeta

Other information calculated from the transaction.

txInfoDepth :: Quantity "block" Natural

Number of slots since the transaction slot.

txInfoTime :: UTCTime

Creation time of the block including this transaction.

txInfoMetadata :: !(Maybe TxMetadata)

Application-specific extension data.

txInfoScriptValidity :: !(Maybe TxScriptValidity)

Tag indicating whether non-native scripts in this transaction passed validation. This is added by the block creator when constructing the block. May be Nothing for pre-Alonzo and pending transactions.

data Direction

The effect of a Transaction on the wallet balance.

Constructors

Outgoing
The wallet balance decreases.

Incoming
The wallet balance increases or stays the same.

data LocalTxSubmissionStatus tx

Information about when a transaction was submitted to the local node. This is used for scheduling resubmissions.

Constructors

LocalTxSubmissionStatus

Fields

txId :: !(Hash "Tx")

submittedTx :: !tx

firstSubmission :: !SlotNo

Time of first successful submission to the local node.

latestSubmission :: !SlotNo

Time of most recent resubmission attempt.

newtype TokenBundleSizeAssessor

A function capable of assessing the size of a token bundle relative to the upper limit of what can be included in a single transaction output.

In general, a token bundle size assessment function f should satisfy the following properties:

Enlarging a bundle that exceeds the limit should also result in a bundle that exceeds the limit: f b1 == TokenBundleSizeExceedsLimit ==> f (b1 add b2) == TokenBundleSizeExceedsLimit

Shrinking a bundle that's within the limit should also result in a bundle that's within the limit: f b1 == TokenBundleWithinLimit ==> f (b1 difference b2) == TokenBundleWithinLimit

Constructors

TokenBundleSizeAssessor

Fields

assessTokenBundleSize :: TokenBundle -> TokenBundleSizeAssessment

data TokenBundleSizeAssessment

Indicates the size of a token bundle relative to the upper limit of what can be included in a single transaction output, defined by the protocol.

Constructors

TokenBundleSizeWithinLimit
Indicates that the size of a token bundle does not exceed the maximum size that can be included in a transaction output.

TokenBundleSizeExceedsLimit
Indicates that the size of a token bundle exceeds the maximum size that can be included in a transaction output.

data TxScriptValidity

Indicates whether or not a transaction is marked as having an invalid script.

Pre-Alonzo era, scripts were not supported.

Constructors

TxScriptValid
The transaction is not marked as having an invalid script.

TxScriptInvalid
The transaction is marked as having an invalid script.

data ScriptWitnessIndex

Identify the location of a ScriptWitness within the context of a TxBody. These are indexes of the objects within the transaction that need or can use script witnesses: inputs, minted assets, withdrawals and certificates. These are simple numeric indices, enumerated from zero. Thus the indices are not stable if the transaction body is modified.

Constructors

ScriptWitnessIndexTxIn !Word
The n'th transaction input, in the order of the TxIds.

ScriptWitnessIndexMint !Word
The n'th minting PolicyId, in the order of the PolicyIds.

ScriptWitnessIndexCertificate !Word
The n'th certificate, in the list order of the certificates.

ScriptWitnessIndexWithdrawal !Word
The n'th withdrawal, in the order of the StakeAddresss.

Serialisation

data SealedTx

SealedTx is a transaction for any hard fork era, possibly incomplete, possibly unsigned, with dual representations to make it convenient to use.

Serialisation/deserialisation is usually done at the application boundaries (e.g. in the API server), and then the wallet core can use it either as a ByteString, or as a Tx.

Construct it with either sealedTxFromCardano or sealedTxFromBytes.

cardanoTxIdeallyNoLaterThan :: AnyCardanoEra -> SealedTx -> InAnyCardanoEra Tx

sealedTxFromBytes :: ByteString -> Either DecoderError SealedTx

Deserialise a transaction to construct a SealedTx.

sealedTxFromBytes'

Arguments

:: AnyCardanoEra
Most recent era

-> ByteString
Serialised transaction

-> Either DecoderError SealedTx

Deserialise a transaction to construct a SealedTx.

sealedTxFromCardano :: InAnyCardanoEra Tx -> SealedTx

Construct a SealedTx from a Cardano.Api transaction.

sealedTxFromCardano' :: IsCardanoEra era => Tx era -> SealedTx

Construct a SealedTx from a Cardano.Api transaction.

sealedTxFromCardanoBody :: IsCardanoEra era => TxBody era -> SealedTx

Construct a SealedTx from a TxBody.

getSerialisedTxParts :: SealedTx -> SerialisedTxParts

Get the serialised transaction body and witnesses from a SealedTx.

unsafeSealedTxFromBytes :: HasCallStack => ByteString -> SealedTx

Only use this for tests.

newtype SerialisedTx

A serialised transaction that may be only partially signed, or even invalid.

Constructors

SerialisedTx

Fields

payload :: ByteString

data SerialisedTxParts

SerialisedTxParts is a serialised transaction body, and a possibly incomplete set of serialised witnesses.

Constructors

SerialisedTxParts

Fields

serialisedTxBody :: ByteString

serialisedTxWitnesses :: [ByteString]

getSealedTxBody :: SealedTx -> InAnyCardanoEra TxBody

getSealedTxWitnesses :: SealedTx -> [InAnyCardanoEra KeyWitness]

persistSealedTx :: SealedTx -> ByteString

Serialise a SealedTx for storage in a database field. The difference between persistSealedTx and $sel:serialisedTx:SealedTx is that this function has a special check for values created by mockSealedTx.

unPersistSealedTx :: ByteString -> Either Text SealedTx

Deserialise a SealedTx which has been stored in a database field. This function includes a special check for mockSealedTx values.

Unit testing helpers

mockSealedTx :: HasCallStack => ByteString -> SealedTx

Construct a SealedTx from a string which need not be a well-formed serialised Cardano transaction.

Be careful using the SealedTx, because any attempt to evaluate its cardanoTx field will crash.

withinEra :: AnyCardanoEra -> AnyCardanoEra -> Bool

a withinEra b is True iff b is the same era as a, or an earlier one.

Functions

fromTransactionInfo :: TransactionInfo -> Tx

Reconstruct a transaction info from a transaction.

inputs :: Tx -> [TxIn]

collateralInputs :: Tx -> [TxIn]

isPending :: TxMeta -> Bool

True if the given metadata refers to a pending transaction

toTxHistory :: TransactionInfo -> (Tx, TxMeta)

Drop time-specific information

txIns :: Set Tx -> Set TxIn

txMetadataIsNull :: TxMetadata -> Bool

Test whether the given metadata map is empty.

txOutCoin :: TxOut -> Coin

txOutAddCoin :: Coin -> TxOut -> TxOut

Increments the Coin value of a TxOut.

Satisfies the following property for all values of c:

>>> txOutSubtractCoin c . txOutAddCoin c == id

txOutSubtractCoin :: Coin -> TxOut -> TxOut

Decrements the Coin value of a TxOut.

Satisfies the following property for all values of c:

>>> txOutSubtractCoin c . txOutAddCoin c == id

If the given Coin is greater than the Coin value of the given TxOut, the resulting TxOut will have a Coin value of zero.

txScriptInvalid :: Tx -> Bool

Returns True if (and only if) the given transaction is marked as having an invalid script.

This function does not actually verify the validity of scripts; it merely checks for the presence or absence of the TxScriptInvalid marker.

Constants

txOutMinCoin :: Coin

The smallest quantity of lovelace that can appear in a transaction output's token bundle.

txOutMaxCoin :: Coin

The greatest quantity of lovelace that can appear in a transaction output's token bundle.

txOutMinTokenQuantity :: TokenQuantity

The smallest token quantity that can appear in a transaction output's token bundle.

txOutMaxTokenQuantity :: TokenQuantity

The greatest token quantity that can appear in a transaction output's token bundle.

Although the ledger specification allows token quantities of unlimited sizes, in practice we'll only see transaction outputs where the token quantities are bounded by the size of a Word64.

txMintBurnMaxTokenQuantity :: TokenQuantity

The greatest quantity of any given token that can be minted or burned in a transaction.

Constraints

data TxConstraints

Provides an abstract cost and size model for transactions.

This allows parts of a transaction to be costed (or sized) individually, without having to compute the cost (or size) of an entire transaction.

Note that the following functions assume one witness is required per input:

$sel:txInputCost:TxConstraints

$sel:txInputSize:TxConstraints

This will lead to slight overestimation in the case of UTxOs that share the same payment key.

Constructors

TxConstraints

Fields

txBaseCost :: Coin

The constant cost of an empty transaction.

txBaseSize :: TxSize

The constant size of an empty transaction.

txInputCost :: Coin

The constant cost of a transaction input, assuming one witness is required per input.

txInputSize :: TxSize

The constant size of a transaction input, assuming one witness is required per input.

txOutputCost :: TokenBundle -> Coin

The variable cost of a transaction output.

txOutputSize :: TokenBundle -> TxSize

The variable size of a transaction output.

txOutputMaximumSize :: TxSize

The maximum size of a transaction output.

txOutputMaximumTokenQuantity :: TokenQuantity

The maximum token quantity that can appear in a transaction output.

txOutputMinimumAdaQuantity :: Address -> TokenMap -> Coin

The variable minimum ada quantity of a transaction output.

txOutputBelowMinimumAdaQuantity :: Address -> TokenBundle -> Bool

Returns True if the given TokenBundle has a Coin value that is below the minimum required.

txRewardWithdrawalCost :: Coin -> Coin

The variable cost of a reward withdrawal.

txRewardWithdrawalSize :: Coin -> TxSize

The variable size of a reward withdrawal.

txMaximumSize :: TxSize

The maximum size of a transaction.

txOutputCoinCost :: TxConstraints -> Coin -> Coin

txOutputCoinSize :: TxConstraints -> Coin -> TxSize

txOutputHasValidSize :: TxConstraints -> TokenBundle -> Bool

txOutputHasValidTokenQuantities :: TxConstraints -> TokenMap -> Bool

newtype TxSize

The size of a transaction, or part of a transaction, in bytes.

Constructors

TxSize

Fields

unTxSize :: Natural

txSizeDistance :: TxSize -> TxSize -> TxSize

Computes the absolute distance between two transaction size quantities.

Queries

txAssetIds :: Tx -> Set AssetId

txOutAssetIds :: TxOut -> Set AssetId

Transformations

txMapAssetIds :: (AssetId -> AssetId) -> Tx -> Tx

txMapTxIds :: (Hash "Tx" -> Hash "Tx") -> Tx -> Tx

txRemoveAssetId :: Tx -> AssetId -> Tx

txOutMapAssetIds :: (AssetId -> AssetId) -> TxOut -> TxOut

txOutRemoveAssetId :: TxOut -> AssetId -> TxOut

Checks

coinIsValidForTxOut :: Coin -> Bool

Conversions (Unsafe)

unsafeCoinToTxOutCoinValue :: HasCallStack => Coin -> Word64

Converts the given Coin value to a value that can be included in a transaction output.

Callers of this function must take responsibility for checking that the given value is:

not smaller than txOutMinCoin

not greater than txOutMaxCoin

This function throws a run-time error if the pre-condition is violated.