Cardano.Wallet.Primitive.Types.Tx.Gen

Documentation

coarbitraryTxIn :: TxIn -> Gen a -> Gen a

genTx :: Gen Tx

genTxHash :: Gen (Hash "Tx")

genTxIndex :: Gen Word32

genTxIn :: Gen TxIn

genTxInFunction :: Gen a -> Gen (TxIn -> a)

genTxInLargeRange :: Gen TxIn

genTxOut :: Gen TxOut

genTxOutCoin :: Gen Coin

Generates coins across the full range allowed in a transaction output.

This generator has a slight bias towards the limits of the range, but otherwise generates values uniformly across the whole range.

This can be useful when testing roundtrip conversions between different types.

genTxOutTokenBundle :: Int -> Gen TokenBundle

genTxScriptValidity :: Gen TxScriptValidity

shrinkTx :: Tx -> [Tx]

shrinkTxHash :: Hash "Tx" -> [Hash "Tx"]

shrinkTxIndex :: Word32 -> [Word32]

shrinkTxIn :: TxIn -> [TxIn]

shrinkTxOut :: TxOut -> [TxOut]

shrinkTxOutCoin :: Coin -> [Coin]

shrinkTxScriptValidity :: TxScriptValidity -> [TxScriptValidity]