Cardano.Wallet.Primitive.Types.UTxO

Description

This module provides the main UTxO data type used by the wallet.

UTxO

newtype UTxO

Constructors

UTxO

Fields

unUTxO :: Map TxIn TxOut

dom :: UTxO -> Set TxIn

Domain of a UTxO = the set of inputs of the utxo.

null :: UTxO -> Bool

size :: UTxO -> Int

balance :: UTxO -> TokenBundle

Compute the balance of a UTxO

isSubsetOf :: UTxO -> UTxO -> Bool

a ⊆ b

empty :: UTxO

disjoint :: UTxO -> UTxO -> Bool

Indicates whether a pair of UTxO sets are disjoint.

excluding :: UTxO -> Set TxIn -> UTxO

ins⋪ u

restrictedBy :: UTxO -> Set TxIn -> UTxO

ins⊲ u

restrictedTo :: UTxO -> Set TxOut -> UTxO

u ⊳ outs

difference :: UTxO -> UTxO -> UTxO

partition :: (TxIn -> Bool) -> UTxO -> (UTxO, UTxO)

Partitions a UTxO set according to a condition.

filter p a == a && filter (not . p) b == b
  where (a,b) = partition p utxo

lookup :: TxIn -> UTxO -> Maybe TxOut

Lookup an input in the UTXO

filter :: (TxIn -> Bool) -> UTxO -> UTxO

Filters a UTxO set according to a condition.

filterByAddressM :: forall f. Monad f => (Address -> f Bool) -> UTxO -> f UTxO

Filters a UTxO set with an indicator function on Address values.

Returns the subset of UTxO entries that have addresses for which the given indicator function returns True.

filterByAddress :: (Address -> Bool) -> UTxO -> UTxO

Filters a UTxO set with an indicator function on Address values.

Returns the subset of UTxO entries that have addresses for which the given indicator function returns True.

filterByAddress f u = runIdentity $ filterByAddressM (pure . f) u filterByAddress (const True) u = u filterByAddress (const False) u = mempty filterByAddress f mempty = mempty filterByAddress f u isSubsetOf u

toList :: UTxO -> [(TxIn, TxOut)]

Converts a UTxO set into a list of UTxO elements.

UTxO delta encoding

data DeltaUTxO

Efficient delta encoding for UTxO.

excluded :: DeltaUTxO -> Set TxIn

First exclude these inputs

received :: DeltaUTxO -> UTxO

Then receive these additional outputs.

excludingD :: UTxO -> Set TxIn -> (DeltaUTxO, UTxO)

Exclude a set of transaction inputs, typically because we spend them.

receiveD :: UTxO -> UTxO -> (DeltaUTxO, UTxO)

Receive additional UTxO / union.

Queries

assetIds :: UTxO -> Set AssetId

txIds :: UTxO -> Set (Hash "Tx")

Transformations

mapAssetIds :: (AssetId -> AssetId) -> UTxO -> UTxO

mapTxIds :: (Hash "Tx" -> Hash "Tx") -> UTxO -> UTxO

Applies a mapping on transaction identifiers to a UTxO set.

If the provided mapping gives rise to a collision within the TxIn key set, then only the smallest TxOut is retained, according to the Ord instance for TxOut.

removeAssetId :: UTxO -> AssetId -> UTxO

UTxO Statistics

data UTxOStatistics

Constructors

UTxOStatistics

Fields

histogram :: ![HistogramBar]

allStakes :: !Word64

boundType :: BoundType

data BoundType

data HistogramBar

Constructors

HistogramBar

Fields

bucketUpperBound :: !Word64

bucketCount :: !Word64

computeStatistics :: (a -> [Word64]) -> BoundType -> [a] -> UTxOStatistics

A more generic function for computing UTxO statistics on some other type of data that maps to UTxO's values.

computeUtxoStatistics :: BoundType -> UTxO -> UTxOStatistics

Compute UtxoStatistics from UTxOs

log10 :: BoundType

Smart-constructor to create bounds using a log-10 scale