Cardano.Wallet.Primitive.Types.UTxOIndex

Description

Provides the UTxOIndex type, which indexes a UTxO set by asset identifier.

The index makes it possible to efficiently compute the subset of a UTxO set containing a particular asset, or to select just a single UTxO containing a particular asset, without having to search linearly through the entire UTxO set.

See the documentation for UTxOIndex for more details.

This module is meant to be imported qualified. For example:

>>> import qualified Cardano.Wallet.Primitive.Types.UTxOIndex as UTxOIndex

Type

data UTxOIndex u

A UTxO set that is indexed by asset identifier.

The index provides a mapping from assets to subsets of the UTxO set.

A UTxO appears in the set for a particular asset if and only if its associated value has a non-zero quantity of that asset.

The index makes it possible to efficiently compute the subset of a UTxO set containing a particular asset, or to select just a single UTxO containing a particular asset, without having to search linearly through the entire UTxO set.

The index also keeps track of the current UTxO balance of all assets, making it possible to efficiently look up the total quantity of a particular asset without having to sum across the entire UTxO set.

The UTxO index data structure has an invariant that can be checked with the checkInvariant function.

Construction

empty :: UTxOIndex u

An index with no entries.

singleton :: Ord u => u -> TokenBundle -> UTxOIndex u

Creates a singleton index from the specified UTxO identifier and value.

fromMap :: Ord u => Map u TokenBundle -> UTxOIndex u

Constructs an index from a map.

Note that this operation is potentially expensive as it must construct an index from scratch, and therefore should only be used sparingly.

fromSequence :: (Foldable f, Ord u) => f (u, TokenBundle) -> UTxOIndex u

Constructs an index from a sequence of entries.

Note that this operation is potentially expensive as it must construct an index from scratch, and therefore should only be used sparingly.

If the given sequence contains more than one mapping for the same UTxO identifier, the mapping that appears latest in the sequence will take precedence, and all others will be ignored.

Deconstruction

toList :: UTxOIndex u -> [(u, TokenBundle)]

Converts an index to a list of its constituent entries.

Consider using fold if your goal is to consume all entries in the output.

toMap :: UTxOIndex u -> Map u TokenBundle

Converts an index into a map.

Consider using fold if your goal is to consume all entries in the output.

Folding

fold :: (a -> u -> TokenBundle -> a) -> a -> UTxOIndex u -> a

Folds strictly over the constituent entries of an index.

Modification

insert :: Ord u => u -> TokenBundle -> UTxOIndex u -> UTxOIndex u

Inserts an entry that maps the given UTxO identifier to the given value.

If the index has an existing value for the specified UTxO identifier, the value referred to by that identifier will be replaced with the specified value.

insertMany :: (Foldable f, Ord u) => f (u, TokenBundle) -> UTxOIndex u -> UTxOIndex u

Inserts multiple entries into an index.

See insert.

delete :: forall u. Ord u => u -> UTxOIndex u -> UTxOIndex u

Deletes the entry corresponding to the given UTxO identifier.

If the index has no existing entry for the specified identifier, the result of applying this function will be equivalent to the identity function.

deleteMany :: (Foldable f, Ord u) => f u -> UTxOIndex u -> UTxOIndex u

Deletes multiple entries from an index.

See delete.

Filtering and partitioning

filter :: Ord u => (u -> Bool) -> UTxOIndex u -> UTxOIndex u

Filters an index.

partition :: Ord u => (u -> Bool) -> UTxOIndex u -> (UTxOIndex u, UTxOIndex u)

Partitions an index.

Queries

assets :: UTxOIndex u -> Set Asset

Returns the complete set of all assets contained in an index.

balance :: UTxOIndex u -> TokenBundle

lookup :: Ord u => u -> UTxOIndex u -> Maybe TokenBundle

Returns the value corresponding to the given UTxO identifier.

If the index has no such identifier, this function returns Nothing.

member :: Ord u => u -> UTxOIndex u -> Bool

Returns True if (and only if) the index has an entry for the given UTxO identifier.

null :: UTxOIndex u -> Bool

Returns True if (and only if) the index is empty.

size :: UTxOIndex u -> Int

Returns the total number of UTxO entries held within the index.

Set operations

difference :: Ord u => UTxOIndex u -> UTxOIndex u -> UTxOIndex u

disjoint :: Ord u => UTxOIndex u -> UTxOIndex u -> Bool

Indicates whether a pair of UTxO indices are disjoint.

Selection

data Asset

A type capable of representing any asset, including both ada and non-ada assets.

TODO: ADP-1449 Move this type away from the UTxOIndex module and replace all usages of it with a type parameter.

Constructors

AssetLovelace

Asset AssetId

data SelectionFilter asset

Specifies a filter for selecting UTxO entries.

Constructors

SelectSingleton asset
Matches UTxOs that contain only the given asset and no other assets.

SelectPairWith asset
Matches UTxOs that contain the given asset and exactly one other asset.

SelectAnyWith asset
Matches UTxOs that contain the given asset and any number of other assets.

SelectAny
Matches all UTxOs regardless of what assets they contain.

selectRandom :: forall m u. (MonadRandom m, Ord u) => UTxOIndex u -> SelectionFilter Asset -> m (Maybe ((u, TokenBundle), UTxOIndex u))

Selects an entry at random from the index according to the given filter.

Returns the selected entry and an updated index with the entry removed.

Returns Nothing if there were no matching entries.

selectRandomWithPriority

Arguments

:: (MonadRandom m, Ord u)

=> UTxOIndex u

-> NonEmpty (SelectionFilter Asset)
A list of selection filters to be traversed in descending order of priority, from left to right.

-> m (Maybe ((u, TokenBundle), UTxOIndex u))

Selects an entry at random from the index according to the given filters.

This function traverses the specified list of filters in descending order of priority, from left to right.

When considering a particular filter:

if the function is able to select a UTxO entry that matches, it terminates with that entry and an updated index with the entry removed.

if the function is not able to select a UTxO entry that matches, it traverses to the next filter available.

This function returns Nothing if (and only if) it traverses the entire list of filters without successfully selecting a UTxO entry.