Cardano.Wallet.StakePool.Metrics

Description

This module can fold over a blockchain to collect metrics about Stake pools.

It interacts with: - Cardano.Wallet.Network which provides the chain - Cardano.Wallet.DB - which can persist the metrics - Cardano.Wallet.Api.Server - which presents the results in an endpoint

Documentation

activityForEpoch :: Word64 -> State -> Map PoolId Int

For a given epoch, and state, this function returns how many blocks each pool produced.

data State

In-memory state keeping track of which pool produced blocks at which slots.

Constructors

State

Fields

tip :: BlockHeader

The blockHeader of the most recently applied block. Used to resume restoration from NetworkLayer.

activity :: Map PoolId [SlotId]

Mapping from pools to the slots where pool produced blocks.

This is needed internally to support rollback, but publicly, only the length of the SlotId-list is likely needed.

applyBlock :: (BlockHeader, PoolId) -> State -> State