Cardano.Wallet.TokenMetadata

Description

A client used to query asset metadata from the Cardano metadata-server.

The OpenAPI specification is here: https://github.com/input-output-hk/metadata-server/blob/master/specifications/api/openapi.yaml

An important consideration is that cardano-wallet should not trust the metadata-server operator to produce correct, valid, authentic, or even non-malicious data.

In future, signatures of property values will be checked to determine authenticity. The exact details are not yet specified.

In any case, we should not rely on the validation that the metadata-server may or may not have applied to the user-supplied metadata.

Associating metadata with assets

fillMetadata :: (Foldable t, Functor t) => TokenMetadataClient IO -> t AssetId -> (Either TokenMetadataError (Maybe AssetMetadata) -> AssetId -> a) -> IO (t a)

Helper for adding metadata to sets of assets.

Token Metadata Client

data TokenMetadataClient m

Represents a client for the metadata server.

newMetadataClient

Arguments

:: Tracer IO TokenMetadataLog
Logging

-> Maybe TokenMetadataServer
URL of metadata server, if enabled.

-> IO (TokenMetadataClient IO)

Construct a TokenMetadataClient for use with getTokenMetadata.

getTokenMetadata :: TokenMetadataClient IO -> [AssetId] -> IO (Either TokenMetadataError [(AssetId, AssetMetadata)])

Fetches metadata for a list of assets using the given client.

data TokenMetadataError

The possible errors which can occur when fetching metadata.

Constructors

TokenMetadataClientError (LoggedException SomeException)
Unhandled exception

TokenMetadataFetchError (LoggedException HttpException)
Error with HTTP request

TokenMetadataJSONParseError ByteString String
Error from aeson decoding of JSON

Logging

data TokenMetadataLog

Constructors

MsgNotConfigured

MsgFetchRequest BatchRequest BracketLog

MsgFetchRequestBody URI ByteString

MsgFetchMetadataMaxSize Int

MsgFetchResult BatchRequest (Either TokenMetadataError BatchResponse)

MsgFetchMetadataTime BatchRequest DiffTime

Generic metadata server client

metadataClient :: Tracer IO TokenMetadataLog -> TokenMetadataServer -> Manager -> BatchRequest -> IO (Either TokenMetadataError BatchResponse)

data BatchRequest

Models a request to the POST metadataquery endpoint of the metadata server -- the only one that we need.

Constructors

BatchRequest

Fields

subjects :: [Subject]

properties :: [PropertyName]

newtype BatchResponse

Models the response from the POST metadataquery endpoint of the metadata server. This should contain properties each subject in the BatchRequest.

Constructors

BatchResponse

Fields

getBatchResponse :: [SubjectProperties]

data SubjectProperties

Property values and signatures for a given subject.

Constructors

SubjectProperties

Fields

subject :: Subject

owner :: Maybe Signature

properties :: (Maybe (Property "name"), Maybe (Property "description"), Maybe (Property "ticker"), Maybe (Property "url"), Maybe (Property "logo"), Maybe (Property "decimals"))

data Property name

A property value and its signatures.

Constructors

Property

Fields

value :: Either (String, Value) (PropertyValue name)

The result of JSON parsing and validating the property value.

signatures :: [Signature]

Zero or more signatures of the property value.

sequenceNumber :: Int

Counter to prevent replaying old signatures.

newtype PropertyName

Metadata property identifier.

Constructors

PropertyName

Fields

unPropertyName :: Text

propertyName :: forall name. KnownSymbol name => Property name -> PropertyName

type family PropertyValue (name :: Symbol) :: Type

The type of a given property name.

newtype Subject

A metadata server subject, which can be any string.

Constructors

Subject

Fields

unSubject :: Text

data Signature

Will be used in future for checking integrity and authenticity of metadata.

Constructors

Signature

Fields

signature :: ByteString

publicKey :: ByteString

Parsing

metadataFromProperties :: SubjectProperties -> Maybe AssetMetadata

Convert metadata server properties response into an AssetMetadata record. Only the values are taken. Signatures are ignored (for now).

Orphan instances

FromJSON AssetDecimals

Instance details

Methods

parseJSON :: Value -> Parser AssetDecimals

parseJSONList :: Value -> Parser [AssetDecimals]

FromJSON AssetURL

Instance details

Methods

parseJSON :: Value -> Parser AssetURL

parseJSONList :: Value -> Parser [AssetURL]

FromJSON AssetLogo

Instance details

Methods

parseJSON :: Value -> Parser AssetLogo

parseJSONList :: Value -> Parser [AssetLogo]