Cardano.Wallet.TokenMetadata.MockServer

Description

A mock metadata-server for testing, metadata requests. Created using the metadata-server Haskell source code as a reference.

Documentation

withMetadataServer :: IO (Server MetadataQueryApi) -> (TokenMetadataServer -> IO a) -> IO a

Start a metadata server on any random port. See withMetadataServerOptions for running a server with middleware and predefined port.

To be used with queryServerStatic.

withMetadataServerOptions :: Middleware -> Maybe Port -> IO (Server MetadataQueryApi) -> (TokenMetadataServer -> IO a) -> IO a

Start a metadata server.

queryServerStatic :: FilePath -> IO (BatchRequest -> Handler BatchResponse)

Serve a json file.

Will filter the json and only serve metadata for the requested subjects.

queryServerReloading :: FilePath -> BatchRequest -> Handler BatchResponse

Like queryServerStatic, except that the JSON file will be reloaded before every request.

Helpers

assetIdFromSubject :: Subject -> AssetId

The reverse of subjectToAssetId

Orphan instances

ToJSON AssetDecimals

Instance details

Methods

toJSON :: AssetDecimals -> Value

toEncoding :: AssetDecimals -> Encoding

toJSONList :: [AssetDecimals] -> Value

toEncodingList :: [AssetDecimals] -> Encoding

ToJSON AssetURL

Instance details

Methods

toJSON :: AssetURL -> Value

toEncoding :: AssetURL -> Encoding

toJSONList :: [AssetURL] -> Value

toEncodingList :: [AssetURL] -> Encoding

ToJSON AssetLogo

Instance details

Methods

toJSON :: AssetLogo -> Value

toEncoding :: AssetLogo -> Encoding

toJSONList :: [AssetLogo] -> Value

toEncodingList :: [AssetLogo] -> Encoding

ToJSON Signature

Instance details

Methods

toJSON :: Signature -> Value

toEncoding :: Signature -> Encoding

toJSONList :: [Signature] -> Value

toEncodingList :: [Signature] -> Encoding

ToJSON SubjectProperties

Instance details

Methods

toJSON :: SubjectProperties -> Value

toEncoding :: SubjectProperties -> Encoding

toJSONList :: [SubjectProperties] -> Value

toEncodingList :: [SubjectProperties] -> Encoding

ToJSON BatchResponse

Instance details

Methods

toJSON :: BatchResponse -> Value

toEncoding :: BatchResponse -> Encoding

toJSONList :: [BatchResponse] -> Value

toEncodingList :: [BatchResponse] -> Encoding

FromJSON BatchRequest

Instance details

Methods

parseJSON :: Value -> Parser BatchRequest

parseJSONList :: Value -> Parser [BatchRequest]

ToJSON (PropertyValue name) => ToJSON (Property name)

Instance details

Methods

toJSON :: Property name -> Value

toEncoding :: Property name -> Encoding

toJSONList :: [Property name] -> Value

toEncodingList :: [Property name] -> Encoding