Cardano.Wallet.Transaction

Description

An extra interface for operation on transactions (e.g. creating witnesses, estimating size...). This makes it possible to decouple those operations from our wallet layer, keeping the implementation flexible to various backends.

Interface

data TransactionLayer k tx

Constructors

TransactionLayer

Fields

mkTransaction :: AnyCardanoEra -> (XPrv, Passphrase "encryption") -> (Address -> Maybe (k 'AddressK XPrv, Passphrase "encryption")) -> ProtocolParameters -> TransactionCtx -> SelectionOf TxOut -> Either ErrMkTransaction (Tx, tx)

Construct a standard transaction

" Standard " here refers to the fact that we do not deal with redemption, multisignature transactions, etc.

This expects as a first argument a mean to compute or lookup private key corresponding to a particular address.

addVkWitnesses :: AnyCardanoEra -> (XPrv, Passphrase "encryption") -> (KeyHash, XPrv, Passphrase "encryption") -> (Address -> Maybe (k 'AddressK XPrv, Passphrase "encryption")) -> (TxIn -> Maybe Address) -> tx -> tx

Add Vk witnesses to a transaction for known inputs.

If inputs can't be resolved, they are simply skipped, hence why this function cannot fail.

mkUnsignedTransaction :: AnyCardanoEra -> XPub -> ProtocolParameters -> TransactionCtx -> SelectionOf TxOut -> Either ErrMkTransaction tx

Construct a standard unsigned transaction

" Standard " here refers to the fact that we do not deal with redemption, multisignature transactions, etc.

The function returns CBOR-ed transaction body to be signed in another step.

calcMinimumCost :: AnyCardanoEra -> ProtocolParameters -> TransactionCtx -> SelectionSkeleton -> Coin

Compute a minimal fee amount necessary to pay for a given selection This also includes necessary deposits.

maxScriptExecutionCost :: ProtocolParameters -> [Redeemer] -> Coin

Compute the maximum execution cost of scripts in a given transaction.

evaluateMinimumFee :: forall era. IsShelleyBasedEra era => ProtocolParameters -> Tx era -> Coin

Evaluate a minimal fee amount necessary to pay for a given tx using ledger's functionality

Will estimate how many witnesses there should be, so it works even for unsigned transactions.

estimateSignedTxSize :: forall era. IsShelleyBasedEra era => ProtocolParameters -> Tx era -> TxSize

Estimate the size of the transaction when fully signed.

evaluateTransactionBalance :: forall era. IsShelleyBasedEra era => Tx era -> ProtocolParameters -> UTxO -> [(TxIn, TxOut, Maybe (Hash "Datum"))] -> Value

Evaluate the balance of a transaction using the ledger. The balance is defined as (value consumed by transaction) - (value produced by transaction). For a transaction to be valid, it must have a balance of zero.

Note that the fee-field of the transaction affects the balance, and is not automatically the minimum fee.

The function takes two UTxOs of different types and merges them. The reason is to workaround a combination of: 1. The wallet UTxO type doesn't support Datum hashes 2. A 'UTxO -> Cardano.UTxO' conversion function is not available in the cardano-wallet-core package, only cardano-wallet. (This package boundary will soon hopefully go away, however)

distributeSurplus :: FeePolicy -> Coin -> TxFeeAndChange [TxOut] -> Either ErrMoreSurplusNeeded (TxFeeAndChange [TxOut])

Distributes a surplus transaction balance between the given change outputs and the given fee. This function is aware of the fact that any increase in a Coin value could increase the size and fee requirement of a transaction.

When comparing the original fee and change outputs to the adjusted fee and change outputs, this function guarantees that:

The number of the change outputs remains constant;

The fee quantity either remains the same or increases.

For each change output:

the ada quantity either remains constant or increases.

non-ada quantities remain the same.

The surplus is conserved: The total increase in the fee and change ada quantities is exactly equal to the surplus.

Any increase in cost is covered: If the total cost has increased by 𝛿c, then the fee value will have increased by at least 𝛿c.

If the cost of distributing the provided surplus is greater than the surplus itself, the function will return ErrMoreSurplusNeeded. If the provided surplus is greater or equal to maximumCostOfIncreasingCoin feePolicy, the function will always return Right.

computeSelectionLimit :: AnyCardanoEra -> ProtocolParameters -> TransactionCtx -> [TxOut] -> SelectionLimit

tokenBundleSizeAssessor :: TokenBundleMaxSize -> TokenBundleSizeAssessor

A function to assess the size of a token bundle.

constraints :: AnyCardanoEra -> ProtocolParameters -> TxConstraints

decodeTx :: AnyCardanoEra -> tx -> (Tx, TokenMapWithScripts, TokenMapWithScripts, [Certificate], Maybe ValidityIntervalExplicit)

Decode an externally-created transaction.

updateTx :: forall era. IsShelleyBasedEra era => Tx era -> TxUpdate -> Either ErrUpdateSealedTx (Tx era)

Update tx by adding additional inputs and outputs

assignScriptRedeemers :: forall era. IsShelleyBasedEra era => ProtocolParameters -> TimeInterpreter (Either PastHorizonException) -> (TxIn -> Maybe (TxOut, Maybe (Hash "Datum"))) -> [Redeemer] -> Tx era -> Either ErrAssignRedeemers (Tx era)

data DelegationAction

Whether the user is attempting any particular delegation action.

Constructors

RegisterKeyAndJoin PoolId

Join PoolId

Quit

data TransactionCtx

Some additional context about a transaction. This typically contains details that are known upfront about the transaction and are used to construct it from inputs selected from the wallet's UTxO.

Constructors

TransactionCtx

Fields

txWithdrawal :: Withdrawal

Withdrawal amount from a reward account, can be zero.

txMetadata :: Maybe TxMetadata

User or application-defined metadata to embed in the transaction.

txValidityInterval :: (Maybe SlotNo, SlotNo)

Transaction optional starting slot and expiry (TTL) slot for which the transaction is valid.

txDelegationAction :: Maybe DelegationAction

An additional delegation to take.

txPlutusScriptExecutionCost :: Coin

Total execution cost of plutus scripts, determined by their execution units and prices obtained from network.

txAssetsToMint :: (TokenMap, Map AssetId (Script KeyHash))

The assets to mint.

txAssetsToBurn :: (TokenMap, Map AssetId (Script KeyHash))

The assets to burn.

txCollateralRequirement :: SelectionCollateralRequirement

The collateral requirement.

txFeePadding :: !Coin

Extra fees. Some parts of a transaction are not representable using cardano-wallet types, which makes it useful to account for them like this. For instance: datums.

defaultTransactionCtx :: TransactionCtx

A default context with sensible placeholder. Can be used to reduce repetition for changing only sub-part of the default context.

data Withdrawal

Constructors

WithdrawalSelf RewardAccount (NonEmpty DerivationIndex) Coin

WithdrawalExternal RewardAccount (NonEmpty DerivationIndex) Coin

NoWithdrawal

withdrawalToCoin :: Withdrawal -> Coin

data TxUpdate

Describes modifications that can be made to a Tx using $sel:updateTx:TransactionLayer.

Constructors

TxUpdate

Fields

extraInputs :: [(TxIn, TxOut)]

extraCollateral :: [TxIn]

Only used in the Alonzo era and later. Will be silently ignored in previous eras.

extraOutputs :: [TxOut]

feeUpdate :: TxFeeUpdate

Set a new fee or use the old one.

data TxFeeUpdate

Method to use when updating the fee of a transaction.

Constructors

UseOldTxFee
Instead of updating the fee, just use the old fee of the Tx (no-op for fee update).

UseNewTxFee Coin
Specify a new fee to use instead.

data TokenMapWithScripts

Constructors

TokenMapWithScripts

Fields

txTokenMap :: !TokenMap

txScripts :: !(Map TokenPolicyId AnyScript)

emptyTokenMapWithScripts :: TokenMapWithScripts

data AnyScript

Constructors

NativeScript !(Script KeyHash)

PlutusScript !PlutusScriptInfo

newtype PlutusScriptInfo

Constructors

PlutusScriptInfo

Fields

languageVersion :: PlutusVersion

data PlutusVersion

Constructors

PlutusVersionV1

PlutusVersionV2

data TxFeeAndChange change

Small helper record to disambiguate between a fee and change Coin values. Used by $sel:distributeSurplus:TransactionLayer.

Constructors

TxFeeAndChange

Fields

fee :: Coin

change :: change

mapTxFeeAndChange

Arguments

:: (Coin -> Coin)
A function to transform the fee

-> (change1 -> change2)
A function to transform the change

-> TxFeeAndChange change1
The original fee and change

-> TxFeeAndChange change2
The transformed fee and change

Manipulates a TxFeeAndChange value.

data ValidityIntervalExplicit

Constructors

ValidityIntervalExplicit

Fields

invalidBefore :: !(Quantity "slot" Word64)

invalidHereafter :: !(Quantity "slot" Word64)

Errors

data ErrSignTx

Possible signing error

Constructors

ErrSignTxAddressUnknown TxIn
We tried to sign a transaction with inputs that are unknown to us?

ErrSignTxUnimplemented
TODO: [ADP-919] Remove ErrSignTxUnimplemented

data ErrMkTransaction

Constructors

ErrMkTransactionNoSuchWallet WalletId

ErrMkTransactionTxBodyError Text
We failed to construct a transaction for some reasons.

ErrMkTransactionInvalidEra AnyCardanoEra
Should never happen, means that that we have programmatically provided an invalid era.

ErrMkTransactionJoinStakePool ErrCannotJoin

ErrMkTransactionQuitStakePool ErrCannotQuit

ErrMkTransactionIncorrectTTL PastHorizonException

data ErrCannotJoin

Constructors

ErrAlreadyDelegating PoolId

ErrNoSuchPool PoolId

data ErrCannotQuit

Constructors

ErrNotDelegatingOrAboutTo

ErrNonNullRewards Coin

newtype ErrUpdateSealedTx

Constructors

ErrExistingKeyWitnesses Int
The SealedTx couldn't not be updated because the *n* existing key-witnesses would have been rendered invalid.

data ErrAssignRedeemers

Constructors

ErrAssignRedeemersScriptFailure Redeemer String
Failed to assign execution units for a particular redeemer. The String indicates the reason of the failure.

TODO: Refine this type to avoid the String and provides a better sum-type of possible errors.

ErrAssignRedeemersTargetNotFound Redeemer
The given redeemer target couldn't be located in the transaction.

ErrAssignRedeemersInvalidData Redeemer String
Redeemer's data isn't a valid Plutus' data.

ErrAssignRedeemersTranslationError (TranslationError StandardCrypto)
Mistranslating of hashes, credentials, certificates etc.

newtype ErrMoreSurplusNeeded

Error for when its impossible for $sel:distributeSurplus:TransactionLayer to distribute the surplus. As long as the surplus is larger than costOfIncreasingCoin, this should never happen.

Constructors

ErrMoreSurplusNeeded Coin