Cardano.Wallet.Util

Description

General utility functions.

Partial functions for "impossible" situations

type HasCallStack = ?callStack :: CallStack

Request a CallStack.

NOTE: The implicit parameter ?callStack :: CallStack is an implementation detail and should not be considered part of the CallStack API, we may decide to change the implementation in the future.

Since: base-4.9.0.0

internalError :: HasCallStack => Builder -> a

Calls the error function, which will usually crash the program.

tina :: HasCallStack => Builder -> [Maybe a] -> a

Take the first Just from a list of Maybe, or die trying. There is no alternative.

invariant

Arguments

:: HasCallStack

=> String
The message

-> a
The value to test

-> (a -> Bool)
The predicate

-> a

Checks whether or not an invariant holds, by applying the given predicate to the given value.

If the invariant does not hold (indicated by the predicate function returning False), throws an error with the specified message.

>>> invariant "not empty" [1,2,3] (not . null)
[1, 2, 3]

>>> invariant "not empty" [] (not . null)
*** Exception: not empty

Handling errors for "impossible" situations.

isInternalError :: ErrorCall -> Maybe String

Tests whether an Exception was caused by internalError.

tryInternalError :: MonadUnliftIO m => a -> m (Either String a)

Evaluates a pure expression to WHNF and handles any occurrence of internalError.

This is intended for use in testing. Don't use this in application code -- that's what normal IO exceptions are for.

String formatting

newtype ShowFmt a

A polymorphic wrapper type with a custom show instance to display data through Buildable instances.

Constructors

ShowFmt

Fields

unShowFmt :: a

mapFirst :: (a -> a) -> [a] -> [a]

Map a function to the first element of a list. Does nothing if the list is empty.

StateT

modifyM :: forall m s. Monad m => (s -> m s) -> StateT s m ()

Effectfully modify the state of a state-monad transformer stack.

HTTP(S) URIs

uriToText :: URI -> Text

parseURI :: Text -> Either TextDecodingError URI