Cardano.Wallet.Version

Description

Provides the package version and git revision which this was compiled from.

It is assumed that all cardano-wallet packages have the same version, that of the core package.

Stack builds will have the git command available to run during compilation.

Nix builds will inject the git revision into the executables after compiling. If the git revision has changed but the sources have not, then no haskell packages will be rebuilt, but the embedded git revision will be updated.

Values computed at compile-time

version :: Version

gitRevision :: GitRevision

The Git revision ID (40 character hex string) of this build.

This requires git to be available when building. Alternatively, the git revision of the cardano-wallet binary can be updated post-build using "Data.FileEmbed.injectWith".

data GitRevision

data Version

A Version represents the version of a software entity.

An instance of Eq is provided, which implements exact equality modulo reordering of the tags in the versionTags field.

An instance of Ord is also provided, which gives lexicographic ordering on the versionBranch fields (i.e. 2.1 > 2.0, 1.2.3 > 1.2.2, etc.). This is expected to be sufficient for many uses, but note that you may need to use a more specific ordering for your versioning scheme. For example, some versioning schemes may include pre-releases which have tags "pre1", "pre2", and so on, and these would need to be taken into account when determining ordering. In some cases, date ordering may be more appropriate, so the application would have to look for date tags in the versionTags field and compare those. The bottom line is, don't always assume that compare and other Ord operations are the right thing for every Version.

Similarly, concrete representations of versions may differ. One possible concrete representation is provided (see showVersion and parseVersion), but depending on the application a different concrete representation may be more appropriate.

Displaying Versions

showVersionAsDate :: Version -> String

Format the Cabal version in the vYYYY-MM-DD style that we use for git tags.

showFullVersion :: Version -> GitRevision -> String

Like showVersionAsDate, but also show the git revision.