Control.Monad.Random.Extra

Description

This module provides functions and types that extend those provided by the Random module hierarchy.

Random number generator seeds

newtype StdGenSeed

A seed for the standard random number generator.

This type is equivalent to the internal state of a StdGen, but provides a representation that is more convenient for construction and serialization.

The number of possible seeds is identical to the number of valid states of the StdGen type, but unlike the StdGen type, whose state has an internal invariant that must not be broken, values of the StdGenSeed type are correct by construction.

Constructors

StdGenSeed

Fields

unStdGenSeed :: Word127

stdGenSeed :: MonadRandom m => m StdGenSeed

Creates a new StdGenSeed from within a random monadic context.

stdGenFromSeed :: StdGenSeed -> StdGen

Converts a StdGenSeed value to a StdGen value.

This function satisfies the following properties:

>>> stdGenFromSeed . stdGenToSeed == id
>>> stdGenToSeed . stdGenFromSeed == id

stdGenToSeed :: StdGen -> StdGenSeed

Converts a StdGen value to a StdGenSeed value.

This function satisfies the following properties:

>>> stdGenFromSeed . stdGenToSeed == id
>>> stdGenToSeed . stdGenFromSeed == id

Non-random contexts

newtype NonRandom a

Provides a stateless context for computations that must be non-random.

This type is useful for testing functions that require a MonadRandom context, but when actual randomness is not required or even desired.

Constructors

NonRandom

Fields

runNonRandom :: a