Data.Quantity

Description

Representation of values with an associated (free) unit of measure. Useful to disambiguate primitive types like Int or String which can be in different bases depending on the context.

Polymorphic Quantity

newtype Quantity (unit :: Symbol) a

Quantity (unit :: Symbol) a is a primitive a multiplied by an unit.

Example:

Instead of providing the unit implicitly as a comment, or a part of a name

>>> a :: Word32 -- in lovelace

we can write

>>> a :: Quantity "lovelace" Word32

which now has a different type from

>>> b :: Quantity "lovelace/byte" Word32

so mixing them up is more difficult.

The unit is mostly a phantom type, but it is also included in the ToJSON/FromJSON instances.

>>> Aeson.encode $ Quantity @"lovelace" 14
{"unit":"lovelace","quantity":14}

Constructors

Quantity

Fields

getQuantity :: a

Percentage

data Percentage

Opaque Haskell type to represent values between 0 and 100 (incl).

data MkPercentageError

Constructors

PercentageOutOfBoundsError

mkPercentage :: Rational -> Either MkPercentageError Percentage

Safe constructor for Percentage

Takes an input in the range [0, 1].

getPercentage :: Percentage -> Rational

clipToPercentage :: Rational -> Percentage

Safe way to make a Percentage by clipping values that are out of bounds.

complementPercentage :: Percentage -> Percentage

The complement is the amount that is missing to make it 100%.

Example: The complementPercentage of 0.7 is 0.3.

percentageToDouble :: Percentage -> Double

Turn a Percentage to a Double (without any extra rounding.)