Data.Time.Text

Description

Utility functions for converting time values to and from text.

Conversion to and from text

utcTimeToText :: TimeFormat -> UTCTime -> Text

Convert the specified time value to text, using the specified time format.

utcTimeFromText :: [TimeFormat] -> Text -> Maybe UTCTime

Attempt to use each of the specified time formats to parse the given text. Returns a time value that corresponds to the first matching format, or Nothing if none of the formats matched.

Time format specification

data TimeFormat

Represents a particular way of representing a moment in time in text.

Constructors

TimeFormat

Fields

timeFormatName :: String

timeFormatPattern :: String

Time formats

iso8601BasicUtc :: TimeFormat

Represents the ISO 8601 basic format (UTC).

iso8601BasicLocal :: TimeFormat

Represents the ISO 8601 basic format (with local timezone).

iso8601ExtendedUtc :: TimeFormat

Represents the ISO 8601 extended format (UTC).

iso8601ExtendedLocal :: TimeFormat

Represents the ISO 8601 extended format (with local timezone).

Time format families

iso8601 :: [TimeFormat]

Represents the ISO 8601 family of formats.

iso8601Basic :: [TimeFormat]

Represents the ISO 8601 basic family of formats.

iso8601Extended :: [TimeFormat]

Represents the ISO 8601 extended family of formats.