Data.Vector.Shuffle

Simple

shuffle :: [a] -> IO [a]

Shuffles a list of elements.

>>> shuffle (outputs coinSel)
[...]

Advanced

mkSeed :: Text -> Int

Generate a random generator seed from a text string

shuffleWith :: RandomGen g => g -> [a] -> IO [a]

Like shuffle, but from a given seed. shuffle will use a randomly generate seed using newStdGen from System.Random.

Properties:

shuffleWith g es == shuffleWith g es

∃Δ> 1. g ≠g', length es > Δ⇒ shuffleWith g es ≠shuffleWith g' es