Network.Ntp

Description

This module provides the Ntp client related settings, types and re-exports used in a number of places throughout codebase.

Documentation

withWalletNtpClient

Arguments

:: IOManager
The global IOManager instance, set up by the application main function.

-> Tracer IO NtpTrace
Logging object

-> (NtpClient -> IO a)
Action to run

-> IO a

Set up a NtpClient and pass it to the given action. The NtpClient is terminated when the callback returns.

getNtpStatus

Arguments

:: NtpClient

-> Bool
When True, will block and force a NTP check instead of using cached results

-> IO ApiNetworkClock

re-exports from ntp-client

data NtpTrace

Constructors

NtpTraceStartNtpClient

NtpTraceRestartDelay Int

NtpTraceRestartingClient

NtpTraceIOError IOError

NtpTraceLookupsFails

NtpTraceClientStartQuery

NtpTraceNoLocalAddr

NtpTraceResult NtpStatus

NtpTraceRunProtocolResults (ResultOrFailure [NtpOffset])

NtpTracePacketSent SockAddr NtpPacket

NtpTracePacketSendError SockAddr IOException

NtpTracePacketDecodeError SockAddr String

NtpTracePacketReceived SockAddr NtpPacket

NtpTraceWaitingForRepliesTimeout IPVersion

data NtpClient

NtpClient which recieves updates of the wall clcok drift every ntpPollDelay. It also allows to force engaging in ntp protocol.

Constructors

NtpClient

Fields

ntpGetStatus :: STM NtpStatus

Query the current NTP status.

ntpQueryBlocking :: IO NtpStatus

Force to update the ntp state, unless an ntp query is already running. This is a blocking operation.

ntpThread :: Async Void

Ntp client thread

Orphan instances

ToText IPVersion

Instance details

Methods

toText :: IPVersion -> Text

ToText NtpTrace

Instance details

Methods

toText :: NtpTrace -> Text

HasPrivacyAnnotation NtpTrace

Instance details

Methods

getPrivacyAnnotation :: NtpTrace -> PrivacyAnnotation

HasSeverityAnnotation NtpTrace

Instance details

Methods

getSeverityAnnotation :: NtpTrace -> Severity