Network.Wai.Middleware.Logging

Middleware

withApiLogger :: Tracer IO ApiLog -> ApiLoggerSettings -> Middleware

Installs a request & response logger on a Wai application.

The logger logs requests' and responses' bodies along with a few other useful piece of information.

Settings

newApiLoggerSettings :: IO ApiLoggerSettings

Create a new opaque ApiLoggerSettings

data ApiLoggerSettings

API logger settings

obfuscateKeys :: (Request -> [Text]) -> ApiLoggerSettings -> ApiLoggerSettings

Define a set of top-level object keys that should be obfuscated for a given request in a JSON format.

data HandlerLog

Tracer events related to the handling of a single request.

Constructors

LogRequestStart

LogRequest Request

LogRequestBody [Text] ByteString
Request content, with list of sensitive json keys.

LogResponse NominalDiffTime Request (Maybe Status)

LogResponseBody ByteString

LogRequestFinish

data ApiLog

API handler trace events are associated with a unique request ID.

Constructors

ApiLog

Fields

requestId :: RequestId

Unique integer associated with the request, for the purpose of tracing.

logMsg :: HandlerLog

Event trace for the handler.

newtype RequestId

Just a wrapper for readability

Constructors

RequestId Integer