UnliftIO.Compat

Description

Extra glue functions to work with MonadUnliftIO exception types. We need this because the retry package uses the generalized exception handler type from Handler. But the Exceptions module has its own definition of exactly the same type.

Handler conversion

coerceHandler :: Handler IO b -> Handler IO b

Convert the generalized handler from Exception type to Catch type

coerceHandlers :: [a -> Handler IO b] -> [a -> Handler IO b]

Convert a list of handler factories from the Exception type to Catch type. Such handlers are used in Recovering for example.

mkRetryHandler :: Exception e => (e -> m Bool) -> [a -> Handler m Bool]

Shortcut for creating a single Retry handler, which doesn't use the RetryStatus info.

Missing combinators

handleIf :: (MonadUnliftIO m, Exception e) => (e -> Bool) -> (e -> m a) -> m a -> m a

A MonadUnliftIO version of handleIf.

Re-export unsafe things

data AsyncCancelled

The exception thrown by cancel to terminate a thread.

Constructors

AsyncCancelled