Cardano.DB.Sqlite

Description

An implementation of the DBLayer which uses Persistent and SQLite.

Documentation

data SqliteContext

Context for the SQLite DBLayer.

Constructors

SqliteContext

Fields

getSqlBackend :: SqlBackend

A handle to the Persistent SQL backend.

runQuery :: forall a. SqlPersistT IO a -> IO a

safely run a query with logging and lock-protection

dbFile :: Maybe FilePath

The actual database file, if any. If none, runs in-memory

trace :: Tracer IO DBLog

A Tracer for logging

chunkSize :: Int

Size of chunks when inserting, updating or deleting many rows at once. We only act on chunkSize values at a time. See also dbChunked.

dbChunked :: ([a] -> SqlPersistT IO b) -> [a] -> SqlPersistT IO ()

Convert a single DB "updateMany" (or similar) query into multiple updateMany queries with smaller lists of values.

This is to prevent too many variables appearing in the SQL statement. SQLITE_MAX_VARIABLE_NUMBER is 999 by default, and we will get a "too many SQL variables" exception if that is exceeded.

We choose a conservative value chunkSize << 999 because there can be multiple variables per row updated.

destroyDBLayer :: SqliteContext -> IO ()

Finalize database statements and close the database connection.

If the database connection is still in use, it will retry for up to a minute, to let other threads finish up.

This function is idempotent: if the database connection has already been closed, calling this function will exit without doing anything.

handleConstraint :: MonadCatch m => e -> m a -> m (Either e a)

Run an action, and convert any Sqlite constraints exception into the given error result. No other exceptions are handled.

startSqliteBackend :: ManualMigration -> Migration -> Tracer IO DBLog -> Maybe FilePath -> IO (Either MigrationError SqliteContext)

Opens the SQLite database connection, sets up query logging and timing, runs schema migrations if necessary.

unsafeRunQuery :: SqliteContext -> SqlPersistT IO a -> IO a

Run a raw query from the outside using an instantiate DB layer. This is completely unsafe because it breaks the abstraction boundary and can have disastrous results on the database consistency.

Manual Migration

newtype ManualMigration

Encapsulates a manual migration action (or sequence of actions) to be performed immediately after an SQL connection is initiated.

Constructors

ManualMigration

Fields

executeManualMigration :: Connection -> IO ()

newtype MigrationError

Error type for when migrations go wrong after opening a database.

Constructors

MigrationError

Fields

getMigrationErrorMessage :: Text

data DBField where

Constructors

DBField :: forall record typ. PersistEntity record => EntityField record typ -> DBField

tableName :: DBField -> Text

fieldName :: DBField -> Text

fieldType :: DBField -> Text

Logging

data DBLog

Constructors

MsgMigrations (Either MigrationError Int)

MsgQuery Text Severity

MsgRun Bool

MsgConnStr Text

MsgClosing (Maybe FilePath)

MsgDatabaseReset

MsgIsAlreadyClosed Text

MsgStatementAlreadyFinalized Text

MsgRemoving Text

MsgManualMigrationNeeded DBField Text

MsgManualMigrationNotNeeded DBField

MsgUpdatingForeignKeysSetting ForeignKeysSetting