Cardano.Pool.DB.Model

Description

An implementation of the production pool database using only pure functions.

These functions and types model the behaviour of the SQLite database backend, and are used for QuickCheck state machine testing, and the MVar database backend.

Model Types

data PoolDatabase

Constructors

PoolDatabase

Fields

pools :: !(Map PoolId [BlockHeader])

Information of what blocks were produced by which stake pools

distributions :: !(Map EpochNo [(PoolId, Quantity "lovelace" Word64)])

Store known stake distributions for epochs

emptyPoolDatabase :: PoolDatabase

Produces an empty model pool production database.

Model Operation Types

type ModelPoolOp a = PoolDatabase -> (Either PoolErr a, PoolDatabase)

newtype PoolErr

Constructors

PointAlreadyExists BlockHeader

Model pool database functions

mCleanPoolProduction :: ModelPoolOp ()

mPutPoolProduction :: BlockHeader -> PoolId -> ModelPoolOp ()

mReadPoolProduction :: EpochNo -> ModelPoolOp (Map PoolId [BlockHeader])

mPutStakeDistribution :: EpochNo -> [(PoolId, Quantity "lovelace" Word64)] -> ModelPoolOp ()

mReadStakeDistribution :: EpochNo -> ModelPoolOp [(PoolId, Quantity "lovelace" Word64)]

mRollbackTo :: SlotId -> ModelPoolOp ()