Cardano.Pool.DB.Sqlite

Description

An implementation of the DBLayer which uses Persistent and SQLite.

Documentation

newDBLayer

Arguments

:: Configuration
Logging configuration

-> Trace IO Text
Logging object

-> Maybe FilePath
Database file location, or Nothing for in-memory database

-> IO (SqliteContext, DBLayer IO)

Sets up a connection to the SQLite database.

Database migrations are run to create tables if necessary.

If the given file path does not exist, it will be created by the sqlite library.

getDBLayer will provide the actual DBLayer implementation. The database should be closed with destroyDBLayer. If you use withDBLayer then both of these things will be handled for you.

withDBLayer

Arguments

:: Configuration
Logging configuration

-> Trace IO Text
Logging object

-> Maybe FilePath
Database file location, or Nothing for in-memory database

-> (DBLayer IO -> IO a)
Action to run.

-> IO a

Runs an action with a connection to the SQLite database.

Database migrations are run to create tables if necessary.

If the given file path does not exist, it will be created by the sqlite library.