Cardano.Pool.DB.Sqlite.TH

Description

Auto-generated Sqlite & Persistent machinery via Template-Haskell. This has been moved into a separate file so that we can treat it slightly differently when computing code-coverage.

Documentation

data PoolProduction

Constructors

PoolProduction

Fields

poolProductionPoolId :: !PoolId

poolProductionSlot :: !SlotId

poolProductionHeaderHash :: !BlockId

poolProductionParentHash :: !BlockId

poolProductionBlockHeight :: !Word32

type PoolProductionId = Key PoolProduction

data StakeDistribution

Constructors

StakeDistribution

Fields

stakeDistributionPoolId :: !PoolId

stakeDistributionEpoch :: !Word64

stakeDistributionStake :: !Word64

type StakeDistributionId = Key StakeDistribution

migrateAll :: Migration