Cardano.Pool.DB.Sqlite.Types

Description

This module contains instances and types necessary for storing wallets in a SQL database with Persistent.

It's in a separate module due to the GHC stage restriction.

The ToJSON/FromJSON and Read instance orphans exist due to class constraints on Persistent functions.

Documentation

sqlSettings' :: MkPersistSettings

Settings for generating the Persistent types.

fromText' :: FromText a => Text -> Either Text a

fromText but with a simpler error type.

fromPersistValueFromText :: FromText a => PersistValue -> Either Text a

fromPersistValue defined in terms of fromText

Orphan instances

Read PoolId

Instance details

Methods

readsPrec :: Int -> ReadS PoolId

readList :: ReadS [PoolId]

readPrec :: ReadPrec PoolId

readListPrec :: ReadPrec [PoolId]

PersistField PoolId

Instance details

Methods

toPersistValue :: PoolId -> PersistValue

fromPersistValue :: PersistValue -> Either Text PoolId

PersistFieldSql PoolId

Instance details

Methods

sqlType :: Proxy PoolId -> SqlType

PathPiece PoolId

Instance details

Methods

fromPathPiece :: Text -> Maybe PoolId

toPathPiece :: PoolId -> Text