Cardano.Pool.Metadata

Description

This module contains a function for downloading stake pool metadata from an external registry.

Types

data StakePoolMetadata

Information about a stake pool. This information is not used directly by cardano-wallet. It is sourced from the stake pool registry and passed straight through to API consumers.

Constructors

StakePoolMetadata

Fields

ticker :: StakePoolTicker

Short human-readable ID for the stake pool.

homepage :: Text

Absolute URL for the stake pool's homepage link.

pledgeAddress :: Text

Bech32-encoded address.

data StakePoolTicker

Very short name for a stake pool.

Fetching metadata

getStakePoolMetadata

Arguments

:: Trace IO RegistryLog
Logging object - use transformTrace to convert to Text.

-> String
URL of a zip archive to fetch and extract.

-> [PoolId]
List of stake pools to get metadata for.

-> IO (Either FetchError [Maybe StakePoolMetadata])

Associate a list of stake pool IDs with their metadata (if present), which is downloaded from the given URL.

The URL must point to a zip archive, with the following stucture:

master.zip
└── testnet-stake-pool-registry-master/
    └── registry/
        ├── pk1afhcpw2tg7nr2m3wr4x8jaa4dv7d09gnv27kwfxpjyvukwxs8qdqwg85xp.json
        └── pk1z4vh8gva25w07x8574uujuveu8gz43fu6qfln3t4prcavrvcphjsk0pdqs.json

The name of the top-level directory can be anything.

Other files in the archive are ignored by getStakePoolMetadata.

The JSON file names are PoolIds.

The required JSON structure is given by the Data.Aeson.FromJSON instance of StakePoolMetadata.

The URL which should be used for incentived testnet is cardanoFoundationRegistryZip.

The returned list will have the same length as the given list of PoolId. If an metadata entry does not exist or could not be parsed, it will be Nothing.

cardanoFoundationRegistryZip :: String

The stake pool registry zipfile download URL for CF.

data FetchError

Things that could go wrong with fetching metadata, which aren't programming errors or expected failures.

Constructors

FetchErrorDownload String
Failed to download the registry archive.

FetchErrorFile String
Failed to store the registry archive in a temporary location.

FetchErrorZipParsingFailed FilePath String
Failed to open the registry archive with the given path, as a zip file, due to the given cause.

Logging

transformTrace :: Trace IO Text -> Trace IO RegistryLog

Converts a text tracer into a stake pool registry tracer.

data RegistryLog

The log messages for getStakePoolMetadata.

Constructors

RegistryLog

Fields

registryLogUrl :: String

registryLogZipFile :: FilePath

registryLogMsg :: RegistryLogMsg

data RegistryLogMsg

Log messages about processing a specific archive.

Constructors

MsgDownloadStarted

MsgDownloadComplete Integer

MsgDownloadError FetchError

MsgExtractFile FilePath

MsgExtractFileResult (Maybe (Either String StakePoolMetadata))