Cardano.Pool.Metrics

Description

This module can fold over a blockchain to collect metrics about Stake pools.

It interacts with: - Cardano.Wallet.Network which provides the chain - Cardano.Pool.DB - which can persist the metrics - Cardano.Wallet.Api.Server - which presents the results in an endpoint

Types

data Block

Constructors

Block

Fields

header :: BlockHeader

producer :: PoolId

Listing stake-pools from the DB

newtype StakePoolLayer m

StakePoolLayer is a thin layer ontop of the DB. It is one value that can easily be passed to the API-server, where it can be used in a simple way.

Constructors

StakePoolLayer

Fields

listStakePools :: ExceptT ErrListStakePools m [(PoolId, (Quantity "lovelace" Word64, Quantity "block" Natural))]

newStakePoolLayer :: DBLayer IO -> NetworkLayer IO t block -> Trace IO Text -> IO (StakePoolLayer IO)

data ErrListStakePools

Constructors

ErrMetricsIsUnsynced (Quantity "percent" Percentage)

ErrListStakePoolsMetricsInconsistency ErrMetricsInconsistency

ErrListStakePoolsErrNetworkTip ErrNetworkTip

Following the chain

monitorStakePools :: Trace IO Text -> NetworkLayer IO t Block -> DBLayer IO -> IO ()

monitorStakePools follows the chain and puts pool productions and stake distributions to a DBLayer, such that the data in the database is always consistent.

The pool productions and stake distrubtions in the db can never be from different forks such that it's safe for readers to access it.

FIXME: This last statement is only true if putStakeProduction and putPoolProduction are running in the same db transaction, which would be the case if we make SqlPersistT the top-level monad for the DBLayer instead of IO.

Combining Metrics

data ErrMetricsInconsistency

Possible errors returned by combineMetrics.

Constructors

ErrProducerNotInDistribution PoolId (Quantity "block" Natural)
Somehow, we tried to combine invalid metrics together and passed a passed a block production that doesn't match the producers found in the stake activity.

Note that the opposite case is okay as we only observe pools that have produced blocks. So it could be the case that a pool exists in the distribution but not in the production! (In which case, we'll assign it a production of '0').

combineMetrics :: Map PoolId (Quantity "lovelace" Word64) -> Map PoolId (Quantity "block" Natural) -> Either ErrMetricsInconsistency (Map PoolId (Quantity "lovelace" Word64, Quantity "block" Natural))

Combines two different sources of data into one:

A stake-distribution map

A pool-production map

If a pool has produced a block without existing in the stake-distribution, i.e it exists in (2) but not (1), this function will return Left ErrMetricsInconsistencyBlockProducerNotInStakeDistr.

If a pool is in (1) but not (2), it simply means it has produced 0 blocks so far.