Cardano.Wallet

Description

Provides wallet layer functions that are used by API layer. Uses both Cardano.Wallet.DB and Cardano.Wallet.Network to realize its role as being intermediary between the three.

Interface

data WalletLayer s t k

Constructors

WalletLayer

Fields

createWallet :: (Show s, NFData s, IsOurs s, DefineTx t) => WalletId -> WalletName -> s -> ExceptT ErrWalletAlreadyExists IO WalletId

Initialise and store a new wallet, returning its ID.

readWallet :: WalletId -> ExceptT ErrNoSuchWallet IO (Wallet s t, WalletMetadata)

Retrieve the wallet state for the wallet with the given ID.

updateWallet :: WalletId -> (WalletMetadata -> WalletMetadata) -> ExceptT ErrNoSuchWallet IO ()

Update a wallet's metadata with the given update function.

updateWalletPassphrase :: WalletKey k => WalletId -> (Passphrase "encryption-old", Passphrase "encryption-new") -> ExceptT ErrUpdatePassphrase IO ()

Change a wallet's passphrase to the given passphrase.

listWallets :: IO [WalletId]

Retrieve a list of known wallet IDs.

removeWallet :: WalletId -> ExceptT ErrNoSuchWallet IO ()

Remove an existing wallet. Note that there's no particular work to be done regarding the restoration worker as it will simply terminate on the next tick when noticing that the corresponding wallet is gone.

restoreWallet :: DefineTx t => WalletId -> ExceptT ErrNoSuchWallet IO ()

Restore a wallet from its current tip up to the network tip.

This function returns immediately, starting a worker thread in the background that will fetch and apply remaining blocks until the network tip is reached or until failure.

listAddresses :: (IsOurs s, CompareDiscovery s, KnownAddresses s, DefineTx t) => WalletId -> ExceptT ErrNoSuchWallet IO [(Address, AddressState)]

List all addresses of a wallet with their metadata. Addresses are ordered from the most-recently-discovered to the oldest known.

createUnsignedTx :: forall e. (DefineTx t, e ~ ErrValidateSelection t) => WalletId -> NonEmpty TxOut -> ExceptT (ErrCreateUnsignedTx e) IO CoinSelection

Prepare a transaction and automatically select inputs from the wallet to cover the requested outputs. Note that this only runs coin selection for the given outputs. In order to construct (and sign) an actual transaction, use $sel:signTx:WalletLayer.

estimateTxFee :: forall e. (DefineTx t, e ~ ErrValidateSelection t) => WalletId -> NonEmpty TxOut -> ExceptT (ErrEstimateTxFee e) IO Fee

Estimate a transaction fee by automatically selecting inputs from the wallet to cover the requested outputs.

listUtxoStatistics :: DefineTx t => WalletId -> ExceptT ErrListUTxOStatistics IO UTxOStatistics

List the wallet's UTxO statistics.

signTx :: (Show s, NFData s, IsOwned s k, GenChange s) => WalletId -> Passphrase "encryption" -> CoinSelection -> ExceptT ErrSignTx IO (Tx t, TxMeta, [TxWitness])

Produce witnesses and construct a transaction from a given selection. Requires the encryption passphrase in order to decrypt the root private key. Note that this doesn't broadcast the transaction to the network. In order to do so, use $sel:submitTx:WalletLayer.

submitTx :: DefineTx t => WalletId -> (Tx t, TxMeta, [TxWitness]) -> ExceptT ErrSubmitTx IO ()

Broadcast a (signed) transaction to the network.

submitExternalTx :: ByteString -> ExceptT ErrSubmitExternalTx IO (Tx t)

Broadcast an externally-signed transaction to the network.

listTransactions :: DefineTx t => WalletId -> Maybe UTCTime -> Maybe UTCTime -> SortOrder -> ExceptT ErrListTransactions IO [TransactionInfo]

List all transactions and metadata from history for a given wallet.

attachPrivateKey :: WalletKey k => WalletId -> (k RootK XPrv, Passphrase "encryption") -> ExceptT ErrNoSuchWallet IO ()

Attach a given private key to a wallet. The private key is necessary for some operations like signing transactions, or generating new accounts.

data BlockchainParameters t

Constructors

BlockchainParameters

Fields

getGenesisBlock :: Block (Tx t)

Very first block.

getGenesisBlockDate :: StartTime

Start time of the chain.

getFeePolicy :: FeePolicy

Policy regarding transaction fee.

getSlotLength :: SlotLength

Length, in seconds, of a slot.

getEpochLength :: EpochLength

Number of slots in a single epoch.

getTxMaxSize :: Quantity "byte" Word16

Maximum size of a transaction (soft or hard limit).

getEpochStability :: Quantity "block" Word32

Length of the suffix of the chain considered unstable

Errors

newtype ErrAdjustForFee

Constructors

ErrCannotCoverFee Word64
UTxO exhausted during fee covering We record what amount missed to cover the fee

data ErrCoinSelection e

Constructors

ErrNotEnoughMoney Word64 Word64
UTxO exhausted during input selection We record the balance of the UTxO as well as the size of the payment we tried to make.

ErrUtxoNotEnoughFragmented Word64 Word64
UTxO is not enough fragmented for the number of transaction outputs We record the number of UTxO entries as well as the number of the outputs of the transaction.

ErrMaximumInputsReached Word64
When trying to construct a transaction, the max number of allowed inputs was reached.

ErrInputsDepleted
When trying to construct a transaction, the available inputs are depleted even when UTxO is properly fragmented and with enough funds to cover payment

ErrInvalidSelection e
Somewhat, we ended up generating an invalid coin selection because of inputs passed down to the coin selection function, or because a target backend has extra-limitations not covered by our coin selection algorithm.

data ErrCreateUnsignedTx e

Errors that can occur when creating an unsigned transaction.

Constructors

ErrCreateUnsignedTxNoSuchWallet ErrNoSuchWallet

ErrCreateUnsignedTxCoinSelection (ErrCoinSelection e)

ErrCreateUnsignedTxFee ErrAdjustForFee

data ErrEstimateTxFee e

Errors that can occur when estimating transaction fees.

Constructors

ErrEstimateTxFeeNoSuchWallet ErrNoSuchWallet

ErrEstimateTxFeeCoinSelection (ErrCoinSelection e)

data ErrListTransactions

Errors that can occur when trying to list transactions.

Constructors

ErrListTransactionsNoSuchWallet ErrNoSuchWallet

ErrListTransactionsStartTimeLaterThanEndTime ErrStartTimeLaterThanEndTime

newtype ErrListUTxOStatistics

Errors that can occur when listing UTxO statistics.

Constructors

ErrListUTxOStatisticsNoSuchWallet ErrNoSuchWallet

newtype ErrMkStdTx

Possible signing error

Constructors

ErrKeyNotFoundForAddress Address
We tried to sign a transaction with inputs that are unknown to us?

data ErrNetworkUnavailable

Network is unavailable

Constructors

ErrNetworkUnreachable Text
Cannot connect to network backend.

ErrNetworkInvalid Text
Network backend reports that the requested network is invalid.

newtype ErrNoSuchWallet

Can't perform given operation because there's no wallet

Constructors

ErrNoSuchWallet WalletId

data ErrPostTx

Error while trying to send a transaction

Constructors

ErrPostTxNetworkUnreachable ErrNetworkUnavailable

ErrPostTxBadRequest Text

ErrPostTxProtocolFailure Text

data ErrDecodeExternalTx

Error while trying to decode externally signed transaction

Constructors

ErrDecodeExternalTxWrongPayload Text

ErrDecodeExternalTxNotSupported

data ErrSignTx

Errors that can occur when signing a transaction.

Constructors

ErrSignTx ErrMkStdTx

ErrSignTxNoSuchWallet ErrNoSuchWallet

ErrSignTxWithRootKey ErrWithRootKey

data ErrStartTimeLaterThanEndTime

Indicates that the specified start time is later than the specified end time.

Constructors

ErrStartTimeLaterThanEndTime

Fields

errStartTime :: UTCTime

errEndTime :: UTCTime

data ErrSubmitTx

Errors that can occur when submitting a signed transaction to the network.

Constructors

ErrSubmitTxNetwork ErrPostTx

ErrSubmitTxNoSuchWallet ErrNoSuchWallet

data ErrSubmitExternalTx

Errors that can occur when submitting an externally-signed transaction to the network.

Constructors

ErrSubmitExternalTxNetwork ErrPostTx

ErrSubmitExternalTxDecode ErrDecodeExternalTx

data ErrUpdatePassphrase

Errors that can occur when trying to change a wallet's passphrase.

Constructors

ErrUpdatePassphraseNoSuchWallet ErrNoSuchWallet

ErrUpdatePassphraseWithRootKey ErrWithRootKey

type family ErrValidateSelection t

A type family for validations that are specific to a particular backend type. This demands an instantiation of the family for a particular backend:

type instance (ErrValidateSelection MyBackend) = MyCustomError

newtype ErrWalletAlreadyExists

Forbidden operation was executed on an already existing wallet

Constructors

ErrWalletAlreadyExists WalletId

data ErrWithRootKey

Errors that can occur when trying to perform an operation on a wallet that requires a private key, but where none is attached to the wallet.

Constructors

ErrWithRootKeyNoRootKey WalletId

ErrWithRootKeyWrongPassphrase WalletId ErrWrongPassphrase

data ErrWrongPassphrase

Indicate a failure when checking for a given Passphrase match

Constructors

ErrWrongPassphrase

Construction

newWalletLayer :: forall s t k. Buildable (Tx t) => Trace IO Text -> BlockchainParameters t -> DBLayer IO s t k -> NetworkLayer t IO -> TransactionLayer t k -> IO (WalletLayer s t k)

Create a new instance of the wallet layer.