Cardano.Wallet

Description

Provides the wallet layer functions that are used by API layer and uses both Cardano.Wallet.DB and Cardano.Wallet.Network to realize its role as being intermediary between the three.

Interface

data WalletLayer s t

Constructors

WalletLayer

Fields

createWallet :: WalletId -> WalletName -> s -> ExceptT ErrWalletAlreadyExists IO WalletId

Initialise and store a new wallet, returning its Id.

readWallet :: WalletId -> ExceptT ErrNoSuchWallet IO (Wallet s t, WalletMetadata)

Retrieve the wallet state for the wallet with the given ID.

updateWallet :: WalletId -> (WalletMetadata -> WalletMetadata) -> ExceptT ErrNoSuchWallet IO ()

Update the wallet metadata with the given update function.

updateWalletPassphrase :: WalletId -> (Passphrase "encryption-old", Passphrase "encryption-new") -> ExceptT ErrUpdatePassphrase IO ()

Change the wallet passphrase to the given passphrase.

listWallets :: IO [WalletId]

Retrieve a list of known wallets IDs.

removeWallet :: WalletId -> ExceptT ErrNoSuchWallet IO ()

Remove an existing wallet. Note that, there's no particular work to be done regarding the restoration worker as it will simply terminate on the next tick when noticing that the corresponding wallet is gone.

restoreWallet :: WalletId -> ExceptT ErrNoSuchWallet IO ()

Restore a wallet from its current tip up to a given target (typically, the network tip).

It returns immediately and fail if the wallet is already beyond the given tip. It starts a worker in background which will fetch and apply remaining blocks until failure or, the target slot is reached.

createUnsignedTx :: WalletId -> CoinSelectionOptions -> NonEmpty TxOut -> ExceptT ErrCreateUnsignedTx IO CoinSelection

Prepare a transaction and automatically select inputs from the wallet to cover the requested outputs. Note that this only run the coin selection for the given outputs. In order to construct (and sign) an actual transaction, have a look at $sel:signTx:WalletLayer.

signTx :: WalletId -> Passphrase "encryption" -> CoinSelection -> ExceptT ErrSignTx IO (Tx, TxMeta, [TxWitness])

Produce witnesses and construct a transaction from a given selection. Requires the encryption passphrase in order to decrypt the root private key. Note that this doesn't broadcast the transaction to the network. In order to do so, have a look at $sel:submitTx:WalletLayer.

submitTx :: WalletId -> (Tx, TxMeta, [TxWitness]) -> ExceptT ErrSubmitTx IO ()

Broadcast a (signed) transaction to the network.

attachPrivateKey :: WalletId -> (Key RootK XPrv, Passphrase "encryption") -> ExceptT ErrNoSuchWallet IO ()

Attach a given private key to a wallet. The private key is necessary for some operations like signing transactions or, generating new accounts.

Errors

data ErrCreateUnsignedTx

Errors occuring when creating an unsigned transaction

Constructors

ErrCreateUnsignedTxNoSuchWallet ErrNoSuchWallet

ErrCreateUnsignedTxCoinSelection CoinSelectionError

ErrCreateUnsignedTxFee ErrAdjustForFee

newtype ErrNoSuchWallet

Can't perform given operation because there's no wallet

Constructors

ErrNoSuchWallet WalletId

data ErrSignTx

Errors occuring when signing a transaction

Constructors

ErrSignTx ErrMkStdTx

ErrSignTxNoSuchWallet ErrNoSuchWallet

ErrSignTxWithRootKey ErrWithRootKey

data ErrSubmitTx

Errors occuring when submitting a signed transaction to the network

Constructors

ErrSubmitTxNetwork ErrPostTx

ErrSubmitTxNoSuchWallet ErrNoSuchWallet

data ErrUpdatePassphrase

Errors occuring when trying to change a wallet's passphrase

Constructors

ErrUpdatePassphraseNoSuchWallet ErrNoSuchWallet

ErrUpdatePassphraseWithRootKey ErrWithRootKey

newtype ErrWalletAlreadyExists

Forbidden operation was executed on an already existing wallet

Constructors

ErrWalletAlreadyExists WalletId

data ErrWithRootKey

Errors occuring when trying to perform an operation on a wallet which requires a private key, but none is attached to the wallet

Constructors

ErrWithRootKeyNoRootKey WalletId

ErrWithRootKeyWrongPassphrase ErrWrongPassphrase

Construction

mkWalletLayer :: forall s t. (IsOwned s, GenChange s, NFData s, Show s, TxId t) => DBLayer IO s t -> NetworkLayer IO -> TransactionLayer -> WalletLayer s t

Create a new instance of the wallet layer.

Helpers

unsafeRunExceptT :: (MonadFail m, Show e) => ExceptT e m a -> m a

Run an ExcepT and throws the error if any. This makes sense only if called after checking for an invariant or, after ensuring that preconditions for meeting the underlying error have been discarded.