Cardano.Wallet

Description

Provides wallet layer functions that are used by API layer. Uses both Cardano.Wallet.DB and Cardano.Wallet.Network to realize its role as being intermediary between the three.

Functions of the wallet layer are often parameterized with variables following the convention below:

s: A state used to keep track of known addresses. Typically, possible values for this parameter are described in AddressDiscovery sub-modules. For instance SeqState or Rnd State.

t: A target backend which captures details specific to a particular chain producer (binary formats, fee policy, networking layer).

k: A key derivation scheme intrisically connected to the underlying discovery state s. This describes how the hierarchical structure of a wallet is defined as well as the relationship between secret keys and public addresses.

Developement

Naming Conventions

Components inside a particular context ctx can be called via dedicated lenses (see Cardano.Wallet#Capabilities). These components are extracted from the context in a where clause according to the following naming convention:

db = ctx ^. dbLayer @s @t @k for the DBLayer.

tr = ctx ^. logger for the Logger.

nw = ctx ^. networkLayer @t for the NetworkLayer.

tl = ctx ^. transactionLayer @t @k for the TransactionLayer.

re = ctx ^. workerRegistry for the WorkerRegistry.

TroubleShooting

• Overlapping instances for HasType (DBLayer IO s t k) ctx
    arising from a use of ‘myFunction’
  Matching instances:

Occurs when a particular function is missing a top-level constraint (because it uses another function that demands such constraint). Here, myFunction needs its surrounding context ctx to have a DBLayer but the constraint is missing from its host function.

Fix: Add "HasDBLayer s t k" as a class-constraint to the surrounding function.

• Overlapping instances for HasType (DBLayer IO s t0 k0) ctx
    arising from a use of ‘myFunction’
  Matching givens (or their superclasses):

Occurs when a function is called in a context where type-level parameters can be inferred. Here, myFunction is called but it is unclear whether the parameter t0 and k0 of its context are the same as the ones from the function at the call-site.

Fix: Add type-applications at the call-site "myFunction @ctx @s @t @k"

WalletLayer

data WalletLayer s t (k :: Depth -> * -> *)

Constructors

WalletLayer (Trace IO Text) (Block (Tx t), BlockchainParameters) (NetworkLayer IO t (Block (Tx t))) (TransactionLayer t k) (DBLayer IO s t k)

Capabilities

Each function in the wallet layer is defined in function of a non-specialized context ctx. That context may require some extra capabilities via class-constraints in the function signature. Capabilities are expressed in the form of a "HasXXX" class-constraints sometimes with extra type parameters.

For example:

listWallets
    :: forall ctx s t k.
        ( HasDBLayer s t k ctx
        )
    => ctx
    -> IO [WalletId]

Requires that the given context has an access to a database layer DBLayer parameterized over the wallet state, a network target and a key derivation scheme. Components are pulled from the context generically (i.e. the concrete ctx must derive Generic) using their associated type. The concrete ctx is therefore expected to be a product-type of all the necessary components.

One can build an interface using only a subset of the wallet layer capabilities and functions, for instance, something to fiddle with wallets and their metadata does not require any networking layer.

type HasDBLayer s t k = HasType (DBLayer IO s t k)

dbLayer :: forall s t k ctx. HasDBLayer s t k ctx => Lens' ctx (DBLayer IO s t k)

type HasLogger = HasType (Trace IO Text)

logger :: forall ctx. HasLogger ctx => Lens' ctx (Trace IO Text)

type HasNetworkLayer t = HasType (NetworkLayer IO t (Block (Tx t)))

This module is only interested in one block-, and tx-type. This constraint hides that choice, for some ease of use.

networkLayer :: forall t ctx. HasNetworkLayer t ctx => Lens' ctx (NetworkLayer IO t (Block (Tx t)))

type HasTransactionLayer t k = HasType (TransactionLayer t k)

transactionLayer :: forall t k ctx. HasTransactionLayer t k ctx => Lens' ctx (TransactionLayer t k)

type HasGenesisData t = HasType (Block (Tx t), BlockchainParameters)

genesisData :: forall t ctx. HasGenesisData t ctx => Lens' ctx (Block (Tx t), BlockchainParameters)

Interface

Wallet

createWallet :: forall ctx s t k. (HasGenesisData t ctx, HasDBLayer s t k ctx, Show s, NFData s, IsOurs s, DefineTx t) => ctx -> WalletId -> WalletName -> s -> ExceptT ErrWalletAlreadyExists IO WalletId

Initialise and store a new wallet, returning its ID.

attachPrivateKey :: forall ctx s t k. HasDBLayer s t k ctx => ctx -> WalletId -> (k RootK XPrv, Passphrase "encryption") -> ExceptT ErrNoSuchWallet IO ()

Attach a given private key to a wallet. The private key is necessary for some operations like signing transactions, or generating new accounts.

listUtxoStatistics :: forall ctx s t k. (HasDBLayer s t k ctx, DefineTx t) => ctx -> WalletId -> ExceptT ErrListUTxOStatistics IO UTxOStatistics

List the wallet's UTxO statistics.

readWallet :: forall ctx s t k. (HasDBLayer s t k ctx, DefineTx t) => ctx -> WalletId -> ExceptT ErrNoSuchWallet IO (Wallet s t, WalletMetadata, Set (Tx t))

Retrieve the wallet state for the wallet with the given ID.

removeWallet :: forall ctx s t k. HasDBLayer s t k ctx => ctx -> WalletId -> ExceptT ErrNoSuchWallet IO ()

Remove an existing wallet. Note that there's no particular work to be done regarding the restoration worker as it will simply terminate on the next tick when noticing that the corresponding wallet is gone.

restoreWallet :: forall ctx s t k. (HasLogger ctx, HasNetworkLayer t ctx, HasDBLayer s t k ctx, DefineTx t) => ctx -> WalletId -> ExceptT ErrNoSuchWallet IO ()

Restore a wallet from its current tip up to the network tip.

This function returns immediately, starting a worker thread in the background that will fetch and apply remaining blocks until the network tip is reached or until failure.

updateWallet :: forall ctx s t k. HasDBLayer s t k ctx => ctx -> WalletId -> (WalletMetadata -> WalletMetadata) -> ExceptT ErrNoSuchWallet IO ()

Update a wallet's metadata with the given update function.

updateWalletPassphrase :: forall ctx s t k. (HasDBLayer s t k ctx, WalletKey k) => ctx -> WalletId -> (Passphrase "encryption-old", Passphrase "encryption-new") -> ExceptT ErrUpdatePassphrase IO ()

Change a wallet's passphrase to the given passphrase.

newtype ErrWalletAlreadyExists

Forbidden operation was executed on an already existing wallet

Constructors

ErrWalletAlreadyExists WalletId

newtype ErrNoSuchWallet

Can't perform given operation because there's no wallet

Constructors

ErrNoSuchWallet WalletId

newtype ErrListUTxOStatistics

Errors that can occur when listing UTxO statistics.

Constructors

ErrListUTxOStatisticsNoSuchWallet ErrNoSuchWallet

data ErrUpdatePassphrase

Errors that can occur when trying to change a wallet's passphrase.

Constructors

ErrUpdatePassphraseNoSuchWallet ErrNoSuchWallet

ErrUpdatePassphraseWithRootKey ErrWithRootKey

Address

listAddresses :: forall ctx s t k. (HasDBLayer s t k ctx, IsOurs s, CompareDiscovery s, KnownAddresses s, DefineTx t) => ctx -> WalletId -> ExceptT ErrNoSuchWallet IO [(Address, AddressState)]

List all addresses of a wallet with their metadata. Addresses are ordered from the most-recently-discovered to the oldest known.

Transaction

createUnsignedTx :: forall ctx s t k e. (HasTransactionLayer t k ctx, HasLogger ctx, HasDBLayer s t k ctx, DefineTx t, e ~ ErrValidateSelection t) => ctx -> WalletId -> NonEmpty TxOut -> ExceptT (ErrCreateUnsignedTx e) IO CoinSelection

Prepare a transaction and automatically select inputs from the wallet to cover the requested outputs. Note that this only runs coin selection for the given outputs. In order to construct (and sign) an actual transaction, use signTx.

estimateTxFee :: forall ctx s t k e. (HasTransactionLayer t k ctx, HasDBLayer s t k ctx, DefineTx t, e ~ ErrValidateSelection t) => ctx -> WalletId -> NonEmpty TxOut -> ExceptT (ErrEstimateTxFee e) IO Fee

Estimate a transaction fee by automatically selecting inputs from the wallet to cover the requested outputs.

forgetPendingTx :: forall ctx s t k. (HasDBLayer s t k ctx, DefineTx t) => ctx -> WalletId -> Hash "Tx" -> ExceptT ErrRemovePendingTx IO ()

Forget pending transaction.

listTransactions :: forall ctx s t k. (HasDBLayer s t k ctx, DefineTx t) => ctx -> WalletId -> Maybe UTCTime -> Maybe UTCTime -> SortOrder -> ExceptT ErrListTransactions IO [TransactionInfo]

List all transactions and metadata from history for a given wallet.

signTx :: forall ctx s t k. (HasTransactionLayer t k ctx, HasDBLayer s t k ctx, Show s, NFData s, IsOwned s k, GenChange s) => ctx -> WalletId -> Passphrase "encryption" -> CoinSelection -> ExceptT ErrSignTx IO (Tx t, TxMeta, [TxWitness])

Produce witnesses and construct a transaction from a given selection. Requires the encryption passphrase in order to decrypt the root private key. Note that this doesn't broadcast the transaction to the network. In order to do so, use submitTx.

submitExternalTx :: forall ctx t k. (HasNetworkLayer t ctx, HasTransactionLayer t k ctx) => ctx -> ByteString -> ExceptT ErrSubmitExternalTx IO (Tx t)

Broadcast an externally-signed transaction to the network.

submitTx :: forall ctx s t k. (HasNetworkLayer t ctx, HasDBLayer s t k ctx, DefineTx t) => ctx -> WalletId -> (Tx t, TxMeta, [TxWitness]) -> ExceptT ErrSubmitTx IO ()

Broadcast a (signed) transaction to the network.

data ErrCreateUnsignedTx e

Errors that can occur when creating an unsigned transaction.

Constructors

ErrCreateUnsignedTxNoSuchWallet ErrNoSuchWallet

ErrCreateUnsignedTxCoinSelection (ErrCoinSelection e)

ErrCreateUnsignedTxFee ErrAdjustForFee

data ErrEstimateTxFee e

Errors that can occur when estimating transaction fees.

Constructors

ErrEstimateTxFeeNoSuchWallet ErrNoSuchWallet

ErrEstimateTxFeeCoinSelection (ErrCoinSelection e)

data ErrSignTx

Errors that can occur when signing a transaction.

Constructors

ErrSignTx ErrMkStdTx

ErrSignTxNoSuchWallet ErrNoSuchWallet

ErrSignTxWithRootKey ErrWithRootKey

newtype ErrMkStdTx

Possible signing error

Constructors

ErrKeyNotFoundForAddress Address
We tried to sign a transaction with inputs that are unknown to us?

newtype ErrAdjustForFee

Constructors

ErrCannotCoverFee Word64
UTxO exhausted during fee covering We record what amount missed to cover the fee

data ErrCoinSelection e

Constructors

ErrNotEnoughMoney Word64 Word64
UTxO exhausted during input selection We record the balance of the UTxO as well as the size of the payment we tried to make.

ErrUtxoNotEnoughFragmented Word64 Word64
UTxO is not enough fragmented for the number of transaction outputs We record the number of UTxO entries as well as the number of the outputs of the transaction.

ErrMaximumInputsReached Word64
When trying to construct a transaction, the max number of allowed inputs was reached.

ErrInputsDepleted
When trying to construct a transaction, the available inputs are depleted even when UTxO is properly fragmented and with enough funds to cover payment

ErrInvalidSelection e
Somewhat, we ended up generating an invalid coin selection because of inputs passed down to the coin selection function, or because a target backend has extra-limitations not covered by our coin selection algorithm.

data ErrSubmitTx

Errors that can occur when submitting a signed transaction to the network.

Constructors

ErrSubmitTxNetwork ErrPostTx

ErrSubmitTxNoSuchWallet ErrNoSuchWallet

data ErrSubmitExternalTx

Errors that can occur when submitting an externally-signed transaction to the network.

Constructors

ErrSubmitExternalTxNetwork ErrPostTx

ErrSubmitExternalTxDecode ErrDecodeSignedTx

data ErrRemovePendingTx

Can't perform removing pending transaction

Constructors

ErrRemovePendingTxNoSuchWallet WalletId

ErrRemovePendingTxNoSuchTransaction (Hash "Tx")

ErrRemovePendingTxTransactionNoMorePending (Hash "Tx")

data ErrPostTx

Error while trying to send a transaction

Constructors

ErrPostTxNetworkUnreachable ErrNetworkUnavailable

ErrPostTxBadRequest Text

ErrPostTxProtocolFailure Text

data ErrDecodeSignedTx

Error while trying to decode externally signed transaction

Constructors

ErrDecodeSignedTxWrongPayload Text

ErrDecodeSignedTxNotSupported

type family ErrValidateSelection t

A type family for validations that are specific to a particular backend type. This demands an instantiation of the family for a particular backend:

type instance (ErrValidateSelection MyBackend) = MyCustomError

data ErrWithRootKey

Errors that can occur when trying to perform an operation on a wallet that requires a private key, but where none is attached to the wallet.

Constructors

ErrWithRootKeyNoRootKey WalletId

ErrWithRootKeyWrongPassphrase WalletId ErrWrongPassphrase

data ErrWrongPassphrase

Indicate a failure when checking for a given Passphrase match

Constructors

ErrWrongPassphrase

data ErrListTransactions

Errors that can occur when trying to list transactions.

Constructors

ErrListTransactionsNoSuchWallet ErrNoSuchWallet

ErrListTransactionsStartTimeLaterThanEndTime ErrStartTimeLaterThanEndTime

data ErrNetworkUnavailable

Network is unavailable

Constructors

ErrNetworkUnreachable Text
Cannot connect to network backend.

ErrNetworkInvalid Text
Network backend reports that the requested network is invalid.

data ErrStartTimeLaterThanEndTime

Indicates that the specified start time is later than the specified end time.

Constructors

ErrStartTimeLaterThanEndTime

Fields

errStartTime :: UTCTime

errEndTime :: UTCTime