Cardano.Wallet.Api.Client

Description

This module provides a half-typed Servant client for the cardano-wallet V2 API.

The functions in this module can be run with "Servant.Client.runClientM".

API Clients

data WalletClient wallet

This data type encapsulates the client functions for all endpoints of the cardano-wallet V2 API.

Constructors

WalletClient

Fields

deleteWallet :: ApiT WalletId -> ClientM ()

getWallet :: ApiT WalletId -> ClientM wallet

getWalletUtxoStatistics :: ApiT WalletId -> ClientM ApiUtxoStatistics

listWallets :: ClientM [wallet]

postWallet :: PostData wallet -> ClientM wallet

putWallet :: ApiT WalletId -> WalletPutData -> ClientM wallet

putWalletPassphrase :: ApiT WalletId -> WalletPutPassphraseData -> ClientM NoContent

forceResyncWallet :: ApiT WalletId -> ApiNetworkTip -> ClientM NoContent

walletClient :: WalletClient ApiWallet

Produces a WalletClient working against the /wallets API.

byronWalletClient :: WalletClient ApiByronWallet

Produces a WalletClient working against the /wallets API.

data TransactionClient

Constructors

TransactionClient

Fields

listTransactions :: ApiT WalletId -> Maybe Iso8601Time -> Maybe Iso8601Time -> Maybe (ApiT SortOrder) -> ClientM [ApiTransactionT Value]

postTransaction :: ApiT WalletId -> PostTransactionDataT Value -> ClientM (ApiTransactionT Value)

postTransactionFee :: ApiT WalletId -> PostTransactionFeeDataT Value -> ClientM ApiFee

postExternalTransaction :: PostExternalTransactionData -> ClientM ApiTxId

deleteTransaction :: ApiT WalletId -> ApiTxId -> ClientM NoContent

transactionClient :: TransactionClient

Produces a 'TransactionClient t' working against the /wallets API.

byronTransactionClient :: TransactionClient

Produces a 'TransactionClient n' working against the /byron-wallets API.

newtype AddressClient

Constructors

AddressClient

Fields

listAddresses :: ApiT WalletId -> Maybe (ApiT AddressState) -> ClientM [Value]

addressClient :: AddressClient

Produces an 'AddressClient n' working against the /wallets API

data StakePoolClient

Constructors

StakePoolClient

Fields

listPools :: ClientM [ApiStakePool]

joinStakePool :: ApiPoolId -> ApiT WalletId -> ApiWalletPassphrase -> ClientM (ApiTransactionT Value)

quitStakePool :: ApiT WalletId -> ApiWalletPassphrase -> ClientM (ApiTransactionT Value)

stakePoolClient :: StakePoolClient

Produces an 'StakePoolsClient n' working against the /stake-pools API

data NetworkClient

Constructors

NetworkClient

Fields

networkInformation :: ClientM ApiNetworkInformation

networkParameters :: ApiEpochNumber -> ClientM ApiNetworkParameters

networkClock :: ClientM ApiNetworkClock

networkClient :: NetworkClient

Produces a NetworkClient