Cardano.Wallet.Api.Client

Description

This module provides a Servant client for the cardano-wallet V2 API.

The functions in this module can be run with "Servant.Client.runClientM".

API endpoints

data WalletClient t

This data type encapsulates the client functions for all endpoints of the cardano-wallet V2 API.

Constructors

WalletClient

Fields

listAddresses :: ApiT WalletId -> Maybe (ApiT AddressState) -> ClientM [ApiAddress t]

deleteWallet :: ApiT WalletId -> ClientM ()

getWallet :: ApiT WalletId -> ClientM ApiWallet

getWalletUtxoStatistics :: ApiT WalletId -> ClientM ApiUtxoStatistics

listWallets :: ClientM [ApiWallet]

postWallet :: WalletOrAccountPostData -> ClientM ApiWallet

putWallet :: ApiT WalletId -> WalletPutData -> ClientM ApiWallet

putWalletPassphrase :: ApiT WalletId -> WalletPutPassphraseData -> ClientM NoContent

forceResyncWallet :: ApiT WalletId -> ApiNetworkTip -> ClientM NoContent

listTransactions :: ApiT WalletId -> Maybe Iso8601Time -> Maybe Iso8601Time -> Maybe (ApiT SortOrder) -> ClientM [ApiTransaction t]

postTransaction :: ApiT WalletId -> PostTransactionData t -> ClientM (ApiTransaction t)

postTransactionFee :: ApiT WalletId -> PostTransactionFeeData t -> ClientM ApiFee

postExternalTransaction :: PostExternalTransactionData -> ClientM ApiTxId

deleteTransaction :: ApiT WalletId -> ApiTxId -> ClientM NoContent

listPools :: ClientM [ApiStakePool]

joinStakePool :: ApiPoolId -> ApiT WalletId -> ApiWalletPassphrase -> ClientM (ApiTransaction t)

quitStakePool :: ApiT WalletId -> ApiWalletPassphrase -> ClientM (ApiTransaction t)

networkInformation :: ClientM ApiNetworkInformation

networkParameters :: ApiEpochNumber -> ClientM ApiNetworkParameters

networkClock :: ClientM ApiNetworkClock

walletClient :: forall t. (DecodeAddress t, EncodeAddress t) => WalletClient t

Produces a WalletClient for the cardano-wallet V2 API.

You should apply a Cardano.Wallet.Primitive.AddressDerivation.NetworkDiscriminant type parameter to this to get a concrete client.

Helper functions

waitForSync :: WalletClient t -> ClientM ()

Poll the wallet server until it reports that it has synced with the network.

waitForRestore :: WalletClient t -> ApiT WalletId -> ClientM ApiWallet

Poll a wallet by ID until it has restored.