Cardano.Wallet.Api.Link

Description

Type-safe endpoint accessors for the wallet API. Under normal circumstances, one would prefer to use WalletClient from Client and not to bother with endpoints at all.

Yet, in some cases (like in black-box testing), one could want to purposely send malformed requests to specific endpoints. Thus, this module facilitates the construction of valid endpoints that'd be accepted by the server, and for which, users are free to send all sort of data as payload, valid or invalid.

This module is meant to be used via qualified imports and with type-applications since all exposed functions are type ambiguous in a variable style of type WalletStyle.

import qualified Cardano.Wallet.Api.Link as Link

For examples:

>>> Link.deleteWallet @'Shelley myWallet
( "DELETE", "/v2/wallets/2512a00e9653fe49a44a5886202e24d77eeb998f" )

>>> Link.getWallet @('Byron 'Icarus) myWallet
( "GET", "/v2/byron-wallets/2512a00e9653fe49a44a5886202e24d77eeb998f" )

Wallets

deleteWallet :: forall (style :: WalletStyle) w. (Discriminate style, HasType (ApiT WalletId) w) => w -> (Method, Text)

getWallet :: forall (style :: WalletStyle) w. (Discriminate style, HasType (ApiT WalletId) w) => w -> (Method, Text)

listWallets :: forall (style :: WalletStyle). Discriminate style => (Method, Text)

postWallet :: PostWallet k => (Method, Text)

putWallet :: forall w. HasType (ApiT WalletId) w => w -> (Method, Text)

putWalletPassphrase :: forall w. HasType (ApiT WalletId) w => w -> (Method, Text)

getUTxOsStatistics :: forall w. HasType (ApiT WalletId) w => w -> (Method, Text)

getMigrationInfo :: forall w. HasType (ApiT WalletId) w => w -> (Method, Text)

migrateWallet :: forall from to. (HasType (ApiT WalletId) from, HasType (ApiT WalletId) to) => from -> to -> (Method, Text)

forceResyncWallet :: forall style w. (HasType (ApiT WalletId) w, Discriminate style) => w -> (Method, Text)

Addresses

listAddresses :: forall w. HasType (ApiT WalletId) w => w -> (Method, Text)

listAddresses' :: forall w. HasType (ApiT WalletId) w => w -> Maybe AddressState -> (Method, Text)

CoinSelections

selectCoins :: forall w. HasType (ApiT WalletId) w => w -> (Method, Text)

Transactions

createTransaction :: forall w. HasType (ApiT WalletId) w => w -> (Method, Text)

listTransactions :: forall (style :: WalletStyle) w. (Discriminate style, HasType (ApiT WalletId) w) => w -> (Method, Text)

listTransactions' :: forall (style :: WalletStyle) w. (Discriminate style, HasType (ApiT WalletId) w) => w -> Maybe Iso8601Time -> Maybe Iso8601Time -> Maybe SortOrder -> (Method, Text)

getTransactionFee :: forall w. HasType (ApiT WalletId) w => w -> (Method, Text)

deleteTransaction :: forall (style :: WalletStyle) w t. (Discriminate style, HasType (ApiT WalletId) w, HasType (ApiT (Hash "Tx")) t) => w -> t -> (Method, Text)

StakePools

listStakePools :: (Method, Text)

joinStakePool :: forall s w. (HasType (ApiT PoolId) s, HasType (ApiT WalletId) w) => s -> w -> (Method, Text)

quitStakePool :: forall s w. (HasType (ApiT PoolId) s, HasType (ApiT WalletId) w) => s -> w -> (Method, Text)

getDelegationFee :: forall w. HasType (ApiT WalletId) w => w -> (Method, Text)

Network

getNetworkInfo :: (Method, Text)

Proxy

postExternalTransaction :: (Method, Text)

class PostWallet k

Minimal complete definition

postWallet

class Discriminate (style :: WalletStyle)

Minimal complete definition

discriminate