Cardano.Wallet.Api.Server

Description

API handlers and server using the underlying wallet layer to provide endpoints reachable through HTTP.

Documentation

data Listen

How the server should listen for incoming requests.

Constructors

ListenOnPort Port
Listen on given TCP port

ListenOnRandomPort
Listen on an unused TCP port, selected at random

data ListenError

Constructors

ListenErrorAddressAlreadyInUse (Maybe Port)

ListenErrorOperationNotPermitted

ListenErrorHostDoesNotExist HostPreference

ListenErrorInvalidAddress HostPreference

data HostPreference

start :: Settings -> Tracer IO ApiLog -> Socket -> Application -> IO ()

Start the application server, using the given settings and a bound socket.

serve :: HasServer api ([] :: [Type]) => Proxy api -> Server api -> Application

server :: forall t n byron icarus shelley. (byron ~ ApiLayer (RndState Mainnet) t ByronKey, icarus ~ ApiLayer (SeqState Mainnet IcarusKey) t IcarusKey, shelley ~ ApiLayer (SeqState n ShelleyKey) t ShelleyKey, DelegationAddress n ShelleyKey, Buildable (ErrValidateSelection t)) => byron -> icarus -> shelley -> StakePoolLayer IO -> NtpClient -> Server (Api n)

A Servant server for our wallet API

byronServer :: forall t n. ApiLayer (RndState Mainnet) t ByronKey -> ApiLayer (SeqState Mainnet IcarusKey) t IcarusKey -> NtpClient -> Server (Api n)

A diminished servant server to serve Byron wallets only.

newApiLayer :: forall ctx s t k. ctx ~ ApiLayer s t k => Tracer IO (WorkerLog WalletId WalletLog) -> (Block, BlockchainParameters, SyncTolerance) -> NetworkLayer IO t Block -> TransactionLayer t k -> DBFactory IO s k -> [WalletId] -> IO ctx

Create a new instance of the wallet layer.

withListeningSocket

Arguments

:: HostPreference
Which host to bind.

-> Listen
Whether to listen on a given port, or random port.

-> (Either ListenError (Port, Socket) -> IO a)
Action to run with listening socket.

-> IO a

Run an action with a TCP socket bound to a port specified by the Listen parameter.

class LiftHandler e where

Lift our wallet layer into servant Handler, by mapping each error to a corresponding servant error.

Minimal complete definition

handler

Methods

liftHandler :: ExceptT e IO a -> Handler a

handler :: e -> ServerError