Cardano.Wallet.Api.Server

Description

API handlers and server using the underlying wallet layer to provide endpoints reachable through HTTP.

Documentation

data Listen

How the server should listen for incoming requests.

Constructors

ListenOnPort Port
Listen on given TCP port

ListenOnRandomPort
Listen on an unused TCP port, selected at random

data ListenError

Constructors

ListenErrorAddressAlreadyInUse (Maybe Port)

ListenErrorOperationNotPermitted

ListenErrorHostDoesNotExist HostPreference

ListenErrorInvalidAddress HostPreference

data HostPreference

start :: forall t (n :: NetworkDiscriminant). (Buildable (ErrValidateSelection t), DefineTx t, DecodeAddress n, EncodeAddress n, PaymentAddress n ByronKey, PaymentAddress n ShelleyKey) => Settings -> Trace IO Text -> Socket -> ApiLayer (RndState n) t ByronKey -> ApiLayer (SeqState n) t ShelleyKey -> StakePoolLayer IO -> IO ()

Start the application server, using the given settings and a bound socket.

newApiLayer :: forall ctx s t k. (ctx ~ ApiLayer s t k, DefineTx t) => Trace IO Text -> (Block (Tx t), BlockchainParameters) -> NetworkLayer IO t (Block (Tx t)) -> TransactionLayer t k -> DBFactory IO s t k -> [WalletId] -> IO ctx

Create a new instance of the wallet layer.

withListeningSocket

Arguments

:: HostPreference
Which host to bind.

-> Listen
Whether to listen on a given port, or random port.

-> (Either ListenError (Port, Socket) -> IO a)
Action to run with listening socket.

-> IO a

Run an action with a TCP socket bound to a port specified by the Listen parameter.

defaultWorkerAfter :: Trace IO Text -> Either SomeException a -> IO ()