Cardano.Wallet.Api.Server

Description

API handlers and server using the underlying wallet layer to provide endpoints reachable through HTTP.

Documentation

data Listen

How the server should listen for incoming requests.

Constructors

ListenOnPort Port
Listen on given TCP port

ListenOnRandomPort
Listen on an unused TCP port, selected at random

start :: forall t. (Buildable (ErrValidateSelection t), DefineTx t, DecodeAddress t, EncodeAddress t, KeyToAddress t RndKey, KeyToAddress t SeqKey) => Settings -> Trace IO Text -> Socket -> ApiLayer (RndState t) t RndKey -> ApiLayer (SeqState t) t SeqKey -> IO ()

Start the application server, using the given settings and a bound socket.

newApiLayer :: forall ctx s t k. (ctx ~ ApiLayer s t k, DefineTx t) => Trace IO Text -> (Block (Tx t), BlockchainParameters) -> NetworkLayer IO t (Block (Tx t)) -> TransactionLayer t k -> DBFactory IO s t k -> [WalletId] -> IO ctx

Create a new instance of the wallet layer.

withListeningSocket

Arguments

:: Listen
Whether to listen on a given port, or random port.

-> ((Port, Socket) -> IO ())
Action to run with listening socket.

-> IO ()

Run an action with a TCP socket bound to a port specified by the Listen parameter.