Cardano.Wallet.Api.Types

Description

API type representations of various types. We define here pretty much all our user-facing types that are mostly composed with internal / primitive types.

This module also define required API instances (JSON, HttpApiData...) for all those types, making sure to match the specification document:

Wallet API Specification

API Types

data ApiAddress t

Constructors

ApiAddress

Fields

id :: !(ApiT Address, Proxy t)

state :: !(ApiT AddressState)

data ApiWallet

Constructors

ApiWallet

Fields

id :: !(ApiT WalletId)

addressPoolGap :: !(ApiT AddressPoolGap)

balance :: !(ApiT WalletBalance)

delegation :: !(ApiT (WalletDelegation (ApiT PoolId)))

name :: !(ApiT WalletName)

passphrase :: !(Maybe (ApiT WalletPassphraseInfo))

state :: !(ApiT WalletState)

data ApiUtxoStatistics

Constructors

ApiUtxoStatistics

Fields

total :: !(Quantity "lovelace" Natural)

scale :: !(ApiT BoundType)

distribution :: !(Map Word64 Word64)

data WalletBalance

Constructors

WalletBalance

Fields

available :: !(Quantity "lovelace" Natural)

total :: !(Quantity "lovelace" Natural)

data WalletPostData

Constructors

WalletPostData

Fields

addressPoolGap :: !(Maybe (ApiT AddressPoolGap))

mnemonicSentence :: !(ApiMnemonicT '[15, 18, 21, 24] "seed")

mnemonicSecondFactor :: !(Maybe (ApiMnemonicT '[9, 12] "generation"))

name :: !(ApiT WalletName)

passphrase :: !(ApiT (Passphrase "encryption"))

newtype WalletPutData

Constructors

WalletPutData

Fields

name :: Maybe (ApiT WalletName)

data WalletPutPassphraseData

Constructors

WalletPutPassphraseData

Fields

oldPassphrase :: !(ApiT (Passphrase "encryption-old"))

newPassphrase :: !(ApiT (Passphrase "encryption-new"))

data PostTransactionData t

Constructors

PostTransactionData

Fields

payments :: !(NonEmpty (AddressAmount t))

passphrase :: !(ApiT (Passphrase "encryption"))

newtype PostTransactionFeeData t

Constructors

PostTransactionFeeData

Fields

payments :: NonEmpty (AddressAmount t)

newtype PostExternalTransactionData

Constructors

PostExternalTransactionData

Fields

payload :: ByteString

data ApiBlockData

Constructors

ApiBlockData

Fields

time :: UTCTime

block :: !(ApiT SlotId)

data ApiTransaction t

Constructors

ApiTransaction

Fields

id :: !(ApiT (Hash "Tx"))

amount :: !(Quantity "lovelace" Natural)

insertedAt :: !(Maybe ApiBlockData)

depth :: !(Quantity "slot" Natural)

direction :: !(ApiT Direction)

inputs :: ![ApiTxInput t]

outputs :: !(NonEmpty (AddressAmount t))

status :: !(ApiT TxStatus)

newtype ApiFee

Constructors

ApiFee

Fields

amount :: Quantity "lovelace" Natural

newtype ApiTxId

Constructors

ApiTxId

Fields

id :: ApiT (Hash "Tx")

data ApiTxInput t

Constructors

ApiTxInput

Fields

source :: !(Maybe (AddressAmount t))

input :: !(ApiT TxIn)

data AddressAmount t

Constructors

AddressAmount

Fields

address :: !(ApiT Address, Proxy t)

amount :: !(Quantity "lovelace" Natural)

data ApiErrorCode

Error codes returned by the API, in the form of snake_cased strings

Constructors

NoSuchWallet

WalletAlreadyExists

NoRootKey

WrongEncryptionPassphrase

MalformedTxPayload

KeyNotFoundForAddress

NotEnoughMoney

UtxoNotEnoughFragmented

TransactionIsTooBig

InputsDepleted

CannotCoverFee

InvalidCoinSelection

NetworkUnreachable

NetworkMisconfigured

NetworkTipNotFound

CreatedInvalidTransaction

RejectedByCoreNode

BadRequest

NotFound

MethodNotAllowed

NotAcceptable

StartTimeLaterThanEndTime

UnsupportedMediaType

UnexpectedError

newtype Iso8601Time

Defines a point in time that can be formatted as and parsed from an ISO 8601-compliant string.

Constructors

Iso8601Time

Fields

getIso8601Time :: UTCTime

Polymorphic Types

newtype ApiT a

Polymorphic wrapper type to put around primitive types and, 3rd party lib types to avoid defining orphan instances and/or, undesirable instances on primitive types. It helps to keep a nice separation of concerns between the API layer and other modules.

Constructors

ApiT

Fields

getApiT :: a

newtype ApiMnemonicT (sizes :: [Nat]) (purpose :: Symbol)

Representation of mnemonics at the API-level, using a polymorphic type in the lengths of mnemonics that are supported (and an underlying purpose). In practice, mnemonics correspond to passphrases or seeds, and although they're nice to manipulate as mnemonics from a user-perspective, carrying around a list of words doesn't really make sense for the business logic, which prefers manipulating scrubbed bytes directly.

data MyWallet
    { mnemonic :: ApiMnemonicT '[15,18,21,24] "root-seed"
    }

Note that the given Nats **have** to be valid mnemonic sizes, otherwise the underlying code won't even compile, with not-so-friendly error messages.

Also, the internal representation holds a [Text] which contains the list of mnemonic words that was parsed. This is only to be able to implement the ToJSON instances and roundtrip, which is a very dubious argument. In practice, we'll NEVER peek at the mnemonic, output them and whatnot.

Constructors

ApiMnemonicT (Passphrase purpose, [Text])

getApiMnemonicT :: ApiMnemonicT sizes purpose -> Passphrase purpose