Cardano.Wallet.DB.MVar

Description

Dummy implementation of the database-layer, using MVar. This may be good for state-machine testing in order to compare it with an implementation on a real data store.

Documentation

newDBLayer :: forall s t. IO (DBLayer IO s t)

Instantiate a new in-memory "database" layer that simply stores data in a local MVar. Data vanishes if the software is shut down.