Cardano.Wallet.DB.Sqlite

Description

An implementation of the DBLayer which uses Persistent and SQLite.

Documentation

data SqliteContext

Context for the SQLite DBLayer.

newDBLayer

Arguments

:: (IsOurs s, NFData s, Show s, PersistState s, PersistTx t, PersistKey k)

=> Configuration
Logging configuration

-> Trace IO Text
Logging object

-> Maybe FilePath
Database file location, or Nothing for in-memory database

-> IO (SqliteContext, DBLayer IO s t k)

Sets up a connection to the SQLite database.

Database migrations are run to create tables if necessary.

If the given file path does not exist, it will be created by the sqlite library.

getDBLayer will provide the actual DBLayer implementation. The database should be closed with destroyDBLayer. If you use withDBLayer then both of these things will be handled for you.

destroyDBLayer :: SqliteContext -> IO ()

Finalize database statements and close the database connection.

withDBLayer

Arguments

:: (IsOurs s, NFData s, Show s, PersistState s, PersistTx t, PersistKey k)

=> Configuration
Logging configuration

-> Trace IO Text
Logging object

-> Maybe FilePath
Database file location, or Nothing for in-memory database

-> (DBLayer IO s t k -> IO a)
Action to run.

-> IO a

Runs an action with a connection to the SQLite database.

Database migrations are run to create tables if necessary.

If the given file path does not exist, it will be created by the sqlite library.

unsafeRunQuery :: SqliteContext -> SqlPersistT IO a -> IO a

Run a raw query from the outside using an instantiate DB layer. This is completely unsafe because it breaks the abstraction boundary and can have disastrous results on the database consistency.

Interfaces

class PersistState s where

Functions for saving/loading the wallet's address discovery state into SQLite.

Methods

insertState :: (WalletId, SlotId) -> s -> SqlPersistT IO ()

Store the state for a checkpoint.

selectState :: (WalletId, SlotId) -> SqlPersistT IO (Maybe s)

Load the state for a checkpoint.

deleteState :: WalletId -> SqlPersistT IO ()

Remove the state for all checkpoints of a wallet.

class DefineTx t => PersistTx t where

Methods

resolvedInputs :: Tx t -> [(TxIn, Maybe Coin)]

mkTx :: Hash "Tx" -> [(TxIn, Maybe Coin)] -> [TxOut] -> Tx t

Extract transaction resolved inputs. This slightly breaks the abstraction boundary of DefineTx which doesn't really force any structure on the resolved inputs. However, here in the DB-Layer, supporting arbitrary shape for the inputs be much more complex and require quite a lot of work. So, we kinda force the format here and only here, and leave the rest of the code with an opaque ResolvedTxIn type.