Cardano.Wallet.DB.Sqlite

Description

An implementation of the DBLayer which uses Persistent and SQLite.

Documentation

newDBLayer

Arguments

:: (IsOurs s, NFData s, Show s, PersistState s, TxId t)

=> Maybe FilePath
Database file location, or Nothing for in-memory database

-> IO (DBLayer IO s t)

Sets up a connection to the SQLite database.

Database migrations are run to create tables if necessary.

If the given file path does not exist, it will be created by the sqlite library.