Cardano.Wallet.DB.Sqlite

Description

An implementation of the DBLayer which uses Persistent and SQLite.

Documentation

newDBLayer

Arguments

:: (IsOurs s, NFData s, Show s, PersistState s, TxId t)

=> Trace IO Text
Logging object

-> Maybe FilePath
Database file location, or Nothing for in-memory database

-> IO (SqlBackend, DBLayer IO s t)

Sets up a connection to the SQLite database.

Database migrations are run to create tables if necessary.

If the given file path does not exist, it will be created by the sqlite library.

runQuery :: SqlBackend -> SqlPersistM a -> IO a

Run a query without error handling. There will be exceptions thrown if it fails.

class PersistState s where

Functions for saving/loading the wallet's address discovery state into SQLite.

Methods

insertState :: (WalletId, SlotId) -> s -> SqlPersistM ()

Store the state for a checkpoint.

selectState :: (WalletId, SlotId) -> SqlPersistM (Maybe s)

Load the state for a checkpoint.

deleteState :: WalletId -> SqlPersistM ()

Remove the state for all checkpoints of a wallet.