Cardano.Wallet.DB.Sqlite.TH

Description

Auto-generated Sqlite & Persistent machinery via Template-Haskell. This has been moved into a separate file so that we can treat it slightly differently when computing code-coverage.

More than 6K lines end-up being generated from the instructions below! As a result, we're going to ignore code-coverage on the following module and, no hand-written functions should be written in this module!

Documentation

data Wallet

Constructors

Wallet

Fields

walTableId :: !WalletId

walTableName :: !Text

walTablePassphraseLastUpdatedAt :: !(Maybe UTCTime)

walTableStatus :: !WalletState

walTableDelegation :: !(Maybe Text)

type WalletId = Key Wallet

data PrivateKey

Constructors

PrivateKey

Fields

privateKeyTableWalletId :: !WalletId

privateKeyTableRootKey :: !ByteString

privateKeyTableHash :: !ByteString

type PrivateKeyId = Key PrivateKey

data TxMeta

Constructors

TxMeta

Fields

txMetaTableTxId :: !TxId

txMetaTableWalletId :: !WalletId

txMetaTableStatus :: !TxStatus

txMetaTableDirection :: !Direction

txMetaTableSlotId :: !SlotId

txMetaTableAmount :: !Natural

type TxMetaId = Key TxMeta

data TxIn

Constructors

TxIn

Fields

txInputTableTxId :: !TxId

txInputTableSourceTxId :: !TxId

txInputTableSourceIndex :: !Word32

type TxInId = Key TxIn

data TxOut

Constructors

TxOut

Fields

txOutputTableTxId :: !TxId

txOutputTableIndex :: !Word32

txOutputTableAddress :: !Address

txOutputTableAmount :: !Coin

type TxOutId = Key TxOut

data Checkpoint

Constructors

Checkpoint

Fields

checkpointTableWalletId :: !WalletId

checkpointTableSlot :: !SlotId

type CheckpointId = Key Checkpoint

data UTxO

Constructors

UTxO

Fields

utxoTableWalletId :: !WalletId

utxoTableCheckpointSlot :: !SlotId

utxoTableInputId :: !TxId

utxoTableInputIndex :: !Word32

utxoTableOutputAddress :: !Address

utxoTableOutputCoin :: !Coin

type UTxOId = Key UTxO

migrateAll :: Migration

fk_checkpoint_utxo :: UTxO -> Key Checkpoint

fk_wallet_checkpoint :: Checkpoint -> Key Wallet

fk_wallet_tx_meta :: TxMeta -> Key Wallet

fk_wallet_private_key :: PrivateKey -> Key Wallet