Cardano.Wallet.DB.Sqlite.Types

Description

This module contains instances and types necessary for storing wallets in a SQL database with Persistent.

It's in a separate module due to the GHC stage restriction.

The ToJSON/FromJSON and Read instance orphans exist due to class constraints on Persistent functions.

Documentation

sqlSettings' :: MkPersistSettings

Settings for generating the Persistent types.

fromText' :: FromText a => Text -> Either Text a

fromText but with a simpler error type.

aesonFromText :: FromText a => String -> Value -> Parser a

Aeson parser defined in terms of fromText

fromPersistValueFromText :: FromText a => PersistValue -> Either Text a

fromPersistValue defined in terms of fromText

fromPersistValueRead :: Read a => PersistValue -> Either Text a

fromPersistValue defined in terms of the Read class

directionToBool :: Direction -> Bool

directionFromBool :: Bool -> Direction

newtype TxId

Constructors

TxId

Fields

getTxId :: Hash "Tx"

newtype BlockId

Constructors

BlockId

Fields

getBlockId :: Hash "BlockHeader"

artificialEpochLength :: EpochLength

As a short-to-medium term solution of persisting SlotId, we use flatSlot with an artificial epochLength. I.e. /not the same epochLength as the blockchain/. This is just for the sake of storing the 64 bit epoch and the 16 bit slot inside a single 64-bit field.

walletStateNum :: WalletState -> Word8

walletStateFromNum :: Word8 -> WalletState

mkCoin :: Word64 -> Either Text Coin

newtype AddressPoolXPub

Constructors

AddressPoolXPub

Fields

getAddressPoolXPub :: SeqKey AccountK XPub

Orphan instances

Read WalletState

Instance details

Methods

readsPrec :: Int -> ReadS WalletState

readList :: ReadS [WalletState]

readPrec :: ReadPrec WalletState

readListPrec :: ReadPrec [WalletState]

Read WalletId

Instance details

Methods

readsPrec :: Int -> ReadS WalletId

readList :: ReadS [WalletId]

readPrec :: ReadPrec WalletId

readListPrec :: ReadPrec [WalletId]

FromJSON SlotId

Instance details

Methods

parseJSON :: Value -> Parser SlotId

parseJSONList :: Value -> Parser [SlotId]

FromJSON WalletId

Instance details

Methods

parseJSON :: Value -> Parser WalletId

parseJSONList :: Value -> Parser [WalletId]

ToJSON SlotId

Instance details

Methods

toJSON :: SlotId -> Value

toEncoding :: SlotId -> Encoding

toJSONList :: [SlotId] -> Value

toEncodingList :: [SlotId] -> Encoding

ToJSON WalletId

Instance details

Methods

toJSON :: WalletId -> Value

toEncoding :: WalletId -> Encoding

toJSONList :: [WalletId] -> Value

toEncodingList :: [WalletId] -> Encoding

PersistField StdGen

Instance details

Methods

toPersistValue :: StdGen -> PersistValue

fromPersistValue :: PersistValue -> Either Text StdGen

PersistField SlotId

Instance details

Methods

toPersistValue :: SlotId -> PersistValue

fromPersistValue :: PersistValue -> Either Text SlotId

PersistField Coin

Instance details

Methods

toPersistValue :: Coin -> PersistValue

fromPersistValue :: PersistValue -> Either Text Coin

PersistField Address

Instance details

Methods

toPersistValue :: Address -> PersistValue

fromPersistValue :: PersistValue -> Either Text Address

PersistField Direction

Instance details

Methods

toPersistValue :: Direction -> PersistValue

fromPersistValue :: PersistValue -> Either Text Direction

PersistField TxStatus

Instance details

Methods

toPersistValue :: TxStatus -> PersistValue

fromPersistValue :: PersistValue -> Either Text TxStatus

PersistField WalletState

Instance details

Methods

toPersistValue :: WalletState -> PersistValue

fromPersistValue :: PersistValue -> Either Text WalletState

PersistField WalletId

Instance details

Methods

toPersistValue :: WalletId -> PersistValue

fromPersistValue :: PersistValue -> Either Text WalletId

PersistField ChangeChain

Instance details

Methods

toPersistValue :: ChangeChain -> PersistValue

fromPersistValue :: PersistValue -> Either Text ChangeChain

PersistField AddressPoolGap

Instance details

Methods

toPersistValue :: AddressPoolGap -> PersistValue

fromPersistValue :: PersistValue -> Either Text AddressPoolGap

PersistFieldSql StdGen

Instance details

Methods

sqlType :: Proxy StdGen -> SqlType

PersistFieldSql SlotId

Instance details

Methods

sqlType :: Proxy SlotId -> SqlType

PersistFieldSql Coin

Instance details

Methods

sqlType :: Proxy Coin -> SqlType

PersistFieldSql Address

Instance details

Methods

sqlType :: Proxy Address -> SqlType

PersistFieldSql Direction

Instance details

Methods

sqlType :: Proxy Direction -> SqlType

PersistFieldSql TxStatus

Instance details

Methods

sqlType :: Proxy TxStatus -> SqlType

PersistFieldSql WalletState

Instance details

Methods

sqlType :: Proxy WalletState -> SqlType

PersistFieldSql WalletId

Instance details

Methods

sqlType :: Proxy WalletId -> SqlType

PersistFieldSql ChangeChain

Instance details

Methods

sqlType :: Proxy ChangeChain -> SqlType

PersistFieldSql AddressPoolGap

Instance details

Methods

sqlType :: Proxy AddressPoolGap -> SqlType

FromHttpApiData WalletId

Instance details

Methods

parseUrlPiece :: Text -> Either Text WalletId

parseHeader :: ByteString -> Either Text WalletId

parseQueryParam :: Text -> Either Text WalletId

ToHttpApiData WalletId

Instance details

Methods

toUrlPiece :: WalletId -> Text

toEncodedUrlPiece :: WalletId -> Builder

toHeader :: WalletId -> ByteString

toQueryParam :: WalletId -> Text

PathPiece WalletId

Instance details

Methods

fromPathPiece :: Text -> Maybe WalletId

toPathPiece :: WalletId -> Text