Cardano.Wallet.DaedalusIPC

Description

Provides a mechanism for Daedalus to discover what port the cardano-wallet server is listening on.

See https://nodejs.org/api/child_process.html#child_process_child_process_spawn_command_args_options for more information about the message protocol.

Documentation

daedalusIPC

Arguments

:: Trace IO Text
Logging object

-> Int
Port number to send to Daedalus

-> IO ()

Start up the Daedalus IPC process. It's called daedalusIPC, but this could be any nodejs program that needs to start cardano-wallet. All it does is reply with a port number when asked, using a very nodejs-specific IPC method.

If the IPC channel was successfully set up, this function won't return until the parent process exits. Otherwise, it will return immediately. Before returning, it will log an message about why it has exited.