Cardano.Wallet.Logging

Description

This module contains utility functions for logging and mapping trace data.

Documentation

transformTextTrace :: ToText a => Trace IO Text -> Trace IO a

Converts a Text trace into any other type of trace that has a ToText instance.

logTrace :: (MonadIO m, DefinePrivacyAnnotation a, DefineSeverity a) => Trace m a -> a -> m ()

Traces some data.

fromLogObject :: Show a => Tracer IO a -> Tracer IO (LogObject a)

Strips out the data that trMessage (or "Cardano.BM.Data.Tracer.toLogObject") adds.

trMessage :: (MonadIO m, DefinePrivacyAnnotation a, DefineSeverity a) => Tracer m (LogObject a) -> Tracer m a

Tracer transformer which converts 'Trace m a' to 'Tracer m a' by wrapping typed log messages into a LogObject.

trMessageText :: (MonadIO m, ToText a, DefinePrivacyAnnotation a, DefineSeverity a) => Tracer m (LogObject Text) -> Tracer m a

Tracer transformer which transforms traced items to their ToText representation and further traces them as a LogObject. If the ToText representation is empty, then no tracing happens.

filterTraceSeverity :: forall m a. Monad m => Severity -> Trace m a -> Trace m a

Tracer transformer which removes tracing below the qgiven severity limit.

stdoutTextTracer :: (MonadIO m, ToText a) => Tracer m a

Creates a tracer that prints any ToText log message. This is useful for debugging functions in the REPL, when you need a Tracer object.

Orphan instances

(MonadIO m, ToText a, DefinePrivacyAnnotation a, DefineSeverity a) => Transformable Text m a

Instance details

Methods

trTransformer :: TracingFormatting -> TracingVerbosity -> Tracer m (LogObject Text) -> Tracer m a