Cardano.Wallet.Logging

Description

This module contains utility functions for logging and mapping trace data.

Documentation

transformTextTrace :: ToText a => Trace IO Text -> Trace IO a

Converts a Text trace into any other type of trace that has a ToText instance.

logTrace :: (MonadIO m, DefinePrivacyAnnotation a, DefineSeverity a) => Trace m a -> a -> m ()

Traces some data.