Cardano.Wallet.Network

Interface

data NetworkLayer t m

Constructors

NetworkLayer

Fields

nextBlocks :: BlockHeader -> ExceptT ErrGetBlock m [Block (Tx t)]

Fetches a contiguous sequence of blocks from the node, starting from the first block available with a slot greater than the given block header.

Blocks are returned in ascending slot order, without skipping blocks.

This function will not necessarily return all blocks available after the given point in time, but will return a reasonably-sized sequence.

It may return the empty list if the node does not have any blocks after the specified starting slot.

networkTip :: ExceptT ErrNetworkTip m (BlockHeader, Quantity "block" Natural)

Get the current network tip from the chain producer and the current chain's height.

postTx :: (Tx t, [TxWitness]) -> ExceptT ErrPostTx m ()

Broadcast a transaction to the chain producer

decodeExternalTx :: ByteString -> ExceptT ErrDecodeExternalTx m (Tx t, [TxWitness])

Decode an externally-signed transaction to the chain producer

Helpers

waitForConnection :: NetworkLayer t IO -> RetryPolicyM IO -> IO ()

Wait until 'networkTip networkLayer' succeeds according to a given retry policy. Throws an exception otherwise.

defaultRetryPolicy :: Monad m => RetryPolicyM m

A default RetryPolicy with a constant delay, but retries for no longer than a minute.

Errors

data ErrNetworkUnavailable

Network is unavailable

Constructors

ErrNetworkUnreachable Text
Cannot connect to network backend.

ErrNetworkInvalid Text
Network backend reports that the requested network is invalid.

data ErrNetworkTip

Error while trying to get the network tip

Constructors

ErrNetworkTipNetworkUnreachable ErrNetworkUnavailable

ErrNetworkTipNotFound

data ErrGetBlock

Error while trying to get one or more blocks

Constructors

ErrGetBlockNetworkUnreachable ErrNetworkUnavailable

ErrGetBlockNotFound (Hash "BlockHeader")

data ErrPostTx

Error while trying to send a transaction

Constructors

ErrPostTxNetworkUnreachable ErrNetworkUnavailable

ErrPostTxBadRequest Text

ErrPostTxProtocolFailure Text

data ErrDecodeExternalTx

Error while trying to decode externally signed transaction

Constructors

ErrDecodeExternalTxWrongPayload Text

ErrDecodeExternalTxNotSupported