Cardano.Wallet.Network

Interface

data NetworkLayer t m

Constructors

NetworkLayer

Fields

nextBlocks :: SlotId -> ExceptT ErrNetworkUnreachable m [Block]

Gets some blocks from the node. It will not necessarily return all the blocks that the node has, but will receive a reasonable-sized chunk. It will never return blocks from before the given slot. It may return an empty list if the node does not have any blocks from after the starting slot.

networkTip :: ExceptT ErrNetworkTip m (Hash "BlockHeader", BlockHeader)

Get the current network tip from the chain producer

postTx :: (Tx, [TxWitness]) -> ExceptT ErrPostTx m ()

Broadcast a transaction to the chain producer

Helpers

waitForConnection :: NetworkLayer t IO -> RetryPolicyM IO -> IO ()

Wait until 'networkTip networkLayer' succeeds according to a given retry policy. Throws an exception otherwise.

defaultRetryPolicy :: Monad m => RetryPolicyM m

Errors

newtype ErrNetworkUnreachable

Network is not reachable

Constructors

ErrNetworkUnreachable Text

data ErrNetworkTip

Error while trying to get the network tip

Constructors

ErrNetworkTipNetworkUnreachable ErrNetworkUnreachable

ErrNetworkTipNotFound

data ErrPostTx

Error while trying to send a transaction

Constructors

ErrPostTxNetworkUnreachable ErrNetworkUnreachable

ErrPostTxBadRequest Text

ErrPostTxProtocolFailure Text