Cardano.Wallet.Network

Interface

data NetworkLayer m target block

Constructors

NetworkLayer

Fields

nextBlocks :: Cursor target -> ExceptT ErrGetBlock m (NextBlocksResult target block)

Starting from the given Cursor, fetches a contiguous sequence of blocks from the node, if they are available. An updated cursor will be returned with a RollFoward result.

Blocks are returned in ascending slot order, without skipping blocks.

If the node does not have any blocks after the specified cursor point, it will return AwaitReply.

If the node has adopted an alternate fork of the chain, it will return RollBackward with a new cursor.

findIntersection :: Cursor target -> m (Maybe BlockHeader)

Attempt to find an intersection between the node's unstable blocks and a given list of headers. This can be useful if we need to know whether we are 'in sync' with the node or, close enough.

initCursor :: [BlockHeader] -> Cursor target

Creates a cursor from the given block header so that nextBlocks can be used to fetch blocks.

cursorSlotId :: Cursor target -> SlotId

Get the slot corresponding to a cursor.

networkTip :: ExceptT ErrNetworkTip m BlockHeader

Get the current network tip from the chain producer

postTx :: (Tx, [TxWitness]) -> ExceptT ErrPostTx m ()

Broadcast a transaction to the chain producer

staticBlockchainParameters :: (Block, BlockchainParameters)

Get the genesis block and blockchain parameters.

Note: The genesis block is a Wallet Block-type and not the block type-parameter used in nextBlocks. This is due to: 1. Cardano.Pool.Metrics wanting to use a block-type specific to non-genesis blocks. 2. The genesis block only being needed in the Wallet (currently).

stakeDistribution :: ExceptT ErrNetworkUnavailable m (EpochNo, Map PoolId (Quantity "lovelace" Word64))

data NextBlocksResult target block

The result of nextBlocks, which is instructions for what the chain consumer should do next.

Constructors

AwaitReply
There are no blocks available from the node, so wait.

RollForward (Cursor target) BlockHeader [block]
Apply the given contiguous non-empty sequence of blocks. Use the updated cursor to get the next batch. The given block header is the current tip of the node.

RollBackward (Cursor target)
The chain consumer must roll back its state, then use the cursor to get the next batch of blocks.

data family Cursor target

A cursor is local state kept by the chain consumer to use as the starting position for nextBlocks. The actual type is opaque and determined by the backend target.

follow

Arguments

:: (Show e0, Show e1)

=> NetworkLayer IO target block
The NetworkLayer used to poll for new blocks.

-> Trace IO Text
Logger trace

-> [BlockHeader]
A list of known tips to start from. Blocks after the tip will be yielded.

-> (NonEmpty block -> BlockHeader -> IO (FollowAction e0))
Callback with blocks and the current tip of the node. follow stops polling and terminates if the callback errors.

-> (SlotId -> IO (FollowAction e1))
Callback with a point of rollback when needed. follow stops polling and terminates if the callback errors.

-> (block -> BlockHeader)
Getter on the abstract block type

-> IO ()

Subscribe to a blockchain and get called with new block (in order)!

data FollowAction err

FollowAction enables the callback of follow to signal if the chain-following should ExitWith, Continue, or if the current callback should be forgotten and retried (Retry).

NOTE: Retry is needed to handle data-races in Metrics, where it is essensial that we fetch the stake distribution while the node-tip

Constructors

ExitWith err
Stop following the chain.

Continue
Continue following the chain.

RetryImmediately
Forget about the blocks in the current callback, and retry immediately.

RetryLater
Like RetryImmediately but only retries after a short delay

Errors

data ErrNetworkUnavailable

Network is unavailable

Constructors

ErrNetworkUnreachable Text
Cannot connect to network backend.

ErrNetworkInvalid Text
Network backend reports that the requested network is invalid.

data ErrNetworkTip

Error while trying to get the network tip

Constructors

ErrNetworkTipNetworkUnreachable ErrNetworkUnavailable

ErrNetworkTipNotFound

data ErrGetBlock

Error while trying to get one or more blocks

Constructors

ErrGetBlockNetworkUnreachable ErrNetworkUnavailable

ErrGetBlockNotFound (Hash "BlockHeader")

data ErrPostTx

Error while trying to send a transaction

Constructors

ErrPostTxNetworkUnreachable ErrNetworkUnavailable

ErrPostTxBadRequest Text

ErrPostTxProtocolFailure Text

Initialization

defaultRetryPolicy :: Monad m => RetryPolicyM m

A default RetryPolicy with a delay that starts short, and that retries for no longer than a minute.

waitForNetwork :: ExceptT ErrNetworkUnavailable IO () -> RetryPolicyM IO -> IO ()

Wait until 'networkTip networkLayer' succeeds according to a given retry policy. Throws an exception otherwise.