Cardano.Wallet.Network.Ports

Description

Provides functions for checking if TCP ports can be connected to, or are available to listen on.

These can be used for: - Waiting until a server in another process has started. - Start servers for testing when there may be multiple test suites running in parallel.

Allocation

data PortNumber

getRandomPort :: IO PortNumber

Find a TCPv4 port which is likely to be free for listening on localhost. This binds a socket, receives an OS-assigned port, then closes the socket.

Note that this is vulnerable to race conditions if another process binds the port returned by getRandomPort before this process does.

Do not use this unless you have no other option.

Status

isPortOpen :: SockAddr -> IO Bool

Checks whether connect() to a given TCPv4 SockAddr succeeds or returns eCONNREFUSED.

Rethrows connection exceptions in all other cases (e.g. when the host is unroutable).

Code courtesy of nh2: https://stackoverflow.com/a/57022572

simpleSockAddr :: (Word8, Word8, Word8, Word8) -> PortNumber -> SockAddr

Creates a SockAttr from host IP and port number.

Example: > simpleSockAddr (127,0,0,1) 8000

Helpers

waitForPort :: RetryPolicyM IO -> PortNumber -> IO Bool

Wait until a TCP port is open to connections according to a given retry policy. Throws an exception if the time out is reached.