Cardano.Wallet.Primitive.AddressDerivation

Description

Primitives for performing address derivation for some given schemes. This is where most of the crypto happens in the wallet and, it is quite important to ensure that the implementations match with other Cardano wallets (like cardano-sl, Yoroi/Icarus, or cardano-cli)

The actual implementations are in the following modules:

Cardano.Wallet.Primitive.AddressDerivation.Sequential

Cardano.Wallet.Primitive.AddressDerivation.Random

Polymorphic / General Purpose Types

data Depth

Key Depth in the derivation path, according to BIP-0039 / BIP-0044

m | purpose' | cointype' | account' | change | address

We do not manipulate purpose, cointype and change paths directly, so they are left out of the sum type.

Constructors

RootK

AccountK

AddressK

newtype Index (derivationType :: DerivationType) (level :: Depth)

A derivation index, with phantom-types to disambiguate derivation type.

let accountIx = Index 'Hardened 'AccountK
let addressIx = Index 'Soft 'AddressK

Constructors

Index

Fields

getIndex :: Word32

data DerivationType

Type of derivation that should be used with the given indexes.

Constructors

Hardened

Soft

data XPub

data XPrv

Backends Interoperability

class WalletKey key => KeyToAddress target (key :: Depth -> * -> *) where

Encoding of addresses for certain key types and backend targets.

Methods

keyToAddress :: key AddressK XPub -> Address

Convert a public key to an Address depending on a particular target.

Note that keyToAddress is ambiguous and requires therefore a type application. See also TxId.

class WalletKey (key :: Depth -> * -> *) where

Methods

changePassphrase :: Passphrase "encryption-old" -> Passphrase "encryption-new" -> key depth XPrv -> key depth XPrv

Re-encrypt a private key using a different passphrase.

*Important**: This function doesn't check that the old passphrase is correct! Caller is expected to have already checked that. Using an incorrect passphrase here will lead to very bad thing.

publicKey :: key depth XPrv -> key depth XPub

Extract the public key part of a private key.

digest :: HashAlgorithm a => key depth XPub -> Digest a

Hash a public key to some other representation.

getRawKey :: key depth raw -> raw

Unwrap the WalletKey to use the XPrv or XPub.

dummyKey :: key AddressK XPub

Produce a fake address key of this scheme, for use in dummyAddress.

class WalletKey key => PersistKey (key :: Depth -> * -> *) where

Operations for saving a WalletKey into a database, and restoring it from a database. The keys should be encoded in hexadecimal strings.

Methods

serializeXPrv :: (key RootK XPrv, Hash "encryption") -> (ByteString, ByteString)

Convert a private key and its password hash into hexadecimal strings suitable for storing in a text file or database column.

deserializeXPrv :: (ByteString, ByteString) -> Either String (key RootK XPrv, Hash "encryption")

The reverse of serializeXPrv. This may fail if the inputs are not valid hexadecimal strings, or if the key is of the wrong length.

dummyAddress :: forall target key. KeyToAddress target key => Address

Produce a fake address of representative size for the target and key type. This can be used in transaction size estimations.

This function is ambiguous, like keyToAddress, and types need to be applied.

Passphrase

newtype Passphrase (purpose :: Symbol)

An encapsulated passphrase. The inner format is free, but the wrapper helps readability in function signatures.

Constructors

Passphrase ScrubbedBytes

class PassphraseMinLength (purpose :: Symbol) where

Methods

passphraseMinLength :: Proxy purpose -> Int

Minimal Length for a passphrase, for lack of better validations

class PassphraseMaxLength (purpose :: Symbol) where

Methods

passphraseMaxLength :: Proxy purpose -> Int

Maximum length for a passphrase

class FromMnemonic (sz :: [Nat]) (purpose :: Symbol) where

Create a passphrase from a mnemonic sentence. This class enables caller to parse text list of variable length into mnemonic sentences.

>>> fromMnemonic @'[12,15,18,21] @"generation" ["toilet", "curse", ... ]
Right (Passphrase <ScrubbedBytes>)

Note that the given Nats **have** to be valid mnemonic sizes, otherwise the underlying code won't even compile, with not-so-friendly error messages.

Methods

fromMnemonic :: [Text] -> Either (FromMnemonicError sz) (Passphrase purpose)

newtype FromMnemonicError (sz :: [Nat])

Error reported from trying to create a passphrase from a given mnemonic

Constructors

FromMnemonicError

Fields

getFromMnemonicError :: String

data ErrWrongPassphrase

Indicate a failure when checking for a given Passphrase match

Constructors

ErrWrongPassphrase

encryptPassphrase :: MonadRandom m => Passphrase purpose -> m (Hash purpose)

Encrypt a Passphrase into a format that is suitable for storing on disk

checkPassphrase :: Passphrase purpose -> Hash purpose -> Either ErrWrongPassphrase ()

Check whether a Passphrase matches with a stored Hash