Cardano.Wallet.Primitive.AddressDerivation

Description

Primitives for performing address derivation for some given schemes. This is where most of the crypto happens in the wallet and, it is quite important to ensure that the implementations match with other Cardano wallets (like cardano-sl, Yoroi/Icarus, or cardano-cli)

The actual implementations are in the following modules:

Cardano.Wallet.Primitive.AddressDerivation.Shelley

Cardano.Wallet.Primitive.AddressDerivation.Byron

HD Derivation

data Depth

Key Depth in the derivation path, according to BIP-0039 / BIP-0044

m | purpose' | cointype' | account' | change | address

We do not manipulate purpose, cointype and change paths directly, so they are left out of the sum type.

Constructors

RootK

AccountK

AddressK

newtype Index (derivationType :: DerivationType) (level :: Depth)

A derivation index, with phantom-types to disambiguate derivation type.

let accountIx = Index 'Hardened 'AccountK
let addressIx = Index 'Soft 'AddressK

Constructors

Index

Fields

getIndex :: Word32

data ChangeChain

Marker for the change chain. In practice, change of a transaction goes onto the addresses generated on the internal chain, whereas the external chain is used for addresses that are part of the advertised targets of a transaction

Constructors

ExternalChain

InternalChain

data DerivationType

Type of derivation that should be used with the given indexes.

Constructors

Hardened

Soft

class HardDerivation (key :: Depth -> * -> *) where

An interface for doing hard derivations from the root private key

Associated Types

type AddressIndexDerivationType key :: DerivationType

Methods

deriveAccountPrivateKey :: Passphrase "encryption" -> key RootK XPrv -> Index Hardened AccountK -> key AccountK XPrv

Derives account private key from the given root private key, using derivation scheme 2 (see cardano-crypto package for more details).

NOTE: The caller is expected to provide the corresponding passphrase (and to have checked that the passphrase is valid). Providing a wrong passphrase will not make the function fail but will instead, yield an incorrect new key that doesn't belong to the wallet.

deriveAddressPrivateKey :: Passphrase "encryption" -> key AccountK XPrv -> ChangeChain -> Index (AddressIndexDerivationType key) AddressK -> key AddressK XPrv

Derives address private key from the given account private key, using derivation scheme 2 (see cardano-crypto package for more details).

It is preferred to use deriveAddressPublicKey whenever possible to avoid having to manipulate passphrases and private keys.

NOTE: The caller is expected to provide the corresponding passphrase (and to have checked that the passphrase is valid). Providing a wrong passphrase will not make the function fail but will instead, yield an incorrect new key that doesn't belong to the wallet.

class HardDerivation key => SoftDerivation (key :: Depth -> * -> *) where

An interface for doing soft derivations from an account public key

Methods

deriveAddressPublicKey :: key AccountK XPub -> ChangeChain -> Index Soft AddressK -> key AddressK XPub

Derives address public key from the given account public key, using derivation scheme 2 (see cardano-crypto package for more details).

This is the preferred way of deriving new sequential address public keys.

Primitive Crypto Types

data XPub

Constructors

XPub

Fields

xpubPublicKey :: !ByteString

xpubChaincode :: !ChainCode

newtype ChainCode

Constructors

ChainCode ByteString

data XPrv

unXPub :: XPub -> ByteString

hex :: ByteArrayAccess bin => bin -> ByteString

Encode a ByteString in base16

fromHex :: ByteArray bout => ByteString -> Either String bout

Decode a ByteString from base16

Network Discrimination

data NetworkDiscriminant

Available network options.

Constructors

Mainnet

Testnet

class NetworkDiscriminantVal (n :: NetworkDiscriminant)

Minimal complete definition

networkDiscriminantVal

networkDiscriminantVal :: NetworkDiscriminantVal n => NetworkDiscriminant

Backends Interoperability

class MkKeyFingerprint key => PaymentAddress (network :: NetworkDiscriminant) key where

Encoding of addresses for certain key types and backend targets.

Methods

paymentAddress :: key AddressK XPub -> Address

Convert a public key to a payment Address valid for the given network discrimination.

Note that paymentAddress is ambiguous and requires therefore a type application.

liftPaymentFingerprint :: KeyFingerprint "payment" key -> Address

Lift a payment fingerprint back into an address.

class PaymentAddress network key => DelegationAddress (network :: NetworkDiscriminant) key where

Methods

delegationAddress

Arguments

:: key AddressK XPub
Payment key

-> key AddressK XPub
Staking key

-> Address

Convert a public key and a staking key to a delegation Address valid for the given network discrimination. Funds sent to this address will be delegated according to the delegation settings attached to the delegation key.

Note that delegationAddress is ambiguous and requires therefore a type application.

class WalletKey (key :: Depth -> * -> *) where

Methods

changePassphrase :: Passphrase "encryption-old" -> Passphrase "encryption-new" -> key depth XPrv -> key depth XPrv

Re-encrypt a private key using a different passphrase.

*Important**: This function doesn't check that the old passphrase is correct! Caller is expected to have already checked that. Using an incorrect passphrase here will lead to very bad thing.

publicKey :: key depth XPrv -> key depth XPub

Extract the public key part of a private key.

digest :: HashAlgorithm a => key depth XPub -> Digest a

Hash a public key to some other representation.

keyTypeDescriptor :: Proxy key -> String

Get a short, human-readable string descriptor that uniquely identifies the specified key type.

getRawKey :: key depth raw -> raw

Unwrap the WalletKey to use the XPrv or XPub.

dummyKey :: key AddressK XPub

Produce a fake address key of this scheme, for use in dummyAddress.

class PersistPrivateKey (key :: * -> *) where

Operations for saving a private key into a database, and restoring it from a database. The keys should be encoded in hexadecimal strings.

Methods

serializeXPrv :: (key XPrv, Hash "encryption") -> (ByteString, ByteString)

Convert a private key and its password hash into hexadecimal strings suitable for storing in a text file or database column.

unsafeDeserializeXPrv :: (ByteString, ByteString) -> (key XPrv, Hash "encryption")

The reverse of serializeXPrv. This may fail if the inputs are not valid hexadecimal strings, or if the key is of the wrong length.

class PersistPublicKey (key :: * -> *) where

Operations for saving a public key into a database, and restoring it from a database. The keys should be encoded in hexadecimal strings.

Methods

serializeXPub :: key XPub -> ByteString

Convert a private key and its password hash into hexadecimal strings suitable for storing in a text file or database column.

unsafeDeserializeXPub :: ByteString -> key XPub

Convert a public key into hexadecimal strings suitable for storing in a text file or database column.

class MkKeyFingerprint (key :: Depth -> * -> *) where

Produce KeyFingerprint for existing types. A fingerprint here uniquely identifies part of an address. It can refer to either the payment key or, if any, the delegation key of an address.

The fingerprint obeys the following rules:

If two addresses are the same, then they have the same fingerprints

It is possible to lift the fingerprint back into an address

This second rule pretty much fixes what can be chosen as a fingerprint for various key types:

For ByronKey, it can only be the address itself!

For ShelleyKey, then the "payment" fingerprint refers to the payment key within a single or grouped address.

Methods

paymentKeyFingerprint :: Address -> KeyFingerprint "payment" key

delegationKeyFingerprint :: Address -> Maybe (KeyFingerprint "delegation" key)

newtype KeyFingerprint (s :: Symbol) key

Something that uniquely identifies a public key. Typically, a hash of that key or the key itself.

Constructors

KeyFingerprint ByteString

dummyAddress :: forall network key. (PaymentAddress network key, WalletKey key) => Address

Produce a fake address of representative size for the target and key type. This can be used in transaction size estimations.

This function is ambiguous, like paymentAddress, and types need to be applied.

Passphrase

newtype Passphrase (purpose :: Symbol)

An encapsulated passphrase. The inner format is free, but the wrapper helps readability in function signatures.

Constructors

Passphrase ScrubbedBytes

class PassphraseMinLength (purpose :: Symbol) where

Methods

passphraseMinLength :: Proxy purpose -> Int

Minimal Length for a passphrase, for lack of better validations

class PassphraseMaxLength (purpose :: Symbol) where

Methods

passphraseMaxLength :: Proxy purpose -> Int

Maximum length for a passphrase

class FromMnemonic (sz :: [Nat]) (purpose :: Symbol) where

Create a passphrase from a mnemonic sentence. This class enables caller to parse text list of variable length into mnemonic sentences.

>>> fromMnemonic @'[12,15,18,21] @"generation" ["toilet", "curse", ... ]
Right (Passphrase <ScrubbedBytes>)

Note that the given Nats **have** to be valid mnemonic sizes, otherwise the underlying code won't even compile, with not-so-friendly error messages.

Methods

fromMnemonic :: [Text] -> Either (FromMnemonicError sz) (Passphrase purpose)

newtype FromMnemonicError (sz :: [Nat])

Error reported from trying to create a passphrase from a given mnemonic

Constructors

FromMnemonicError

Fields

getFromMnemonicError :: String

data ErrWrongPassphrase

Indicate a failure when checking for a given Passphrase match

Constructors

ErrWrongPassphrase

encryptPassphrase :: MonadRandom m => Passphrase purpose -> m (Hash purpose)

Encrypt a Passphrase into a format that is suitable for storing on disk

checkPassphrase :: Passphrase purpose -> Hash purpose -> Either ErrWrongPassphrase ()

Check whether a Passphrase matches with a stored Hash