Cardano.Wallet.Primitive.AddressDerivation.Icarus

Description

Implementation of address derivation for Icarus keys. This uses the Byron derivation for addresses, but on top of the derivation scheme V2.

Types

newtype IcarusKey (depth :: Depth) key

A cryptographic key for sequential-scheme address derivation, with phantom-types to disambiguate key types.

let rootPrivateKey = IcarusKey 'RootK XPrv
let accountPubKey = IcarusKey 'AccountK XPub
let addressPubKey = IcarusKey 'AddressK XPub

Constructors

IcarusKey

Fields

getKey :: key

Generation and derivation

generateKeyFromSeed

Arguments

:: SomeMnemonic
The root mnemonic

-> Passphrase "encryption"
Master encryption passphrase

-> IcarusKey RootK XPrv

Generate a root key from a corresponding seed. The seed should be at least 16 bytes.

generateKeyFromHardwareLedger

Arguments

:: SomeMnemonic
The root mnemonic

-> Passphrase "encryption"
Master encryption passphrase

-> IcarusKey RootK XPrv

Hardware Ledger devices generates keys from mnemonic using a different approach (different from the rest of Cardano).

It is a combination of:

SLIP 0010

BIP 0032

BIP 0039

RFC 8032

What seems to be arbitrary changes from Ledger regarding the calculation of the initial chain code and generation of the root private key.

unsafeGenerateKeyFromSeed

Arguments

:: SomeMnemonic
The root mnemonic

-> Passphrase "encryption"
Master encryption passphrase

-> IcarusKey depth XPrv

Generate a new key from seed. Note that the depth is left open so that the caller gets to decide what type of key this is. This is mostly for testing, in practice, seeds are used to represent root keys, and one should use generateKeyFromSeed.

minSeedLengthBytes :: Int

The minimum seed length for generateKeyFromSeed and unsafeGenerateKeyFromSeed.