Cardano.Wallet.Primitive.AddressDerivation.Random

Description

Implementation of address derivation for the random scheme, as implemented by the legacy Cardano wallets.

For full documentation of the key derivation schemes, see the Cardano.Crypto.Wallet module, and the implementation in cardano-crypto.

Key derivation and generation

deriveAccountPrivateKey :: Passphrase "encryption" -> Key RootK XPrv -> Index Hardened AccountK -> Key AccountK XPrv

Derives account private key from the given root private key, using derivation scheme 1.

NOTE: The caller is expected to provide the corresponding passphrase (and to have checked that the passphrase is valid). Providing a wrong passphrase will not make the function fail but will instead, yield an incorrect new key that doesn't belong to the wallet.

deriveAddressPrivateKey :: Passphrase "encryption" -> Key AccountK XPrv -> Index Soft AddressK -> Key AddressK XPrv

Derives address private key from the given account private key, using derivation scheme 1.

NOTE: The caller is expected to provide the corresponding passphrase (and to have checked that the passphrase is valid). Providing a wrong passphrase will not make the function fail but will instead, yield an incorrect new key that doesn't belong to the wallet.

generateKeyFromSeed :: Passphrase "seed" -> Passphrase "encryption" -> Key RootK XPrv

Generate a root key from a corresponding seed. The seed should be at least 16 (minSeedLengthBytes) bytes.

unsafeGenerateKeyFromSeed :: Passphrase "seed" -> Passphrase "encryption" -> Key depth XPrv

Generate a new key from seed. Note that the depth is left open so that the caller gets to decide what type of key this is. This is mostly for testing, in practice, seeds are used to represent root keys, and one should use generateKeyFromSeed.

minSeedLengthBytes :: Int

The amount of entropy carried by a BIP-39 12-word mnemonic is 16 bytes.

Address encoding/decoding

encodeDerivationPath :: Key RootK XPub -> Index Hardened AccountK -> Index Soft AddressK -> Encoding

Encode the derivation path (account index and address index) part of a random HD address.

This is the opposite of decodeDerivationPath.

NOTE: The caller must ensure that the key length is 32 bytes, which can be done by using the generateKeyFromSeed and publicKey functions.

decodeDerivationPath :: Key RootK XPub -> Decoder s (Maybe (Index Hardened AccountK, Index Soft AddressK))

Decode the HD random derivation path section of an address. If the public key is incorrect, the decode result will be Nothing.

This is the opposite of encodeDerivationPath.

decodeAddressDerivationPath :: Key RootK XPub -> Decoder s (Maybe (Index Hardened AccountK, Index Soft AddressK))

Decode the Addr Root + Attributes + Type section of a HD random scheme address, and return the derivation path.

It will fail to parse if the address is not in the random scheme. If the public key is incorrect, the decode result will be Nothing.