Cardano.Wallet.Primitive.AddressDerivation.Sequential

Description

Implementation of address derivation for the sequential schemes, as implemented by Yoroi/Icarus and cardano-cli.

Documentation

data ChangeChain

Marker for the change chain. In practice, change of a transaction goes onto the addresses generated on the internal chain, whereas the external chain is used for addresses that are part of the advertised targets of a transaction

Constructors

ExternalChain

InternalChain

generateKeyFromSeed

Arguments

:: (Passphrase "seed", Passphrase "generation")
The actual seed and its recovery / generation passphrase

-> Passphrase "encryption"

-> Key RootK XPrv

Generate a root key from a corresponding seed The seed should be at least 16 bytes

unsafeGenerateKeyFromSeed

Arguments

:: (Passphrase "seed", Passphrase "generation")
The actual seed and its recovery / generation passphrase

-> Passphrase "encryption"

-> Key depth XPrv

Generate a new key from seed. Note that the depth is left open so that the caller gets to decide what type of key this is. This is mostly for testing, in practice, seeds are used to represent root keys, and one should use generateKeyFromSeed.

deriveAccountPrivateKey :: Passphrase "encryption" -> Key RootK XPrv -> Index Hardened AccountK -> Key AccountK XPrv

Derives account private key from the given root private key, using derivation scheme 2 (see cardano-crypto package for more details).

NOTE: The caller is expected to provide the corresponding passphrase (and to have checked that the passphrase is valid). Providing a wrong passphrase will not make the function fail but will instead, yield an incorrect new key that doesn't belong to the wallet.

deriveAddressPrivateKey :: Passphrase "encryption" -> Key AccountK XPrv -> ChangeChain -> Index Soft AddressK -> Key AddressK XPrv

Derives address private key from the given account private key, using derivation scheme 2 (see cardano-crypto package for more details).

It is preferred to use deriveAddressPublicKey whenever possible to avoid having to manipulate passphrases and private keys.

NOTE: The caller is expected to provide the corresponding passphrase (and to have checked that the passphrase is valid). Providing a wrong passphrase will not make the function fail but will instead, yield an incorrect new key that doesn't belong to the wallet.

deriveAddressPublicKey :: Key AccountK XPub -> ChangeChain -> Index Soft AddressK -> Key AddressK XPub

Derives address public key from the given account public key, using derivation scheme 2 (see cardano-crypto package for more details).

This is the preferred way of deriving new sequential address public keys.

minSeedLengthBytes :: Int

The minimum seed length for generateKeyFromSeed and unsafeGenerateKeyFromSeed.