Cardano.Wallet.Primitive.AddressDerivation.Shelley

Description

Implementation of address derivation for Shelley Keys. Shelley really means Jörmungandr here.

Types

newtype ShelleyKey (depth :: Depth) key

A cryptographic key for sequential-scheme address derivation, with phantom-types to disambiguate key types.

let rootPrivateKey = ShelleyKey 'RootK XPrv
let accountPubKey = ShelleyKey 'AccountK XPub
let addressPubKey = ShelleyKey 'AddressK XPub

Constructors

ShelleyKey

Fields

getKey :: key

Constants

minSeedLengthBytes :: Int

The minimum seed length for generateKeyFromSeed and unsafeGenerateKeyFromSeed.

publicKeySize :: Int

Size, in bytes, of a public key (without chain code)

addrSingleSize :: Int

addrGroupedSize :: Int

class KnownNetwork (n :: NetworkDiscriminant) where

Embed some constants into a network type.

Minimal complete definition

addrSingle, addrGrouped, addrAccount

Methods

addrSingle

Arguments

:: Word8
Address discriminant byte for single addresses, this is the first byte of every addresses using the Shelley format carrying only a spending key.

addrGrouped

Arguments

:: Word8
Address discriminant byte for grouped addresses, this is the first byte of every addresses using the Shelley format carrying both a spending and an account key.

addrAccount

Arguments

:: Word8
Address discriminant byte for account addresses, this is the first byte of every addresses using the Shelley format carrying only an account key.

knownDiscriminants :: [Word8]

Generation and derivation

generateKeyFromSeed

Arguments

:: (Passphrase "seed", Passphrase "generation")
The actual seed and its recovery / generation passphrase

-> Passphrase "encryption"

-> ShelleyKey RootK XPrv

Generate a root key from a corresponding seed. The seed should be at least 16 bytes.

unsafeGenerateKeyFromSeed

Arguments

:: (Passphrase "seed", Passphrase "generation")
The actual seed and its recovery / generation passphrase

-> Passphrase "encryption"

-> ShelleyKey depth XPrv

Generate a new key from seed. Note that the depth is left open so that the caller gets to decide what type of key this is. This is mostly for testing, in practice, seeds are used to represent root keys, and one should use generateKeyFromSeed.

Address

decodeShelleyAddress :: forall n. (KnownNetwork n, NetworkDiscriminantVal n) => ByteString -> Either TextDecodingError Address

Verify the structure of a payload decoded from a Bech32 text string