Cardano.Wallet.Primitive.AddressDerivation.Shelley

Description

Implementation of address derivation for the sequential schemes, as implemented by Yoroi/Icarus and cardano-cli.

ShelleyKey types

newtype ShelleyKey (depth :: Depth) key

A cryptographic key for sequential-scheme address derivation, with phantom-types to disambiguate key types.

let rootPrivateKey = ShelleyKey 'RootK XPrv
let accountPubKey = ShelleyKey 'AccountK XPub
let addressPubKey = ShelleyKey 'AddressK XPub

Constructors

ShelleyKey

Fields

getKey :: key

data ChangeChain

Marker for the change chain. In practice, change of a transaction goes onto the addresses generated on the internal chain, whereas the external chain is used for addresses that are part of the advertised targets of a transaction

Constructors

ExternalChain

InternalChain

ShelleyKey generation and derivation

generateKeyFromSeed

Arguments

:: (Passphrase "seed", Passphrase "generation")
The actual seed and its recovery / generation passphrase

-> Passphrase "encryption"

-> ShelleyKey RootK XPrv

Generate a root key from a corresponding seed. The seed should be at least 16 bytes.

unsafeGenerateKeyFromSeed

Arguments

:: (Passphrase "seed", Passphrase "generation")
The actual seed and its recovery / generation passphrase

-> Passphrase "encryption"

-> ShelleyKey depth XPrv

Generate a new key from seed. Note that the depth is left open so that the caller gets to decide what type of key this is. This is mostly for testing, in practice, seeds are used to represent root keys, and one should use generateKeyFromSeed.

minSeedLengthBytes :: Int

The minimum seed length for generateKeyFromSeed and unsafeGenerateKeyFromSeed.

deriveAccountPrivateKey :: Passphrase "encryption" -> ShelleyKey RootK XPrv -> Index Hardened AccountK -> ShelleyKey AccountK XPrv

Derives account private key from the given root private key, using derivation scheme 2 (see cardano-crypto package for more details).

NOTE: The caller is expected to provide the corresponding passphrase (and to have checked that the passphrase is valid). Providing a wrong passphrase will not make the function fail but will instead, yield an incorrect new key that doesn't belong to the wallet.

deriveAddressPrivateKey :: Passphrase "encryption" -> ShelleyKey AccountK XPrv -> ChangeChain -> Index Soft AddressK -> ShelleyKey AddressK XPrv

Derives address private key from the given account private key, using derivation scheme 2 (see cardano-crypto package for more details).

It is preferred to use deriveAddressPublicKey whenever possible to avoid having to manipulate passphrases and private keys.

NOTE: The caller is expected to provide the corresponding passphrase (and to have checked that the passphrase is valid). Providing a wrong passphrase will not make the function fail but will instead, yield an incorrect new key that doesn't belong to the wallet.

deriveAddressPublicKey :: ShelleyKey AccountK XPub -> ChangeChain -> Index Soft AddressK -> ShelleyKey AddressK XPub

Derives address public key from the given account public key, using derivation scheme 2 (see cardano-crypto package for more details).

This is the preferred way of deriving new sequential address public keys.

Passphrase

changePassphraseSeq :: Passphrase "encryption-old" -> Passphrase "encryption-new" -> ShelleyKey depth XPrv -> ShelleyKey depth XPrv

Re-encrypt a private key using a different passphrase.

*Important**: This function doesn't check that the old passphrase is correct! Caller is expected to have already checked that. Using an incorrect passphrase here will lead to very bad thing.

Storing and retrieving keys

deserializeXPubSeq :: ByteString -> Either String (ShelleyKey purpose XPub)

The reverse of serializeXPub. This will fail if the input is not hexadecimal string of the correct length.

serializeXPubSeq :: ShelleyKey purpose XPub -> ByteString

Convert a public key into a hexadecimal string suitable for storing in a text file or database column.

Encoding / Decoding

decodeShelleyAddress :: forall n. (KnownNetwork n, NetworkDiscriminantVal n) => ByteString -> Either TextDecodingError Address

Verify the structure of a payload decoded from a Bech32 text string