Cardano.Wallet.Primitive.AddressDiscovery

Description

This module contains primitives necessary to perform address discovery. So far, we're only interested in address following a sequential derivation scheme as specified in BIP-0044. Later, we may introduce backward compatibility with random address scheme from the legacy Cardano wallets.

Abstractions

class IsOurs s where

This abstraction exists to give us the ability to keep the wallet business logic agnostic to the address derivation and discovery mechanisms.

This is needed because two different address schemes lives on Cardano:

A hierarchical random scheme: rather custom made, with several flaws; this is the original and now legacy address scheme.

A hierarchical sequential scheme: a new scheme based on the BIP-0044 specification, which is better suited for our present needs.

In practice, we will need a wallet that can support both, even if not at the same time, and this little abstraction can buy us this without introducing too much overhead.

Methods

isOurs

Arguments

:: Address

-> s

-> (Bool, s)
Checks whether an address is ours or not.

class IsOurs s => IsOwned s where

More powerful than isOurs, this abstractions offer the underlying state the ability to find / compute the address private key corresponding to a given known address.

Requiring IsOwned as a constraint supposed that there is a way to recover the root private key of a particular wallet. This isn't true for externally owned wallet which would delegate its key management to a third party (like a hardware Ledger or Trezor).

Methods

isOwned

Arguments

:: s

-> (Key RootK XPrv, Passphrase "encryption")

-> Address

-> Maybe (Key AddressK XPrv, Passphrase "encryption")
Derive the private key corresponding to an address. Careful, this operation can be costly. Note that the state is discarded from this function as we do not intend to discover any addresses from this operation; This is merely a lookup from known addresses.

class GenChange s where

Abstracting over change address generation. In theory, this is only needed for sending transactions on a wallet following a particular scheme. This abstractions allows for defining an heuristic to pick new change address. For instance, in BIP-44, change addresses belong to a particular change chain (also called "Internal Chain").

Methods

genChange

Arguments

:: s

-> (Address, s)
Generate a new change address for the given scheme. The rules for generating a new change address depends on the underlying scheme.

Sequential Derivation

Address Pool Gap

data AddressPoolGap

Maximum number of consecutive undiscovered addresses allowed

newtype MkAddressPoolGapError

Possible errors when casting to an AddressPoolGap

Constructors

ErrGapOutOfRange Word8

defaultAddressPoolGap :: AddressPoolGap

A default AddressPoolGap, as suggested in BIP-0044

getAddressPoolGap :: AddressPoolGap -> Word8

mkAddressPoolGap :: Word8 -> Either MkAddressPoolGapError AddressPoolGap

Smart constructor for AddressPoolGap

Address Pool

data AddressPool target (chain :: ChangeChain)

An AddressPool which keeps track of sequential addresses within a given Account and change chain. See mkAddressPool to create a new or existing pool:

>>> mkAddressPool xpub gap changeChain mempty
AddressPool { }

gap :: AddressPool target chain -> AddressPoolGap

The actual gap for the pool. This can't change for a given pool.

addresses :: AddressPool t chain -> [Address]

Get all addresses in the pool, sorted from the first address discovered, up until the next one.

In practice, we always have:

mkAddressPool key g cc (addresses pool) == pool

changeChain :: forall (chain :: ChangeChain). Typeable chain => ChangeChain

Bring a ChangeChain type back to the term-level. This requires a type application and either a scoped type variable, or an explicit passing of a ChangeChain.

>>> changeChain @'ExternalChain
ExternalChain

>>> changeChain @chain
...

accountPubKey :: AddressPool target chain -> Key AccountK XPub

Corresponding key for the pool (a pool is tied to only one account)

mkAddressPool :: forall t chain. (KeyToAddress t, Typeable chain) => Key AccountK XPub -> AddressPoolGap -> [Address] -> AddressPool t chain

Create a new Address pool from a list of addresses. Note that, the list is expected to be ordered in sequence (first indexes, first in the list).

The pool will grow from the start if less than g :: AddressPoolGap are given, such that, there are always g undiscovered addresses in the pool.

lookupAddress :: forall t chain. (KeyToAddress t, Typeable chain) => Address -> AddressPool t chain -> (Maybe (Index Soft AddressK), AddressPool t chain)

Lookup an address in the pool. When we find an address in a pool, the pool may be amended if the address was discovered near the edge. It is also possible that the pool is not amended at all - this happens in the case that an address is discovered far from the edge.

Pending Change Indexes

data PendingIxs

An ordered set of pending indexes. This keep track of indexes used

emptyPendingIxs :: PendingIxs

An empty pending set of change indexes.

NOTE: We do not define a Monoid instance here because there's no rational of combining two pending sets.

State

data SeqState t

A state to keep track of sequential addresses as described in BIP-44

Internally, the state keeps track of a few things for us and is it is parameterized by a type t which represents a target backend. This enables the state to be agnostic to the underlying address format.

Constructors

SeqState

Fields

internalPool :: !(AddressPool t InternalChain)

Addresses living on the InternalChain

externalPool :: !(AddressPool t ExternalChain)

Addresses living on the ExternalChain

pendingChangeIxs :: !PendingIxs

Indexes from the internal pool that have been used in pending transactions. The list is maintained sorted in descending order (cf: PendingIxs)

mkSeqState :: KeyToAddress t => (Key RootK XPrv, Passphrase "encryption") -> AddressPoolGap -> SeqState t

Construct a Sequential state for a wallet.