Cardano.Wallet.Primitive.AddressDiscovery.Random

Description

An implementation of address discovery for the random address scheme as used by the legacy Cardano wallets.

State

data RndState target

HD random address discovery state and key material for AD.

Constructors

RndState

Fields

rndKey :: RndKey RootK XPrv

The wallet root key.

accountIndex :: Index Hardened AccountK

The account index used for address _generation_ in this wallet. Note that addresses will be _discovered_ from any and all account indices, regardless of this value.

addresses :: Map DerivationPath Address

The addresses which have so far been discovered, and their derivation paths.

pendingAddresses :: Map DerivationPath Address

The addresses which have been generated for use as change addresses, and their derivation paths. Once addresses have been discovered they are removed from this set and added to addresses.

gen :: StdGen

The state of the RNG.

mkRndState :: RndKey RootK XPrv -> Int -> RndState t

Initialize the HD random address discovery state from a root key and RNG seed.