Cardano.Wallet.Primitive.AddressDiscovery.Random

Description

An implementation of address discovery for the random address scheme as used by the legacy Cardano wallets.

State

newtype RndState

Constructors

RndState ByteString

mkRndState :: RndState