Cardano.Wallet.Primitive.AddressDiscovery.Sequential

Description

An implementation of address discovery for the sequential address derivation scheme specified in BIP-0044.

The management of _accounts_ is left-out for this implementation focuses on a single account. In practice, one wants to manage a set of pools, one per account.

Sequential Derivation

Address Pool Gap

data AddressPoolGap

Maximum number of consecutive undiscovered addresses allowed

newtype MkAddressPoolGapError

Possible errors when casting to an AddressPoolGap

Constructors

ErrGapOutOfRange Integer

defaultAddressPoolGap :: AddressPoolGap

A default AddressPoolGap, as suggested in BIP-0044

getAddressPoolGap :: AddressPoolGap -> Word8

mkAddressPoolGap :: Integer -> Either MkAddressPoolGapError AddressPoolGap

Smart constructor for AddressPoolGap

Address Pool

data AddressPool target (chain :: ChangeChain)

An AddressPool which keeps track of sequential addresses within a given Account and change chain. See mkAddressPool to create a new or existing pool:

>>> mkAddressPool xpub gap changeChain mempty
AddressPool { }

gap :: AddressPool target chain -> AddressPoolGap

The actual gap for the pool. This can't change for a given pool.

addresses :: AddressPool t chain -> [Address]

Get all addresses in the pool, sorted from the first address discovered, up until the next one.

In practice, we always have:

mkAddressPool key g cc (addresses pool) == pool

changeChain :: forall (c :: ChangeChain). Typeable c => ChangeChain

Bring a ChangeChain type back to the term-level. This requires a type application and either a scoped type variable, or an explicit passing of a ChangeChain.

>>> changeChain @'ExternalChain
ExternalChain

>>> changeChain @chain
...

accountPubKey :: AddressPool target chain -> SeqKey AccountK XPub

Corresponding key for the pool (a pool is tied to only one account)

mkAddressPool :: forall t c. (KeyToAddress t SeqKey, Typeable c) => SeqKey AccountK XPub -> AddressPoolGap -> [Address] -> AddressPool t c

Create a new Address pool from a list of addresses. Note that, the list is expected to be ordered in sequence (first indexes, first in the list).

The pool will grow from the start if less than g :: AddressPoolGap are given, such that, there are always g undiscovered addresses in the pool.

lookupAddress :: forall t c. (KeyToAddress t SeqKey, Typeable c) => Address -> AddressPool t c -> (Maybe (Index Soft AddressK), AddressPool t c)

Lookup an address in the pool. When we find an address in a pool, the pool may be amended if the address was discovered near the edge. It is also possible that the pool is not amended at all - this happens in the case that an address is discovered far from the edge.

Pending Change Indexes

data PendingIxs

An ordered set of pending indexes. This keep track of indexes used

emptyPendingIxs :: PendingIxs

An empty pending set of change indexes.

NOTE: We do not define a Monoid instance here because there's no rational of combining two pending sets.

pendingIxsToList :: PendingIxs -> [Index Soft AddressK]

pendingIxsFromList :: [Index Soft AddressK] -> PendingIxs

Construct a PendingIxs from a list, ensuring that it is a set of indexes in descending order.

State

data SeqState t

A state to keep track of sequential addresses as described in BIP-44

Internally, the state keeps track of a few things for us and is it is parameterized by a type t which represents a target backend. This enables the state to be agnostic to the underlying address format.

Constructors

SeqState

Fields

internalPool :: !(AddressPool t InternalChain)

Addresses living on the InternalChain

externalPool :: !(AddressPool t ExternalChain)

Addresses living on the ExternalChain

pendingChangeIxs :: !PendingIxs

Indexes from the internal pool that have been used in pending transactions. The list is maintained sorted in descending order (cf: PendingIxs)

mkSeqState :: KeyToAddress t SeqKey => (SeqKey RootK XPrv, Passphrase "encryption") -> AddressPoolGap -> SeqState t

Construct a Sequential state for a wallet.