Cardano.Wallet.Primitive.Model

Description

Here we find the "business logic" to manage a Cardano wallet. This is a direct implementation of the model from the Formal Specification for a Cardano Wallet Note that, this module is purposedly agnostic to how blocks are retrieved or how various types are serialized.

This is really about how the wallet keep track of its internal state, and its UTxO (where the relationship is defined via the IsOurs abstraction to allow this core code to be used with any sort of derivation scheme).

All those functions are pure and there's no reason to shove in any sort of side-effects in here :)

Type

data Wallet s t

An opaque wallet type, see initWallet and applyBlocks to construct and update wallets.

Internally, this keeps track or a few things including:

UTxOs

Pending transaction

Transaction history

TODO: Known & used addresses

The Wallet is paremeterized over two types:

s: A _state_ used to keep track of known addresses. The business logic doesn't know how to answer the question 'Is this address ours?', so we expect this state to be able to answer that for us.

t: A target backend. This makes the wallet fairly agnostic to the type of binary representation used by the underlying target network and it allows us to re-use the same logic to provide a wallet backend for multiple backends (for instance, Byron or Shelley) which may have divergence in their binary formats. For the sake of this module, we only care about one particular super-power, and its the ability to compute transaction id (which is intrinsically linked to the transaction's binary format).

A few examples to make it concrete:

Wallet RndState Byron
Wallet SeqState Shelley
Wallet SeqState Bitcoin

Construction & Modification

initWallet

Arguments

:: (IsOurs s, NFData s, Show s, DefineTx t)

=> Block (Tx t)
The genesis block

-> s

-> Wallet s t

Create a an empty wallet and apply the given genesis block

The wallet tip will be set to the header of the applied genesis block.

updateState :: (IsOurs s, NFData s, Show s) => s -> Wallet s t -> Wallet s t

Update the state of an existing Wallet model

applyBlock :: forall s t. DefineTx t => Block (Tx t) -> Wallet s t -> (Map (Hash "Tx") (Tx t, TxMeta), Wallet s t)

Apply Block is the only way to make the wallet evolve. It returns a new updated wallet state, as well as the set of all our transaction discovered while applying the block.

applyBlocks :: forall s t. DefineTx t => NonEmpty (Block (Tx t)) -> Wallet s t -> NonEmpty (Map (Hash "Tx") (Tx t, TxMeta), Wallet s t)

Apply multiple blocks in sequence to an existing wallet, returning a list of intermediate wallet states.

Each intermediate wallet state is paired with the set of transactions that belong to that state but not the previous state.

For an original wallet state w and a list of blocks b such that:

b = [b1, b2, ..., bn]

Returns the following list of updates:

[ (t b1, w + b1)
, (t b2, w + b1 + b2)
, ...
, (t bn, w + b1 + b2 + ... + bn) ]

Where:

(t bi) is the set of transactions contained within block bi.

(w + bi) is the wallet state after applying block bi to wallet w.

newPending :: (Tx t, TxMeta) -> Wallet s t -> Wallet s t

unsafeInitWallet

Arguments

:: (IsOurs s, NFData s, Show s, DefineTx t)

=> UTxO
Unspent tx outputs belonging to this wallet

-> Set (Tx t, TxMeta)
Pending outgoing transactions

-> BlockHeader
Header of the latest applied block (current tip)

-> s
Address discovery state

-> Quantity "block" Natural
Block height

-> Wallet s t

Constructs a wallet from the exact given state. Using this function instead of initWallet and applyBlock allows the wallet invariants to be broken. Therefore it should only be used in the special case of loading wallet checkpoints from the database (where it is assumed a valid wallet was stored into the database).

Accessors

currentTip :: Wallet s t -> BlockHeader

Get the wallet current tip

getState :: Wallet s t -> s

Get the wallet current state

availableBalance :: DefineTx t => Wallet s t -> Natural

Available balance = balance . availableUTxO

totalBalance :: DefineTx t => Wallet s t -> Natural

Total balance = balance . totalUTxO

totalUTxO :: forall s t. DefineTx t => Wallet s t -> UTxO

Total UTxO = availableUTxO <> changeUTxO

availableUTxO :: forall s t. DefineTx t => Wallet s t -> UTxO

Available UTxO = pending ⋪ utxo

utxo :: Wallet s t -> UTxO

Actual utxo

getPending :: Wallet s t -> Set (Tx t, TxMeta)

Get the set of pending transactions

blockHeight :: Wallet s t -> Quantity "block" Natural

Get the block height of the chain.

A new wallet from initWallet (which already has the genesis block applied) has a block height of 0.