Cardano.Wallet.Primitive.Model

Description

Here we find the "business logic" to manage a Cardano wallet. This is a direct implementation of the model from the Formal Specification for a Cardano Wallet Note that, this module is purposedly agnostic to how blocks are retrieved or how various types are serialized.

This is really about how the wallet keep track of its internal state, and its UTxO (where the relationship is defined via the IsOurs abstraction to allow this core code to be used with any sort of derivation scheme).

All those functions are pure and there's no reason to shove in any sort of side-effects in here :)

Type

data Wallet s

An opaque wallet type, see initWallet and applyBlock to construct and update wallets.

Internally, this keeps track or a few things including:

UTxOs

Pending transaction

Transaction history

TODO: Known & used addresses

Construction & Modification

initWallet :: (IsOurs s, NFData s, Show s) => s -> Wallet s

Create an empty wallet from an initial state

updateState :: (IsOurs s, NFData s, Show s) => s -> Wallet s -> Wallet s

Update the state of an existing Wallet model

applyBlock :: Block -> Wallet s -> (Map (Hash "Tx") (Tx, TxMeta), Wallet s)

Apply Block is the only way to make the wallet evolve. It returns a new updated wallet state, as well as the set of all our transaction discovered while applying the block.

applyBlocks :: [Block] -> Wallet s -> (Map (Hash "Tx") (Tx, TxMeta), Wallet s)

Helper to apply multiple blocks in sequence to an existing wallet. It's basically just a foldl' applyBlock over the given blocks.

newPending :: Tx -> Wallet s -> Wallet s

Accessors

currentTip :: Wallet s -> SlotId

Get the wallet current tip

getState :: Wallet s -> s

Get the wallet current state

availableBalance :: Wallet s -> Natural

Available balance = balance . availableUTxO

totalBalance :: Wallet s -> Natural

Total balance = balance . totalUTxO

totalUTxO :: Wallet s -> UTxO

Total UTxO = availableUTxO <> changeUTxO

availableUTxO :: Wallet s -> UTxO

Available UTxO = pending ⋪ utxo