Cardano.Wallet.Primitive.Model

Description

Here we find the "business logic" to manage a Cardano wallet. This is a direct implementation of the model from the Formal Specification for a Cardano Wallet Note that, this module is purposedly agnostic to how blocks are retrieved or how various types are serialized.

This is really about how the wallet keep track of its internal state, and its UTxO (where the relationship is defined via the IsOurs abstraction to allow this core code to be used with any sort of derivation scheme).

All those functions are pure and there's no reason to shove in any sort of side-effects in here :)

Type

data Wallet s

An opaque wallet type, see initWallet, updateState, applyBlock, and applyBlocks to construct and update wallets.

Internally, this keeps track or a few things including:

UTxOs

Transaction history

Known & used addresses, via address discovery state

Blockchain parameters

The Wallet is parameterized over two types:

s: A _state_ used to keep track of known addresses. The business logic doesn't know how to answer the question 'Is this address ours?', so we expect this state to be able to answer that for us.

t: A target backend. This makes the wallet fairly agnostic to the type of binary representation used by the underlying target network and it allows us to re-use the same logic to provide a wallet backend for multiple backends (for instance, Byron or Shelley) which may have divergence in their binary formats. For the sake of this module, we only care about one particular super-power, and its the ability to compute transaction id (which is intrinsically linked to the transaction's binary format).

A few examples to make it concrete:

Wallet RndState Byron
Wallet SeqState Shelley
Wallet SeqState Bitcoin

data BlockchainParameters

Constructors

BlockchainParameters

Fields

getGenesisBlockHash :: Hash "Genesis"

Hash of the very first block

getGenesisBlockDate :: StartTime

Start time of the chain.

getFeePolicy :: FeePolicy

Policy regarding transaction fee.

getSlotLength :: SlotLength

Length, in seconds, of a slot.

getEpochLength :: EpochLength

Number of slots in a single epoch.

getTxMaxSize :: Quantity "byte" Word16

Maximum size of a transaction (soft or hard limit).

getEpochStability :: Quantity "block" Word32

Length of the suffix of the chain considered unstable

Construction & Modification

initWallet

Arguments

:: (IsOurs s, NFData s, Show s)

=> Block
The genesis block

-> BlockchainParameters
Initial blockchain parameters

-> s
Initial address discovery state

-> ([(Tx, TxMeta)], Wallet s)

Create a an empty wallet and apply the given genesis block

The wallet tip will be set to the header of the applied genesis block.

updateState :: (IsOurs s, NFData s, Show s) => s -> Wallet s -> Wallet s

Update the state of an existing Wallet model

applyBlock :: Block -> Wallet s -> ([(Tx, TxMeta)], Wallet s)

Apply Block is the primary way of making the wallet evolve. It returns the updated wallet state, as well as a set of all transactions belonging to the wallet discovered while applying the block.

applyBlocks :: NonEmpty Block -> Wallet s -> NonEmpty ([(Tx, TxMeta)], Wallet s)

Apply multiple blocks in sequence to an existing wallet, returning a list of intermediate wallet states.

Each intermediate wallet state is paired with the set of transactions that belong to that state but not the previous state.

For an original wallet state w and a list of blocks b such that:

b = [b1, b2, ..., bn]

Returns the following list of updates:

[ (t b1, w + b1)
, (t b2, w + b1 + b2)
, ...
, (t bn, w + b1 + b2 + ... + bn) ]

Where:

(t bi) is the set of transactions contained within block bi.

(w + bi) is the wallet state after applying block bi to wallet w.

unsafeInitWallet

Arguments

:: (IsOurs s, NFData s, Show s)

=> UTxO
Unspent tx outputs belonging to this wallet

-> BlockHeader
Header of the latest applied block (current tip)

-> s
Address discovery state

-> BlockchainParameters
Blockchain parameters

-> Wallet s

Constructs a wallet from the exact given state. Using this function instead of initWallet and applyBlock allows the wallet invariants to be broken. Therefore it should only be used in the special case of loading wallet checkpoints from the database (where it is assumed a valid wallet was stored into the database).

Accessors

currentTip :: Wallet s -> BlockHeader

Get the wallet current tip

getState :: Wallet s -> s

Get the wallet current state

availableBalance :: Set Tx -> Wallet s -> Natural

Available balance = balance . availableUTxO

totalBalance :: Set Tx -> Wallet s -> Natural

Total balance = balance . totalUTxO

totalUTxO :: Set Tx -> Wallet s -> UTxO

Total UTxO = availableUTxO <> changeUTxO

availableUTxO :: Set Tx -> Wallet s -> UTxO

Available UTxO = pending ⋪ utxo

utxo :: Wallet s -> UTxO

Actual utxo

blockchainParameters :: Wallet s -> BlockchainParameters

Get the current chain parameters.

Parameters may change over time via protocol updates, so we keep track of them as part of the wallet checkpoints.

Auxiliary

slotParams :: BlockchainParameters -> SlotParameters