Cardano.Wallet.Primitive.Types

Description

This module contains the core primitive of a Wallet. This is roughly a Haskell translation of the Formal Specification for a Cardano Wallet

It doesn't contain any particular business-logic code, but define a few primitive operations on Wallet core types as well.

Block

data Block tx

Constructors

Block

Fields

header :: !BlockHeader

transactions :: ![tx]

data BlockHeader

Constructors

BlockHeader

Fields

slotId :: SlotId

prevBlockHash :: !(Hash "BlockHeader")

Tx

class (NFData (Tx t), Show (Tx t), Ord (Tx t), Buildable (Tx t)) => DefineTx t where

An abstraction for computing transaction id. The target is an open-type that can be used to discriminate on. For instance:

instance DefineTx (HttpBridge network) where
  txId _ = {- ... -}
  ,,,

Note that txId is ambiguous and requires therefore a type application. Likely, a corresponding target would be found in scope (requires however ScopedTypeVariables).

For example, assuming there's a type t in scope, one can simply do:

txId @t tx

Associated Types

type Tx t :: *

Methods

txId :: Tx t -> Hash "Tx"

inputs :: Tx t -> [TxIn]

Compute a transaction id; assumed to be effectively injective. It returns an hex-encoded 64-byte hash.

NOTE: This is a rather expensive operation

outputs :: Tx t -> [TxOut]

Get transaction's inputs, ordered

data TxIn

Constructors

TxIn

Fields

inputId :: !(Hash "Tx")

inputIx :: !Word32

data TxOut

Constructors

TxOut

Fields

address :: !Address

coin :: !Coin

data TxMeta

Constructors

TxMeta

Fields

status :: !TxStatus

direction :: !Direction

slotId :: !SlotId

amount :: !(Quantity "lovelace" Natural)

data Direction

The effect of a Transaction on the wallet balance.

Constructors

Outgoing
The wallet balance decreases.

Incoming
The wallet balance increases or stays the same.

data TxStatus

Constructors

Pending

InLedger

Invalidated

newtype TxWitness

TxWitness is proof that transaction inputs are allowed to be spent

Constructors

TxWitness

Fields

unWitness :: ByteString

data TransactionInfo

Full expanded and resolved information about a transaction, suitable for presentation to the user.

Constructors

TransactionInfo

Fields

txInfoId :: !(Hash "Tx")

Transaction ID of this transaction

txInfoInputs :: ![(TxIn, Maybe TxOut)]

Transaction inputs and (maybe) corresponding outputs of the source. Source information can only be provided for outgoing payments.

txInfoOutputs :: ![TxOut]

Payment destination.

txInfoMeta :: !TxMeta

Other information calculated from the transaction.

txInfoDepth :: Quantity "slot" Natural

Number of slots since the transaction slot.

txInfoTime :: UTCTime

Creation time of the block including this transaction.

txIns :: forall t. DefineTx t => Set (Tx t) -> Set TxIn

isPending :: TxMeta -> Bool

True if the given tuple refers to a pending transaction

Address

newtype Address

Representation of Cardano addresses. Addresses are basically a human-friendly representation of public keys. Historically in Cardano, there exists different sort of addresses, and new ones are to come. So far, we can distinguish between three types of addresses:

Byron Random addresses, which holds a payload with derivation path details

Byron Sequential addresses, also known as Icarus'style addresses

Shelley base addresses, see also implementation-decisions/address

For more details, see also About Address Derivation

Shelley base addresses can be declined into two types:

Single Addresses: which only holds a public spending key

Group Addresses: which hold both a spending and delegation keys

It'll therefore seem legitimate to represent addresses as:

data Address
  = ByronAddress !ByteString
  | SingleAddress !XPub
  | GroupAddress !XPub XPub

However, there's a major drawback to this approach: we have to consider all three constructors everywhere, and make sure we test every function using them three despite having no need for such fine-grained representation.

Indeed, from the wallet core code, addresses are nothing more than an opaque bunch of bytes that can be compared with each others. When signing transactions, we have to lookup addresses anyway and therefore, can re-derive their corresponding public keys. The only moment the distinction between address type matters is when it comes to representing addresses at the edge of the application (the API layer). And here, this is precisely where we need to also what target backend we're connected to. Different backends use different encodings which may not be compatible.

Therefore, for simplicity, it's easier to consider addresses as "bytes", and only peak into these bytes whenever we need to do something with them. This makes it fairly clear that addresses are just an opaque string for the wallet layer and that the underlying encoding is rather agnostic to the underlying backend.

Constructors

Address

Fields

unAddress :: ByteString

data AddressState

Denotes if an address has been previously used or not... whether that be in the output of a transaction on the blockchain or one in our pending set.

Constructors

Used

Unused

class EncodeAddress t where

An abstract class to allow encoding of addresses depending on the target backend used.

Methods

encodeAddress :: Proxy t -> Address -> Text

class DecodeAddress t where

An abstract class to allow decoding of addresses depending on the target backend used.

Methods

decodeAddress :: Proxy t -> Text -> Either TextDecodingError Address

Coin

newtype Coin

Coins are stored as Lovelace (reminder: 1 Lovelace = 1e6 ADA)

Constructors

Coin

Fields

getCoin :: Word64

isValidCoin :: Coin -> Bool

UTxO

newtype UTxO

Constructors

UTxO

Fields

getUTxO :: Map TxIn TxOut

balance :: UTxO -> Natural

Compute the balance of a UTxO

balance' :: [(TxIn, TxOut)] -> Word64

Compute the balance of a unwrapped UTxO

pickRandom :: MonadRandom m => UTxO -> m (Maybe (TxIn, TxOut), UTxO)

Pick a random element from a UTxO, returns Nothing if the UTxO is empty. Otherwise, returns the selected entry and, the UTxO minus the selected one.

excluding :: UTxO -> Set TxIn -> UTxO

ins⋪ u

isSubsetOf :: UTxO -> UTxO -> Bool

a ⊆ b

restrictedBy :: UTxO -> Set TxIn -> UTxO

ins⊲ u

restrictedTo :: UTxO -> Set TxOut -> UTxO

u ⊳ outs

class Dom a where

Allows us to define the "domain" of any type — UTxO in particular — and use dom to refer to the inputs of an utxo.

This is the terminology used in the Formal Specification for a Cardano Wallet uses.

Associated Types

type DomElem a :: *

Methods

dom :: a -> Set (DomElem a)

data UTxOStatistics

Constructors

UTxOStatistics

Fields

histogram :: ![HistogramBar]

allStakes :: !Word64

boundType :: BoundType

data HistogramBar

Constructors

HistogramBar

Fields

bucketUpperBound :: !Word64

bucketCount :: !Word64

data BoundType

computeUtxoStatistics :: BoundType -> UTxO -> UTxOStatistics

Compute UtxoStatistics from UTxOs

log10 :: BoundType

Smart-constructor to create bounds using a log-10 scale

Slotting

data SlotId

A slot identifier is the combination of an epoch and slot.

Constructors

SlotId

Fields

epochNumber :: !Word64

slotNumber :: !Word16

data SlotParameters

The essential parameters necessary for performing slot arithmetic.

Constructors

SlotParameters

Fields

getEpochLength :: EpochLength

getSlotLength :: SlotLength

getGenesisBlockDate :: StartTime

newtype SlotLength

Duration of a single slot.

Constructors

SlotLength NominalDiffTime

newtype EpochLength

Number of slots in a single epoch

Constructors

EpochLength Word16

newtype StartTime

Blockchain start time

Constructors

StartTime UTCTime

slotRatio

Arguments

:: EpochLength

-> SlotId
Numerator

-> SlotId
Denominator

-> Quantity "percent" Percentage

Compute the approximate ratio / progress between two slots. This is an approximation for a few reasons, one of them being that we hard code the epoch length as a static number whereas it may vary in practice.

flatSlot :: EpochLength -> SlotId -> Word64

Convert a SlotId to the number of slots since genesis.

fromFlatSlot :: EpochLength -> Word64 -> SlotId

Convert a flatSlot index to SlotId.

slotStartTime :: SlotParameters -> SlotId -> UTCTime

The time that a slot begins.

slotCeiling :: SlotParameters -> UTCTime -> SlotId

For the given time t, determine the ID of the earliest slot with start time s such that 't ≤ s'.

slotFloor :: SlotParameters -> UTCTime -> Maybe SlotId

For the given time t, determine the ID of the latest slot with start time s such that 's ≤ t'.

slotAt :: SlotParameters -> UTCTime -> Maybe SlotId

For the given time t, determine the ID of the unique slot with start time s and end time e such that 's ≤ t ≤ e'.

slotDifference :: SlotParameters -> SlotId -> SlotId -> Quantity "slot" Natural

slotDifference a b is how many slots a is after b. The result is non-negative, and if b > a then this function returns zero.

slotMinBound :: SlotId

Returns the earliest slot.

slotPred :: SlotParameters -> SlotId -> Maybe SlotId

Return the slot immediately before the given slot.

slotSucc :: SlotParameters -> SlotId -> SlotId

Return the slot immediately after the given slot.

slotRangeFromTimeRange :: SlotParameters -> Range UTCTime -> Maybe (Range SlotId)

Transforms the given inclusive time range into an inclusive slot range.

This function returns a slot range if (and only if) the specified time range intersects with the life of the blockchain.

If, on the other hand, the specified time range terminates before the start of the blockchain, this function returns Nothing.

Wallet Metadata

data WalletMetadata

Additional information about a wallet that can't simply be derived from the blockchain like Wallet s t is.

Whereas Wallet s t in Primitive can be updated using applyBlock, WalletMetadata is not*.

) Except for possibly WalletMetadata and $sel:delegation:WalletMetadata...

Constructors

WalletMetadata

Fields

name :: !WalletName

creationTime :: !UTCTime

passphraseInfo :: !(Maybe WalletPassphraseInfo)

status :: !WalletState

delegation :: !(WalletDelegation PoolId)

newtype WalletId

Constructors

WalletId

Fields

getWalletId :: Digest Blake2b_160

newtype WalletName

Length-restricted name of a wallet

Constructors

WalletName

Fields

getWalletName :: Text

walletNameMinLength :: Int

Calling 'fromText @WalletName' on shorter longer string will fail.

walletNameMaxLength :: Int

Calling 'fromText @WalletName' on a longer string will fail.

data WalletState

Constructors

Ready

Restoring !(Quantity "percent" Percentage)

data WalletDelegation poolId

Constructors

NotDelegating

Delegating !poolId

newtype WalletPassphraseInfo

Constructors

WalletPassphraseInfo

Fields

lastUpdatedAt :: UTCTime

data WalletBalance

Constructors

WalletBalance

Fields

available :: !(Quantity "lovelace" Natural)

total :: !(Quantity "lovelace" Natural)

Stake Pools

newtype PoolId

Represent stake pool identifier. Note that the internal representation is left open currently, until we figure out a better type for those.

Constructors

PoolId

Fields

getPoolId :: Text

Querying

data SortOrder

Represents a sort order, applicable to the results returned by a query.

Constructors

Ascending
Sort in ascending order.

Descending
Sort in descending order.

Ranges

data Range a

Represents a range of values.

A range is defined by two optional bounds:

an inclusive lower bound

an inclusive upper bound

There are four cases:

Value 	 Range Represented 	 Membership Function

Range (Just x) (Just y) 	 [ x, y ] 	 \p -> p >= x && p <= y

Range (Just x) Nothing 	 [ x, ∞ ) 	 \p -> p >= x

Range Nothing (Just y) 	 (−∞, y ] 	 \p -> p <= y

Range Nothing Nothing 	 (−∞, ∞ ) 	 \p -> True

Constructors

Range

Fields

inclusiveLowerBound :: Maybe a

inclusiveUpperBound :: Maybe a

data RangeBound a

Represents a range boundary.

Constructors

NegativeInfinity

InclusiveBound a

PositiveInfinity

wholeRange :: Range a

The range that includes everything.

isAfterRange :: Ord a => a -> Range a -> Bool

Returns True if (and only if) the given range has an upper bound and the specified value is greater than the upper bound.

isBeforeRange :: Ord a => a -> Range a -> Bool

Returns True if (and only if) the given range has a lower bound and the specified value is smaller than the lower bound.

isSubrangeOf :: Ord a => Range a -> Range a -> Bool

Returns True if (and only if) the first given range is a subrange of the second given range.

isWithinRange :: Ord a => a -> Range a -> Bool

Returns True if (and only if) the given value is not smaller than the lower bound (if present) of the given range and is not greater than the upper bound (if present) of the given range.

mapRangeLowerBound :: (a -> a) -> Range a -> Range a

Apply a function to the lower bound of a range.

mapRangeUpperBound :: (a -> a) -> Range a -> Range a

Apply a function to the upper bound of a range.

rangeIsFinite :: Range a -> Bool

Returns True if (and only if) the given range has both a lower and upper bound.

rangeIsSingleton :: Eq a => Range a -> Bool

Returns True if (and only if) the range covers exactly one value.

rangeIsValid :: Ord a => Range a -> Bool

Returns True if (and only if) the lower bound of a range is not greater than its upper bound.

rangeHasLowerBound :: Range a -> Bool

Returns True if (and only if) the given range has a lower bound.

rangeHasUpperBound :: Range a -> Bool

Returns True if (and only if) the given range has an upper bound.

rangeLowerBound :: Range a -> RangeBound a

Get the lower bound of a Range.

rangeUpperBound :: Range a -> RangeBound a

Get the upper bound of a Range.

ProtocolMagic

newtype ProtocolMagic

Magic constant associated to a given network

Constructors

ProtocolMagic Int32

Polymorphic

newtype Hash (tag :: Symbol)

Constructors

Hash

Fields

getHash :: ByteString

newtype ShowFmt a

A polymorphic wrapper type with a custom show instance to display data through Buildable instances.

Constructors

ShowFmt a

invariant

Arguments

:: String
The message

-> a
The value to test

-> (a -> Bool)
The predicate

-> a

Checks whether or not an invariant holds, by applying the given predicate to the given value.

If the invariant does not hold (indicated by the predicate function returning False), throws an error with the specified message.

>>> invariant "not empty" [1,2,3] (not . null)
[1, 2, 3]

>>> invariant "not empty" [] (not . null)
*** Exception: not empty

distance :: (Ord a, Num a) => a -> a -> a

Compute distance between two numeric values |a - b|