Cardano.Wallet.Primitive.Types

Description

This module contains the core primitive of a Wallet. This is roughly a Haskell translation of the Formal Specification for a Cardano Wallet

It doesn't contain any particular business-logic code, but define a few primitive operations on Wallet core types as well.

Block

data Block

Constructors

Block

Fields

header :: !BlockHeader

transactions :: ![Tx]

data BlockHeader

Constructors

BlockHeader

Fields

slotId :: SlotId

prevBlockHash :: !(Hash "BlockHeader")

Tx

data Tx

Constructors

Tx

Fields

inputs :: ![TxIn]

NOTE: Order of inputs matters in the transaction representation. The transaction id is computed from the binary representation of a tx, for which inputs are serialized in a specific order.

outputs :: ![TxOut]

NOTE: Order of outputs matter in the transaction representations. Outputs are used as inputs for next transactions which refer to them using their indexes. It matters also for serialization.

class TxId target where

An abstraction for computing transaction id. The target is an open-type that can be used to discriminate on. For instance:

instance TxId HttpBridge where
  txId _ = {- ... -}

Note that txId is ambiguous and requires therefore a type application. Likely, a corresponding target would be found in scope (requires however ScopedTypeVariables).

For example, assuming there's a type target in scope, one can simply do:

txId @target tx

Methods

txId :: Tx -> Hash "Tx"

data TxIn

Constructors

TxIn

Fields

inputId :: !(Hash "Tx")

inputIx :: !Word32

data TxOut

Constructors

TxOut

Fields

address :: !Address

coin :: !Coin

data TxMeta

Constructors

TxMeta

Fields

status :: !TxStatus

direction :: !Direction

slotId :: !SlotId

amount :: !(Quantity "lovelace" Natural)

data Direction

The effect of a Transaction on the wallet balance.

Constructors

Outgoing
The wallet balance decreases.

Incoming
The wallet balance increases or stays the same.

data TxStatus

Constructors

Pending

InLedger

Invalidated

data TxWitness

TxWitness is proof that transaction inputs are allowed to be spent

Constructors

PublicKeyWitness ByteString (Hash "signature")
A signature of a transaction by the owner of the address of an input.

TODO: Use the XPub type instead of ByteString

ScriptWitness ByteString
Related to Plutus

RedeemWitness ByteString
Used to redeem ADA from the pre-sale

txIns :: Set Tx -> Set TxIn

Address

newtype Address

Representation of Cardano addresses. Addresses are basically a human-friendly representation of public keys. Historically in Cardano, there exists different sort of addresses, and new ones are to come. So far, we can distinguish between three types of addresses:

Byron Random addresses, which holds a payload with derivation path details

Byron Sequential addresses, also known as Icarus'style addresses

Shelley base addresses, see also implementation-decisions/address

For more details, see also About Address Derivation

Shelley base addresses can be declined into two types:

Single Addresses: which only holds a public spending key

Group Addresses: which hold both a spending and delegation keys

It'll therefore seem legitimate to represent addresses as:

data Address
  = ByronAddress !ByteString
  | SingleAddress !XPub
  | GroupAddress !XPub XPub

However, there's a major drawback to this approach: we have to consider all three constructors everywhere, and make sure we test every function using them three despite having no need for such fine-grained representation.

Indeed, from the wallet core code, addresses are nothing more than an opaque bunch of bytes that can be compared with each others. When signing transactions, we have to lookup addresses anyway and therefore, can re-derive their corresponding public keys. The only moment the distinction between address type matters is when it comes to representing addresses at the edge of the application (the API layer). And here, this is precisely where we need to also what target backend we're connected to. Different backends use different encodings which may not be compatible.

Therefore, for simplicity, it's easier to consider addresses as "bytes", and only peak into these bytes whenever we need to do something with them. This makes it fairly clear that addresses are just an opaque string for the wallet layer and that the underlying encoding is rather agnostic to the underlying backend.

Constructors

Address

Fields

getAddress :: ByteString

data AddressState

Denotes if an address has been previously used or not... whether that be in the output of a transaction on the blockchain or one in our pending set.

Constructors

Used

Unused

class EncodeAddress t where

An abstract class to allow encoding of addresses depending on the target backend used.

Methods

encodeAddress :: Proxy t -> Address -> Text

class DecodeAddress t where

An abstract class to allow decoding of addresses depending on the target backend used.

Methods

decodeAddress :: Proxy t -> Text -> Either TextDecodingError Address

Coin

newtype Coin

Coins are stored as Lovelace (reminder: 1 Lovelace = 1e6 ADA)

Constructors

Coin

Fields

getCoin :: Word64

isValidCoin :: Coin -> Bool

UTxO

newtype UTxO

Constructors

UTxO

Fields

getUTxO :: Map TxIn TxOut

balance :: UTxO -> Natural

Compute the balance of a UTxO

balance' :: [(TxIn, TxOut)] -> Word64

Compute the balance of a unwrapped UTxO

pickRandom :: MonadRandom m => UTxO -> m (Maybe (TxIn, TxOut), UTxO)

Pick a random element from a UTxO, returns Nothing if the UTxO is empty. Otherwise, returns the selected entry and, the UTxO minus the selected one.

excluding :: UTxO -> Set TxIn -> UTxO

ins⋪ u

isSubsetOf :: UTxO -> UTxO -> Bool

a ⊆ b

restrictedBy :: UTxO -> Set TxIn -> UTxO

ins⊲ u

restrictedTo :: UTxO -> Set TxOut -> UTxO

u ⊳ outs

class Dom a where

Allows us to define the "domain" of any type — UTxO in particular — and use dom to refer to the inputs of an utxo.

This is the terminology used in the Formal Specification for a Cardano Wallet uses.

Associated Types

type DomElem a :: *

Methods

dom :: a -> Set (DomElem a)

Slotting

data SlotId

A slot identifier is the combination of an epoch and slot.

Constructors

SlotId

Fields

epochNumber :: !Word64

slotNumber :: !Word16

slotRatio

Arguments

:: SlotId
Numerator

-> SlotId
Denominator

-> Quantity "percent" Percentage

Compute the approximate ratio / progress between two slots. This is an approximation for a few reasons, one of them being that we hard code the epoch length as a static number whereas it may vary in practice.

flatSlot :: SlotId -> Word64

Convert a SlotId to the number of slots since genesis.

fromFlatSlot :: Word64 -> SlotId

Convert a flatSlot index to SlotId.

Wallet Metadata

data WalletMetadata

Additional information about a wallet that can't simply be derived from the blockchain like Wallet s t is.

Whereas Wallet s t in Primitive can be updated using applyBlock, WalletMetadata is not*.

) Except for possibly WalletMetadata and $sel:delegation:WalletMetadata...

Constructors

WalletMetadata

Fields

name :: !WalletName

passphraseInfo :: !(Maybe WalletPassphraseInfo)

status :: !WalletState

delegation :: !(WalletDelegation PoolId)

newtype WalletId

Constructors

WalletId

Fields

getWalletId :: Digest Blake2b_160

newtype WalletName

Length-restricted name of a wallet

Constructors

WalletName

Fields

getWalletName :: Text

walletNameMinLength :: Int

Calling 'fromText @WalletName' on shorter longer string will fail.

walletNameMaxLength :: Int

Calling 'fromText @WalletName' on a longer string will fail.

data WalletState

Constructors

Ready

Restoring !(Quantity "percent" Percentage)

data WalletDelegation poolId

Constructors

NotDelegating

Delegating !poolId

newtype WalletPassphraseInfo

Constructors

WalletPassphraseInfo

Fields

lastUpdatedAt :: UTCTime

data WalletBalance

Constructors

WalletBalance

Fields

available :: !(Quantity "lovelace" Natural)

total :: !(Quantity "lovelace" Natural)

Stake Pools

newtype PoolId

Represent stake pool identifier. Note that the internal representation is left open currently, until we figure out a better type for those.

Constructors

PoolId

Fields

getPoolId :: Text

Polymorphic

newtype Hash (tag :: Symbol)

Constructors

Hash

Fields

getHash :: ByteString

newtype ShowFmt a

A polymorphic wrapper type with a custom show instance to display data through Buildable instances.

Constructors

ShowFmt a

invariant

Arguments

:: String
A title / message to throw in case of violation

-> a

-> (a -> Bool)

-> a

Check whether an invariants holds or not.

>>> invariant "not empty" [1,2,3] (not . null)
[1, 2, 3]

>>> invariant "not empty" [] (not . null)
*** Exception: not empty

distance :: (Ord a, Num a) => a -> a -> a

Compute distance between two numeric values |a - b|