Cardano.Wallet.Registry

Worker Registry

data WorkerRegistry key resource

A registry to keep track of worker threads and acquired resources.

empty :: Ord key => IO (WorkerRegistry key resource)

Construct a new empty registry

keys :: WorkerRegistry key resource -> IO [key]

Get all registered keys in the registry

lookup :: (MonadIO m, Ord key) => WorkerRegistry key resource -> key -> m (Maybe (Worker key resource))

Lookup the registry for a given worker

register :: forall resource ctx key msg. (Ord key, key ~ WorkerKey ctx, msg ~ WorkerMsg ctx, HasLogger (WorkerLog key msg) ctx, HasWorkerCtx resource ctx) => WorkerRegistry key resource -> ctx -> key -> MkWorker key resource msg ctx -> IO (Maybe (Worker key resource))

Register a new worker for a given key.

A worker maintains an acquired resource. It expects a task as an argument and will terminate as soon as its task is over. In practice, we provide a never-ending task that keeps the worker alive forever.

Returns Nothing if the worker fails to acquire the necessary resource or terminates unexpectedly before entering its main action.

unregister :: Ord key => WorkerRegistry key resource -> key -> IO ()

Unregister a worker from the registry, terminating the running task.

Worker

data Worker key resource

A worker which holds and manipulate a paticular acquired resource. That resource can be, for example, a handle to a database connection.

data MkWorker key resource msg ctx

See newWorker

Constructors

MkWorker

Fields

workerBefore :: WorkerCtx ctx -> key -> IO ()

A task to execute before the main worker's task. When creating a worker, this task is guaranteed to have terminated once newWorker returns.

workerMain :: WorkerCtx ctx -> key -> IO ()

A task for the worker, possibly infinite

workerAfter :: Tracer IO (WorkerLog key msg) -> Either SomeException () -> IO ()

Action to run when the worker exits

workerAcquire :: (resource -> IO ()) -> IO ()

A bracket-style factory to acquire a resource

defaultWorkerAfter :: Tracer IO (WorkerLog key msg) -> Either SomeException a -> IO ()

workerThread :: Worker key resource -> ThreadId

workerId :: Worker key resource -> key

workerResource :: Worker key resource -> resource

Context

class HasType resource (WorkerCtx ctx) => HasWorkerCtx resource ctx where

A class to link an existing context to a worker context.

Associated Types

type WorkerCtx ctx :: *

type WorkerMsg ctx :: *

type WorkerKey ctx :: *

Methods

hoistResource :: resource -> (WorkerMsg ctx -> WorkerLog (WorkerKey ctx) (WorkerMsg ctx)) -> ctx -> WorkerCtx ctx

Logging

data WorkerLog key msg

Constructors

MsgFinished

MsgThreadKilled

MsgUserInterrupt

MsgUnhandledException Text

MsgFromWorker key msg