Cardano.Wallet.Registry

Worker Registry

data WorkerRegistry key resource

A registry to keep track of worker threads and acquired resources.

empty :: Ord key => IO (WorkerRegistry key resource)

Construct a new empty registry

insert :: Ord key => WorkerRegistry key resource -> Worker key resource -> IO ()

Register a new worker

keys :: WorkerRegistry key resource -> IO [key]

Get all registered keys in the registry

lookup :: (MonadIO m, Ord key) => WorkerRegistry key resource -> key -> m (Maybe (Worker key resource))

Lookup the registry for a given worker

remove :: Ord key => WorkerRegistry key resource -> key -> IO ()

Remove a worker from the registry (and cancel any running task)

Worker

data Worker key resource

A worker which holds and manipulate a paticular acquired resource. That resource can be, for example, a handle to a database connection.

data MkWorker key resource ctx

See newWorker

Constructors

MkWorker

Fields

workerBefore :: WorkerCtx ctx -> key -> IO ()

A task to execute before the main worker's task. When creating a worker, this task is guaranteed to have terminated once newWorker returns.

workerMain :: WorkerCtx ctx -> key -> IO ()

A task for the worker, possibly infinite

workerAfter :: Trace IO Text -> Either SomeException () -> IO ()

Action to run when the worker exits

workerAcquire :: (resource -> IO ()) -> IO ()

A bracket-style factory to acquire a resource

newWorker :: forall key resource ctx. (HasLogger ctx, HasWorkerCtx resource ctx, ToText key) => ctx -> key -> MkWorker key resource ctx -> IO (Maybe (Worker key resource))

Create a new worker for a given key. Workers maintain an acquired resource. They expect a task as argument and will terminate as soon as their task is over; so in practice, we provide a never-ending task that keeps the worker alive forever.

Returns Nothing if the worker fails to acquire the necessary resource or terminate unexpectedly before entering its main action.

>>> newWorker ctx k withDBLayer restoreWallet
worker<thread#1234>

workerThread :: Worker key resource -> ThreadId

workerId :: Worker key resource -> key

workerResource :: Worker key resource -> resource

Context

class HasType resource (WorkerCtx ctx) => HasWorkerCtx resource ctx where

A class to link an existing context to a worker context.

Associated Types

type WorkerCtx ctx :: *

Methods

hoistResource :: resource -> ctx -> WorkerCtx ctx