Cardano.Wallet.Unsafe

Documentation

unsafeFromHex :: HasCallStack => ByteString -> ByteString

Decode an hex-encoded ByteString into raw bytes, or fail.

unsafeDecodeAddress :: (HasCallStack, DecodeAddress t) => Proxy t -> Text -> Address

Decode a bech32-encoded Text into an Address, or fail.

unsafeDecodeHex :: HasCallStack => Get a -> ByteString -> a

Run a decoder on a hex-encoded ByteString, or fail.

unsafeRunExceptT :: (MonadFail m, Show e) => ExceptT e m a -> m a

Run an ExceptT and throws the error if any. This makes sense only if called after checking for an invariant or, after ensuring that preconditions for meeting the underlying error have been discarded.

unsafeXPrv :: HasCallStack => ByteString -> XPrv

Build a XPrv from an hex-encoded bytestring

unsafeMkMnemonic :: forall mw n csz. (ConsistentEntropy n mw csz, EntropySize mw ~ n, HasCallStack) => [Text] -> Mnemonic mw

Build Mnemonic from literals

unsafeDeserialiseCbor :: HasCallStack => (forall s. Decoder s a) -> ByteString -> a

CBOR deserialise without error handling - handy for prototypes or testing.