Cardano.Wallet.Version

Description

Provides the package version and git revision which this was compiled from.

It is assumed that all cardano-wallet packages have the same version, that of the core package.

Stack builds will have the git command available to run during compilation.

Nix builds will inject the git revision into the executables after compiling. If the git revision has changed but the sources have not, then no haskell packages will be rebuilt, but the embedded git revision will be updated.

Values computed at compile-time

version :: Version

gitRevision :: GitRevision

The Git revision ID (40 character hex string) of this build.

This requires git do be available when building. Alternatively, the git revision of the cardano-wallet binary can be updated post-build using "Data.FileEmbed.injectWith".

Displaying Versions

showVersion :: Version -> String

Provides one possible concrete representation for Version. For a version with versionBranch = [1,2,3] and versionTags = ["tag1","tag2"], the output will be 1.2.3-tag1-tag2.

showFullVersion :: Version -> GitRevision -> String

Like showVersion, but also show the git revision.