Cardano.Wallet.Version.TH

Description

Template Haskell function for getting the git revision from the local repo. This is a separate module due to the GHC stage restriction.

Documentation

gitRevFromGit :: Q Exp

Git revision found by running git rev-parse. If git could not be executed, then this will be an empty string.