Network.Ntp

Description

This module provides the Ntp client related settings, types and re-exports used in a number of places throught codebase.

Documentation

withWalletNtpClient

Arguments

:: IOManager
The global IOManager instance, set up by the application main function.

-> Tracer IO NtpTrace
Logging object

-> (NtpClient -> IO a)
Action to run

-> IO a

Set up a NtpClient and pass it to the given action. The NtpClient is terminated when the callback returns.

getNtpStatus :: NtpClient -> IO ApiNetworkClock

re-exports from ntp-client

data NtpTrace

Constructors

NtpTraceStartNtpClient

NtpTraceRestartDelay !Int

NtpTraceRestartingClient

NtpTraceIOError !IOError

NtpTraceLookupsFails

NtpTraceClientStartQuery

NtpTraceNoLocalAddr

NtpTraceResult !NtpStatus

NtpTraceRunProtocolResults !(ResultOrFailure [NtpOffset])

NtpTracePacketSent !SockAddr !NtpPacket

NtpTracePacketSendError !SockAddr !IOException

NtpTracePacketDecodeError !SockAddr !String

NtpTracePacketReceived SockAddr !NtpPacket

NtpTraceWaitingForRepliesTimeout !IPVersion

data NtpClient

Constructors

NtpClient

Fields

ntpGetStatus :: STM NtpStatus

ntpQueryBlocking :: IO NtpStatus

ntpThread :: Async Void

Orphan instances

ToText IPVersion

Instance details

Methods

toText :: IPVersion -> Text

ToText NtpTrace

Instance details

Methods

toText :: NtpTrace -> Text

HasPrivacyAnnotation NtpTrace

Instance details

Methods

getPrivacyAnnotation :: NtpTrace -> PrivacyAnnotation

HasSeverityAnnotation NtpTrace

Instance details

Methods

getSeverityAnnotation :: NtpTrace -> Severity