Network.Ntp

Description

This module provides the Ntp client related settings, types and re-exports used in a number of places throught codebase.

Documentation

ntpSettings :: NtpSettings

getNtpStatus :: NtpClient -> IO ApiNetworkClock

re-exports from ntp-client

withNtpClient :: Tracer IO NtpTrace -> NtpSettings -> (NtpClient -> IO a) -> IO a

data NtpTrace

Constructors

NtpTraceStartNtpClient

NtpTraceTriggerUpdate

NtpTraceRestartDelay !Int

NtpTraceRestartingClient

NtpTraceClientSleeping

NtpTraceIOError !IOError

NtpTraceLookupServerFailed !String

NtpTraceClientStartQuery

NtpTraceNoLocalAddr

NtpTraceIPv4IPv6NoReplies

NtpTraceReportPolicyQueryFailed

NtpTraceQueryResult !Microsecond

NtpTraceRunProtocolError !IPVersion !IOError

NtpTraceRunProtocolNoResult !IPVersion

NtpTraceRunProtocolSuccess !IPVersion

NtpTraceSocketOpen !IPVersion

NtpTraceSocketClosed !IPVersion

NtpTracePacketSent !IPVersion

NtpTracePacketSentError !IPVersion !IOError

NtpTracePacketDecodeError !IPVersion !String

NtpTracePacketReceived !IPVersion

NtpTraceWaitingForRepliesTimeout !IPVersion

data NtpClient

Constructors

NtpClient

Fields

ntpGetStatus :: STM NtpStatus

ntpQueryBlocking :: IO NtpStatus

ntpThread :: Async Void

Orphan instances

ToText NtpTrace

Instance details

Methods

toText :: NtpTrace -> Text

ToText IPVersion

Instance details

Methods

toText :: IPVersion -> Text

DefinePrivacyAnnotation NtpTrace

Instance details

Methods

definePrivacyAnnotation :: NtpTrace -> PrivacyAnnotation

DefineSeverity NtpTrace

Instance details

Methods

defineSeverity :: NtpTrace -> Severity