Network.Wai.Middleware.Logging

Middleware

withApiLogger :: Trace IO Text -> ApiLoggerSettings -> Middleware

Installs a request & response logger on a Wai application.

The logger logs requests' and responses' bodies along with a few other useful piece of information.

Settings

newApiLoggerSettings :: IO ApiLoggerSettings

Create a new opaque ApiLoggerSettings

data ApiLoggerSettings

API logger settings

obfuscateKeys :: (Request -> [Text]) -> ApiLoggerSettings -> ApiLoggerSettings

Define a set of top-level object keys that should be obfuscated for a given request in a JSON format.