Network.Wai.Middleware.Logging

Documentation

newtype ApiLoggerSettings

API logger settings

Constructors

ApiLoggerSettings

Fields

obfuscateKeys :: Request -> [Text]

For a given request, obfuscate the values associated with the given keys from a JSON object payload.

withApiLogger :: Trace IO Text -> ApiLoggerSettings -> Middleware

Installs a request & response logger on a Wai application.

The logger logs requests' and responses' bodies along with a few other useful piece of information.