Ouroboros.Network.Client.Wallet

Description

Ouroboros mini-protocols clients for implementing cardano-wallet. These clients implement the logic and lift away concerns related to concrete data-type representation so that the code can be re-used / shared between Byron and Shelley.

ChainSyncFollowTip

chainSyncFollowTip

Arguments

:: Monad m

=> (Tip block -> m ())
Callback for when the tip changes.

-> ChainSyncClient (Serialised block) (Tip block) m Void

Client for the 'Chain Sync' mini-protocol, which provides notifications when the node tip changes.

This is used in the same way as chainSyncWithBlocks, except that only one of these clients is necessary, rather than one client per wallet.

ChainSyncWithBlocks

data ChainSyncCmd block (m :: * -> *)

We interact with the NetworkClient via a commands instrumenting the client to move within the state-machine protocol. Commands are sent from a parent thread via a shared TQueue.

MAIN THREAD | NETWORK CLIENT THREAD | *---------------* | | | | | Wallet Engine | | | | | *---------------* | | ^ | v | | *---------------* | *----------------* | | | | | | Network Layer |TQueue ]===| Network Client | | | | | | *---------------* | *----------------* | | ^ | v | | (ChainSync + TxSubmission)

The NetworkClient is idling most of the time and blocking on the TQueue while waiting for commands. Upon receiving a command, it interprets it by sending the corresponding instruction to the node and responding via a given callback.

See also send for invoking commands.

Constructors

CmdFindIntersection [Point block] (Maybe (Point block) -> m ())

CmdNextBlocks (NextBlocksResult (Point block) block -> m ())

chainSyncWithBlocks

Arguments

:: (Monad m, MonadSTM m, HasHeader block)

=> Quantity "block" Word32
Epoch stability (a.k.a k)

-> (Tip block -> BlockHeader)
Convert an abstract tip to a concrete BlockHeader

TODO: We probably need a better type for representing Tip as well!

-> TQueue m (ChainSyncCmd block m)
We use a TQueue as a communication channel to drive queries from outside of the network client to the client itself. Requests are pushed to the queue which are then transformed into messages to keep the state-machine moving.

-> ChainSyncClientPipelined block (Tip block) m Void

Client for the 'Chain Sync' mini-protocol.

A corresponding Channel can be obtained using a MuxInitiatorApplication constructor. Once started, the client simply runs ad-infinitum but one may interact with it via a TQueue of commands / messages used to move inside the state-machine.

In a typical usage, chainSyncWithBlocks would be executed in a forked thread and given a TQueue over which the parent thread as control.

>>> forkIO $ void $ chainSyncWithBlocks tr queue channel
()
>>> writeTQueue queue ...

Agency ------------------------------------------------------------------------- Client has agency* | Idle Server has agency* | Intersect, Next

A peer has agency if it is expected to send the next message.

-----------* | Intersect |◀══════════════════════════════╗

-----------* FindIntersect ║ │ ║ │ *---------* *------* │ Intersect.{Found,NotFound} | |═════════════▶| Done | └───────────────────────────────╼| | MsgDone *------* | Idle | ╔═══════════════════════════════════| | ║ RequestNext | |⇦ START ║ *---------* ▼ ╿

------* Roll.{Backward,Forward} │ | Next |────────────────────────────────────┘

------*

LocalTxSubmission

data LocalTxSubmissionCmd tx err (m :: * -> *)

Sending command to the localTxSubmission client. See also ChainSyncCmd.

Constructors

CmdSubmitTx tx (Maybe err -> m ())

localTxSubmission

Arguments

:: (MonadThrow m, MonadTimer m)

=> TQueue m (LocalTxSubmissionCmd tx err m)
We use a TQueue as a communication channel to drive queries from outside of the network client to the client itself. Requests are pushed to the queue which are then transformed into messages to keep the state-machine moving.

-> LocalTxSubmissionClient tx err m Void

Client for the 'Local Tx Submission' mini-protocol.

A corresponding Channel can be obtained using a MuxInitiatorApplication constructor.

Agency ------------------------------------------------------------------------- Client has agency* | Idle Server has agency* | Busy * A peer has agency if it is expected to send the next message.

-----------* | Busy |◀══════════════════════════════╗

-----------* SubmitTx ║ │ │ ║ │ │ *---------* *------* │ │ AcceptTx | |═════════════▶| Done | │ └────────────────────────────╼| | MsgDone *------* │ RejectTx | Idle | └──────────────────────────────────╼| | | |⇦ START

---------*

LocalStateQuery

data LocalStateQueryCmd block state (m :: * -> *)

Command to send to the localStateQuery client. See also ChainSyncCmd.

Constructors

CmdQueryLocalState (Point block) (Query block state) (LocalStateQueryResult state -> m ())

type LocalStateQueryResult state = Either AcquireFailure state

Shorthand for the possible outcomes of acquiring local state parameters.

localStateQuery

Arguments

:: (MonadThrow m, MonadTimer m)

=> TQueue m (LocalStateQueryCmd block state m)
We use a TQueue as a communication channel to drive queries from outside of the network client to the client itself. Requests are pushed to the queue which are then transformed into messages to keep the state-machine moving.

-> LocalStateQueryClient block (Query block) m Void

Client for the 'Local State Query' mini-protocol.

A corresponding Channel can be obtained using a MuxInitiatorApplication constructor.

Agency ------------------------------------------------------------------------- Client has agency* | Idle, Acquired Server has agency* | Acquiring, Querying * A peer has agency if it is expected to send the next message.

┌───────────────┐ Done ┌───────────────┐ ┌──────▶│ Idle ├─────────────▶│ Done │ │ └───┬───────────┘ └───────────────┘ │ │ ▲ │ Acquire │ │ │ │ │ Failure │ ▼ │ │ ┌───────────┴───┐ Result │ │ Acquiring │◀─────────────────────┐ │ └───┬───────────┘ │ Release│ │ ▲ │ │ │ │ │ │ Acquired ▼ │ ReAcquire │ │ ┌───────────┴───┐ ┌────────┴───────┐ └───────┤ Acquired │────────────>│ Querying │ └───────────────┘ └────────────────┘

Helpers

send :: (MonadSTM m, MonadAsync m, MonadTimer m) => TQueue m (cmd m) -> ((a -> m ()) -> cmd m) -> m (Either ErrNetworkUnavailable a)

Helper function to easily send commands to the node's client and read responses back.

>>> queue `send` CmdNextBlocks
RollForward cursor nodeTip blocks

>>> queue `send` CmdNextBlocks
AwaitReply